### 2024-11-18 #7500 @editorialbot
[PRE REVIEW]: NoiseModelling: an open-source software for outdoor noise mapping
**Submitting author:** @Nitnelav (<a href="http://orcid.org/0000-0001-5287-9668">Valentin LE BESCOND</a>)
**Repository:** https://github.com/Universite-Gustave-Eiffel/NoiseModelling
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 5.0.0
