+++ Docs/conf.py
@@ -178,7 +178,8 @@ texinfo_documents = [
 # -- Options for intersphinx extension ---------------------------------------
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/': None}
+intersphinx_mapping = {'python': ('https://docs.python.org/3',
+                                  'python-inv.txt')}
 
 # -- Options for todo extension ----------------------------------------------
 
