+++ paper/paper.bib
@@ -43,6 +43,7 @@
 	title = {{PostGIS}},
 	isbn = {978-0-387-30858-6},
 	url = {http://www.springer.com/computer/database+management+%26+information+retrieval/book/978-0-387-30858-6},
+	doi = {10.1007/springerreference_62668},
 	urldate = {2024-11-04},
 	publisher = {Springer},
 	author = {Strobl, Christian},
