+++ paper/paper.bib
@@ -39,18 +39,16 @@
 	file = {Snapshot:/home/aumond/Zotero/storage/IBVPUAUT/halshs-01179756.html:text/html},
 }
 
-@incollection{strobl_postgis_2008,
-	title = {{PostGIS}},
-	isbn = {978-0-387-30858-6},
-	url = {http://www.springer.com/computer/database+management+%26+information+retrieval/book/978-0-387-30858-6},
-	doi = {10.1007/springerreference_62668},
-	urldate = {2024-11-04},
-	publisher = {Springer},
-	author = {Strobl, Christian},
-	editor = {Shekhar, Shashi and Xiong, Hui},
-	year = {2008},
-	pages = {891--898},
-	file = {Snapshot:/home/aumond/Zotero/storage/EAU7SKU3/58176.html:text/html},
+@software{mckenna_postgis_2022,
+  author       = {McKenna, Jeff and
+                  PostGIS Team},
+  title        = {PostGIS 3.2.0, The Olivier Courtin Release},
+  month        = jan,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {3.2.0},
+  doi          = {10.5281/zenodo.5879632},
+  url          = {https://doi.org/10.5281/zenodo.5879632}
 }
 
 @article{chen_environmental_2023,
