### 2024-11-19 28124cd
restore postgis test in github workflow
