+++ noisemodelling-tutorial-01/src/test/java/org/noise_planet/nmtutorial01/PostgisTest.java
@@ -1,149 +1,37 @@
 package org.noise_planet.nmtutorial01;
 
+import org.h2gis.postgis_jts_osgi.DataSourceFactoryImpl;
+import org.h2gis.utilities.JDBCUtilities;
+import org.junit.Test;
+import org.postgresql.util.PSQLException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.net.ConnectException;
+import java.sql.Connection;
+import java.util.*;
+
 public class PostgisTest {
     Logger LOGGER = LoggerFactory.getLogger(PostgisTest.class);
-/*
+
     @Test
     public void testPostgisNoiseModelling1() throws Exception {
         DataSourceFactoryImpl dataSourceFactory = new DataSourceFactoryImpl();
         Properties p = new Properties();
         p.setProperty("serverName", "localhost");
         p.setProperty("portNumber", "5432");
-        p.setProperty("databaseName", "postgres");
+        p.setProperty("databaseName", "orbisgis_db");
         p.setProperty("user", "orbisgis");
         p.setProperty("password", "orbisgis");
         try(Connection connection = JDBCUtilities.wrapConnection(dataSourceFactory.createDataSource(p).getConnection())) {
-            Statement sql = connection.createStatement();
-
-            // Clean DB
-
-            sql.execute("DROP TABLE IF EXISTS BUILDINGS");
-            sql.execute("DROP TABLE IF EXISTS LW_ROADS");
-            sql.execute("DROP TABLE IF EXISTS RECEIVERS");
-            sql.execute("DROP TABLE IF EXISTS DEM");
-
-            // Import BUILDINGS
-
-            LOGGER.info("Import buildings");
-
-            GeoJsonRead.importTable(connection, Main.class.getResource("buildings.geojson").getFile(), "BUILDINGS","UTF-8",false);
-
-            // Import noise source
-
-            LOGGER.info("Import noise source");
-
-            GeoJsonRead.importTable(connection, Main.class.getResource("lw_roads.geojson").getFile(), "lw_roads","UTF-8",false);
-            // Set primary key
-            sql.execute("ALTER TABLE lw_roads ADD CONSTRAINT lw_roads_pk PRIMARY KEY (\"PK\");");
-
-            // Import BUILDINGS
-
-            LOGGER.info("Import evaluation coordinates");
-
-            GeoJsonRead.importTable(connection, Main.class.getResource("receivers.geojson").getFile(), "receivers","UTF-8",false);
-            // Set primary key
-            sql.execute("ALTER TABLE receivers ADD CONSTRAINT RECEIVERS_pk PRIMARY KEY (\"PK\");");
-
-            // Import MNT
-
-            LOGGER.info("Import digital elevation model");
-
-            GeoJsonRead.importTable(connection, Main.class.getResource("dem_lorient.geojson").getFile(), "dem","UTF-8",false);
-
-            // Init NoiseModelling
-            PointNoiseMap pointNoiseMap = new PointNoiseMap("buildings", "lw_roads", "receivers");
-
-            pointNoiseMap.setMaximumPropagationDistance(160.0d);
-            pointNoiseMap.setSoundReflectionOrder(0);
-            pointNoiseMap.setComputeHorizontalDiffraction(true);
-            pointNoiseMap.setComputeVerticalDiffraction(true);
-            // Building height field name
-            pointNoiseMap.setHeightField("HEIGHT");
-            // Point cloud height above sea level POINT(X Y Z)
-            pointNoiseMap.setDemTable("DEM");
-            // Do not propagate for low emission or far away sources.
-            // error in dB
-            pointNoiseMap.setMaximumError(0.1d);
-
-            // Init custom input in order to compute more than just attenuation
[…]
