### 2024-11-19 92a95f5
fix compute deltaE ok <= TC11
