+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/PathFinder.java
@@ -828,7 +828,6 @@ public class PathFinder {
         }
         List<Coordinate> pts = convexHullPoints;
 
-        double e = 0;
         Coordinate src = cutProfile.getSource().getCoordinate();
 
         // Move then check reflection height if there is diffraction on the path
@@ -966,9 +965,6 @@ public class PathFinder {
             path.setPoints2DGround(segmentGroundPoints);
             segments.add(path);
             if (i != pts.size() - 1) {
-                if (i != 1) {
-                    e += path.d;
-                }
                 PointPath pt = points.get(points.size() - 1);
                 pt.type = DIFH;
                 pt.bodyBarrier = bodyBarrier;
@@ -982,7 +978,6 @@ public class PathFinder {
                 }
             }
         }
-        pathParameters.e = e;
 
         if(points.isEmpty()) {
             return null;
@@ -1011,7 +1006,12 @@ public class PathFinder {
                 if(segments.isEmpty()) {
                     segments.add(pathParameters.getSRSegment());
                 }
-                pathParameters.deltaH = segments.get(0).d + e + segments.get(segments.size()-1).d - srPath.dc;
+                // Compute cumulated distance between the first diffraction and the last diffraction point
+                pathParameters.e = 0;
+                for(int idPoint = 1; idPoint < points.size() - 2; idPoint++) {
+                    pathParameters.e += points.get(idPoint).coordinate.distance(points.get(idPoint+1).coordinate);
+                }
+                pathParameters.deltaH = segments.get(0).d + pathParameters.e + segments.get(segments.size()-1).d - srPath.dc;
             } else {
                 segments.addAll(rayleighSegments);
                 points.addAll(1, rayleighPoints);
@@ -1028,21 +1028,26 @@ public class PathFinder {
         SegmentPath seg1 = segments.get(0);
         SegmentPath seg2 = segments.get(segments.size()-1);
 
-        double dSO0 = src.distance(c0);
-        double dOnR = cn.distance(rcv);
+        double dSO0 = seg1.d;
+        double dOnR = seg2.d;
         LineSegment sr = new LineSegment(src, rcv);
 
         LineSegment sPrimeR = new LineSegment(seg1.sPrime, rcv);
         double dSPrimeR = seg1.sPrime.distance(rcv);
         double dSPrimeO = seg1.sPrime.distance(c0);
-        pathParameters.deltaSPrimeRH = sPrimeR.orientationIndex(c0)*(dSPrimeO + e + dOnR - dSPrimeR);
-        pathParameters.deltaSPrimeRF = toCurve(dSPrimeO, dSPrimeR) + toCurve(e, dSPrimeR) + toCurve(dOnR, dSPrimeR) - toCurve(dSPrimeR, dSPrimeR);
+        // Compute cumulated distance between the first diffraction and the last diffraction point
+        pathParameters.e = 0;
+        for(int idPoint = 1; idPoint < points.size() - 2; idPoint++) {
+            pathParameters.e += points.get(idPoint).coordinate.distance(points.get(idPoint+1).coordinate);
+        }
+        pathParameters.deltaSPrimeRH = sPrimeR.orientationIndex(c0)*(dSPrimeO + pathParameters.e + dOnR - dSPrimeR);
[…]
