### 2024-11-20 6df6fa6
Update .readthedocs.yaml
