+++ .readthedocs.yaml
@@ -11,3 +11,11 @@ sphinx:
 formats:
   - pdf
   - htmlzip
+
+# Optionally declare the Python requirements required to build your docs
+python:
+   install:
+   - requirements: docs/requirements.txt
+   - method: pip
+     path: .
+     
