+++ .readthedocs.yaml
@@ -15,7 +15,7 @@ formats:
 # Optionally declare the Python requirements required to build your docs
 python:
    install:
-   - requirements: docs/requirements.txt
+   - requirements: Docs/requirements.txt
    - method: pip
      path: .
      
