### 2024-11-20 974b2cf
recompute Z interpolation in diffraction points between R and S. Maybe not the thing to do ?
