+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/PathFinder.java
@@ -500,120 +500,6 @@ public class PathFinder {
             return cnossosPath;
         }
         return null;
-
-        /**
-        List<Coordinate> coords = toDirectLine(coordinates);
-        if (!coordinates.isEmpty()) {
-            if (coordinates.size() > 2) {
-                List<Coordinate> topoPts = new ArrayList<>();
-                topoPts.add(coordinates.get(0));
-                double g = 0;
-                double d = 0;
-                List<CutPoint> allCutPoints = new ArrayList<>();
-                CutProfile mainProfile = new CutProfile();
-                for(int i=0; i<coordinates.size()-1; i++) {
-                    CutProfile profile = data.profileBuilder.getProfile(coordinates.get(i), coordinates.get(i+1), data.gS, false);
-                    mainProfile.addCutPoints(profile.getCutPoints().subList(1, profile.getCutPoints().size()));
-                    profile.setSrcOrientation(orientation);
-                    double dist = coordinates.get(i).distance(coordinates.get(i+1));
-                    g+=profile.getGPath()*dist;
-                    d+=dist;
-                    topoPts.addAll(profile.getCutPoints().stream()
-                            .filter(cut -> cut.getType().equals(BUILDING) || cut.getType().equals(TOPOGRAPHY) || cut.getType().equals(RECEIVER))
-                            .map(CutPoint::getCoordinate)
-                            .collect(Collectors.toList()));
-                    allCutPoints.addAll(profile.getCutPoints());
-
-                }
-                double res = g/d;
-                List<Integer> toRemove = new ArrayList<>();
-                for(int i=0; i<coordinates.size()-1; i++) {
-                    Coordinate c0 = coordinates.get(i);
-                    Coordinate c1 = coordinates.get(i+1);
-                    boolean between = false;
-                    for(int j=0; j<topoPts.size(); j++) {
-                        Coordinate p = topoPts.get(j);
-                        if(p.equals(c1)) {
-                            break;
-                        }
-                        if(p.equals(c0)) {
-                            between = true;
-                        }
-                        if(between && p.z > c1.z) {
-                            toRemove.add(j);
-                        }
-                    }
-                }
-                Collections.sort(toRemove);
-                Collections.reverse(toRemove);
-                for(int i : toRemove) {
-                    topoPts.remove(i);
-                }
-                //Set z value
-                // TODO do not call getZGround, we have it with getProfile
-                for (final Coordinate pt : topoPts) {
-                    coordinates.forEach(c -> {
-                        if (c.equals(pt) && c.z == pt.z) {
-                            pt.z = data.profileBuilder.getZGround(pt);
-                        }
-                    });
-                }
-                //Filter same pts
-                toRemove = new ArrayList<>();
-                for(int i=0; i<topoPts.size()-1; i++) {
-                    Coordinate pi0 = topoPts.get(i);
-                    Coordinate pi1 = topoPts.get(i+1);
-                    if(pi0.equals(pi1) && pi0.z == pi1.z) {
-                        toRemove.add(i);
-                    }
-                }
-                Collections.sort(toRemove);
-                Collections.reverse(toRemove);
-                for(int i : toRemove) {
-                    topoPts.remove(i);
-                }
-
-                List<Coordinate> groundPts = toDirectLine(topoPts);
-                PointPath src = new PointPath(coords.get(0), data.profileBuilder.getZ(coordinates.get(0)), new ArrayList<>(), SRCE);
[…]
