### 2024-11-21 0d9d31e
find accurate approximation of the buffer used in unit test add more test in unit test
