+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -1766,13 +1766,12 @@ public class AttenuationCnossosTest {
         double[] expectedDeltaGroundORF = new double[]{-1.18, -0.96, -0.81, -0.71, -0.65, -0.61, -0.60, -0.59};
         double[] expectedADiffF = new double[]{3.36, 4.33, 5.69, 7.50, 9.74, 12.30, 15.06, 17.94};
 
-        //Disabled because only diffraction
-        /*double[] expectedWH = new double[]{1.1e-04, 6.0e-04, 3.4e-03, Double.NaN, Double.NaN, 0.53, 2.70, 12.70};
+        double[] expectedWH = new double[]{1.1e-04, 6.0e-04, 3.4e-03, Double.NaN, Double.NaN, 0.53, 2.70, 12.70};
         double[] expectedCfH = new double[]{200.89, 217.45, 220.41, Double.NaN, Double.NaN, 1.88, 0.37, 0.08};
         double[] expectedAGroundH = new double[]{-1.32, -1.32, -1.32, Double.NaN, -Double.NaN, -1.32, -1.32, -1.32};
         double[] expectedWF = new double[]{0.00, 0.00, 0.00, 0.01, 0.08, 0.42, 2.16, 10.35};
         double[] expectedCfF = new double[]{199.59, 214.11, 225.39, 131.90, 22.89, 2.42, 0.46, 0.10};
-        double[] expectedAGroundF = new double[]{-1.32, -1.32, -1.29, -1.05, -1.32, -1.32, -1.32, -1.32};*/
+        double[] expectedAGroundF = new double[]{-1.32, -1.32, -1.29, -1.05, -1.32, -1.32, -1.32, -1.32};
 
         double[] expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
         double[] expectedAAtm = new double[]{0.02, 0.08, 0.20, 0.37, 0.71, 1.88, 6.36, 22.70};
@@ -1783,6 +1782,7 @@ public class AttenuationCnossosTest {
         double[] expectedLF = new double[]{32.85, 31.83, 30.35, 28.36, 25.78, 22.06, 14.81, -4.41};
         double[] expectedL = new double[]{32.70, 31.58, 29.99, 27.89, 24.36, 21.46, 14.18, -5.05};
         double[] expectedLA = sumArray(expectedL,A_WEIGHTING);
+
         //Actual values
         CnossosPath proPath = propDataOut.getPropagationPaths().get(0);
         double[] actualDeltaDiffSRH = proPath.aBoundaryH.deltaDiffSR;
@@ -1804,12 +1804,12 @@ public class AttenuationCnossosTest {
         double[] actualADiffF = proPath.aBoundaryF.aDiff;
 
         //Disabled because only diffraction
-        /*double[] actualWH = proPath.groundAttenuation.wH;
+        double[] actualWH = proPath.groundAttenuation.wH;
         double[] actualCfH = proPath.groundAttenuation.cfH;
         double[] actualAGroundH = proPath.groundAttenuation.aGroundH;
         double[] actualWF = proPath.groundAttenuation.wF;
         double[] actualCfF = proPath.groundAttenuation.cfF;
-        double[] actualAGroundF = proPath.groundAttenuation.aGroundF;*/
+        double[] actualAGroundF = proPath.groundAttenuation.aGroundF;
 
         double[] actualAlphaAtm = propDataOut.genericMeteoData.getAlpha_atmo();
         double[] actualAAtm = proPath.aAtm;
@@ -1862,151 +1862,9 @@ public class AttenuationCnossosTest {
[…]
