### 2024-11-22 8392435
resetore assert
