### 2024-11-22 dee8255
remove unused code
