+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/path/MirrorReceiverVisitor.java
@@ -1,137 +1,0 @@
-/**
- * NoiseModelling is a library capable of producing noise maps. It can be freely used either for research and education, as well as by experts in a professional use.
- * <p>
- * NoiseModelling is distributed under GPL 3 license. You can read a copy of this License in the file LICENCE provided with this software.
- * <p>
- * Official webpage : http://noise-planet.org/noisemodelling.html
- * Contact: contact@noise-planet.org
- */
-
-package org.noise_planet.noisemodelling.pathfinder.path;
-
-import org.locationtech.jts.algorithm.LineIntersector;
-import org.locationtech.jts.algorithm.RobustLineIntersector;
-import org.locationtech.jts.geom.Coordinate;
-import org.locationtech.jts.geom.LineSegment;
-import org.locationtech.jts.index.ItemVisitor;
-import org.locationtech.jts.triangulate.quadedge.Vertex;
-import org.noise_planet.noisemodelling.pathfinder.profilebuilder.Wall;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class MirrorReceiverVisitor implements ItemVisitor {
-    List<MirrorReceiver> result = new ArrayList<>();
-    List<Wall> buildWalls;
-
-    public Coordinate getReceiver() {
-        return receiver;
-    }
-
-    public Coordinate getSource() {
-        return source;
-    }
-
-    Coordinate source;
-    Coordinate receiver;
-    LineSegment sourceReceiverSegment;
-    double maximumDistanceFromSegment;
-    double maximumPropagationDistance;
-    int visitedNode = 0;
-
-    public MirrorReceiverVisitor(List<Wall> buildWalls, Coordinate source, Coordinate receiver,
-                                 double maximumDistanceFromSegment,
-                                 double maximumPropagationDistance) {
-        this.buildWalls = buildWalls;
-        this.source = source;
-        this.receiver = receiver;
-        this.sourceReceiverSegment = new LineSegment(source, receiver);
-        this.maximumDistanceFromSegment = maximumDistanceFromSegment;
-        this.maximumPropagationDistance = maximumPropagationDistance;
-    }
-    public MirrorReceiverVisitor(Coordinate source, Coordinate receiver) {
-
-        this.source = source;
-        this.receiver = receiver;
-    }
-
-
-    public void visitItemNoWall(Object item) {
-        visitedNode++;
-
-        MirrorReceiver receiverImage = (MirrorReceiver) item;
-
-        // Add mirrored positions to the respective lists
-        source = receiverImage.getReceiverPos();
-        receiver = receiverImage.getParentMirror().getReceiverPos();
-    }
-
-
-    /**
-     *
-     * @param item the index item to be visited
-     */
-    @Override
-    public void visitItem(Object item) {
-        visitedNode++;
-        // try to excluded walls without taking into account the topography and other factors
-
-        MirrorReceiver receiverImage = (MirrorReceiver) item;
-        // Check propagation distance
-        if(receiverImage.getReceiverPos().distance3D(source) < maximumPropagationDistance) {
-            // Check distance of walls
-            MirrorReceiver currentReceiverImage = receiverImage;
-            Coordinate reflectionPoint = source;
-            while (currentReceiverImage != null) {
-                final Wall currentWall = currentReceiverImage.getWall();
-                final LineSegment currentWallLineSegment = currentWall.getLineSegment();
-                if (currentWallLineSegment.distance(sourceReceiverSegment) > maximumDistanceFromSegment) {
-                    return;
-                }
-                // Check if reflection is placed on the wall segment
[…]
