+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -5023,14 +5023,19 @@ public class AttenuationCnossosTest {
         //Run computation
         computeRays.run(propDataOut);
 
-        exportRays("target/tc18.json", propDataOut);
+        assertEquals(2, propDataOut.getPropagationPaths().size());
 
+        double[] L = addArray(propDataOut.getVerticesSoundLevel().get(0).value, new double[]{93-26.2,93-16.1,93-8.6,93-3.2,93,93+1.2,93+1.0,93-1.1});
+
+        // Table 195
         double[] expectedWH = new double[]{0.00, 0.00, 0.00, 0.02, 0.11, 0.58, 2.94, 13.68};
         double[] expectedCfH = new double[]{201.47,218.76,217.74,102.89,14.04,1.71,0.34,0.07};
         double[] expectedAGroundH = new double[]{-1.26,-1.26,-1.26,2.12,-1.26,-1.26,-1.26,-1.26};
         double[] expectedWF = new double[]{0.00, 0.00, 0.00, 0.01, 0.08, 0.42, 2.16, 10.35};
         double[] expectedCfF = new double[]{199.62,214.14,225.42,131.91,22.89,2.42,0.46,0.10};
         double[] expectedAGroundF = new double[]{-1.26,-1.26,-1.26,-1.26,-1.26,-1.26,-1.26,-1.26};
+
+        // Table 196
         double[] expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
         double[] expectedAAtm = new double[]{0.02, 0.08,0.20,0.37,0.71,1.88,6.37,22.73};
         double[] expectedADiv = new double[]{56.78,56.78,56.78,56.78,56.78,56.78,56.78,56.78};
@@ -5042,8 +5047,145 @@ public class AttenuationCnossosTest {
         double[] expectedLA = new double[]{11.26,21.30,28.68,32.54,36.77,36.80,32.11,13.65};
 
         //Actual values
-        //Actual values - right lateral
         CnossosPath proPath = propDataOut.getPropagationPaths().get(0);
+
+        double[] actualWH = proPath.groundAttenuation.wH;
+        double[] actualCfH = proPath.groundAttenuation.cfH;
+        double[] actualAGroundH = proPath.groundAttenuation.aGroundH;
+        double[] actualWF = proPath.groundAttenuation.wF;
+        double[] actualCfF = proPath.groundAttenuation.cfF;
+        double[] actualAGroundF = proPath.groundAttenuation.aGroundF;
+
+        double[] actualAlphaAtm = propDataOut.genericMeteoData.getAlpha_atmo();
+        double[] actualAAtm = proPath.aAtm;
+        double[] actualADiv = proPath.aDiv;
+        double[] actualABoundaryH = proPath.double_aBoundaryH;
+        double[] actualABoundaryF = proPath.double_aBoundaryF;
+        double[] actualLH = addArray(proPath.aGlobalH, SOUND_POWER_LEVELS);
+        double[] actualLF = addArray(proPath.aGlobalF, SOUND_POWER_LEVELS);
+        double[] actualL = addArray(proPath.aGlobal, SOUND_POWER_LEVELS);
+        double[] actualLA = addArray(actualL, A_WEIGHTING);
+
[…]
