### 2024-11-22 bbe51b6
Remove reflexion from e computation. Fix variable for delta H computation
