### 2024-11-22 af2edc8
TC18 0.5 dB error
