+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -5149,129 +5149,13 @@ public class AttenuationCnossosTest {
         assertDoubleArrayEquals("DeltaDiffSRF - vertical plane", expectedDeltaDiffSRF, actualDeltaDiffSRF, ERROR_EPSILON_LOW);
         assertDoubleArrayEquals("AGroundSOF - vertical plane", expectedAGroundSOF, actualAGroundSOF, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("AGroundORF - vertical plane", expectedAGroundORF, actualAGroundORF, ERROR_EPSILON_LOWEST);
-        assertDoubleArrayEquals("DeltaDiffSPrimeRF - vertical plane", expectedDeltaDiffSPrimeRF, actualDeltaDiffSPrimeRF, ERROR_EPSILON_LOW);
+        assertDoubleArrayEquals("DeltaDiffSPrimeRF - vertical plane", expectedDeltaDiffSPrimeRF, actualDeltaDiffSPrimeRF, ERROR_EPSILON_MEDIUM);
         assertDoubleArrayEquals("DeltaDiffSRPrimeF - vertical plane", expectedDeltaDiffSRPrimeF, actualDeltaDiffSRPrimeF, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("DeltaGroundSOF - vertical plane", expectedDeltaGroundSOF, actualDeltaGroundSOF, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("DeltaGroundORF - vertical plane", expectedDeltaGroundORF, actualDeltaGroundORF, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("actualADiffF - vertical plane", expectedADiffF, actualADiffF, ERROR_EPSILON_LOW);
 
-
-
-
-
-
-        /*
-        actualAlphaAtm = propDataOut.genericMeteoData.getAlpha_atmo();
-        actualAAtm = proPath.aAtm;
-        actualADiv = proPath.aDiv;
-        actualABoundaryH = proPath.double_aBoundaryH;
-        actualABoundaryF = proPath.double_aBoundaryF;
-
-        double[] actualLA = addArray(actualL, A_WEIGHTING);
-
-
-        double[] actualDeltaDiffSRH = proPath.aBoundaryH.deltaDiffSR;
-        double[] actualAGroundSOH = proPath.aBoundaryH.aGroundSO;
-        double[] actualAGroundORH = proPath.aBoundaryH.aGroundOR;
-        double[] actualDeltaDiffSPrimeRH = proPath.aBoundaryH.deltaDiffSPrimeR;
-        double[] actualDeltaDiffSRPrimeH = proPath.aBoundaryH.deltaDiffSRPrime;
-        double[] actualDeltaGroundSOH = proPath.aBoundaryH.deltaGroundSO;
-        double[] actualDeltaGroundORH = proPath.aBoundaryH.deltaGroundOR;
-        double[] actualADiffH = proPath.aBoundaryH.aDiff;
-
-        double[] actualDeltaDiffSRF = proPath.aBoundaryF.deltaDiffSR;
-        double[] actualAGroundSOF = proPath.aBoundaryF.aGroundSO;
-        double[] actualAGroundORF = proPath.aBoundaryF.aGroundOR;
-        double[] actualDeltaDiffSPrimeRF = proPath.aBoundaryF.deltaDiffSPrimeR;
-        double[] actualDeltaDiffSRPrimeF = proPath.aBoundaryF.deltaDiffSRPrime;
-        double[] actualDeltaGroundSOF = proPath.aBoundaryF.deltaGroundSO;
[…]
