### 2024-11-22 21f7980
remove function argument
