### 2024-11-22 3e753ca
New Dynamic Tests with associated tutorial
