+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseEmissionMaker.java
@@ -333,41 +333,41 @@ public class NoiseEmissionMaker extends Scene {
         int way = 3; // default value 2-way road
 
         // Read fields
-        if(sourceFields.containsKey("LV_SPD_"+period)) {
-            lv_speed = rs.getDouble(sourceFields.get("LV_SPD_"+period));
+        if(sourceFields.containsKey("LV_SPD"+period)) {
+            lv_speed = rs.getDouble(sourceFields.get("LV_SPD"+period));
         }
-        if(sourceFields.containsKey("MV_SPD_"+period)) {
-            mv_speed = rs.getDouble(sourceFields.get("MV_SPD_"+period));
+        if(sourceFields.containsKey("MV_SPD"+period)) {
+            mv_speed = rs.getDouble(sourceFields.get("MV_SPD"+period));
         }
-        if(sourceFields.containsKey("HGV_SPD_"+period)) {
-            hgv_speed = rs.getDouble(sourceFields.get("HGV_SPD_"+period));
+        if(sourceFields.containsKey("HGV_SPD"+period)) {
+            hgv_speed = rs.getDouble(sourceFields.get("HGV_SPD"+period));
         }
-        if(sourceFields.containsKey("WAV_SPD_"+period)) {
-            wav_speed = rs.getDouble(sourceFields.get("WAV_SPD_"+period));
+        if(sourceFields.containsKey("WAV_SPD"+period)) {
+            wav_speed = rs.getDouble(sourceFields.get("WAV_SPD"+period));
         }
-        if(sourceFields.containsKey("WBV_SPD_"+period)) {
-            wbv_speed = rs.getDouble(sourceFields.get("WBV_SPD_"+period));
+        if(sourceFields.containsKey("WBV_SPD"+period)) {
+            wbv_speed = rs.getDouble(sourceFields.get("WBV_SPD"+period));
         }
-        if(sourceFields.containsKey("LV_"+period)) {
-            lvPerHour = rs.getDouble(sourceFields.get("LV_"+period));
+        if(sourceFields.containsKey("LV"+period)) {
+            lvPerHour = rs.getDouble(sourceFields.get("LV"+period));
         }
-        if(sourceFields.containsKey("MV_"+period)) {
-            mvPerHour = rs.getDouble(sourceFields.get("MV_"+period));
+        if(sourceFields.containsKey("MV"+period)) {
+            mvPerHour = rs.getDouble(sourceFields.get("MV"+period));
         }
-        if(sourceFields.containsKey("HGV_"+period)) {
-            hgvPerHour = rs.getDouble(sourceFields.get("HGV_"+period));
+        if(sourceFields.containsKey("HGV"+period)) {
+            hgvPerHour = rs.getDouble(sourceFields.get("HGV"+period));
         }
-        if(sourceFields.containsKey("WAV_"+period)) {
-            wavPerHour = rs.getDouble(sourceFields.get("WAV_"+period));
+        if(sourceFields.containsKey("WAV"+period)) {
+            wavPerHour = rs.getDouble(sourceFields.get("WAV"+period));
         }
-        if(sourceFields.containsKey("WBV_"+period)) {
-            wbvPerHour = rs.getDouble(sourceFields.get("WBV_"+period));
+        if(sourceFields.containsKey("WBV"+period)) {
+            wbvPerHour = rs.getDouble(sourceFields.get("WBV"+period));
         }
         if(sourceFields.containsKey("PVMT")) {
             roadSurface= rs.getString(sourceFields.get("PVMT"));
         }
-        if(sourceFields.containsKey("TEMP_"+period)) {
-            temperature = rs.getDouble(sourceFields.get("TEMP_"+period));
+        if(sourceFields.containsKey("TEMP"+period)) {
[…]
