+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/PathFinder.java
@@ -349,21 +349,6 @@ public class PathFinder {
         return 2*max(1000, 8*d)* asin(mn/(2*max(1000, 8*d)));
     }
 
-
-    /**
-     *
-     * @param src
-     * @param rcv
-     * @param meanPlane
-     * @param gPath
-     * @param gS
-     * @return
-     */
-    private static SegmentPath computeSegment(Coordinate src, Coordinate rcv, double[] meanPlane, double gPath, double gS) {
-        return computeSegment(src, src.y, rcv, rcv.y, meanPlane, gPath, gS);
-    }
-
-
     /**
      * Compute the segment path
      * @param src
@@ -372,31 +357,27 @@ public class PathFinder {
      * @return the calculated segment
      */
     private static SegmentPath computeSegment(Coordinate src, Coordinate rcv, double[] meanPlane) {
-        return computeSegment(src, src.y, rcv, rcv.y, meanPlane, 0, 0);
+        return computeSegment(src, rcv, meanPlane, 0, 0);
     }
 
     /**
      * Compute the segment path with more attribute
      * @param src
-     * @param sz
      * @param rcv
-     * @param rz
      * @param meanPlane
      * @param gPath
      * @param gS
      * @return the computed segment path
      */
 
-    private static SegmentPath computeSegment(Coordinate src, double sz, Coordinate rcv, double rz, double[] meanPlane, double gPath, double gS) {
+    private static SegmentPath computeSegment(Coordinate src, Coordinate rcv, double[] meanPlane, double gPath, double gS) {
         SegmentPath seg = new SegmentPath();
-        Coordinate srcZ = new Coordinate(src.x, sz);
-        Coordinate rcvZ = new Coordinate(rcv.x, rz);
-        Coordinate sourcePointOnMeanPlane = projectPointOnLine(srcZ, meanPlane[0], meanPlane[1]);
-        Coordinate receiverPointOnMeanPlane = projectPointOnLine(rcvZ, meanPlane[0], meanPlane[1]);
-        Vector2D sourceToProjectedPoint = Vector2D.create(srcZ, sourcePointOnMeanPlane);
-        Vector2D receiverToProjectedPoint = Vector2D.create(rcvZ, receiverPointOnMeanPlane);
-        seg.s = srcZ;
-        seg.r = rcvZ;
+        Coordinate sourcePointOnMeanPlane = projectPointOnLine(src, meanPlane[0], meanPlane[1]);
+        Coordinate receiverPointOnMeanPlane = projectPointOnLine(rcv, meanPlane[0], meanPlane[1]);
+        Vector2D sourceToProjectedPoint = Vector2D.create(src, sourcePointOnMeanPlane);
+        Vector2D receiverToProjectedPoint = Vector2D.create(rcv, receiverPointOnMeanPlane);
+        seg.s = src;
+        seg.r = rcv;
         seg.sMeanPlane = sourcePointOnMeanPlane;
         seg.rMeanPlane = receiverPointOnMeanPlane;
         seg.sPrime = Vector2D.create(sourcePointOnMeanPlane).add(sourceToProjectedPoint).toCoordinate();
@@ -404,8 +385,8 @@ public class PathFinder {
 
         seg.d = src.distance(rcv);
         seg.dp =sourcePointOnMeanPlane.distance(receiverPointOnMeanPlane);
-        seg.zsH = srcZ.distance(sourcePointOnMeanPlane);
-        seg.zrH = rcvZ.distance(receiverPointOnMeanPlane);
+        seg.zsH = src.distance(sourcePointOnMeanPlane);
+        seg.zrH = rcv.distance(receiverPointOnMeanPlane);
         seg.a = meanPlane[0];
[…]
