### 2024-11-26 e230431
remove reflection absorption class as it is cached but should be stored for homogenous and favorable
