+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -4545,22 +4545,17 @@ public class AttenuationCnossosTest {
         actualDeltaGroundSO = proPath.aBoundaryH.deltaGroundSO;
         actualDeltaGroundOR = proPath.aBoundaryH.deltaGroundOR;
         actualADiff = proPath.aBoundaryH.aDiff;
-
-        actualAlphaAtm = propDataOut.genericMeteoData.getAlpha_atmo();
         actualAAtm = proPath.aAtm;
         actualADiv = proPath.aDiv;
         actualABoundaryH = proPath.double_aBoundaryH;
         actualABoundaryF = proPath.double_aBoundaryF;
         actualLH = addArray(proPath.aGlobalH, SOUND_POWER_LEVELS);
         actualLF = addArray(proPath.aGlobalF, SOUND_POWER_LEVELS);
-        actualLA = addArray(actualLF, A_WEIGHTING);
+        actualL = addArray(proPath.aGlobal, SOUND_POWER_LEVELS);
+        actualLA = addArray(actualL, A_WEIGHTING);
         double[] reflexionLA = actualLA;
 
         double[] LA = sumDbArray(directLA,reflexionLA);
-        double[] diffL = diffArray(expectedL, actualL);
-        double[] diffLa = diffArray(new double[]{14.02, 23.84, 30.95, 33.86, 38.37, 38.27, 33.25, 15.28},LA);
-        double[] valL = getMaxValeurAbsolue(diffL);
-        double[] valLA = getMaxValeurAbsolue(diffLa);
 
         assertDoubleArrayEquals("WH - reflexion", expectedWH, actualWH, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("CfH - reflexion", expectedCfH, actualCfH, ERROR_EPSILON_LOWEST);
@@ -4577,8 +4572,6 @@ public class AttenuationCnossosTest {
         assertDoubleArrayEquals("DeltaGroundSO - reflexion", expectedDeltaGroundSO, actualDeltaGroundSO, ERROR_EPSILON_VERY_LOW);
         assertDoubleArrayEquals("DeltaGroundOR - reflexion", expectedDeltaGroundOR, actualDeltaGroundOR, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("ADiff - reflexion", expectedADiff, actualADiff, ERROR_EPSILON_VERY_LOW);
-
-        assertDoubleArrayEquals("AlphaAtm - reflexion", expectedAlphaAtm, actualAlphaAtm, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("AAtm - reflexion", expectedAAtm, actualAAtm, ERROR_EPSILON_LOW);
         assertDoubleArrayEquals("ADiv - reflexion", expectedADiv, actualADiv, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("ABoundaryH - reflexion", expectedABoundaryH, actualABoundaryH, ERROR_EPSILON_LOW);
@@ -4741,7 +4734,7 @@ public class AttenuationCnossosTest {
         double[] expectedDeltaDiffSRF = new double[]{7.22,8.76,10.80,13.24,15.93,18.77,21.69,24.65};
         double[] expectedAGroundSOF = new double[]{-0.43,-0.43, -0.43, -0.43, -0.43, -0.43, -0.43, -0.43};
         double[] expectedAGroundORF  = new double[]{-1.90,-1.90, -1.90, -1.90, -1.90, -1.90,-1.90,-1.90};
[…]
