+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -4596,12 +4596,12 @@ public class AttenuationCnossosTest {
                         new Coordinate(170, 60, 15)}, Arrays.asList(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.5), 1)
 
                 .addWall(new Coordinate[]{
-                        new Coordinate(87, 50),
-                        new Coordinate(92, 32)}, 12, Arrays.asList(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.5), 2);
+                        new Coordinate(87, 50, 12),
+                        new Coordinate(92, 32, 12)}, Arrays.asList(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.5), 2);
 
         addGroundAttenuationTC5(builder);
         addTopographicTC5Model(builder);
-
+        builder.setzBuildings(true);
         builder.finishFeeding();
 
         //Propagation data building
@@ -4783,12 +4783,12 @@ public class AttenuationCnossosTest {
         assertDoubleArrayEquals("ADiv - reflexion", expectedADiv, actualADiv, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("ABoundaryH - reflexion", expectedABoundaryH, actualABoundaryH, ERROR_EPSILON_VERY_LOW);
         assertDoubleArrayEquals("ABoundaryF - reflexion", expectedABoundaryF, actualABoundaryF, ERROR_EPSILON_VERY_LOW);
-        assertDoubleArrayEquals("LH - reflexion", expectedLH, actualLH, ERROR_EPSILON_VERY_HIGH);
-        assertDoubleArrayEquals("LF - reflexion", expectedLF, actualLF, ERROR_EPSILON_VERY_HIGH);
-        assertDoubleArrayEquals("L - reflexion", expectedL, actualL, ERROR_EPSILON_VERY_HIGH);
-        assertDoubleArrayEquals("LA - reflexion", expectedLA, actualLA, ERROR_EPSILON_VERY_HIGH);
+        assertDoubleArrayEquals("LH - reflexion", expectedLH, actualLH, ERROR_EPSILON_VERY_LOW);
+        assertDoubleArrayEquals("LF - reflexion", expectedLF, actualLF, ERROR_EPSILON_VERY_LOW);
+        assertDoubleArrayEquals("L - reflexion", expectedL, actualL, ERROR_EPSILON_VERY_LOW);
+        assertDoubleArrayEquals("LA - reflexion", expectedLA, actualLA, ERROR_EPSILON_VERY_LOW);
 
-        assertArrayEquals(  new double[]{11.69,21.77,28.93,32.71,36.83,36.83,32.12,13.66},L, ERROR_EPSILON_LOW);
+        assertArrayEquals(  new double[]{11.69,21.77,28.93,32.71,36.83,36.83,32.12,13.66},L, ERROR_EPSILON_VERY_LOW);
     }
 
     /**
@@ -5081,6 +5081,7 @@ public class AttenuationCnossosTest {
         ProfileBuilder profileBuilder = new ProfileBuilder();
 
         addTopographicTC5Model(profileBuilder);
+        addGroundAttenuationTC5(profileBuilder);
 
         profileBuilder.finishFeeding();
 
@@ -5175,7 +5176,6 @@ public class AttenuationCnossosTest {
     /**
      * TC21 - Building on ground with spatially varying heights and acoustic properties
      */
-    // ToDo erreur potentielle Cnossos : chemin direct
     @Test
     public void TC21() throws IOException {
         //Profile building
@@ -5191,8 +5191,8 @@ public class AttenuationCnossosTest {
[…]

[…]
