### 2024-11-27 3b9b753
fix retrodiff on favorable condition with horizontal diffraction, TC18 pass with < 0.1 dB
