+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/PathFinderTest.java
@@ -14,13 +14,11 @@ import org.cts.op.CoordinateOperationException;
 import org.junit.After;
 import org.junit.Test;
 import org.locationtech.jts.algorithm.CGAlgorithms3D;
-import org.locationtech.jts.geom.Coordinate;
-import org.locationtech.jts.geom.Envelope;
-import org.locationtech.jts.geom.GeometryFactory;
-import org.locationtech.jts.geom.LineSegment;
+import org.locationtech.jts.geom.*;
 import org.locationtech.jts.math.Plane3D;
 import org.locationtech.jts.math.Vector2D;
 import org.locationtech.jts.math.Vector3D;
+import org.locationtech.jts.operation.distance.DistanceOp;
 import org.locationtech.jts.operation.distance3d.PlanarPolygon3D;
 import org.noise_planet.noisemodelling.pathfinder.cnossos.CnossosPath;
 //import org.noise_planet.noisemodelling.pathfinder.path.CnossosPathParameters;
@@ -1753,12 +1751,14 @@ public class PathFinderTest {
     @Test
     public void TC21()  {
         //Profile building
+
+        // the rounding of the unit test input data lead to errors. We had to move two vertex to match with the expected intersection
         ProfileBuilder profileBuilder = new ProfileBuilder()
                 .addBuilding(new Coordinate[]{
                         new Coordinate(167.2, 39.5, 11.5),
-                        new Coordinate(151.6, 48.5, 11.5),
+                        new Coordinate(151.575, 48.524, 11.5),
                         new Coordinate(141.1, 30.3, 11.5),
-                        new Coordinate(156.7, 21.33, 11.5), // the rounding of the unit test input data lead to errors. We had to add 0.03 m to y value
+                        new Coordinate(156.657, 21.3409, 11.5),
                         new Coordinate(159.7, 26.5, 11.5),
                         new Coordinate(151.0, 31.5, 11.5),
                         new Coordinate(155.5, 39.3, 11.5),
@@ -1770,6 +1770,16 @@ public class PathFinderTest {
 
         profileBuilder.setzBuildings(true);
         profileBuilder.finishFeeding();
+
+        //
+        //Coordinate source = new Coordinate(10, 10, 1);
+        //Coordinate receiver = new Coordinate(200, 25, 14);
+        //
+        //GeometryFactory gf = new GeometryFactory();
+        //Geometry g1 = gf.createPoint(source).buffer(147.10, 128);
+        //Geometry g2 = gf.createPoint(receiver).buffer(43.49, 128);
+        //Coordinate[] pts = DistanceOp.nearestPoints(g1, g2);
+        //System.out.println(gf.createMultiPointFromCoords(pts));
 
         //Propagation data building
         Scene rayData = new ProfileBuilderDecorator(profileBuilder)
@@ -1788,6 +1798,8 @@ public class PathFinderTest {
 
         //Run computation
         computeRays.run(propDataOut);
+
+        assertEquals(3, propDataOut.getPropagationPaths().size());
 
         // Test R-CRIT table 235
         Coordinate D = propDataOut.getPropagationPaths().get(0).getSegmentList().get(0).r;
@@ -1820,6 +1832,8 @@ public class PathFinderTest {
             new Coordinate(175.54, 10),
             new Coordinate(190.59, 10));
 
+
+        /* Table 225 */
         List<Coordinate> expectedZProfileSO = Arrays.asList(
                 new Coordinate(0.0, 0.0),
                 new Coordinate(110.34, 0.0),
@@ -1831,6 +1845,20 @@ public class PathFinderTest {
                 new Coordinate(147.26, 5.66),
[…]
