+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -5178,17 +5178,19 @@ public class AttenuationCnossosTest {
 
     /**
      * TC21 - Building on ground with spatially varying heights and acoustic properties
+     * ERROR_EPSILON_MEDIUM because we add favorable contribution for left and right paths but not in the reference document
      */
     @Test
     public void TC21() throws IOException {
         //Profile building
+        // the rounding of the unit test input data lead to errors. We had to move two vertex to match with the expected intersection
         ProfileBuilder profileBuilder = new ProfileBuilder()
                 .setzBuildings(true)
                 .addBuilding(new Coordinate[]{
                         new Coordinate(167.2, 39.5, 11.5),
-                        new Coordinate(151.6, 48.5, 11.5),
+                        new Coordinate(151.575, 48.524, 11.5),
                         new Coordinate(141.1, 30.3, 11.5),
-                        new Coordinate(156.7, 21.3, 11.5),
+                        new Coordinate(156.657, 21.3409, 11.5),
                         new Coordinate(159.7, 26.5, 11.5),
                         new Coordinate(151.0, 31.5, 11.5),
                         new Coordinate(155.5, 39.3, 11.5),
@@ -5296,8 +5298,6 @@ public class AttenuationCnossosTest {
         double[] actualLF = addArray(proPath.aGlobalF, SOUND_POWER_LEVELS);
         double[] actualL = addArray(proPath.aGlobal, SOUND_POWER_LEVELS);
         double[] actualLA = addArray(actualL, A_WEIGHTING);
-        double[] directLA = actualLA;
-        double[] diffLVertical = diffArray(expectedL, actualL);
 
         assertDoubleArrayEquals("WH - vertical plane", expectedWH, actualWH, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("CfH - vertical plane", expectedCfH, actualCfH, ERROR_EPSILON_LOWEST);
@@ -5347,12 +5347,50 @@ public class AttenuationCnossosTest {
         expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
         expectedAAtm = new double[]{0.02, 0.08, 0.20, 0.37, 0.70, 1.85, 6.26, 22.33};
         expectedADiv = new double[]{56.62, 56.62, 56.62, 56.62, 56.62, 56.62, 56.62, 56.62};
-        expectedABoundaryH = new double[]{18.85, 21.74, 24.71, 27.70, 30.70, 33.70, 36.71, 39.72};
-        expectedABoundaryF = new double[]{NaN, NaN, NaN, NaN, NaN, NaN, NaN, NaN};
-        expectedLH = new double[]{18.62, 15.68, 12.48, 9.08, 6.07, 1.86, -5.79, -25.71};
-        expectedLF = new double[]{NaN, NaN, NaN, NaN, NaN, NaN, NaN, NaN};
-        expectedL = new double[]{18.62, 15.68, 12.48, 9.08, 6.07, 1.86, -5.79, -25.71};
+        expectedLH = new double[]{32.63, 32.56, 32.43, 32.22, 31.82, 30.53, 25.85, 9.29};
         expectedLA = new double[]{3.42, 13.45, 20.82, 26.01, 28.81, 28.72, 23.84, 5.18};
+
[…]
