### 2024-11-27 369994a
TC26 pass < 0.5 dB error due to favorable contribution on reflexion path (ignored/skipped in test case reference)
