+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -6957,6 +6957,7 @@ public class AttenuationCnossosTest {
 
     /**
      * TC26 – Road source with influence of retrodiffraction
+     * Issue we compute and add favorable contribution, on reflexion path but not in test case reference
      * */
     @Test
     public void TC26() throws IOException {
@@ -6970,10 +6971,8 @@ public class AttenuationCnossosTest {
                         new Coordinate(74.0, 52.0, 6),
                         new Coordinate(130.0, 60.0, 8)}, Arrays.asList(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.5), -1)
 
-               // .addGroundEffect(factory.toGeometry(new Envelope(0, 50, -10, 100)), 0.0)
-                .addGroundEffect(0.0, 50.0, -10.0, 100.0, 0.0)
-                .addGroundEffect(50.0, 150.0, -10.0, 100.0, 0.5)
-               // .addGroundEffect(factory.toGeometry(new Envelope(50, 150, -10, 100)), 0.5)
+                .addGroundEffect(factory.toGeometry(new Envelope(0, 50, -10, 100)), 0.0)
+                .addGroundEffect(factory.toGeometry(new Envelope(50, 150, -10, 100)), 0.5)
                 .setzBuildings(true)
                 .finishFeeding();
 
@@ -6998,6 +6997,8 @@ public class AttenuationCnossosTest {
         PathFinder computeRays = new PathFinder(rayData);
         computeRays.setThreadCount(1);
         computeRays.run(propDataOut);
+
+        assertEquals(2, propDataOut.getPropagationPaths().size());
 
         //Expected values
         //Path0 : vertical plane
@@ -7044,12 +7045,12 @@ public class AttenuationCnossosTest {
         double[] valL = getMaxValeurAbsolue(diffL);
         double[] valLA = getMaxValeurAbsolue(diffLa);
 
-        assertDoubleArrayEquals("WH - lateral left", expectedWH, actualWH, ERROR_EPSILON_LOWEST);
-        assertDoubleArrayEquals("CfH - lateral left", expectedCfH, actualCfH, ERROR_EPSILON_LOWEST);
-        assertDoubleArrayEquals("AGroundH - lateral left", expectedAGroundH, actualAGroundH, ERROR_EPSILON_LOWEST);
-        assertDoubleArrayEquals("WF - lateral left", expectedWF, actualWF, ERROR_EPSILON_LOWEST);
-        assertDoubleArrayEquals("CfF - lateral left", expectedCfF, actualCfF, ERROR_EPSILON_VERY_LOW);
-        assertDoubleArrayEquals("AGroundF - lateral left", expectedAGroundF, actualAGroundF, ERROR_EPSILON_VERY_LOW);
+        assertDoubleArrayEquals("WH - vertical plane", expectedWH, actualWH, ERROR_EPSILON_LOWEST);
+        assertDoubleArrayEquals("CfH - vertical plane", expectedCfH, actualCfH, ERROR_EPSILON_LOWEST);
+        assertDoubleArrayEquals("AGroundH - vertical plane", expectedAGroundH, actualAGroundH, ERROR_EPSILON_LOWEST);
+        assertDoubleArrayEquals("WF - vertical plane", expectedWF, actualWF, ERROR_EPSILON_LOWEST);
+        assertDoubleArrayEquals("CfF - vertical plane", expectedCfF, actualCfF, ERROR_EPSILON_VERY_LOW);
+        assertDoubleArrayEquals("AGroundF - vertical plane", expectedAGroundF, actualAGroundF, ERROR_EPSILON_VERY_LOW);
[…]
