### 2024-11-27 32dfd18
TC27 aGroundF issues..
