### 2024-11-28 296bb6a
add new tests TC27
