### 2024-11-28 d7c1d59
TC28 15 dB error, issue with favorable path
