### 2024-11-28 0a5d05f
Fix source line split when there is no topo
