### 2024-11-28 da402a0
fix sql for postgres, table location issues with postgis connection
