+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapLoader.java
@@ -16,6 +16,7 @@ import org.h2gis.utilities.SpatialResultSet;
 import org.h2gis.utilities.TableLocation;
 import org.h2gis.utilities.dbtypes.DBTypes;
 import org.h2gis.utilities.dbtypes.DBUtils;
+import org.h2gis.utilities.wrapper.ConnectionWrapper;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.geom.prep.PreparedPolygon;
 import org.locationtech.jts.io.WKTWriter;
@@ -287,16 +288,17 @@ public abstract class NoiseMapLoader {
      */
     protected void fetchCellDem(Connection connection, Envelope fetchEnvelope, ProfileBuilder mesh) throws SQLException {
         if(!demTable.isEmpty()) {
+            DBTypes dbType = DBUtils.getDBType(connection.unwrap(Connection.class));
             List<String> geomFields = getGeometryColumnNames(connection,
-                    TableLocation.parse(demTable));
+                    TableLocation.parse(demTable, dbType));
             if(geomFields.isEmpty()) {
                 throw new SQLException("Digital elevation model table \""+demTable+"\" must exist and contain a POINT field");
             }
             String topoGeomName = geomFields.get(0);
             try (PreparedStatement st = connection.prepareStatement(
-                    "SELECT " + TableLocation.quoteIdentifier(topoGeomName) + " FROM " +
+                    "SELECT " + TableLocation.quoteIdentifier(topoGeomName, dbType) + " FROM " +
                             demTable + " WHERE " +
-                            TableLocation.quoteIdentifier(topoGeomName) + " && ?::geometry")) {
+                            TableLocation.quoteIdentifier(topoGeomName, dbType) + " && ?::geometry")) {
                 st.setObject(1, geometryFactory.toGeometry(fetchEnvelope));
                 try (SpatialResultSet rs = st.executeQuery().unwrap(SpatialResultSet.class)) {
                     while (rs.next()) {
@@ -320,14 +322,15 @@ public abstract class NoiseMapLoader {
     protected void fetchCellSoilAreas(Connection connection, Envelope fetchEnvelope, ProfileBuilder builder)
             throws SQLException {
         if(!soilTableName.isEmpty()){
+            DBTypes dbType = DBUtils.getDBType(connection.unwrap(Connection.class));
             double startX = Math.floor(fetchEnvelope.getMinX() / groundSurfaceSplitSideLength) * groundSurfaceSplitSideLength;
             double startY = Math.floor(fetchEnvelope.getMinY() / groundSurfaceSplitSideLength) * groundSurfaceSplitSideLength;
             String soilGeomName = getGeometryColumnNames(connection,
-                    TableLocation.parse(soilTableName)).get(0);
+                    TableLocation.parse(soilTableName, dbType)).get(0);
             try (PreparedStatement st = connection.prepareStatement(
-                    "SELECT " + TableLocation.quoteIdentifier(soilGeomName) + ", G FROM " +
+                    "SELECT " + TableLocation.quoteIdentifier(soilGeomName, dbType) + ", G FROM " +
                             soilTableName + " WHERE " +
-                            TableLocation.quoteIdentifier(soilGeomName) + " && ?::geometry")) {
+                            TableLocation.quoteIdentifier(soilGeomName, dbType) + " && ?::geometry")) {
                 st.setObject(1, geometryFactory.toGeometry(fetchEnvelope));
                 try (SpatialResultSet rs = st.executeQuery().unwrap(SpatialResultSet.class)) {
                     while (rs.next()) {
@@ -400,24 +403,25 @@ public abstract class NoiseMapLoader {
         Geometry envGeo = geometryFactory.toGeometry(fetchEnvelope);
[…]
