### 2024-11-29 21abe60
fix sql for postgres it is working locally
