### 2024-11-29 89c32cf
github action was not catching wps issues of gradle
