+++ .github/workflows/CI.yml
@@ -27,9 +27,12 @@ jobs:
           server-username: MAVEN_USERNAME
           server-password: MAVEN_PASSWORD
       - name: Building
-        run: |
-          mvn test install -B
-          cd wps_scripts && ./gradlew test --info --stacktrace
+        run: mvn test install -B
+      - name: Setup Gradle
+        uses: gradle/actions/setup-gradle@v4
+      - name: Build and test WPS with Gradle
+        run:  ./gradlew build --info --stacktrace
+        working-directory: ./wps_scripts
       - name: Deploy
         env:
           MAVEN_USERNAME: ${{ secrets.OSSRH_USERNAME }}
