### 2024-11-29 3a6ed54
change threading of noisemodelling to be able to see exception when one of the thread has failed
