+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/ThreadPathFinder.java
@@ -13,9 +13,16 @@ import org.h2gis.api.ProgressVisitor;
 import org.noise_planet.noisemodelling.pathfinder.path.PointPath;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.utils.profiler.ReceiverStatsMetric;
+
+import java.util.concurrent.Callable;
+
 import static org.noise_planet.noisemodelling.pathfinder.PathFinder.LOGGER;
 
-public final class ThreadPathFinder implements Runnable {
+/**
+ * A Thread class to evaluate all receivers cut planes.
+ * Return true if the computation is done without issues
+ */
+public final class ThreadPathFinder implements Callable<Boolean> {
     int startReceiver; // Included
     int endReceiver; // Excluded
     PathFinder propagationProcess;
@@ -48,7 +55,7 @@ public final class ThreadPathFinder implements Runnable {
      * Executes the computation of ray paths for each receiver in the specified range.
      */
     @Override
-    public void run() {
+    public Boolean call() throws Exception {
         try {
             for (int idReceiver = startReceiver; idReceiver < endReceiver; idReceiver++) {
                 if (visitor != null) {
@@ -83,5 +90,6 @@ public final class ThreadPathFinder implements Runnable {
             }
             throw ex;
         }
+        return true;
     }
 }
