+++ .github/workflows/CI.yml
@@ -42,7 +42,7 @@ jobs:
       - name: Clean
         run: rm -f $HOME/.gradle/caches/modules-2/modules-2.lock
       - name: Cache
-        uses: actions/cache@v2
+        uses: actions/cache@v4
         with:
           path: |
             $HOME/.m2/
