### 2024-12-02 cac6c5e
replace deprecated version of junit with latest version of junit. simplify dependency xml. Upgrade H2GIS / H2
