### 2024-12-02 ed77ad7
handle issue when converting SRID of a table. Use postgis compatible function to convert srid
