+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -11,7 +11,7 @@ assignees: ''
 A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
 
 **Describe the solution you'd like**
-A clea[Import_OSM.groovy](../../../../data/bench_test/NoiseModellingV5/noisemodelling/wps/Import_and_Export/Import_OSM.groovy)r and concise description of what you want to happen.
+A clear and concise description of what you want to happen.
 
 **Describe alternatives you've considered**
 A clear and concise description of any alternative solutions or features you've considered.
