### 2024-12-03 2143079
add unit test issue with path state
