### 2024-12-04 9ed58d5
move Cnossos attenuation stuff from pathfinder to propagation package
