### 2024-12-05 e3dd0fb
Rework cutpoint to remove attributes on unrelated cutpoints type. Prepare for export in json format
