+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/profilebuilder/ProfileBuilder.java
@@ -907,24 +907,23 @@ public class ProfileBuilder {
      * @return Cutting profile.
      */
     public CutProfile getProfile(Coordinate sourceCoordinate, Coordinate receiverCoordinate, double gS, boolean stopAtObstacleOverSourceReceiver) {
-        CutProfile profile = new CutProfile();
+        CutPointSource sourcePoint  = new CutPointSource(sourceCoordinate);
+        CutPointReceiver receiverPoint = new CutPointReceiver(receiverCoordinate);
+
+        CutProfile profile = new CutProfile(sourcePoint, receiverPoint);
 
         // Add sourceCoordinate
-        CutPoint sourcePoint = profile.addSource(sourceCoordinate);
         int groundAbsorptionIndex = getIntersectingGroundAbsorption(FACTORY.createPoint(sourceCoordinate));
         if(groundAbsorptionIndex >= 0) {
             sourcePoint.setGroundCoefficient(groundAbsorptions.get(groundAbsorptionIndex).getCoefficient());
         } else {
             sourcePoint.setGroundCoefficient(gS);
         }
-        // Add receiver point
-        profile.addReceiver(receiverCoordinate);
 
         //Fetch topography evolution between sourceCoordinate and receiverCoordinate
         if(topoTree != null) {
             addTopoCutPts(sourceCoordinate, receiverCoordinate, profile, stopAtObstacleOverSourceReceiver);
             if(stopAtObstacleOverSourceReceiver && profile.hasTopographyIntersection) {
-                profile.setCutPoints(Arrays.asList(profile.getSource(), profile.getReceiver()));
                 return profile;
             }
         } else {
@@ -937,23 +936,8 @@ public class ProfileBuilder {
             LineSegment fullLine = new LineSegment(sourceCoordinate, receiverCoordinate);
             addGroundBuildingCutPts(fullLine, profile, stopAtObstacleOverSourceReceiver);
             if(stopAtObstacleOverSourceReceiver && profile.hasBuildingIntersection) {
-                profile.setCutPoints(Arrays.asList(profile.getSource(), profile.getReceiver()));
                 return profile;
             }
-        }
-
-        //Sort all the cut point from sourceCoordinate to receiverCoordinate positions
-        profile.sort(sourceCoordinate);
-
-        // Remove all points before source and after receiver
-        // because objects may be at the exact location of receiver or source
-        while(!profile.getCutPoints().isEmpty() &&
-                !profile.getCutPoints().get(0).type.equals(SOURCE)) {
-            profile.getCutPoints().remove(0);
-        }
-        while(!profile.getCutPoints().isEmpty() &&
-                !profile.getCutPoints().get(profile.getCutPoints().size() - 1).type.equals(RECEIVER)) {
-            profile.getCutPoints().remove(profile.getCutPoints().size() - 1);
         }
 
         // Propagate ground coefficient for unknown coefficients
@@ -961,10 +945,11 @@ public class ProfileBuilder {
         for (CutPoint cutPoint : profile.cutPoints) {
             if(Double.isNaN(cutPoint.groundCoefficient)) {
                 cutPoint.setGroundCoefficient(currentCoefficient);
-            } else if (cutPoint.getType().equals(GROUND_EFFECT)) {
+            } else if (cutPoint instanceof CutPointGroundEffect) {
                 currentCoefficient = cutPoint.getGroundCoefficient();
             }
         }
+
         // Compute the interpolation of Z ground for intermediate points
         CutPoint previousZGround = sourcePoint;
         int nextPointIndex = 0;
@@ -1037,9 +1022,12 @@ public class ProfileBuilder {
         // Split line into segments for structures based on RTree in order to limit the number of queries
[…]
