+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/TestPathFinder.java
@@ -14,11 +14,12 @@ import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.io.ParseException;
 import org.locationtech.jts.io.WKTReader;
-import org.noise_planet.noisemodelling.pathfinder.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
 import org.noise_planet.noisemodelling.pathfinder.utils.geometry.JTSUtility;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -29,9 +30,7 @@ import static org.noise_planet.noisemodelling.pathfinder.PathFinder.splitLineStr
 
 
 public class TestPathFinder {
-    //private static final Logger LOGGER = LoggerFactory.getLogger(TestComputeCnossosRays.class);
-    //private static final double EPSILON = 1e0;
-
+    private static final Logger LOGGER = LoggerFactory.getLogger(TestPathFinder.class);
 
     @Test
     public void testMeanPlane() {
@@ -162,179 +161,6 @@ public class TestPathFinder {
     /**
      * Test vertical edge diffraction ray computation
      *
-     * @throws ParseException
-     */
-    @Test
-    public void TestComputeHorizontalEdgeDiffraction() throws ParseException {
-        GeometryFactory factory = new GeometryFactory();
-        WKTReader wktReader = new WKTReader(factory);
-        //Scene dimension
-        Envelope cellEnvelope = new Envelope(new Coordinate(316705, 6706347, 0.), new Coordinate(316828, 6706469, 0.));
-        //Create obstruction test object
-        ProfileBuilder profileBuilder = new ProfileBuilder();
-        profileBuilder.addBuilding(wktReader.read("POLYGON ((316759.81 6706397.4101739, 316759.81 6706403.99033324, 316765.7002829955 6706404.19000385, 316778.88539775345 6706404.489665548, 316777.49275745824 6706400.809116197, 316765.598667453 6706399.209910818, 316765.5966749492 6706399.209431015, 316765.594822107 6706399.208555082, 316765.59318675095 6706399.207319812, 316765.5918375706 6706399.205777088, 316765.59083123534 6706399.203991711, 316765.59021001414 6706399.202038671, 316765.58999999997 6706399.2, 316765.58999999997 6706397.509829072, 316759.81 6706397.4101739))"), 16.68046);
[…]
