+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/TestWallReflection.java
@@ -8,16 +8,20 @@
  */
 package org.noise_planet.noisemodelling.pathfinder;
 
+import org.checkerframework.checker.units.qual.A;
 import org.h2.tools.Csv;
 import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.io.ParseException;
 import org.locationtech.jts.io.WKTReader;
-import org.noise_planet.noisemodelling.pathfinder.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.path.MirrorReceiver;
 import org.noise_planet.noisemodelling.pathfinder.path.MirrorReceiversCompute;
-import org.noise_planet.noisemodelling.pathfinder.path.PointPath;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPoint;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReflection;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.Wall;
 import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
@@ -28,6 +32,7 @@ import java.sql.SQLException;
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class TestWallReflection {
@@ -99,102 +104,108 @@ public class TestWallReflection {
 
         assertEquals(4, mirrorResults.size());
 
-        List<CnossosPath> CnossosPaths = computeRays.computeReflexion(receiver,
-                inputData.sourceGeometries.get(0).getCoordinate(),
-                new Orientation(), receiverMirrorIndex);
+        PathFinderVisitor pathFinderVisitor = new PathFinderVisitor(true, inputData);
 
+        computeRays.computeReflexion(new PathFinder.ReceiverPointInfo(1, receiver),
+                new PathFinder.SourcePointInfo(1, inputData.sourceGeometries.get(0).getCoordinate(), 1.0,
+                new Orientation()), receiverMirrorIndex, pathFinderVisitor);
+
+        List<CutProfile> profiles = new ArrayList<>(pathFinderVisitor.cutProfiles);
         // Only one second order reflexion propagation path must be found
-        assertEquals(1, CnossosPaths.size());
+        assertEquals(1, profiles.size());
+
         // Check expected values for the propagation path
-        CnossosPath firstPath = CnossosPaths.get(0);
-        var it = firstPath.getPointList().iterator();
+        CutProfile firstPath = profiles.get(0);
+        var it = firstPath.cutPoints.iterator();
         assertTrue(it.hasNext());
-        PointPath current = it.next();
-        assertEquals(PointPath.POINT_TYPE.SRCE ,current.type);
-        assertEquals(0.0, current.coordinate.x, 1e-12);
-        current = it.next();
-        assertEquals(PointPath.POINT_TYPE.REFL ,current.type);
-        assertEquals(38.68, current.coordinate.x, 0.02);
-        current = it.next();
-        assertEquals(PointPath.POINT_TYPE.REFL ,current.type);
-        assertEquals(53.28, current.coordinate.x, 0.02);
-        current = it.next();
[…]
