+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/profilebuilder/CutPointGroundEffect.java
@@ -25,4 +25,14 @@ public class CutPointGroundEffect extends CutPoint {
         this.groundCoefficient = groundAbsorptionCoefficient;
         this.processedWallIndex = processedWallIndex;
     }
+
+    @Override
+    public String toString() {
+        return "CutPointGroundEffect{" +
+                "groundCoefficient=" + groundCoefficient +
+                ", zGround=" + zGround +
+                ", coordinate=" + coordinate +
+                ", processedWallIndex=" + processedWallIndex +
+                '}';
+    }
 }
