+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/TestWallReflection.java
@@ -91,17 +91,6 @@ public class TestWallReflection {
         List<MirrorReceiver> mirrorResults = receiverMirrorIndex.findCloseMirrorReceivers(inputData.
                 sourceGeometries.get(0).getCoordinate());
 
-        try {
-            try (FileWriter fileWriter = new FileWriter("target/testNReflexion_testVisibilityCone.csv")) {
-                StringBuilder sb = new StringBuilder();
-                receiverMirrorIndex.exportVisibility(sb, inputData.maxSrcDist, inputData.maxRefDist,
-                        0, mirrorResults, true);
-                fileWriter.write(sb.toString());
-            }
-        } catch (IOException ex) {
-            //ignore
-        }
-
         assertEquals(4, mirrorResults.size());
 
         PathFinderVisitor pathFinderVisitor = new PathFinderVisitor(true, inputData);
@@ -116,6 +105,8 @@ public class TestWallReflection {
 
         // Check expected values for the propagation path
         CutProfile firstPath = profiles.get(0);
+        // S->Ref->Ref->R
+        assertEquals(4, firstPath.cutPoints.size());
         var it = firstPath.cutPoints.iterator();
         assertTrue(it.hasNext());
         CutPoint current = it.next();
@@ -125,87 +116,95 @@ public class TestWallReflection {
                 current.coordinate, 1e-12);
         current = it.next();
         assertInstanceOf(CutPointReflection.class, current);
+        PathFinderTest.assert3DCoordinateEquals("",
+                new Coordinate(599102.81, 646245.83, 2.9), current.coordinate, 0.01);
         current = it.next();
         assertInstanceOf(CutPointReflection.class, current);
-        current = it.next();
-        assertInstanceOf(CutPointReflection.class, current);
+        PathFinderTest.assert3DCoordinateEquals("",
+                new Coordinate(599092.38, 646235.61, 3.61), current.coordinate, 0.01);
         current = it.next();
         assertInstanceOf(CutPointReceiver.class, current);
     }
-//
-//    @Test
-//    public void testNReflexionWithDem() throws ParseException, IOException, SQLException {
-//        GeometryFactory factory = new GeometryFactory();
-//
-//        //Create profile builder
-//        ProfileBuilder profileBuilder = new ProfileBuilder();
-//        profileBuilder.setzBuildings(false); // building Z is height not altitude
-//        Csv csv = new Csv();
-//        WKTReader wktReader = new WKTReader();
-//        try(ResultSet rs = csv.read(new FileReader(
-//                        TestWallReflection.class.getResource("testNReflexionBuildings.csv").getFile()),
-//                new String[]{"geom", "id"})) {
-//            assertTrue(rs.next()); //skip column name
-//            while(rs.next()) {
-//                profileBuilder.addBuilding(wktReader.read(rs.getString(1)), 10, rs.getInt(2));
-//            }
-//        }
-//        profileBuilder.addTopographicPoint(new Coordinate(598962.08,646370.83,500.00));
-//        profileBuilder.addTopographicPoint(new Coordinate(599252.92,646370.11,500.00));
-//        profileBuilder.addTopographicPoint(new Coordinate(599254.37,646100.19,500.00));
-//        profileBuilder.addTopographicPoint(new Coordinate(598913.00,646104.52,500.00));
-//        profileBuilder.finishFeeding();
-//        assertEquals(5, profileBuilder.getBuildingCount());
-//        Scene inputData = new Scene(profileBuilder);
-//        inputData.addReceiver(new Coordinate(599093.85,646227.90, 504));
[…]
