### 2024-12-10 dbc0a30
fix json deserialization
