+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/PathFinderTest.java
@@ -12,38 +12,30 @@ package org.noise_planet.noisemodelling.pathfinder;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
-import com.fasterxml.jackson.databind.ser.FilterProvider;
-import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
-import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
-import org.cts.crs.CRSException;
-import org.cts.op.CoordinateOperationException;
 import org.junit.jupiter.api.Test;
 import org.locationtech.jts.algorithm.CGAlgorithms3D;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.math.Vector3D;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPoint;
-import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReflection;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointWall;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilderDecorator;
-import org.noise_planet.noisemodelling.pathfinder.utils.documents.KMLDocument;
+import org.noise_planet.noisemodelling.pathfinder.utils.geometry.CoordinateMixin;
+import org.noise_planet.noisemodelling.pathfinder.utils.geometry.LineSegmentMixin;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.xml.stream.XMLStreamException;
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 import java.util.*;
 
-import static java.lang.Double.NaN;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -69,8 +61,26 @@ public class PathFinderTest {
      */
     public static final double DELTA_PLANES = 0.1;
 
+    private void assertCutProfile(String utName, CutProfile cutProfile) throws IOException {
+        String testCaseFileName = utName + ".json";
+        if(overwriteTestCase) {
+            URL resourcePath = PathFinder.class.getResource("test_cases");
+            if(resourcePath != null) {
+                File destination = new File(resourcePath.getFile(), testCaseFileName);
+                try (FileWriter utFile = new FileWriter(destination)){
+                    utFile.write(cutProfileAsJson(cutProfile));
+                }
+                LOGGER.warn("{} written in \n{}", testCaseFileName, destination);
+            }
+        }
+        assertCutProfile(PathFinder.class.getResourceAsStream("test_cases/"+testCaseFileName),
+                cutProfile);
+    }
+
     public static String cutProfileAsJson(CutProfile cutProfile) throws JsonProcessingException {
         ObjectMapper mapper = new ObjectMapper();
[…]
