### 2024-12-10 aad2eb9
move pathfinder test to AttenuationCnossosTest
