+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/PathFinderTest.java
@@ -426,2219 +426,1157 @@ public class PathFinderTest {
         assertCutProfile("TC07", propDataOut.cutProfiles.getFirst());
     }
 
-//    /**
-//     * Test TC08 -- Flat ground with spatially varying acoustic properties and short barrier
-//     */
-//    @Test
-//    public void TC08() {
-//
-//        GeometryFactory factory = new GeometryFactory();
-//
-//        //Create profile builder
-//        ProfileBuilder profileBuilder = new ProfileBuilder()
-//
-//                // Add building
-//                .addWall(new Coordinate[]{
-//                                new Coordinate(175, 50, 0),
-//                                new Coordinate(190, 10, 0)},
-//                        6, 1)
-//                // Add ground effect
-//                .addGroundEffect(factory.toGeometry(new Envelope(0, 50, -250, 250)), 0.9)
-//                .addGroundEffect(factory.toGeometry(new Envelope(50, 150, -250, 250)), 0.5)
-//                .addGroundEffect(factory.toGeometry(new Envelope(150, 225, -250, 250)), 0.2)
-//
-//                .finishFeeding();
-//
-//
-//        //Propagation data building
-//        Scene rayData = new ProfileBuilderDecorator(profileBuilder)
-//                .addSource(10, 10, 1)
-//                .addReceiver(200, 50, 4)
-//                .setGs(0.9)
-//                .hEdgeDiff(true)
-//                .vEdgeDiff(true)
-//                .build();
-//
-//        //Out and computation settings
-//        PathFinderVisitor propDataOut = new PathFinderVisitor(true);
-//        PathFinder computeRays = new PathFinder(rayData);
-//        computeRays.setThreadCount(1);
-//
-//        //Run computation
-//        computeRays.run(propDataOut);
-//
-//        //Expected values
-//
-//        /*Table 41 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.00, 0.00));
-//        expectedZProfile.add(new Coordinate(170.49, 0.00));
-//        expectedZProfile.add(new Coordinate(170.49, 6.00));
-//        expectedZProfile.add(new Coordinate(170.49, 0.00));
-//        expectedZProfile.add(new Coordinate(194.16, 0.00));
-//
-//        /* Table 42 */
-//        Coordinate expectedSPrime =new Coordinate(0.00,-1.00);
-//        Coordinate expectedRPrime =new Coordinate(194.16,-4.00);
-//
-//        assertEquals(3, propDataOut.getCutPlanes().size());
-//
-//
-//        if(!profileBuilder.getWalls().isEmpty()){
-//            assertMirrorPoint(expectedSPrime,expectedRPrime,propDataOut.getCutPlanes().get(0).getSegmentList().get(0).sPrime,propDataOut.getCutPlanes().get(0).getSegmentList().get(propDataOut.getCutPlanes().get(0).getSegmentList().size()-1).rPrime);
-//        }
-//
-//        /* Table 43 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b    zs    zr      dp    Gp   Gp'
-//                {0.00, 0.00, 1.00, 6.00, 170.49, 0.55, 0.61},
-//                {0.00, 0.00, 6.00, 4.00, 023.68, 0.20,  NaN}
-//        };
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a     b    zs    zr      dp    Gp   Gp'
-//                {0.00, 0.00, 1.00, 4.00, 221.23, 0.46, 0.46}
-//        };
-//        double [][] segmentsMeanPlanes2 = new double[][]{
[…]
