+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -13,12 +13,15 @@ package org.noise_planet.noisemodelling.jdbc;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import org.junit.jupiter.api.Test;
+import org.locationtech.jts.algorithm.CGAlgorithms3D;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.io.ParseException;
 import org.locationtech.jts.io.WKTReader;
 import org.locationtech.jts.math.Vector3D;
+import org.noise_planet.noisemodelling.pathfinder.PathFinderVisitor;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.PathFinder;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.delaunay.LayerDelaunayError;
 import org.noise_planet.noisemodelling.propagation.cnossos.PointPath;
@@ -1961,6 +1964,53 @@ public class AttenuationCnossosTest {
         //Run computation
         computeRays.run(propDataOut);
 
+
+        //Expected values
+
+        /*Table 41 */
+        List<Coordinate> expectedZProfile = new ArrayList<>();
+        expectedZProfile.add(new Coordinate(0.00, 0.00));
+        expectedZProfile.add(new Coordinate(170.49, 0.00));
+        expectedZProfile.add(new Coordinate(170.49, 6.00));
+        expectedZProfile.add(new Coordinate(170.49, 0.00));
+        expectedZProfile.add(new Coordinate(194.16, 0.00));
+
+        /* Table 42 */
+        Coordinate expectedSPrime =new Coordinate(0.00,-1.00);
+        Coordinate expectedRPrime =new Coordinate(194.16,-4.00);
+
+        assertEquals(3, propDataOut.getPropagationPaths().size());
+
+
+        if(!profileBuilder.getWalls().isEmpty()){
+            assertMirrorPoint(expectedSPrime,expectedRPrime,
+                    propDataOut.getPropagationPaths().get(0).getSegmentList().get(0).sPrime,
+                    propDataOut.getPropagationPaths().get(0).getSegmentList().
+                            get(propDataOut.getPropagationPaths().get(0).getSegmentList().size()-1).rPrime);
+        }
+
+        /* Table 43 */
+        double [][] segmentsMeanPlanes0 = new double[][]{
+                //  a     b    zs    zr      dp    Gp   Gp'
+                {0.00, 0.00, 1.00, 6.00, 170.49, 0.55, 0.61},
+                {0.00, 0.00, 6.00, 4.00, 023.68, 0.20,  NaN}
+        };
+        double [][] segmentsMeanPlanes1 = new double[][]{
+                //  a     b    zs    zr      dp    Gp   Gp'
+                {0.00, 0.00, 1.00, 4.00, 221.23, 0.46, 0.46}
+        };
+        double [][] segmentsMeanPlanes2 = new double[][]{
+                //  a     b    zs    zr      dp    Gp   Gp'
+                {0.00, 0.00, 1.00, 4.00, 194.78, 0.51, 0.51}
+        };
+
+        //Assertion
+
+        assertZProfil(expectedZProfile, propDataOut.getPropagationPaths().get(0).getCutProfile().computePts2DGround());
+        assertPlanes(segmentsMeanPlanes0, propDataOut.getPropagationPaths().get(0).getSegmentList());
[…]
