### 2024-12-10 d56904a
complete pathfinder test with json
