### 2024-12-10 d25636c
fix code, now need to remove duplicate points from CutProfile and move it to computePtsGround
