+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -198,7 +198,7 @@ public class AttenuationCnossosTest {
         attData.setTemperature(TEMPERATURE);
 
         //Run computation
-        Attenuation propDataOut0 = new Attenuation(true, true, attData);
+        Attenuation propDataOut0 = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays0 = new PathFinder(rayData);
         computeRays0.setThreadCount(1);
         computeRays0.run(propDataOut0);
@@ -206,7 +206,7 @@ public class AttenuationCnossosTest {
 
         // Barrier, no interaction
         rayData.setBodyBarrier(false);
-        Attenuation propDataOut1 = new Attenuation(true, true, attData);
+        Attenuation propDataOut1 = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays1 = new PathFinder(rayData);
         computeRays1.setThreadCount(1);
         computeRays1.run(propDataOut1);
@@ -235,7 +235,7 @@ public class AttenuationCnossosTest {
         rayData2.setComputeVerticalDiffraction(true);
         rayData2.setBodyBarrier(true);
 
-        Attenuation propDataOut2 = new Attenuation(true, true, attData);
+        Attenuation propDataOut2 = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays2 = new PathFinder(rayData2);
         computeRays2.run(propDataOut2);
         double[] values2 = propDataOut2.receiversAttenuationLevels.pop().value;
@@ -258,7 +258,7 @@ public class AttenuationCnossosTest {
         rayData3.setComputeHorizontalDiffraction(false);
         rayData3.setComputeVerticalDiffraction(true);
         rayData3.setBodyBarrier(false);
-        Attenuation propDataOut3 = new Attenuation(true, true, attData);
+        Attenuation propDataOut3 = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays3 = new PathFinder(rayData3);
         computeRays3.run(propDataOut3);
         double[] values3 = propDataOut3.receiversAttenuationLevels.pop().value;
@@ -315,14 +315,14 @@ public class AttenuationCnossosTest {
         attData.setTemperature(TEMPERATURE);
 
         //Run computation
-        Attenuation propDataOut0 = new Attenuation(true, true, attData);
+        Attenuation propDataOut0 = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays0 = new PathFinder(rayData);
         computeRays0.setThreadCount(1);
         rayData.reflexionOrder=0;
         computeRays0.run(propDataOut0);
         double[] values0 = propDataOut0.receiversAttenuationLevels.pop().value;
 
-        Attenuation propDataOut1 = new Attenuation(true, true, attData);
+        Attenuation propDataOut1 = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays1 = new PathFinder(rayData);
         computeRays1.setThreadCount(1);
         rayData.reflexionOrder=1;
@@ -376,7 +376,7 @@ public class AttenuationCnossosTest {
         attData.setTemperature(TEMPERATURE);
 
         //Out and computation settings
-        Attenuation propDataOut = new Attenuation(true, true, attData);
+        Attenuation propDataOut = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays = new PathFinder(rayData);
         computeRays.setThreadCount(1);
 
@@ -427,7 +427,7 @@ public class AttenuationCnossosTest {
         attData.setTemperature(TEMPERATURE);
 
         //Out and computation settings
-        Attenuation propDataOut = new Attenuation(true, true, attData);
+        Attenuation propDataOut = new Attenuation(true, true, attData, rayData);
         PathFinder computeRays = new PathFinder(rayData);
[…]
