+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/CnossosPathBuilder.java
@@ -7,12 +7,15 @@ import org.locationtech.jts.geom.*;
 import org.locationtech.jts.math.Vector2D;
 import org.locationtech.jts.math.Vector3D;
 import org.locationtech.jts.triangulate.quadedge.Vertex;
-import org.noise_planet.noisemodelling.pathfinder.PathFinder;
-import org.noise_planet.noisemodelling.pathfinder.path.MirrorReceiver;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPoint;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReflection;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointTopography;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointVEdgeDiffraction;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointWall;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
-import org.noise_planet.noisemodelling.pathfinder.profilebuilder.Wall;
 import org.noise_planet.noisemodelling.pathfinder.utils.geometry.JTSUtility;
 import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
 
@@ -24,8 +27,6 @@ import java.util.stream.Collectors;
 import static java.lang.Math.*;
 import static java.lang.Math.max;
 import static org.noise_planet.noisemodelling.propagation.cnossos.PointPath.POINT_TYPE.*;
-import static org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder.IntersectionType.*;
-import static org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder.IntersectionType.V_EDGE_DIFFRACTION;
 import static org.noise_planet.noisemodelling.pathfinder.utils.geometry.GeometryUtils.projectPointOnLine;
 
 /**
@@ -33,12 +34,13 @@ import static org.noise_planet.noisemodelling.pathfinder.utils.geometry.Geometry
  */
 public class CnossosPathBuilder {
     public static final double ALPHA0 = 2e-4;
+    private static final double EPSILON = 1e-7;
 
     public static void computeRayleighDiff(SegmentPath srSeg, CutProfile cutProfile, CnossosPath pathParameters,
                                      LineSegment dSR, List<SegmentPath> segments, List<PointPath> points,
                                      List<Coordinate> pts2D, Coordinate[] pts2DGround, List<Integer> cut2DGroundIndex,
                                            List<Integer> frequencyTable) {
-        final List<CutPoint> cuts = cutProfile.getCutPoints();
+        final List<CutPoint> cuts = cutProfile.cutPoints;
 
         Coordinate src = pts2D.get(0);
         Coordinate rcv = pts2D.get(pts2D.size() - 1);
@@ -198,7 +200,7 @@ public class CnossosPathBuilder {
      * @param bodyBarrier
      * @return The cnossos path or null
      */
-    public static CnossosPath computeHEdgeDiffraction(CutProfile cutProfile , boolean bodyBarrier, List<Integer> frequencyTable) {
+    public static CnossosPath computeAttenuationFromCutProfile(CutProfile cutProfile , boolean bodyBarrier, List<Integer> frequencyTable, double gS) {
[…]
