+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/Utils.java
@@ -137,11 +137,6 @@ public class Utils {
             }
             wjSources.add(sl);
         }
-
-        @Override
-        public double[] getMaximalSourcePower(int sourceId) {
-            return wjSources.get(sourceId);
-        }
     }
 
 
