### 2024-12-10 3e4a506
insert primary key in cut profile for wall and buildings
