### 2024-12-11 1fa1e7f
save wall/building primary key
