+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/PathFinder.java
@@ -597,6 +597,9 @@ public class PathFinder {
     private void insertReflectionPointAttributes(CutPoint sourceOrReceiverPoint, List<CutPoint> mainProfileCutPoints, MirrorReceiver mirrorReceiver) {
         CutPointReflection reflectionPoint = new CutPointReflection(sourceOrReceiverPoint,
                 mirrorReceiver.getWall().getLineSegment(), mirrorReceiver.getWall().getAlphas());
+        if(mirrorReceiver.wall.primaryKey >= 0) {
+            reflectionPoint.wallPk = mirrorReceiver.wall.primaryKey;
+        }
         mainProfileCutPoints.add(reflectionPoint);
     }
 
