### 2024-12-11 dac6271
do not duplicate wall and reflection point anymore
