### 2024-12-11 7506f38
simplify function signature
