+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/AttenuationVisitor.java
@@ -11,6 +11,8 @@ package org.noise_planet.noisemodelling.propagation;
 
 import org.noise_planet.noisemodelling.pathfinder.IComputePathsOut;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
@@ -45,39 +47,33 @@ public class AttenuationVisitor implements IComputePathsOut {
         CnossosPath cnossosPath = CnossosPathBuilder.computeAttenuationFromCutProfile(cutProfile, scene.isBodyBarrier(),
                 scene.freq_lvl, scene.gS);
         if(cnossosPath != null) {
-            addPropagationPaths(cutProfile.getSource().id, cutProfile.getSource().li, cutProfile.getReceiver().id,
-                    Collections.singletonList(cnossosPath));
+            addPropagationPaths(cutProfile.getSource(), cutProfile.getReceiver(), Collections.singletonList(cnossosPath));
         }
         return PathSearchStrategy.CONTINUE;
     }
 
     /**
      * Get propagation path result
-     * @param sourceId Source identifier
-     * @param sourceLi Source power per meter coefficient
+     * @param source Source identifier
+     * @param receiver Receiver identifier
      * @param path Propagation path result
      */
-    public double[] addPropagationPaths(long sourceId, double sourceLi, long receiverId, List<CnossosPath> path) {
-        double[] aGlobalMeteo = multiThreadParent.computeCnossosAttenuation(attenuationCnossosParameters, sourceId, sourceLi, receiverId, path);
+    public double[] addPropagationPaths(CutPointSource source, CutPointReceiver receiver, List<CnossosPath> path) {
+        double[] aGlobalMeteo = multiThreadParent.computeCnossosAttenuation(attenuationCnossosParameters, source.id, source.li, path);
         multiThreadParent.rayCount.addAndGet(path.size());
         if(keepRays) {
-            if(multiThreadParent.inputData != null && sourceId < multiThreadParent.inputData.sourcesPk.size() &&
-                    receiverId < multiThreadParent.inputData.receiversPk.size()) {
-                for(CnossosPath pathParameter : path) {
-                    // Copy path content in order to keep original ids for other method calls
-                    //CnossosPathParameters pathParametersPk = new CnossosPathParameters(pathParameter);
-                    pathParameter.setIdReceiver(multiThreadParent.inputData.receiversPk.get((int)receiverId).intValue());
-                    pathParameter.setIdSource(multiThreadParent.inputData.sourcesPk.get((int)sourceId).intValue());
-                    pathParameter.setSourceOrientation(pathParameter.getSourceOrientation());
-                    pathParameter.setGs(pathParameter.getGs());
-                    pathParameters.add(pathParameter);
-                }
-            } else {
-                pathParameters.addAll(path);
+            for(CnossosPath pathParameter : path) {
+                // Copy path content in order to keep original ids for other method calls
+                //CnossosPathParameters pathParametersPk = new CnossosPathParameters(pathParameter);
+                pathParameter.setIdReceiver(receiver.receiverPk == -1 ? receiver.id : receiver.receiverPk);
+                pathParameter.setIdSource(source.sourcePk == -1 ? source.id : source.sourcePk);
[…]
