### 2024-12-12 b0f1538
fix attenuation
