+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/Attenuation.java
@@ -77,8 +77,8 @@ public class Attenuation implements IComputePathsOut {
      * No more propagation paths will be pushed for this receiver identifier
      * @param receiverId
      */
-
-    public void finalizeReceiver(long receiverId) {
+    @Override
+    public void finalizeReceiver(int receiverId) {
 
     }
 
@@ -100,8 +100,8 @@ public class Attenuation implements IComputePathsOut {
 
     /**
      * Get propagation path result
-     * @param sourceId Source identifier
-     * @param sourceLi Source power per meter coefficient
+     * @param source Source identifier
+     * @param receiver receiver identifier
      * @param path Propagation path result
      */
     public double[] addPropagationPaths(CutPointSource source, CutPointReceiver receiver, List<CnossosPath> path) {
@@ -112,8 +112,8 @@ public class Attenuation implements IComputePathsOut {
         }
         double[] aGlobalMeteo = computeCnossosAttenuation(genericMeteoData, source.id, source.li, path);
         if (aGlobalMeteo != null && aGlobalMeteo.length > 0) {
-            receiversAttenuationLevels.add(new SourceReceiverAttenuation(receiver.receiverPk == -1 ? receiver.id : receiver.receiverPk,
-                    source.sourcePk == -1 ? source.id : source.sourcePk, aGlobalMeteo));
+            receiversAttenuationLevels.add(new SourceReceiverAttenuation(receiver,
+                    source, aGlobalMeteo));
             return aGlobalMeteo;
         } else {
             return new double[0];
@@ -125,7 +125,6 @@ public class Attenuation implements IComputePathsOut {
      * @param data
      * @param sourceId
      * @param sourceLi
-     * @param receiverId
      * @param pathParameters
      * @return double list of attenuation
      */
