### 2024-12-13 acd21c9
attenuation cnossos test ok
