+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -17,6 +17,7 @@ import org.locationtech.jts.algorithm.CGAlgorithms3D;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.io.ParseException;
 import org.locationtech.jts.io.WKTReader;
+import org.locationtech.jts.math.Vector2D;
 import org.locationtech.jts.math.Vector3D;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.PathFinder;
@@ -6618,44 +6619,86 @@ public class AttenuationCnossosTest {
 
     }
 
+    private static Coordinate from3DVector(Vector3D vector3D) {
+        return new Coordinate(vector3D.getX(), vector3D.getY(), vector3D.getZ());
+    }
+
+    private static Vector3D to3DVector(Vector2D vector2D) {
+        return new Vector3D(vector2D.getX(), vector2D.getY(), 0);
+    }
+
+    /**
+     * Move linesegment to match the expected distance from source
+     * @param sourceReceiver
+     * @param segmentToMove
+     * @param expectedDistance
+     */
+    private static void fixLineSegment(LineSegment sourceReceiver, LineSegment segmentToMove, double expectedDistance) {
+        Coordinate[] closestPoints = sourceReceiver.closestPoints(segmentToMove);
+        // create a translation vector to fix the distance
+        Vector3D fixVector = to3DVector(Vector2D.create(sourceReceiver.p0, sourceReceiver.p1).normalize()
+                .multiply(expectedDistance-closestPoints[0].distance(sourceReceiver.p0)));
+        segmentToMove.p0 = from3DVector(Vector3D.create(segmentToMove.p0).add(fixVector));
+        segmentToMove.p1 = from3DVector(Vector3D.create(segmentToMove.p1).add(fixVector));
+    }
+
+    public static void makeParallel(LineSegment reference, LineSegment toEdit) {
+        // edit second point position in order to have the second line parallel to the first line
+        toEdit.p1 =
+                Vector2D.create(reference.p0, reference.p1).normalize()
+                        .multiply(toEdit.getLength())
+                        .add(Vector2D.create(toEdit.p0)).toCoordinate();
+        toEdit.p1.z = toEdit.p0.z;
+    }
 
     public static void addTopographicTC23Model(ProfileBuilder profileBuilder) {
         // Create parallel lines for the slope edge because unit test table values are rounded and
-        // the rounding make the lines non-parallel
+        // the rounding make the lines non-parallel and at the wrong distance
 
-        // base line left
-        Vector3D v1 = new Vector3D(new Coordinate(46.27, 36.28, 0),
+        // we will use expected distance on Z Profile to construct the lines
+        Coordinate source = new Coordinate(38, 14, 1);
+        Coordinate receiver = new Coordinate(107, 25.95, 4);
+        LineSegment sourceReceiver = new LineSegment(source, receiver);
+        Double[] expectedDistance = new Double[]{14.21, 22.64, 23.98, 32.3};
+
+        // base line for constructing expected results
+        LineSegment leftTopographicVerticalLine = new LineSegment(new Coordinate(46.27, 36.28, 0),
[…]
