### 2024-12-13 ab7ef91
attenuation cnossos test ok, fix pk and index
