### 2024-12-13 3569a70
Fix failed merge with geometry in output tables
