+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapParameters.java
@@ -69,6 +69,28 @@ public class NoiseMapParameters {
     Boolean sqlOutputFileCompression = true;
     Boolean dropResultsTable = true;
     public boolean computeLAEQOnly = false;
+
+    /**
+     * If true the position of the receiver (with the altitude if available) will be exported into the results tables
+     */
+    boolean exportReceiverPosition = false;
+
+    /**
+     * @return If true the position of the receiver (with the altitude if available) will be exported into the results
+     * tables
+     */
+    public boolean isExportReceiverPosition() {
+        return exportReceiverPosition;
+    }
+
+    /**
+     * @param exportReceiverPosition If true the position of the receiver (with the altitude if available) will be
+     *                               exported into the results tables
+     */
+    public void setExportReceiverPosition(boolean exportReceiverPosition) {
+        this.exportReceiverPosition = exportReceiverPosition;
+    }
+
     /**
      * @param maximumRaysOutputCount if export rays, do not keep more than this number of rays per computation area (0 infinite)
      */
