### 2025-01-06 1ae408e
fix coordinate and srid in output tables
