### 2025-01-06 2b38db9
always fetch srid in NoiseMapMaker
