### 2025-01-06 be6347b
do not look for path in the unit test of attenuation computation, use only generated json vertical profiles
