### 2025-01-07 04005aa
prepare file for deviation computation from cnossos tests
