### 2025-01-07 30c6104
read json and evaluate attenuation
