### 2025-01-08 650b0bf
generate main report table
