+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -1,4 +1,3 @@
-
 /**
  * NoiseModelling is a library capable of producing noise maps. It can be freely used either for research and education, as well as by experts in a professional use.
  * <p>
@@ -127,56 +126,6 @@ public class AttenuationCnossosTest {
         }
 
     }
-
-
-    public static double[] twoDgtAftrComma(double[] valeurs) {
-        return Arrays.stream(valeurs)
-                .map(nombre -> Double.parseDouble(String.format(Locale.US, "%.2f", nombre)))
-                .toArray();
-    }
-   /* public static FileWriter file(String fileName) throws IOException {
-        FileWriter writer = new FileWriter(fileName);
-        writer.write("Rapport de Tests\n");
-        writer.write("================\n\n");
-        writer.write(".. list-table:: Liste des tests\n");
-        writer.write("   :widths: 10 20 20 25 30\n\n");
-        writer.write("   * - Test Case\n");
-        writer.write("     - Without lateral diffraction\n");
-        writer.write("       - Yes/No\n");
-        writer.write("     - With lateral diffraction\n");
-        writer.write("       - Yes/No\n");
-        writer.write("     - Largest Deviation\n");
-        writer.write("       - dB / Hz\n");
-        writer.write("     - Details\n");
-
-
-        return writer;
-    }
-
-    private static final FileWriter writer;
-
-    static {
-        try {
-            writer = file("/home/maguettte/IdeaProjects/NoiseModelling/source/Rapport.rst");
-            System.out.println("debut");
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    
-
-    /*@Before
-    public void setUp() {
-        System.out.println("avant");
-        try {
-            writer
-        } catch (IOException e) {
-            System.out.println("Erreur lors de l'ouverture du fichier : " + e.getMessage());
-        }
-    }*/
-
-
 
     /**
      * Test body-barrier effect
@@ -570,7 +519,7 @@ public class AttenuationCnossosTest {
     @Test
     public void TC01() throws IOException {
 
-        Attenuation propDataOut = computeCnossosPath("TC01");
+        Attenuation propDataOut = computeCnossosPath("TC01_Direct");
 
         //Expected values
         double[][][] pts = new double[][][]{
@@ -646,7 +595,7 @@ public class AttenuationCnossosTest {
     public void TC02() throws IOException {
 
         //Out and computation settings
-        Attenuation propDataOut =  computeCnossosPath("TC02");
+        Attenuation propDataOut =  computeCnossosPath("TC02_Direct");
 
         //Expected values
         double[][][] pts = new double[][][]{
@@ -721,7 +670,7 @@ public class AttenuationCnossosTest {
     public void TC03() throws IOException {
 
         //Out and computation settings
-        Attenuation propDataOut =  computeCnossosPath( "TC03");
+        Attenuation propDataOut =  computeCnossosPath( "TC03_Direct");
 
         //Expected values
         double[][][] pts = new double[][][]{
@@ -795,7 +744,7 @@ public class AttenuationCnossosTest {
     public void TC04() throws IOException {
 
[…]
