+++ noisemodelling-tutorial-01/src/main/java/org/noise_planet/nmtutorial01/GenerateReferenceDeviation.java
@@ -22,20 +22,44 @@ import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosPar
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Iterator;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.IntStream;
 
 public class GenerateReferenceDeviation {
-
+    private static final Logger LOGGER = LoggerFactory.getLogger(GenerateReferenceDeviation.class);
+    private static final List<Integer> FREQ_LVL = Arrays.asList(Scene.asOctaveBands(Scene.DEFAULT_FREQUENCIES_THIRD_OCTAVE));
     private static final double[] SOUND_POWER_LEVELS = new double[]{93, 93, 93, 93, 93, 93, 93, 93};
+    private static final double[] A_WEIGHTING = new double[]{-26.2, -16.1, -8.6, -3.2, 0.0, 1.2, 1.0, -1.1};
+
     private static final double HUMIDITY = 70;
     private static final double TEMPERATURE = 10;
+    private static final String CHECKED = "☑";
+    private static final String UNCHECKED = "□";
+    private static final String REPORT_HEADER = "Final results\n" +
+            "================\n" +
+            "\n" +
+            ".. list-table:: Tests list\n" +
+            "   :widths: 10 20 20 25 30\n" +
+            "\n" +
+            "   * - Test Case\n" +
+            "     - Do not the deviate more than ±0,1 dB\n" +
+            "       - Yes/No\n" +
+            "     - Do not the deviate more than ±0,1 dB neglecting lateral diffraction\n" +
+            "       - Yes/No\n" +
+            "     - Largest Deviation\n" +
+            "       - dB / Hz\n" +
+            "     - Details\n";
 
     private static CutProfile loadCutProfile(String utName) throws IOException {
         String testCaseFileName = utName + ".json";
         try(InputStream inputStream = PathFinder.class.getResourceAsStream("test_cases/"+testCaseFileName)) {
+            if(inputStream == null) {
+                throw new IOException("Document " + testCaseFileName + " not found");
+            }
             ObjectMapper mapper = new ObjectMapper();
             return mapper.readValue(inputStream, CutProfile.class);
         }
@@ -89,6 +113,78 @@ public class GenerateReferenceDeviation {
         return doubleArray;
     }
 
+    private static DeviationResult computeDeviation(double[] expected, double[] actual) {
+        assert expected.length == actual.length;
+        assert expected.length == FREQ_LVL.size();
+        int largestDiffIndex = IntStream.range(0, expected.length)
+                .boxed()
+                .max(Comparator.comparingDouble(i -> Math.abs(expected[i] - actual[i])))
+                .orElse(-1);
+        if(largestDiffIndex >= 0) {
+            return new DeviationResult(Math.abs(expected[largestDiffIndex] - actual[largestDiffIndex]),
+                    FREQ_LVL.get(largestDiffIndex));
+        } else {
+            return new DeviationResult(0, 0);
+        }
+    }
+
+    private static void addUTDeviation(String utName, StringBuilder sb, JsonNode expectedValues, Attenuation actual) {
+        double[] expectedLA = asArray(expectedValues.get("LA"));
+        double[] expectedLAWithoutLateral = asArray(expectedValues.get("LA_WL"));
+        double[] actualLA = addArray(SOUND_POWER_LEVELS, addArray(actual.receiversAttenuationLevels.getFirst().value,
+                A_WEIGHTING));
[…]
