### 2025-01-08 baebfce
generate main report table in single file
