+++ noisemodelling-tutorial-01/src/main/java/org/noise_planet/nmtutorial01/GenerateReferenceDeviation.java
@@ -19,6 +19,7 @@ import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilderD
 import org.noise_planet.noisemodelling.propagation.Attenuation;
 import org.noise_planet.noisemodelling.propagation.AttenuationVisitor;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
+import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -39,19 +40,20 @@ public class GenerateReferenceDeviation {
     private static final double TEMPERATURE = 10;
     private static final String CHECKED = "☑";
     private static final String UNCHECKED = "□";
-    private static final String REPORT_HEADER = "Final results\n" +
-            "================\n" +
+    private static final String REPORT_HEADER = "Conformity to ISO 17534-1:2015\n" +
+            "==============================\n" +
             "\n" +
-            ".. list-table:: Tests list\n" +
+            "Conformity table\n" +
+            "^^^^^^^^^^^^^^^^\n" +
+            "| Conform - Do not the deviate more than ±0,1 dB \n" +
+            "| NLD Conform - Do not the deviate more than ±0,1 dB neglecting lateral diffraction\n" +
+            ".. list-table::\n" +
             "   :widths: 10 20 20 25 30\n" +
             "\n" +
             "   * - Test Case\n" +
-            "     - Do not the deviate more than ±0,1 dB\n" +
-            "       - Yes/No\n" +
-            "     - Do not the deviate more than ±0,1 dB neglecting lateral diffraction\n" +
-            "       - Yes/No\n" +
+            "     - Conform ? \n" +
+            "     - NLD Conform ?\n" +
             "     - Largest Deviation\n" +
-            "       - dB / Hz\n" +
             "     - Details\n";
 
     private static CutProfile loadCutProfile(String utName) throws IOException {
@@ -141,23 +143,23 @@ public class GenerateReferenceDeviation {
                 "     - %s\n" +
                 "     - %s\n" +
                 "     - %s\n" +
-                "     - :doc:`Detail <%s>`\n",
+                "     - `%s`_\n",
                 utName,
                 lADeviation.deviation <= 0.1 ? CHECKED : UNCHECKED,
                 lADeviationWithoutLateral.deviation <= 0.1 ? CHECKED : UNCHECKED,
                 lADeviation.deviation > lADeviationWithoutLateral.deviation ?
-                        String.format(Locale.ROOT, "%.2f dB / %d Hz",
+                        String.format(Locale.ROOT, "%.2f dB @ %d Hz",
                                 lADeviation.deviation,
                                 lADeviation.frequency)
                         :
-                        String.format(Locale.ROOT, "%.2f dB / %d Hz",
+                        String.format(Locale.ROOT, "%.2f dB @ %d Hz",
                                 lADeviationWithoutLateral.deviation,
                                 lADeviationWithoutLateral.frequency),
                 utName));
     }
 
-    private static void addUTDeviationDetails(String utName, StringBuilder sb, JsonNode expectedValues, Attenuation actual) {
-        double[] actualLH = addArray(actual.getPropagationPaths().get(0).aGlobalH, SOUND_POWER_LEVELS);
+    private static void addUTDeviationDetails(String utName, StringBuilder sb, JsonNode expectedValues, CnossosPath actual) {
+        double[] actualLH = addArray(actual.aGlobalH, SOUND_POWER_LEVELS);
         double[] expectedLH = asArray(expectedValues.get("LH"));
         DeviationResult lhDeviation = computeDeviation(expectedLH, actualLH);
         sb.append(String.format(Locale.ROOT, "\n\n%s \n" +
[…]
