+++ Docs/TC26_D.rst
@@ -1,52 +1,0 @@
-TC26
-
-Vertical Plane 
-
-================
-
-.. list-table::
-   :widths: 25 25 25
-
-   * - Parameters
-     - Maximum Difference
-     - Frequency
-   * - WH
-     - 0.09
-     - 8000
-   * - CfH
-     - 8.6
-     - 500
-   * - AGroundH
-     - 0.02
-     - 63
-   * - WF
-     - 0.0
-     - 2000
-   * - CfF
-     - 7.69
-     - 250
-   * - AGroundF
-     - 0.38
-     - 4000
-   * - AlphaAtm
-     - 0.0
-     - 2000
-   * - AAtm
-     - 1.69
-     - 8000
-   * - ADiv
-     - 1.22
-     - 63
-   * - ABoundaryH
-     - 11.19
-     - 8000
-   * - ABoundaryF
-     - 6.73
-     - 2000
-
-   * - LH
-     - 9.02
-     - 2000
-   * - LF
-     - 5.38
-     - 2000
