+++ Docs/TC07_D.rst
@@ -1,78 +1,0 @@
-TC07
-================
-
-.. list-table::
-   :widths: 25 25 25
-
-   * - Parameters
-     - Maximum Difference
-     - Frequency
-   * - DeltaDiffSRH
-     - 0.0
-     - 1000
-   * - AGroundSOH
-     - 0.07
-     - 1000
-   * - AGroundORH
-     - 0.0
-     - 63
-   * - DeltaDiffSPrimeRH
-     - 0.0
-     - 250
-   * - DeltaDiffSRPrimeH
-     - 0.0
-     - 500
-   * - DeltaGroundSOH
-     - 0.06
-     - 1000
-   * - DeltaGroundORH
-     - 0.0
-     - 63
-   * - ADiffH
-     - 0.06
-     - 1000
-   * - DeltaDiffSRF
-     - 0.0
-     - 2000
-   * - AGroundSOF
-     - 0.0
-     - 63
-   * - AGroundORF
-     - 0.0
-     - 63
-   * - DeltaDiffSPrimeRF
-     - 0.0
-     - 63
-   * - DeltaDiffSRPrimeF
-     - 0.0
-     - 8000
-   * - DeltaGroundSOF
-     - 0.01
-     - 63
-   * - DeltaGroundORF
-     - 0.0
-     - 125
-   * - ADiffF
-     - 0.0
-     - 8000
-   * - AlphaAtm
-     - 0.0
-     - 2000
-   * - AAtm
-     - 0.0
-     - 500
-   * - ADiv
-     - 0.02
-     - 63
-   * - ABoundaryH
-     - 0.06
-     - 1000
-   * - ABoundaryF
-     - 0.0
-     - 8000
-   * - LH
-     - 0.06
-     - 1000
-   * - LF
-     - 0.0
-     - 250

[…]
