+++ Docs/Cnossos_Report.rst
@@ -1,0 +1,1320 @@
+Conformity to ISO 17534-1:2015
+==============================
+.. This document has been generated with noisemodelling-tutorial-01/src/main/java/org/noise_planet/nmtutorial01/GenerateReferenceDeviation.java
+
+Conformity table
+^^^^^^^^^^^^^^^^
+| Conform - Do not the deviate more than ±0,1 dB 
+| NLD Conform - Do not the deviate more than ±0,1 dB neglecting lateral diffraction
+.. list-table::
+   :widths: 10 20 20 25 30
+
+   * - Test Case
+     - Conform ? 
+     - NLD Conform ?
+     - Largest Deviation
+     - Details
+   * - TC01
+     - ☑
+     - ☑
+     - 0.00 dB @ 2000 Hz
+     - `TC01`_
+   * - TC02
+     - ☑
+     - ☑
+     - 0.07 dB @ 1000 Hz
+     - `TC02`_
+   * - TC03
+     - ☑
+     - ☑
+     - 0.05 dB @ 500 Hz
+     - `TC03`_
+   * - TC04
+     - ☑
+     - ☑
+     - 0.07 dB @ 1000 Hz
+     - `TC04`_
+   * - TC05
+     - ☑
+     - ☑
+     - 0.00 dB @ 1000 Hz
+     - `TC05`_
+   * - TC06
+     - ☑
+     - ☑
+     - 0.06 dB @ 500 Hz
+     - `TC06`_
+   * - TC07
+     - ☑
+     - ☑
+     - 0.02 dB @ 1000 Hz
+     - `TC07`_
+   * - TC08
+     - ☑
+     - ☑
+     - 0.02 dB @ 1000 Hz
+     - `TC08`_
+   * - TC09
+     - ☑
+     - ☑
+     - 0.01 dB @ 250 Hz
+     - `TC09`_
+   * - TC10
+     - ☑
+     - ☑
+     - 0.02 dB @ 250 Hz
+     - `TC10`_
+   * - TC11
+     - ☑
+     - ☑
+     - 0.02 dB @ 125 Hz
+     - `TC11`_
+   * - TC12
+     - ☑
+     - ☑
+     - 0.04 dB @ 500 Hz
+     - `TC12`_
+   * - TC13
+     - ☑
+     - ☑
+     - 0.01 dB @ 500 Hz
+     - `TC13`_
+   * - TC14
+     - ☑
+     - ☑
+     - 0.04 dB @ 8000 Hz
+     - `TC14`_
+   * - TC15
+     - ☑
+     - ☑
+     - 0.06 dB @ 250 Hz
+     - `TC15`_
+   * - TC16
+     - ☑
+     - ☑
+     - 0.00 dB @ 2000 Hz
+     - `TC16`_
+   * - TC17
+     - ☑
+     - ☑
+     - 0.05 dB @ 500 Hz
+     - `TC17`_
+   * - TC18
+     - ☑
+     - ☑
+     - 0.02 dB @ 500 Hz
+     - `TC18`_
+   * - TC19
+     - □
+     - ☑
+     - 0.77 dB @ 250 Hz
+     - `TC19`_
+   * - TC20
+     - ☑
+     - ☑
+     - 0.00 dB @ 8000 Hz
+     - `TC20`_
+   * - TC21
+     - □
+     - ☑
+     - 0.81 dB @ 63 Hz
+     - `TC21`_
+   * - TC22
+     - ☑
+     - ☑
+     - 0.00 dB @ 125 Hz
+     - `TC22`_
+   * - TC23
+     - □
+     - □
+     - 2.13 dB @ 4000 Hz
+     - `TC23`_
+   * - TC24
[…]
