### 2025-01-10 8bc8199
begin work skipping non-significant sources
