### 2025-01-13 f858cba
evaluation of error, todo collect noise levels
