+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/PathFinderVisitor.java
@@ -64,7 +64,7 @@ public class PathFinderVisitor implements IComputePathsOut {
     }
 
     @Override
-    public void startReceiver(Collection<PathFinder.SourcePointInfo> sourceList) {
+    public void startReceiver(PathFinder.ReceiverPointInfo receiver, Collection<PathFinder.SourcePointInfo> sourceList) {
 
     }
 
