+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -63,7 +63,7 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.noise_planet.noisemodelling.jdbc.Utils.*;
 import static org.noise_planet.noisemodelling.pathfinder.path.Scene.DEFAULT_FREQUENCIES_THIRD_OCTAVE;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.dbaToW;
-import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.multArray;
+import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.multiplicationArray;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.sumArray;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.sumDbArray;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.wToDba;
@@ -6155,76 +6155,84 @@ public class AttenuationCnossosTest {
         }
     }
 
+
+    private double testIgnoreNonSignificantSourcesParam(Connection connection, double maxError) throws SQLException, IOException {
+        // Init NoiseModelling
+        NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker("BUILDINGS",
+                "LW_ROADS", "RECEIVERS");
+
+        noiseMapByReceiverMaker.setMaximumPropagationDistance(5000.0);
+        noiseMapByReceiverMaker.setSoundReflectionOrder(1);
+        noiseMapByReceiverMaker.setThreadCount(1);
+        noiseMapByReceiverMaker.setComputeHorizontalDiffraction(true);
+        noiseMapByReceiverMaker.setComputeVerticalDiffraction(true);
+        // Building height field name
+        noiseMapByReceiverMaker.setHeightField("HEIGHT");
+
+
+        // Init custom input in order to compute more than just attenuation
+        // LW_ROADS contain Day Evening Night emission spectrum
+        NoiseMapParameters noiseMapParameters = new NoiseMapParameters(NoiseMapParameters.INPUT_MODE.INPUT_MODE_LW_DEN);
+        noiseMapParameters.setExportRaysMethod(NoiseMapParameters.ExportRaysMethods.TO_MEMORY);
+
+        noiseMapParameters.setComputeLDay(false);
+        noiseMapParameters.setComputeLEvening(false);
+        noiseMapParameters.setComputeLNight(false);
+        noiseMapParameters.setComputeLDEN(true);
+        noiseMapParameters.keepAbsorption = true;
+        noiseMapParameters.setMaximumError(maxError);
+
+        NoiseMapMaker noiseMapMaker = new NoiseMapMaker(connection, noiseMapParameters);
+
+        noiseMapByReceiverMaker.setPropagationProcessDataFactory(noiseMapMaker);
+        noiseMapByReceiverMaker.setComputeRaysOutFactory(noiseMapMaker);
+
+        RootProgressVisitor progressLogger = new RootProgressVisitor(1, true, 1);
+
+        noiseMapByReceiverMaker.initialize(connection, new EmptyProgressVisitor());
+
+        noiseMapParameters.getPropagationProcessPathData(NoiseMapParameters.TIME_PERIOD.DAY).setTemperature(20);
+        noiseMapParameters.getPropagationProcessPathData(NoiseMapParameters.TIME_PERIOD.EVENING).setTemperature(16);
[…]
