### 2025-01-13 1451c83
collect noise levels (non aggregate)
