+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapInStack.java
@@ -75,35 +75,10 @@ public class NoiseMapInStack implements IComputePathsOut {
     double[] sumLevels(List<double[]> wjSources, List<Attenuation.SourceReceiverAttenuation> receiverAttenuationLevels) {
         double[] levels = new double[noiseMapComputeRaysOut.dayPathData.freq_lvl.size()];
         for (Attenuation.SourceReceiverAttenuation lvl : receiverAttenuationLevels) {
-            if(wjSources.size() > lvl.sourceIndex && lvl.sourceIndex >= 0) {
+            if(wjSources.size() > lvl.source.id && lvl.source.id >= 0) {
                 levels = sumArray(levels,
-                        dbaToW(sumArray(wToDba(wjSources.get(lvl.sourceIndex)), lvl.value)));
+                        dbaToW(sumArray(wToDba(wjSources.get(lvl.source.id)), lvl.value)));
             }
-        }
-        return levels;
-    }
-
-
-    /**
-     * Processes the attenuation levels for a receiver and pushes the result into a concurrent linked deque.
-     * @param receiverIndex              the index of the receiver in memory list
-     * @param receiverPK                 the primary key of the receiver.
-     * @param wjSources                  the list of source attenuation levels.
-     * @param receiverAttenuationLevels  the list of attenuation levels from receiver to sources.
-     * @param result                     the concurrent linked deque to push the result into.
-     * @param feedStack                  {@code true} if the result should be pushed into the result stack, {@code false} otherwise.
-     * @return the computed attenuation levels for the receiver.
-     */
-    double[] processAndPushResult(int receiverIndex, long receiverPK, List<double[]> wjSources, List<Attenuation.SourceReceiverAttenuation> receiverAttenuationLevels, ConcurrentLinkedDeque<Attenuation.SourceReceiverAttenuation> result, boolean feedStack) {
-        double[] levels = sumLevels(wjSources, receiverAttenuationLevels);
-        if(feedStack) {
-            pushInStack(result, new Attenuation.SourceReceiverAttenuation(receiverPK,
-                    receiverIndex,
-                    -1,
-                    -1,
-                    wToDba(levels),
-                    noiseMapComputeRaysOut.inputData.receivers.get(receiverIndex)
-            ));
         }
         return levels;
     }
@@ -118,7 +93,8 @@ public class NoiseMapInStack implements IComputePathsOut {
 
     private NoiseMapParameters.TimePeriodParameters computeLdenAttenuation(CnossosPath cnossosPath) {
         NoiseMapParameters.TimePeriodParameters denWAttenuation =
-                new NoiseMapParameters.TimePeriodParameters(new double[0], new double[0], new double[0]);
+                new NoiseMapParameters.TimePeriodParameters(cnossosPath.getCutProfile().getSource(),
+                        cnossosPath.getCutProfile().getReceiver(),new double[0], new double[0], new double[0]);
         CutPointSource source = cnossosPath.getCutProfile().getSource();
         List<CnossosPath> cnossosPaths = Collections.singletonList(cnossosPath);
         if (noiseMapParameters.computeLDay || noiseMapParameters.computeLDEN) {
@@ -146,24 +122,29 @@ public class NoiseMapInStack implements IComputePathsOut {
     }
 
     public static double[] computeLden(NoiseMapParameters.TimePeriodParameters denWAttenuation,
-                                       double[] wjSourcesD, double[] wjSourcesE, double[] wjSourcesN) {
-        double[] level = new double[0];
+                                       double[] wjSourcesD, double[] wjSourcesE, double[] wjSourcesN, NoiseMapParameters.TimePeriodParameters denWLevel) {
+        double[] ldenLevel = new double[0];
[…]
