### 2025-01-13 1c8fff5
aggregate collected noise levels
