### 2025-01-14 f7d79d7
skip next irrelevant noise sources, unit test ok
