+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapInStack.java
@@ -31,7 +31,6 @@ import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicator
 public class NoiseMapInStack implements IComputePathsOut {
     NoiseMap noiseMapComputeRaysOut;
     NoiseMapParameters noiseMapParameters;
-    AttenuationVisitor[] lDENAttenuationVisitor = new AttenuationVisitor[3];
     public List<CnossosPath> pathParameters = new ArrayList<CnossosPath>();
     /**
      * Collected attenuation per receiver
@@ -45,7 +44,7 @@ public class NoiseMapInStack implements IComputePathsOut {
      * Favorable Free Field cumulated global power at receiver, only used to stop looking for far sources
      */
     Map<Integer, Double> maximumWjExpectedSplAtReceiver = new HashMap<>();
-    double sumMaximumWjExpectedSplAtReceiver = 0;
+    double sumMaximumRemainingWjExpectedSplAtReceiver = 0;
     public static final double DAY_RATIO = 12. / 24.;
     public static final double EVENING_RATIO = 4. / 24.;
     public static final double NIGHT_RATIO = 8. / 24.;
@@ -57,13 +56,6 @@ public class NoiseMapInStack implements IComputePathsOut {
     public NoiseMapInStack(NoiseMap multiThreadParent) {
         this.noiseMapComputeRaysOut = multiThreadParent;
         this.noiseMapParameters = multiThreadParent.noiseEmissionMaker.noiseMapParameters;
-        lDENAttenuationVisitor[0] = new AttenuationVisitor(multiThreadParent, multiThreadParent.dayPathData);
-        lDENAttenuationVisitor[1] = new AttenuationVisitor(multiThreadParent, multiThreadParent.eveningPathData);
-        lDENAttenuationVisitor[2] = new AttenuationVisitor(multiThreadParent, multiThreadParent.nightPathData);
-        for (AttenuationVisitor attenuationVisitor : lDENAttenuationVisitor) {
-            attenuationVisitor.keepRays = false;
-        }
-
     }
 
     /**
@@ -182,15 +174,15 @@ public class NoiseMapInStack implements IComputePathsOut {
 
                 double currentLevelAtReceiver = wToDba(sumArray(wjAtReceiver));
                 // replace unknown value of expected power for this source
-                if(maximumWjExpectedSplAtReceiver.containsKey(source.id)) {
-                    sumMaximumWjExpectedSplAtReceiver -= maximumWjExpectedSplAtReceiver.get(source.id);
-                    maximumWjExpectedSplAtReceiver.remove(source.id);
+                int sourceHashCode = source.coordinate.hashCode();
+                if(maximumWjExpectedSplAtReceiver.containsKey(sourceHashCode)) {
+                    sumMaximumRemainingWjExpectedSplAtReceiver -= maximumWjExpectedSplAtReceiver.get(sourceHashCode);
+                    maximumWjExpectedSplAtReceiver.remove(sourceHashCode);
                 }
-                sumMaximumWjExpectedSplAtReceiver += sumArray(lden);
-                double maximumExpectedLevelInDb = wToDba(sumMaximumWjExpectedSplAtReceiver);
+                double maximumExpectedLevelInDb = wToDba(sumArray(wjAtReceiver) + sumMaximumRemainingWjExpectedSplAtReceiver);
                 double dBDiff = maximumExpectedLevelInDb - currentLevelAtReceiver;
                 if (dBDiff < noiseMapParameters.maximumError) {
-                    strategy = PathSearchStrategy.SKIP_RECEIVER;
+                    strategy = PathSearchStrategy.PROCESS_SOURCE_BUT_SKIP_RECEIVER;
                 }
[…]
