### 2025-01-14 d5e3519
fix lden computation
