+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapInStack.java
@@ -9,6 +9,7 @@
 
 package org.noise_planet.noisemodelling.jdbc;
 
+import org.locationtech.jts.geom.Coordinate;
 import org.noise_planet.noisemodelling.pathfinder.IComputePathsOut;
 import org.noise_planet.noisemodelling.pathfinder.PathFinder;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
@@ -18,7 +19,6 @@ import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.noise_planet.noisemodelling.propagation.Attenuation;
-import org.noise_planet.noisemodelling.propagation.AttenuationVisitor;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPathBuilder;
 
 import java.util.*;
@@ -45,9 +45,10 @@ public class NoiseMapInStack implements IComputePathsOut {
      */
     Map<Integer, Double> maximumWjExpectedSplAtReceiver = new HashMap<>();
     double sumMaximumRemainingWjExpectedSplAtReceiver = 0;
+    //
     public static final double DAY_RATIO = 12. / 24.;
-    public static final double EVENING_RATIO = 4. / 24.;
-    public static final double NIGHT_RATIO = 8. / 24.;
+    public static final double EVENING_RATIO = 4. / 24. * dbaToW(5.0);
+    public static final double NIGHT_RATIO = 8. / 24. * dbaToW(10.0);
 
     /**
      * Constructs a NoiseMapInStack object with a multi-threaded parent NoiseMap instance.
@@ -67,9 +68,9 @@ public class NoiseMapInStack implements IComputePathsOut {
     double[] sumLevels(List<double[]> wjSources, List<Attenuation.SourceReceiverAttenuation> receiverAttenuationLevels) {
         double[] levels = new double[noiseMapComputeRaysOut.dayPathData.freq_lvl.size()];
         for (Attenuation.SourceReceiverAttenuation lvl : receiverAttenuationLevels) {
-            if(wjSources.size() > lvl.source.id && lvl.source.id >= 0) {
+            if(wjSources.size() > lvl.source.sourceIndex && lvl.source.sourceIndex >= 0) {
                 levels = sumArray(levels,
-                        dbaToW(sumArray(wToDba(wjSources.get(lvl.source.id)), lvl.value)));
+                        dbaToW(sumArray(wToDba(wjSources.get(lvl.source.sourceIndex)), lvl.value)));
             }
         }
         return levels;
@@ -84,9 +85,11 @@ public class NoiseMapInStack implements IComputePathsOut {
     }
 
     private NoiseMapParameters.TimePeriodParameters computeLdenAttenuation(CnossosPath cnossosPath) {
+        PathFinder.SourcePointInfo sourceInfo = new PathFinder.SourcePointInfo(cnossosPath.getCutProfile().getSource());
+        PathFinder.ReceiverPointInfo receiverInfo = new PathFinder.ReceiverPointInfo(cnossosPath.getCutProfile().getReceiver());
         NoiseMapParameters.TimePeriodParameters denWAttenuation =
-                new NoiseMapParameters.TimePeriodParameters(cnossosPath.getCutProfile().getSource(),
-                        cnossosPath.getCutProfile().getReceiver(),new double[0], new double[0], new double[0]);
+                new NoiseMapParameters.TimePeriodParameters(sourceInfo,
+                        receiverInfo, new double[0], new double[0], new double[0]);
[…]
