+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapParameters.java
@@ -10,6 +10,7 @@
 package org.noise_planet.noisemodelling.jdbc;
 
 
+import org.noise_planet.noisemodelling.pathfinder.PathFinder;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
@@ -303,13 +304,14 @@ public class NoiseMapParameters {
      * representing the noise levels for different time periods.
      */
     public static class TimePeriodParameters {
-        public CutPointSource source = null;
-        public CutPointReceiver receiver = null;
+        public PathFinder.SourcePointInfo source = null;
+        public PathFinder.ReceiverPointInfo receiver = null;
         public double [] dayLevels = new double[0];
         public double [] eveningLevels = new double[0];
         public double [] nightLevels = new double[0];
 
-        public TimePeriodParameters(CutPointSource source, CutPointReceiver receiver, double[] dayLevels, double[] eveningLevels, double[] nightLevels) {
+        public TimePeriodParameters(PathFinder.SourcePointInfo source, PathFinder.ReceiverPointInfo receiver,
+                                    double[] dayLevels, double[] eveningLevels, double[] nightLevels) {
             this.source = source;
             this.receiver = receiver;
             this.dayLevels = dayLevels;
