### 2025-01-14 d99df0a
restore unit test. Check missing wj for day evening or night
