+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapInStack.java
@@ -170,9 +170,9 @@ public class NoiseMapInStack implements IComputePathsOut {
                 // Add power to evaluate potential error if ignoring remaining sources
                 NoiseEmissionMaker noiseEmissionMaker = noiseMapComputeRaysOut.noiseEmissionMaker;
                 double[] lden = computeLden(denWAttenuation,
-                        noiseEmissionMaker.wjSourcesD.get(source.id),
-                        noiseEmissionMaker.wjSourcesE.get(source.id),
-                        noiseEmissionMaker.wjSourcesN.get(source.id), new NoiseMapParameters.TimePeriodParameters());
+                        getSpectrum(noiseEmissionMaker.wjSourcesD, source.id),
+                        getSpectrum(noiseEmissionMaker.wjSourcesE, source.id),
+                        getSpectrum(noiseEmissionMaker.wjSourcesN, source.id), new NoiseMapParameters.TimePeriodParameters());
                 addGlobalReceiverLevel(lden);
 
                 double currentLevelAtReceiver = wToDba(sumArray(wjAtReceiver));
@@ -215,9 +215,9 @@ public class NoiseMapInStack implements IComputePathsOut {
                         scene.freq_lvl, scene.gS);
                 if (cnossosPath != null) {
                     double[] wjReceiver = computeLden(computeLdenAttenuation(cnossosPath),
-                            noiseEmissionMaker.wjSourcesD.get(sourcePointInfo.sourceIndex),
-                            noiseEmissionMaker.wjSourcesE.get(sourcePointInfo.sourceIndex),
-                            noiseEmissionMaker.wjSourcesN.get(sourcePointInfo.sourceIndex),
+                            getSpectrum(noiseEmissionMaker.wjSourcesD, sourcePointInfo.sourceIndex),
+                            getSpectrum(noiseEmissionMaker.wjSourcesE, sourcePointInfo.sourceIndex),
+                            getSpectrum(noiseEmissionMaker.wjSourcesN, sourcePointInfo.sourceIndex),
                             new NoiseMapParameters.TimePeriodParameters());
                     double globalReceiver = sumArray(wjReceiver);
                     sumMaximumRemainingWjExpectedSplAtReceiver += globalReceiver;
@@ -316,10 +316,10 @@ public class NoiseMapInStack implements IComputePathsOut {
     @Override
     public void finalizeReceiver(PathFinder.ReceiverPointInfo receiver) {
         if(!this.pathParameters.isEmpty()) {
-            if(noiseMapParameters.getExportRaysMethod() == org.noise_planet.noisemodelling.jdbc.NoiseMapParameters.ExportRaysMethods.TO_RAYS_TABLE) {
+            if(noiseMapParameters.getExportRaysMethod() == NoiseMapParameters.ExportRaysMethods.TO_RAYS_TABLE) {
                 // Push propagation rays
                 pushInStack(noiseMapComputeRaysOut.attenuatedPaths.rays, this.pathParameters);
-            } else if(noiseMapParameters.getExportRaysMethod() == org.noise_planet.noisemodelling.jdbc.NoiseMapParameters.ExportRaysMethods.TO_MEMORY
+            } else if(noiseMapParameters.getExportRaysMethod() == NoiseMapParameters.ExportRaysMethods.TO_MEMORY
                     && (noiseMapParameters.getMaximumRaysOutputCount() == 0 ||
                     noiseMapComputeRaysOut.propagationPathsSize.get() < noiseMapParameters.getMaximumRaysOutputCount())){
                 int newRaysSize = noiseMapComputeRaysOut.propagationPathsSize.addAndGet(this.pathParameters.size());
@@ -420,106 +420,5 @@ public class NoiseMapInStack implements IComputePathsOut {
             }
         }
         receiverAttenuationPerSource.clear();
[…]
