### 2025-01-15 a0ca227
Add statistics in profiler Add optional parameter to reduce delaunay fence area restore statics profiler, cut count
