### 2025-01-15 7f8a769
fix maximum distance missing on lateral diffraction
