### 2025-01-16 3f41136
fix missing reset of summed power at receiver level for maxdberror algorithm. Add some more profile data. Do not reset strategy if there is no reflection.
