### 2025-01-16 d22912e
Simplify computation of expected noise level at receiver on the dbError preparation step.
