+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/PointPath.java
@@ -39,7 +39,7 @@ public class PointPath {
          */
         SRCE,
         /**
-         * Reflection on > 15° obstacle
+         * Reflection on {@literal >} 15° obstacle
          */
         REFL,
         /**
