+++ .github/workflows/static.yml
@@ -35,6 +35,12 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+      - name: Set up JDK 11
+        uses: actions/setup-java@v1
+        with:
+          java-version: 11
+      - name: Generate Javadoc
+        run: mvn javadoc:aggregate
       - uses: actions/setup-python@v5
       - name: Install dependencies
         run: |
