### 2025-01-17 3637979
generate cnossos report in static website github workflow
