+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/RegressionTest.java
@@ -3,6 +3,8 @@ package org.noise_planet.noisemodelling.jdbc;
 import org.h2gis.api.EmptyProgressVisitor;
 import org.h2gis.api.ProgressVisitor;
 import org.h2gis.functions.factory.H2GISDBFactory;
+import org.h2gis.functions.io.fgb.FGBRead;
+import org.h2gis.functions.io.fgb.fileTable.FGBDriver;
 import org.h2gis.utilities.JDBCUtilities;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -13,7 +15,9 @@ import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFuncti
 import org.noise_planet.noisemodelling.pathfinder.utils.profiler.RootProgressVisitor;
 import org.noise_planet.noisemodelling.propagation.Attenuation;
 
+import java.io.IOException;
 import java.sql.Connection;
+import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.*;
 
@@ -36,15 +40,64 @@ public class RegressionTest {
         }
     }
 
-    /**
-     * Got reflection index out of bound exception in this scenario in the past (source->reflection->h diffraction->receiver)
-     */
-    @Test
-    public void testScenarioOutOfBoundException() throws Exception {
-        try(Statement st = connection.createStatement()) {
-            st.execute(getRunScriptRes("regression_nan/lw_roads.sql"));
+    public static NoiseMapMaker runPropagation(Connection connection, NoiseMapByReceiverMaker noiseMapByReceiverMaker) throws SQLException, IOException {
+        // Init NoiseModelling
+        // Building height field name
+        noiseMapByReceiverMaker.setHeightField("HEIGHT");
 
-            // Init NoiseModelling
+
+        // Init custom input in order to compute more than just attenuation
+        // LW_ROADS contain Day Evening Night emission spectrum
+        NoiseMapParameters noiseMapParameters = new NoiseMapParameters(NoiseMapParameters.INPUT_MODE.INPUT_MODE_LW_DEN);
+        noiseMapParameters.setExportRaysMethod(NoiseMapParameters.ExportRaysMethods.TO_MEMORY);
+
+        noiseMapParameters.setComputeLDay(false);
+        noiseMapParameters.setComputeLEvening(false);
+        noiseMapParameters.setComputeLNight(false);
+        noiseMapParameters.setComputeLDEN(true);
+        noiseMapParameters.keepAbsorption = true;
+
+        NoiseMapMaker noiseMapMaker = new NoiseMapMaker(connection, noiseMapParameters);
+
+        noiseMapByReceiverMaker.setPropagationProcessDataFactory(noiseMapMaker);
+        noiseMapByReceiverMaker.setComputeRaysOutFactory(noiseMapMaker);
+
+        RootProgressVisitor progressLogger = new RootProgressVisitor(1, true, 1);
+
+        noiseMapByReceiverMaker.initialize(connection, new EmptyProgressVisitor());
+
+        noiseMapParameters.getPropagationProcessPathData(NoiseMapParameters.TIME_PERIOD.DAY).setTemperature(20);
+        noiseMapParameters.getPropagationProcessPathData(NoiseMapParameters.TIME_PERIOD.EVENING).setTemperature(16);
+        noiseMapParameters.getPropagationProcessPathData(NoiseMapParameters.TIME_PERIOD.NIGHT).setTemperature(10);
+
+        noiseMapByReceiverMaker.setGridDim(1);
+
+        // Set of already processed receivers
+        Set<Long> receivers = new HashSet<>();
+
+        // Fetch cell identifiers with receivers
[…]
