### 2025-01-20 2d51589
Extend dem domain retrieval on cells in order to cover far away sources (limit of propagation distance). Extend dem domain out of maximum propagation distance with average altitude in order to not raise NaN issues with altitude.
