### 2025-01-20 ca72f74
some print changes
