### 2025-01-23 c0df232
refactoring to separate lden computing
