+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/LdenNoiseMapLoaderTest.java
@@ -30,12 +30,7 @@ import org.noise_planet.noisemodelling.pathfinder.utils.profiler.RootProgressVis
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
 import org.noise_planet.noisemodelling.propagation.Attenuation;
 
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.sql.Connection;
-import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -43,13 +38,13 @@ import java.util.concurrent.atomic.AtomicInteger;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.noise_planet.noisemodelling.jdbc.Utils.getRunScriptRes;
 
-public class NoiseMapByReceiverMakerTest {
+public class LdenNoiseMapLoaderTest {
 
     private Connection connection;
 
     @BeforeEach
     public void tearUp() throws Exception {
-        connection = JDBCUtilities.wrapConnection(H2GISDBFactory.createSpatialDataBase(NoiseMapByReceiverMakerTest.class.getSimpleName(), true, ""));
+        connection = JDBCUtilities.wrapConnection(H2GISDBFactory.createSpatialDataBase(LdenNoiseMapLoaderTest.class.getSimpleName(), true, ""));
     }
 
     @AfterEach
@@ -62,29 +57,29 @@ public class NoiseMapByReceiverMakerTest {
     @Test
     public void testGroundSurface() throws Exception {
         try(Statement st = connection.createStatement()) {
-            st.execute(String.format("CALL SHPREAD('%s', 'LANDCOVER2000')", NoiseMapByReceiverMakerTest.class.getResource("landcover2000.shp").getFile()));
+            st.execute(String.format("CALL SHPREAD('%s', 'LANDCOVER2000')", LdenNoiseMapLoaderTest.class.getResource("landcover2000.shp").getFile()));
             st.execute(getRunScriptRes("scene_with_landcover.sql"));
-            NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker("BUILDINGS", "ROADS_GEOM", "RECEIVERS");
-            noiseMapByReceiverMaker.setComputeHorizontalDiffraction(true);
-            noiseMapByReceiverMaker.setComputeVerticalDiffraction(true);
-            noiseMapByReceiverMaker.setSoundReflectionOrder(1);
-            noiseMapByReceiverMaker.setReceiverHasAbsoluteZCoordinates(false);
-            noiseMapByReceiverMaker.setSourceHasAbsoluteZCoordinates(false);
-            noiseMapByReceiverMaker.setHeightField("HEIGHT");
-            noiseMapByReceiverMaker.setSoilTableName("LAND_G");
-            noiseMapByReceiverMaker.setComputeVerticalDiffraction(true);
-            noiseMapByReceiverMaker.initialize(connection, new EmptyProgressVisitor());
+            LdenNoiseMapLoader ldenNoiseMapLoader = new LdenNoiseMapLoader("BUILDINGS", "ROADS_GEOM", "RECEIVERS");
+            ldenNoiseMapLoader.setComputeHorizontalDiffraction(true);
+            ldenNoiseMapLoader.setComputeVerticalDiffraction(true);
+            ldenNoiseMapLoader.setSoundReflectionOrder(1);
+            ldenNoiseMapLoader.setReceiverHasAbsoluteZCoordinates(false);
[…]
