+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationCnossosTest.java
@@ -12,7 +12,6 @@ package org.noise_planet.noisemodelling.jdbc;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import org.checkerframework.checker.units.qual.C;
 import org.h2gis.api.EmptyProgressVisitor;
 import org.h2gis.api.ProgressVisitor;
 import org.h2gis.functions.factory.H2GISDBFactory;
@@ -932,7 +931,7 @@ public class AttenuationCnossosTest {
         List<Integer> res1 = new ArrayList<>(3) ;
         List<Integer> res2 = new ArrayList<>(3);
 
-        for(int f : propDataOut.inputData.freq_lvl) {
+        for(int f : propDataOut.inputData.frequencyArray) {
             if(deltaD > -(340./f) / 20) {
                 res1.add(1);
             }
@@ -3454,7 +3453,7 @@ public class AttenuationCnossosTest {
         List<Integer> res1 = new ArrayList<>(3) ;
         List<Integer> res2 = new ArrayList<>(3);
 
-        for(int f : propDataOut.inputData.freq_lvl) {
+        for(int f : propDataOut.inputData.frequencyArray) {
             if(-deltaD > -(340./f) / 20) {
                 res1.add(1);
             }
@@ -4362,7 +4361,7 @@ public class AttenuationCnossosTest {
         List<Integer> res1 = new ArrayList<>(3) ;
         List<Integer> res2 = new ArrayList<>(3);
 
-        for(int f : propDataOut.inputData.freq_lvl) {
+        for(int f : propDataOut.inputData.frequencyArray) {
             if(deltaD > -(340./f) / 20) {
                 res1.add(1);
             }
@@ -6165,55 +6164,55 @@ public class AttenuationCnossosTest {
 
     private double testIgnoreNonSignificantSourcesParam(Connection connection, double maxError) throws SQLException, IOException {
         // Init NoiseModelling
-        NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker("BUILDINGS",
+        LdenNoiseMapLoader ldenNoiseMapLoader = new LdenNoiseMapLoader("BUILDINGS",
                 "LW_ROADS", "RECEIVERS");
 
-        noiseMapByReceiverMaker.setMaximumPropagationDistance(5000.0);
-        noiseMapByReceiverMaker.setSoundReflectionOrder(1);
-        noiseMapByReceiverMaker.setThreadCount(1);
-        noiseMapByReceiverMaker.setComputeHorizontalDiffraction(true);
-        noiseMapByReceiverMaker.setComputeVerticalDiffraction(true);
+        ldenNoiseMapLoader.setMaximumPropagationDistance(5000.0);
+        ldenNoiseMapLoader.setSoundReflectionOrder(1);
+        ldenNoiseMapLoader.setThreadCount(1);
+        ldenNoiseMapLoader.setComputeHorizontalDiffraction(true);
+        ldenNoiseMapLoader.setComputeVerticalDiffraction(true);
         // Building height field name
-        noiseMapByReceiverMaker.setHeightField("HEIGHT");
+        ldenNoiseMapLoader.setHeightField("HEIGHT");
 
 
         // Init custom input in order to compute more than just attenuation
         // LW_ROADS contain Day Evening Night emission spectrum
-        NoiseMapParameters noiseMapParameters = new NoiseMapParameters(NoiseMapParameters.INPUT_MODE.INPUT_MODE_LW_DEN);
[…]
