### 2025-01-23 91d2f87
fix refactoring to separate lden computing
