+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AtmosphericAttenuationTest.java
@@ -10,6 +10,8 @@ package org.noise_planet.noisemodelling.propagation;
 
 import org.junit.jupiter.api.Test;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
+import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
 
 import java.util.Arrays;
@@ -23,8 +25,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  */
 public class AtmosphericAttenuationTest {
     private static final double EPSILON = 0.1;
-    private static final List<Double> freq_lvl_exact = Arrays.asList(AttenuationCnossosParameters.asOctaveBands(
-            Scene.DEFAULT_FREQUENCIES_EXACT_THIRD_OCTAVE));
+    private static final List<Double> freq_lvl_exact = Arrays.asList(AcousticIndicatorsFunctions.asOctaveBands(
+            ProfileBuilder.DEFAULT_FREQUENCIES_EXACT_THIRD_OCTAVE));
 
     @Test
     public void atmoTestMinus20degree() {
