### 2025-01-23 0143c6c
next refactoring to compute lden or only attenuation with optionally time dependent noise level.
