### 2025-01-24 2dde79d
pathfinder and propagation ok, Gs specific to cnossos
