+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapLoader.java
@@ -28,6 +28,7 @@ import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.Wall;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.WallAbsorption;
+import org.noise_planet.noisemodelling.pathfinder.utils.profiler.ProfilerThread;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,7 +40,7 @@ import java.util.*;
 import static org.h2gis.utilities.GeometryTableUtilities.getGeometryColumnNames;
 import static org.h2gis.utilities.GeometryTableUtilities.getSRID;
 /**
- * Common attributes for propagation of sound sources.
+ * Common attributes for subdivision of the domain
  * @author Nicolas Fortin
  */
 public abstract class NoiseMapLoader {
@@ -52,7 +53,6 @@ public abstract class NoiseMapLoader {
     private String alphaFieldName = "G";
     protected final String buildingsTableName;
     protected final String sourcesTableName;
-    protected String soilTableName = "";
     // Digital elevation model table. (Contains points or triangles)
     protected String demTable = "";
     protected String sound_lvl_field = "DB_M";
@@ -78,6 +78,7 @@ public abstract class NoiseMapLoader {
 
     protected String heightField = "HEIGHT";
     protected GeometryFactory geometryFactory;
+    protected ProfilerThread profilerThread;
 
     // Initialised attributes
     /**
@@ -90,6 +91,16 @@ public abstract class NoiseMapLoader {
         this.buildingsTableName = buildingsTableName;
         this.sourcesTableName = sourcesTableName;
     }
+
+    /**
+     * Computation stacks and timing are collected by this class in order
+     * to profile the execution of the simulation
+     * @param profilerThread Instance of ProfilerThread
+     */
+    public void setProfilerThread(ProfilerThread profilerThread) {
+        this.profilerThread = profilerThread;
+    }
+
 
     /**
      * The table shall contain the following fields :
@@ -279,71 +290,6 @@ public abstract class NoiseMapLoader {
             }
         }
     }
-
-    /**
-     * Fetches soil areas data for the specified cell envelope and adds them to the profile builder.
-     * @param connection         the database connection to use for querying the soil areas data.
-     * @param fetchEnvelope      the envelope representing the cell to fetch soil areas data for.
-     * @param builder            the profile builder to which the soil areas data will be added.
-     * @throws SQLException      if an SQL exception occurs while fetching the soil areas data.
-     */
-    protected void fetchCellSoilAreas(Connection connection, Envelope fetchEnvelope, ProfileBuilder builder)
-            throws SQLException {
-        if(!soilTableName.isEmpty()){
-            DBTypes dbType = DBUtils.getDBType(connection.unwrap(Connection.class));
-            double startX = Math.floor(fetchEnvelope.getMinX() / groundSurfaceSplitSideLength) * groundSurfaceSplitSideLength;
-            double startY = Math.floor(fetchEnvelope.getMinY() / groundSurfaceSplitSideLength) * groundSurfaceSplitSideLength;
-            String soilGeomName = getGeometryColumnNames(connection,
-                    TableLocation.parse(soilTableName, dbType)).get(0);
-            try (PreparedStatement st = connection.prepareStatement(
-                    "SELECT " + TableLocation.quoteIdentifier(soilGeomName, dbType) + ", G FROM " +
[…]
