+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/SceneWithAttenuation.java
@@ -1,0 +1,162 @@
+package org.noise_planet.noisemodelling.propagation;
+
+import org.h2gis.utilities.JDBCUtilities;
+import org.h2gis.utilities.SpatialResultSet;
+import org.locationtech.jts.geom.Geometry;
+import org.noise_planet.noisemodelling.emission.directivity.DirectivitySphere;
+import org.noise_planet.noisemodelling.pathfinder.path.Scene;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
+import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
+
+import java.io.IOException;
+import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+
+public class SceneWithAttenuation extends Scene {
+    public static final double DEFAULT_GS = 0.0;
+
+    /**
+     * Attenuation and other attributes relative to direction on sphere
+     */
+    public Map<Integer, DirectivitySphere> directionAttributes = new HashMap<>();
+
+    /**
+     * Link between sources PK and DirectivitySphere specified in linked with directionAttributes
+     */
+    public Map<Long, Integer> sourceEmissionAttenuation = new HashMap<>();
+
+    /**
+     * Link between sources PK and gs ground factor of the source area
+     */
+    public Map<Long, Double> sourceGs = new HashMap<>();
+
+    /**
+     * Cached source table fields
+     */
+    Map<String, Integer> sourceFieldNames = new HashMap<>();
+
+    public SceneWithAttenuation(ProfileBuilder profileBuilder) {
+        super(profileBuilder);
+    }
+
+    public SceneWithAttenuation() {
+    }
+
+    /**
+     * Retrieves the ground speed of the noise source at the specified index.
+     * @param srcIndex
+     * @return the ground speed of the noise source at the specified index.
+     */
+    public double getSourceGs(int srcIndex){
+        return sourceGs.get(sourcesPk.get(srcIndex));
+    }
+
+    /**
+     * Add geometry with additional attributes
+     * @param pk Unique source identifier
+     * @param geom Source geometry
+     * @param gs Additional attributes
+     */
+
+    public void addSource(Long pk, Geometry geom, Double gs) {
+        addSource(pk, geom);
+        sourceGs.put(pk, gs);
+    }
+
+    /**
+     * Sets the direction attributes for the receiver.
+     * @param directionAttributes
+     */
+    public void setDirectionAttributes(Map<Integer, DirectivitySphere> directionAttributes) {
+        this.directionAttributes = directionAttributes;
+        // Check if the directivities contain all required frequencies
+        directionAttributes.forEach((integer, directivitySphere) -> {
+            profileBuilder.frequencyArray.forEach(frequency->{
+                if(!directivitySphere.coverFrequency(frequency)) {
+                    throw new IllegalArgumentException(
+                            String.format(Locale.ROOT,
+                                    "The provided DirectivitySphere does not handle %d Hertz", frequency));
+                }
+            });
+        });
+    }
+
+    /**
+     * Add geometry with additional attributes
+     * @param pk Unique source identifier
+     * @param geom Source geometry
+     * @param rs Additional attributes fetched from database
+     */
+    public void addSource(Long pk, Geometry geom, SpatialResultSet rs) throws SQLException, IOException {
+        if(sourceFieldNames.isEmpty()) {
[…]
