### 2025-01-24 6f17650
move cnossos attenuation test to propagation project
