### 2025-01-24 1cac024
merge concept of lden and time period in jdbc
