+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapMaker.java
@@ -28,8 +28,8 @@ import java.util.*;
 /**
  * Create SQL Tables from a stream of noise levels
  */
-public class NoiseMapMaker implements PropagationProcessDataFactory, IComputeRaysOutFactory, ProfilerThread.Metric {
-    LdenNoiseMapParameters ldenNoiseMapParameters;
+public class NoiseMapMaker implements NoiseMapByReceiverMaker.PropagationProcessDataFactory, NoiseMapByReceiverMaker.IComputeRaysOutFactory, ProfilerThread.Metric {
+    NoiseMapParameters noiseMapParameters;
     NoiseMapWriter noiseMapWriter;
     Thread tableWriterThread;
     Connection connection;
@@ -37,12 +37,17 @@ public class NoiseMapMaker implements PropagationProcessDataFactory, IComputeRay
     static final int WRITER_CACHE = 65536;
     AttenuatedPaths AttenuatedPaths = new AttenuatedPaths();
     int srid;
+    List<String> noiseSource = Arrays.asList("ROLLING","TRACTIONA", "TRACTIONB","AERODYNAMICA","AERODYNAMICB","BRIDGE");
 
 
+    /**
+     * Attenuation and other attributes relative to direction on sphere
+     */
+    public Map<Integer, DirectivitySphere> directionAttributes = new HashMap<>();
 
 
-    public NoiseMapMaker(Connection connection, LdenNoiseMapParameters LdenNoiseMapParameters) {
-        this.ldenNoiseMapParameters = LdenNoiseMapParameters;
+    public NoiseMapMaker(Connection connection, NoiseMapParameters NoiseMapParameters) {
+        this.noiseMapParameters = NoiseMapParameters;
         this.connection = connection;
     }
 
@@ -65,34 +70,109 @@ public class NoiseMapMaker implements PropagationProcessDataFactory, IComputeRay
         return AttenuatedPaths;
     }
 
+    /**
+     * Inserts directivity attributes for noise sources for trains into the directionAttributes map.
+     */
+
+    public void insertTrainDirectivity() {
+        directionAttributes.clear();
+        directionAttributes.put(0, new OmnidirectionalDirection());
+        int i=1;
+        for(String typeSource : noiseSource) {
+            directionAttributes.put(i, new RailwayCnossosDirectivitySphere(new LineSource(typeSource)));
+            i++;
+        }
+    }
 
     /**
      * Initializes the NoiseMap parameters and attenuation data based on the input mode specified in the NoiseMap parameters.
      * @param connection   the database connection to be used for initialization.
-     * @param ldenNoiseMapLoader the noise map by receiver maker object associated with the computation process.
+     * @param noiseMapByReceiverMaker the noise map by receiver maker object associated with the computation process.
      * @throws SQLException
      */
 
     @Override
-    public void initialize(Connection connection, LdenNoiseMapLoader ldenNoiseMapLoader) throws SQLException {
-        if(JDBCUtilities.tableExists(connection, ldenNoiseMapLoader.getSourcesTableName())) {
-            this.srid = GeometryTableUtilities.getSRID(connection, ldenNoiseMapLoader.getSourcesTableName());
+    public void initialize(Connection connection, NoiseMapByReceiverMaker noiseMapByReceiverMaker) throws SQLException {
+        if(JDBCUtilities.tableExists(connection, noiseMapByReceiverMaker.getSourcesTableName())) {
+            this.srid = GeometryTableUtilities.getSRID(connection, noiseMapByReceiverMaker.getSourcesTableName());
         }
-        ldenNoiseMapParameters.initialize(connection, ldenNoiseMapLoader);
[…]
