+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/AttenuationOutputSingleThread.java
@@ -1,0 +1,465 @@
+/**
+ * NoiseModelling is a library capable of producing noise maps. It can be freely used either for research and education, as well as by experts in a professional use.
+ * <p>
+ * NoiseModelling is distributed under GPL 3 license. You can read a copy of this License in the file LICENCE provided with this software.
+ * <p>
+ * Official webpage : http://noise-planet.org/noisemodelling.html
+ * Contact: contact@noise-planet.org
+ */
+
+package org.noise_planet.noisemodelling.jdbc;
+
+import org.noise_planet.noisemodelling.pathfinder.IComputePathsOut;
+import org.noise_planet.noisemodelling.pathfinder.PathFinder;
+import org.noise_planet.noisemodelling.pathfinder.path.Scene;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
+import org.noise_planet.noisemodelling.propagation.AttenuationComputeOutput;
+import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossos;
+import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
+import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
+import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPathBuilder;
+
+import java.util.*;
+import java.util.concurrent.ConcurrentLinkedDeque;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.*;
+import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.wToDba;
+
+
+public class AttenuationOutputSingleThread implements IComputePathsOut {
+    AttenuationOutputMultiThread attenuationOutputMultiThreadComputeRaysOut;
+    NoiseMapParameters noiseMapParameters;
+    public List<CnossosPath> pathParameters = new ArrayList<CnossosPath>();
+
+    /**
+     * Collected attenuation per receiver
+     */
+    Map<Integer, NoiseMapParameters.TimePeriodParameters> receiverAttenuationPerSource = new HashMap<>();
+
+    /**
+     * MaxError DB Processing variable
+     * Current, cumulated global power at receiver, only used to stop looking for far sources
+     */
+    double[] wjAtReceiver = new double[0];
+
+    /**
+     * MaxError DB Processing variable
+     * Favorable Free Field cumulated global power at receiver, only used to stop looking for far sources
+     */
+    Map<Integer, Double> maximumWjExpectedSplAtReceiver = new HashMap<>();
+    /**
+     * MaxError DB Processing variable
+     * Next Free Field cumulated global power at receiver, only used to stop looking for far sources
+     */
+    double sumMaximumRemainingWjExpectedSplAtReceiver = 0;
+
+    public AtomicInteger cutProfileCount = new AtomicInteger(0);
+
+    /**
+     * Constructs a NoiseMapInStack object with a multi-threaded parent NoiseMap instance.
+     * @param multiThreadParent
+     */
+    public AttenuationOutputSingleThread(AttenuationOutputMultiThread multiThreadParent) {
+        this.attenuationOutputMultiThreadComputeRaysOut = multiThreadParent;
+        this.noiseMapParameters = multiThreadParent.noiseEmissionMaker.noiseMapParameters;
+    }
+
+    /**
[…]
