### 2025-01-27 12dcfa0
continue period computation
