+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/AttenuationOutputSingleThread.java
@@ -29,15 +29,19 @@ import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicator
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.wToDba;
 
 
+/**
+ * Managed by a single thread, process all incoming vertical profile, compute attenuation and push on appropriate stack
+ * for exporting result values in a thread safe way
+ */
 public class AttenuationOutputSingleThread implements IComputePathsOut {
     AttenuationOutputMultiThread attenuationOutputMultiThreadComputeRaysOut;
-    NoiseMapParameters noiseMapParameters;
+    NoiseMapDatabaseParameters noiseMapDatabaseParameters;
     public List<CnossosPath> pathParameters = new ArrayList<CnossosPath>();
 
     /**
      * Collected attenuation per receiver
      */
-    Map<Integer, NoiseMapParameters.TimePeriodParameters> receiverAttenuationPerSource = new HashMap<>();
+    Map<Integer, NoiseMapDatabaseParameters.TimePeriodParameters> receiverAttenuationPerSource = new HashMap<>();
 
     /**
      * MaxError DB Processing variable
@@ -59,12 +63,13 @@ public class AttenuationOutputSingleThread implements IComputePathsOut {
     public AtomicInteger cutProfileCount = new AtomicInteger(0);
 
     /**
-     * Constructs a NoiseMapInStack object with a multi-threaded parent NoiseMap instance.
+     * Constructs a NoiseMapInStack object with a multithreaded parent NoiseMap instance.
+     * This class is not thread-safe
      * @param multiThreadParent
      */
     public AttenuationOutputSingleThread(AttenuationOutputMultiThread multiThreadParent) {
         this.attenuationOutputMultiThreadComputeRaysOut = multiThreadParent;
-        this.noiseMapParameters = multiThreadParent.noiseEmissionMaker.noiseMapParameters;
+        this.noiseMapDatabaseParameters = multiThreadParent.sceneWithEmission.noiseMapParameters;
     }
 
     /**
@@ -92,27 +97,27 @@ public class AttenuationOutputSingleThread implements IComputePathsOut {
         }
     }
 
-    private NoiseMapParameters.TimePeriodParameters computeFastLdenAttenuation(PathFinder.SourcePointInfo sourceInfo,
-                                                                               PathFinder.ReceiverPointInfo receiverInfo) {
+    private NoiseMapDatabaseParameters.TimePeriodParameters computeFastLdenAttenuation(PathFinder.SourcePointInfo sourceInfo,
+                                                                                       PathFinder.ReceiverPointInfo receiverInfo) {
         // For the quick attenuation evaluation
         // only take account of geometric dispersion and atmospheric attenuation
         double distance = Math.max(1.0, sourceInfo.position.distance3D(receiverInfo.position));
         // 3 dB gain as we consider source G path is equal to 0
         double attenuationDivGeom = AttenuationCnossos.getADiv(distance) - 3;
-        NoiseMapParameters.TimePeriodParameters denWAttenuation =
-                new NoiseMapParameters.TimePeriodParameters(sourceInfo,
+        NoiseMapDatabaseParameters.TimePeriodParameters denWAttenuation =
+                new NoiseMapDatabaseParameters.TimePeriodParameters(sourceInfo,
                         receiverInfo, new double[0], new double[0], new double[0]);
-        if (noiseMapParameters.computeLDay || noiseMapParameters.computeLDEN) {
[…]
