+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/DefaultCutPlaneProcessing.java
@@ -1,0 +1,7 @@
+package org.noise_planet.noisemodelling.jdbc.output;
+
+import org.noise_planet.noisemodelling.jdbc.NoiseMapByReceiverMaker;
+
+public class DefaultCutPlaneProcessing implements NoiseMapByReceiverMaker.IComputeRaysOutFactory {
+
+}
