### 2025-01-28 2e22f8f
separate code for fetching data in tables
