+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseMapLoader.java
@@ -21,6 +21,8 @@ import org.locationtech.jts.geom.prep.PreparedPolygon;
 import org.locationtech.jts.io.WKTWriter;
 import org.noise_planet.noisemodelling.emission.directivity.DirectivityRecord;
 import org.noise_planet.noisemodelling.emission.directivity.DiscreteDirectivitySphere;
+import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
+import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
 import org.noise_planet.noisemodelling.jdbc.utils.CellIndex;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.Building;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
@@ -44,15 +46,12 @@ import static org.h2gis.utilities.GeometryTableUtilities.getSRID;
 public abstract class NoiseMapLoader {
     // When computing cell size, try to keep propagation distance away from the cell
     // inferior to this ratio (in comparison with cell width)
-    AttenuationCnossosParameters attenuationCnossosParametersDay = new AttenuationCnossosParameters();
-    AttenuationCnossosParameters attenuationCnossosParametersEvening = new AttenuationCnossosParameters();
-    AttenuationCnossosParameters attenuationCnossosParametersNight = new AttenuationCnossosParameters();
+
     Logger logger = LoggerFactory.getLogger(NoiseMapLoader.class);
     private static final int DEFAULT_FETCH_SIZE = 300;
     protected int fetchSize = DEFAULT_FETCH_SIZE;
     protected static final double MINIMAL_BUFFER_RATIO = 0.3;
-    private String alphaFieldName = "G";
-    protected final String buildingsTableName;
+    protected DefaultTableLoader.BuildingTableParameters buildingTableParameters = new DefaultTableLoader.BuildingTableParameters();
     protected final String sourcesTableName;
     protected String sourcesEmissionTableName = "";
     protected String soilTableName = "";
@@ -65,9 +64,6 @@ public abstract class NoiseMapLoader {
     protected double maximumPropagationDistance = 750;
     protected double maximumReflectionDistance = 100;
     protected double gs = 0;
-    /** if true take into account z value on Buildings Polygons
-     * In this case, z represent the altitude (from the sea to the top of the wall) */
-    protected boolean zBuildings = false;
     // Soil areas are splited by the provided size in order to reduce the propagation time
     protected double groundSurfaceSplitSideLength = 200;
     protected int soundReflectionOrder = 2;
@@ -76,10 +72,7 @@ public abstract class NoiseMapLoader {
     public boolean verbose = true;
     protected boolean computeHorizontalDiffraction = true;
     protected boolean computeVerticalDiffraction = true;
-    /** TODO missing reference to the SIGMA value of materials */
-    protected double wallAbsorption = 100000;
 
-    protected String heightField = "HEIGHT";
     protected GeometryFactory geometryFactory;
 
     // Initialised attributes
@@ -90,64 +83,8 @@ public abstract class NoiseMapLoader {
     protected Envelope mainEnvelope = new Envelope();
 
     public NoiseMapLoader(String buildingsTableName, String sourcesTableName) {
-        this.buildingsTableName = buildingsTableName;
+        this.buildingTableParameters.buildingsTableName = buildingsTableName;
         this.sourcesTableName = sourcesTableName;
-    }
-
-
-    /**
-     * Retrieves the propagation process path data for the specified time period.
-     * @param time_period the time period for which to retrieve the propagation process path data.
-     * @return the attenuation Cnossos parameters for the specified time period.
-     */
[…]
