### 2025-01-30 44baa6d
read emission level and traffic from jdbc tables
