+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/input/SceneWithEmission.java
@@ -1,9 +1,14 @@
 package org.noise_planet.noisemodelling.jdbc.input;
 
+import org.h2gis.utilities.SpatialResultSet;
+import org.locationtech.jts.geom.Geometry;
+import org.noise_planet.noisemodelling.jdbc.EmissionTableGenerator;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
+import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.noise_planet.noisemodelling.propagation.SceneWithAttenuation;
 
-import java.nio.ByteBuffer;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.*;
 
 /**
@@ -11,7 +16,9 @@ import java.util.*;
  * This is input data, not thread safe, never update anything here during propagation
  */
 public class SceneWithEmission extends SceneWithAttenuation {
-
+    /** Old style DEN columns traffic period  */
+    Map<String, Integer> sourceFieldsCache = new HashMap<>();
+    Map<String, Integer> sourceEmissionFieldsCache = new HashMap<>();
 
     /**
      * Set of unique power spectrum. Key is the hash of the spectrum. Power spectrum in energetic e = pow(10, dbVal / 10.0)
@@ -21,25 +28,85 @@ public class SceneWithEmission extends SceneWithAttenuation {
     //  For each source primary key give the map between period and source power spectrum hash value
     public Map<Long, ArrayList<PeriodEmission>> wjSources = new HashMap<>();
 
-    public SceneWithEmission(ProfileBuilder profileBuilder) {
+    public SceneInputSettings sceneInputSettings = new SceneInputSettings();
+
+    public SceneWithEmission(ProfileBuilder profileBuilder, SceneInputSettings sceneInputSettings) {
         super(profileBuilder);
+        this.sceneInputSettings = sceneInputSettings;
     }
 
     public SceneWithEmission() {
-    }
-
-    public static byte[] convertDoubleArrayToByteArray(double[] doubleArray) {
-        ByteBuffer byteBuffer = ByteBuffer.allocate(doubleArray.length * Double.BYTES);
-        for (double value : doubleArray) {
-            byteBuffer.putDouble(value);
-        }
-        return byteBuffer.array();
     }
 
     int updatePowerSpectrumSet(double[] wj) {
         int hashCode = Arrays.hashCode(wj);
         powerSpectrum.putIfAbsent(hashCode, wj);
         return hashCode;
+    }
+
+    public void processTrafficFlowDEN(Long pk, SpatialResultSet rs) throws SQLException {
+        // Source table PK, GEOM, LV_D, LV_E, LV_N ...
+        double[][] lw = EmissionTableGenerator.computeLw(rs, sceneInputSettings.coefficientVersion, sourceFieldsCache);
+        for (EmissionTableGenerator.STANDARD_PERIOD period : EmissionTableGenerator.STANDARD_PERIOD.values()) {
+            addSourceEmission(pk, EmissionTableGenerator.STANDARD_PERIOD_VALUE[period.ordinal()], lw[period.ordinal()]);
+        }
+    }
+
+    /**
+     * @param pk Source primary key
+     * @param rs Emission source table PK_SOURCE, PERIOD, LV, HV ..
+     * @throws SQLException
+     */
+    public void processTrafficFlow(Long pk, ResultSet rs) throws SQLException {
+        String period = rs.getString("PERIOD");
+        // Use geometry as default slope (if field slope is not provided
+        double defaultSlope = 0;
+        if(!sourceEmissionFieldsCache.containsKey("SLOPE")) {
+            int sourceIndex = sourcesPk.indexOf(pk);
+            if(sourceIndex >= 0) {
[…]
