+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/EmissionTableGenerator.java
@@ -22,38 +22,48 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.dbaToW;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.sumArray;
 
-public class LdenTableGenerator {
-    public Map<String, Integer> sourceFields = null;
-    public List<Integer> frequencyArray = Arrays.asList(AcousticIndicatorsFunctions.asOctaveBands(ProfileBuilder.DEFAULT_FREQUENCIES_THIRD_OCTAVE));
+/**
+ * Create emission table from traffic data (RAIL or ROADS)
+ */
+public class EmissionTableGenerator {
+    public static final List<Integer> roadOctaveFrequencyBands = Arrays.asList(AcousticIndicatorsFunctions.asOctaveBands(ProfileBuilder.DEFAULT_FREQUENCIES_THIRD_OCTAVE));
+    public enum STANDARD_PERIOD {DAY, EVENING, NIGHT, DAY_EVENING_NIGHT}
+    public static final String[] STANDARD_PERIOD_VALUE = new String[] {"D", "E", "N", "DEN"};
 
     public static final double DAY_RATIO = 12. / 24.;
     public static final double EVENING_RATIO = 4. / 24. * dbaToW(5.0);
     public static final double NIGHT_RATIO = 8. / 24. * dbaToW(10.0);
 
     /**
-     * Retrieves the emissions for the specified period from the given result set
-     * @param rs result set of source
-     * @param period D or E or N
-     * @param slope Gradient percentage of road from -12 % to 12 %
-     * @param coefficientVersion Cnossos coefficient version  (1 = 2015, 2 = 2020)
-     * @return Emission spectrum in dB
-     *
+     * Cache table fields in upper case in Map
+     * @param sourceFieldsCache map
+     * @param rs table to load
+     * @throws SQLException If error
      */
-    public double[] getEmissionFromTrafficTable(ResultSet rs, String period, double slope, int coefficientVersion) throws SQLException, IOException {
-        if (sourceFields == null) {
-            sourceFields = new HashMap<>();
+    public static void cacheFields(Map<String, Integer> sourceFieldsCache, ResultSet rs) throws SQLException {
+        if (sourceFieldsCache.isEmpty()) {
             int fieldId = 1;
             for (String fieldName : JDBCUtilities.getColumnNames(rs.getMetaData())) {
-                sourceFields.put(fieldName.toUpperCase(), fieldId++);
+                sourceFieldsCache.put(fieldName.toUpperCase(), fieldId++);
             }
         }
+    }
+    /**
+     * Retrieves the emissions for the specified period from the given result set
+     * @param rs result set of source
+     * @param period optional column name to add after attribute fields _D or _E or _N
+     * @param slope Default, gradient percentage of road from -12 % to 12 %
+     * @param coefficientVersion Cnossos coefficient version  (1 = 2015, 2 = 2020)
+     * @return Emission spectrum in dB
+     */
+    public static double[] getEmissionFromTrafficTable(ResultSet rs, String period, double slope, int coefficientVersion, Map<String, Integer> sourceFieldsCache) throws SQLException {
+        cacheFields(sourceFieldsCache, rs);
         // Set default values
[…]
