### 2025-01-30 23863e4
collect/compute noise level on cutprofile visitor
