+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/NoiseMapWriter.java
@@ -15,10 +15,12 @@ import org.locationtech.jts.geom.PrecisionModel;
 import org.noise_planet.noisemodelling.jdbc.NoiseMapByReceiverMaker;
 import org.noise_planet.noisemodelling.jdbc.NoiseMapDatabaseParameters;
 import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
+import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
 import org.noise_planet.noisemodelling.jdbc.utils.StringPreparedStatements;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
 import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.noise_planet.noisemodelling.propagation.AttenuationComputeOutput;
+import org.noise_planet.noisemodelling.propagation.ReceiverNoiseLevel;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,6 +33,7 @@ import java.sql.Statement;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ConcurrentLinkedDeque;
+import java.util.stream.DoubleStream;
 import java.util.zip.GZIPOutputStream;
 
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.*;
@@ -47,25 +50,29 @@ public class NoiseMapWriter implements Runnable {
     private Connection connection;
     NoiseMapByReceiverMaker noiseMapByReceiverMaker;
     ResultsCache resultsCache;
-    public List<Double> aWeightingArray = Arrays.asList(AcousticIndicatorsFunctions.asOctaveBands(ProfileBuilder.DEFAULT_FREQUENCIES_A_WEIGHTING_THIRD_OCTAVE));
     boolean started = false;
     Writer o;
     int srid;
+    public List<Integer> frequencyArray = Arrays.asList(AcousticIndicatorsFunctions.asOctaveBands(ProfileBuilder.DEFAULT_FREQUENCIES_THIRD_OCTAVE));
+    public double[] aWeightingArray = Arrays.stream(
+                    asOctaveBands(ProfileBuilder.DEFAULT_FREQUENCIES_A_WEIGHTING_THIRD_OCTAVE)).
+            mapToDouble(value -> value).toArray();
 
     /**
      * Constructs a new NoiseMapWriter object with the specified parameters.
      * @param connection        the database connection used for writing data
-     * @param noiseMapDatabaseParameters the parameters defining the noise map computation
+     * @param noiseMapByReceiverMaker the parameters defining the noise map computation
      * @param ResultsCache   the attenuated paths containing computed noise data
-     * @param srid the spatial reference identifier (SRID) for geometric data
      */
     public NoiseMapWriter(Connection connection, NoiseMapByReceiverMaker noiseMapByReceiverMaker, ResultsCache ResultsCache) {
         this.connection = connection;
         this.noiseMapByReceiverMaker = noiseMapByReceiverMaker;
         this.resultsCache = ResultsCache;
         this.srid = noiseMapByReceiverMaker.getGeometryFactory().getSRID();
-        if(noiseMapByReceiverMaker.getPropagationProcessDataFactory() instanceof DefaultCutPlaneProcessing) {
-            aWeightingArray = ((DefaultTableLoader)noiseMapByReceiverMaker.getPropagationProcessDataFactory()).aWeightingArray;
[…]
