+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/TimePeriodParametersLdenAttenuationOutputMultiThreadLoaderFactoryTest.java
@@ -29,6 +29,7 @@ import org.noise_planet.noisemodelling.emission.railway.RailWayParameters;
 import org.noise_planet.noisemodelling.emission.railway.cnossos.RailwayCnossos;
 import org.noise_planet.noisemodelling.emission.utils.Utils;
 import org.noise_planet.noisemodelling.jdbc.LdenNoiseMapParameters.INPUT_MODE;
+import org.noise_planet.noisemodelling.jdbc.output.AttenuationOutputMultiThread;
 import org.noise_planet.noisemodelling.jdbc.railway.RailWayLWGeom;
 import org.noise_planet.noisemodelling.jdbc.railway.RailWayLWIterator;
 import org.noise_planet.noisemodelling.jdbc.utils.CellIndex;
