### 2025-01-31 f176308
period on cnossos propagation (was already there in cnossos path)
