+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/AttenuationOutputSingleThread.java
@@ -20,6 +20,8 @@ import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.propagation.AttenuationComputeOutput;
 import org.noise_planet.noisemodelling.propagation.ReceiverNoiseLevel;
+import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossos;
+import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPathBuilder;
@@ -67,32 +69,18 @@ public class AttenuationOutputSingleThread implements IComputePathsOut {
 
     public AtomicInteger cutProfileCount = new AtomicInteger(0);
 
+    ProgressVisitor progressVisitor;
+
     /**
      * Constructs a NoiseMapInStack object with a multithreaded parent NoiseMap instance.
      * This class is not thread-safe
      * @param multiThreadParent
      */
-    public AttenuationOutputSingleThread(AttenuationOutputMultiThread multiThreadParent) {
+    public AttenuationOutputSingleThread(AttenuationOutputMultiThread multiThreadParent, ProgressVisitor progressVisitor) {
         this.multiThread = multiThreadParent;
         this.noiseMapDatabaseParameters = multiThreadParent.noiseMapDatabaseParameters;
+        this.progressVisitor = progressVisitor;
     }
-//
-//    /**
-//     * Energetic sum of VerticeSL attenuation with WJ sources
-//     * @param wjSources
-//     * @param receiverAttenuationLevels
-//     * @return
-//     */
-//    double[] sumLevels(List<double[]> wjSources, List<ReceiverNoiseLevel> receiverAttenuationLevels) {
-//        double[] levels = new double[0];
-//        for (ReceiverNoiseLevel lvl : receiverAttenuationLevels) {
-//            if(wjSources.size() > lvl.source.sourceIndex && lvl.source.sourceIndex >= 0) {
-//                levels = sumArray(levels,
-//                        dbaToW(sumArray(wToDba(wjSources.get(lvl.source.sourceIndex)), lvl.levels)));
-//            }
-//        }
-//        return levels;
-//    }
 
     private void addGlobalReceiverLevel(double[] wjLevel) {
         if(wjAtReceiver.length != wjLevel.length) {
@@ -101,93 +89,33 @@ public class AttenuationOutputSingleThread implements IComputePathsOut {
             wjAtReceiver = AcousticIndicatorsFunctions.sumArray(wjAtReceiver, wjLevel);
         }
     }
-//
-//    private NoiseMapDatabaseParameters.TimePeriodParameters computeFastLdenAttenuation(PathFinder.SourcePointInfo sourceInfo,
-//                                                                                       PathFinder.ReceiverPointInfo receiverInfo) {
-//        // For the quick attenuation evaluation
-//        // only take account of geometric dispersion and atmospheric attenuation
-//        double distance = Math.max(1.0, sourceInfo.position.distance3D(receiverInfo.position));
-//        // 3 dB gain as we consider source G path is equal to 0
-//        double attenuationDivGeom = AttenuationCnossos.getADiv(distance) - 3;
-//        NoiseMapDatabaseParameters.TimePeriodParameters denWAttenuation =
[…]
