+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/AttenuationOutputMultiThread.java
@@ -51,7 +51,7 @@ public class AttenuationOutputMultiThread extends AttenuationComputeOutput {
      */
     @Override
     public IComputePathsOut subProcess(ProgressVisitor visitor) {
-        return new AttenuationOutputSingleThread(this);
+        return new AttenuationOutputSingleThread(this, visitor);
     }
 
 }
