### 2025-01-31 #7500 @kyleniemeyer
@Nitnelav unfortunately, the waitlist is a function of JOSS editor availability, on a first-in first-out basis, so your submission will be in the queue until someone becomes available to handle it.
