### 2025-02-03 0f23d9e
noise level processing at receiver level looks ok
