+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/AttenuationOutputSingleThread.java
@@ -10,7 +10,6 @@
 package org.noise_planet.noisemodelling.jdbc.output;
 
 import org.h2gis.api.ProgressVisitor;
-import org.noise_planet.noisemodelling.jdbc.NoiseEmissionMaker;
 import org.noise_planet.noisemodelling.jdbc.NoiseMapDatabaseParameters;
 import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
 import org.noise_planet.noisemodelling.pathfinder.IComputePathsOut;
@@ -18,7 +17,6 @@ import org.noise_planet.noisemodelling.pathfinder.PathFinder;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointReceiver;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutPointSource;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
-import org.noise_planet.noisemodelling.propagation.AttenuationComputeOutput;
 import org.noise_planet.noisemodelling.propagation.ReceiverNoiseLevel;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossos;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
@@ -39,14 +37,15 @@ import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicator
  * for exporting result values in a thread safe way. It processes the receiver one at a time.
  */
 public class AttenuationOutputSingleThread implements IComputePathsOut {
+    private static final int UNKNOWN_SOURCE_ID = -1;
     AttenuationOutputMultiThread multiThread;
-    NoiseMapDatabaseParameters noiseMapDatabaseParameters;
+    NoiseMapDatabaseParameters dbSettings;
     public List<CnossosPath> pathParameters = new ArrayList<CnossosPath>();
 
     /**
      * Collected attenuation/noise level on the current receiver
      */
-    List<ReceiverNoiseLevel> receiverAttenuationList = new LinkedList<>();
+    Map<Integer, TimePeriodParameters> receiverAttenuationList = new HashMap<>();
 
     /**
      * MaxError DB Processing variable
@@ -78,7 +77,7 @@ public class AttenuationOutputSingleThread implements IComputePathsOut {
      */
     public AttenuationOutputSingleThread(AttenuationOutputMultiThread multiThreadParent, ProgressVisitor progressVisitor) {
         this.multiThread = multiThreadParent;
-        this.noiseMapDatabaseParameters = multiThreadParent.noiseMapDatabaseParameters;
+        this.dbSettings = multiThreadParent.noiseMapDatabaseParameters;
         this.progressVisitor = progressVisitor;
     }
 
@@ -115,6 +114,24 @@ public class AttenuationOutputSingleThread implements IComputePathsOut {
             pathParameters.add(new CnossosPath(proPathParameters));
         }
         return attenuation;
+    }
+
+    /**
+     * Update internal map with new attenuation
+     * @param noiseLevel
+     */
+    private void processNoiseLevel(ReceiverNoiseLevel noiseLevel) {
+        int keyToUpdate = UNKNOWN_SOURCE_ID;
+        if(!dbSettings.isMergeSources()) {
+            keyToUpdate = noiseLevel.source.sourceIndex;
+        }
+        TimePeriodParameters periodParameters = new TimePeriodParameters(
+                dbSettings.isMergeSources() ? new PathFinder.SourcePointInfo() : noiseLevel.source,
+                noiseLevel.period, noiseLevel.levels);
+
+        receiverAttenuationList.merge(keyToUpdate, periodParameters,
[…]
