+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/DelaunayReceiversMaker.java
@@ -10,9 +10,7 @@
 
 package org.noise_planet.noisemodelling.jdbc;
 
-import org.h2gis.utilities.GeometryTableUtilities;
-import org.h2gis.utilities.JDBCUtilities;
-import org.h2gis.utilities.TableLocation;
+import org.h2gis.utilities.*;
 import org.h2gis.utilities.dbtypes.DBTypes;
 import org.h2gis.utilities.dbtypes.DBUtils;
 import org.locationtech.jts.densify.Densifier;
@@ -21,6 +19,8 @@ import org.locationtech.jts.io.WKTWriter;
 import org.locationtech.jts.operation.buffer.BufferOp;
 import org.locationtech.jts.operation.buffer.BufferParameters;
 import org.locationtech.jts.simplify.TopologyPreservingSimplifier;
+import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
+import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
 import org.noise_planet.noisemodelling.pathfinder.delaunay.Triangle;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
 import org.noise_planet.noisemodelling.pathfinder.delaunay.LayerDelaunay;
@@ -32,15 +32,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.sql.Statement;
+import java.sql.*;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
+
+import static org.h2gis.utilities.GeometryTableUtilities.getGeometryColumnNames;
 
 /**
  * Create input receivers built from Delaunay for contructing a NoiseMap rendering.
@@ -398,6 +397,61 @@ public class DelaunayReceiversMaker extends NoiseMapLoader {
         this.epsilon = epsilon;
     }
 
+
+
+    /**
+     * Fetch source geometries and power
+     * @param connection Active connection
+     * @param fetchEnvelope Fetch envelope
+     * @param scene (Out) Propagation process input data
+     * @throws SQLException
+     */
+    public void fetchCellSource(Connection connection, Envelope fetchEnvelope, boolean doIntersection, List<Geometry> sourceGeometries)
+            throws SQLException {
+
+        DBTypes dbType = DBUtils.getDBType(connection.unwrap(Connection.class));
+        TableLocation sourceTableIdentifier = TableLocation.parse(sourcesTableName, dbType);
+        List<String> geomFields = getGeometryColumnNames(connection, sourceTableIdentifier);
+        if (geomFields.isEmpty()) {
+            throw new SQLException(String.format("The table %s does not exists or does not contain a geometry field", sourceTableIdentifier));
+        }
+        String sourceGeomName = geomFields.get(0);
+        Geometry domainConstraint = geometryFactory.toGeometry(fetchEnvelope);
+        Tuple<String, Integer> primaryKey = JDBCUtilities.getIntegerPrimaryKeyNameAndIndex(
+                connection.unwrap(Connection.class), new TableLocation(sourcesTableName, dbType));
+        int pkIndex = primaryKey.second();
+        if (pkIndex < 1) {
+            throw new IllegalArgumentException(String.format("Source table %s does not contain a primary key", sourceTableIdentifier));
+        }
+        try (PreparedStatement st = connection.prepareStatement("SELECT * FROM " + sourcesTableName + " WHERE "
[…]
