+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/NoiseMapByReceiverMakerTest.java
@@ -20,15 +20,12 @@ import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.io.WKTWriter;
-import org.noise_planet.noisemodelling.jdbc.Utils.JDBCComputeRaysOut;
-import org.noise_planet.noisemodelling.jdbc.Utils.JDBCPropagationData;
 import org.noise_planet.noisemodelling.pathfinder.*;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.GroundAbsorption;
 import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
 import org.noise_planet.noisemodelling.pathfinder.utils.profiler.RootProgressVisitor;
 import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
-import org.noise_planet.noisemodelling.propagation.Attenuation;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
@@ -58,324 +55,324 @@ public class NoiseMapByReceiverMakerTest {
             connection.close();
         }
     }
-
-    @Test
-    public void testGroundSurface() throws Exception {
-        try(Statement st = connection.createStatement()) {
-            st.execute(String.format("CALL SHPREAD('%s', 'LANDCOVER2000')", NoiseMapByReceiverMakerTest.class.getResource("landcover2000.shp").getFile()));
-            st.execute(getRunScriptRes("scene_with_landcover.sql"));
-            NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker("BUILDINGS", "ROADS_GEOM", "RECEIVERS");
-            noiseMapByReceiverMaker.setComputeHorizontalDiffraction(true);
-            noiseMapByReceiverMaker.setComputeVerticalDiffraction(true);
-            noiseMapByReceiverMaker.setSoundReflectionOrder(1);
-            noiseMapByReceiverMaker.setReceiverHasAbsoluteZCoordinates(false);
-            noiseMapByReceiverMaker.setSourceHasAbsoluteZCoordinates(false);
-            noiseMapByReceiverMaker.setHeightField("HEIGHT");
-            noiseMapByReceiverMaker.setSoilTableName("LAND_G");
-            noiseMapByReceiverMaker.setComputeVerticalDiffraction(true);
-            noiseMapByReceiverMaker.initialize(connection, new EmptyProgressVisitor());
-
-            noiseMapByReceiverMaker.setComputeRaysOutFactory(new JDBCComputeRaysOut(false));
-            noiseMapByReceiverMaker.setPropagationProcessDataFactory(new JDBCPropagationData());
-
-            Set<Long> receivers = new HashSet<>();
-            noiseMapByReceiverMaker.setThreadCount(1);
-            RootProgressVisitor progressVisitor = new RootProgressVisitor(noiseMapByReceiverMaker.getGridDim() * noiseMapByReceiverMaker.getGridDim(), true, 5);
-            double expectedMaxArea = Math.pow(noiseMapByReceiverMaker.getGroundSurfaceSplitSideLength(), 2);
-            for(int i=0; i < noiseMapByReceiverMaker.getGridDim(); i++) {
-                for(int j=0; j < noiseMapByReceiverMaker.getGridDim(); j++) {
-                    IComputePathsOut out = noiseMapByReceiverMaker.evaluateCell(connection, i, j, progressVisitor, receivers);
-                    if(out instanceof Attenuation) {
[…]
