+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/AttenuationVisitor.java
@@ -10,7 +10,7 @@
 package org.noise_planet.noisemodelling.propagation;
 
 import org.h2gis.api.ProgressVisitor;
-import org.noise_planet.noisemodelling.pathfinder.IComputePathsOut;
+import org.noise_planet.noisemodelling.pathfinder.CutPlaneVisitor;
 import org.noise_planet.noisemodelling.pathfinder.PathFinder;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
@@ -24,7 +24,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 /**
  * Receive vertical cut plane, compute the attenuation corresponding to this plane
  */
-public class AttenuationVisitor implements IComputePathsOut {
+public class AttenuationVisitor implements CutPlaneVisitor {
     public AttenuationComputeOutput multiThreadParent;
     public List<ReceiverNoiseLevel> receiverAttenuationLevels = new ArrayList<>();
     public List<CnossosPath> pathParameters = new ArrayList<CnossosPath>();
@@ -116,12 +116,4 @@ public class AttenuationVisitor implements IComputePathsOut {
         receiverAttenuationLevels.clear();
     }
 
-    /**
-     *
-     * @return an instance of the interface IComputePathsOut
-     */
-    @Override
-    public IComputePathsOut subProcess(ProgressVisitor visitor) {
-        return multiThreadParent.subProcess(visitor);
-    }
 }
