+++ wps_scripts/noisemodelling_jnius.py
@@ -18,7 +18,7 @@ RunnerMain = autoclass('org.noisemodelling.runner.Main')
 LoggerFactory = autoclass('org.slf4j.LoggerFactory')
 PropagationDataBuilder = autoclass('org.noise_planet.noisemodelling.pathfinder.PropagationDataBuilder')
 CnossosPaths = autoclass('org.noise_planet.noisemodelling.pathfinder.cnossos.CnossosPaths')
-IComputePathsOut = autoclass('org.noise_planet.noisemodelling.pathfinder.IComputePathsOut')
+IComputePathsOut = autoclass('org.noise_planet.noisemodelling.pathfinder.CutPlaneVisitor')
 ProfileBuilder = autoclass('org.noise_planet.noisemodelling.pathfinder.ProfileBuilder')
 ProfilerThread = autoclass('org.noise_planet.noisemodelling.pathfinder.utils.ProfilerThread')
 ComputeRaysOutAttenuation = autoclass('org.noise_planet.noisemodelling.propagation.ComputeRaysOutAttenuation')

[…]
