+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseEmissionMaker.java
@@ -9,36 +9,7 @@
 
 package org.noise_planet.noisemodelling.jdbc;
 
-import org.apache.commons.codec.digest.DigestUtils;
-import org.h2.security.SHA256;
-import org.h2gis.functions.spatial.convert.ST_Force3D;
-import org.h2gis.functions.spatial.edit.ST_UpdateZ;
-import org.h2gis.utilities.JDBCUtilities;
-import org.h2gis.utilities.SpatialResultSet;
-import org.locationtech.jts.geom.Coordinate;
-import org.locationtech.jts.geom.Geometry;
-import org.locationtech.jts.geom.LineString;
-import org.noise_planet.noisemodelling.emission.directivity.DirectivitySphere;
-import org.noise_planet.noisemodelling.emission.railway.RailWayParameters;
-import org.noise_planet.noisemodelling.emission.road.cnossos.RoadCnossos;
-import org.noise_planet.noisemodelling.emission.road.cnossos.RoadCnossosParameters;
-import org.noise_planet.noisemodelling.emission.utils.Utils;
-import org.noise_planet.noisemodelling.jdbc.railway.RailWayLWGeom;
-import org.noise_planet.noisemodelling.jdbc.railway.RailWayLWIterator;
-import org.noise_planet.noisemodelling.pathfinder.path.Scene;
-import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
-import org.noise_planet.noisemodelling.propagation.SceneWithAttenuation;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.*;
-
-import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.dbaToW;
+import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.dBToW;
 
 /**
  * Read source database and compute the sound emission spectrum of roads sources
