### 2025-02-03 0921596
fix missing wall alpha on diffraction on horizontal edges
