+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/CnossosPathBuilder.java
@@ -408,6 +408,9 @@ public class CnossosPathBuilder {
                 PointPath pt = points.get(points.size() - 1);
                 pt.type = DIFH;
                 pt.bodyBarrier = bodyBarrier;
+                if(cutPt1 instanceof CutPointWall) {
+                    pt.alphaWall = ((CutPointWall) cutPt1).wallAlpha;
+                }
             }
         }
 
