### 2025-02-04 6be0780
parse old style lw source table option
