+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/AttenuationComputeOutputCnossosTest.java
@@ -9,6 +9,7 @@
 
 package org.noise_planet.noisemodelling.jdbc;
 
+import org.h2gis.api.EmptyProgressVisitor;
 import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.*;
 import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
@@ -26,6 +27,8 @@ import org.slf4j.LoggerFactory;
 
 import java.io.FileWriter;
 import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -500,91 +503,52 @@ public class AttenuationComputeOutputCnossosTest {
 //    }
 //
 //
-//    private double testIgnoreNonSignificantSourcesParam(Connection connection, double maxError) throws SQLException, IOException {
-//        // Init NoiseModelling
-//        LdenNoiseMapLoader ldenNoiseMapLoader = new LdenNoiseMapLoader("BUILDINGS",
-//                "LW_ROADS", "RECEIVERS");
-//
-//        ldenNoiseMapLoader.setMaximumPropagationDistance(5000.0);
-//        ldenNoiseMapLoader.setSoundReflectionOrder(1);
-//        ldenNoiseMapLoader.setThreadCount(1);
-//        ldenNoiseMapLoader.setComputeHorizontalDiffraction(true);
-//        ldenNoiseMapLoader.setComputeVerticalDiffraction(true);
-//        // Building height field name
-//        ldenNoiseMapLoader.setHeightField("HEIGHT");
-//
-//
-//        // Init custom input in order to compute more than just attenuation
-//        // LW_ROADS contain Day Evening Night emission spectrum
-//        LdenNoiseMapParameters ldenNoiseMapParameters = new LdenNoiseMapParameters(LdenNoiseMapParameters.INPUT_MODE.INPUT_MODE_LW_DEN);
-//        ldenNoiseMapParameters.setExportRaysMethod(LdenNoiseMapParameters.ExportRaysMethods.TO_MEMORY);
-//
-//        ldenNoiseMapParameters.setComputeLDay(false);
-//        ldenNoiseMapParameters.setComputeLEvening(false);
-//        ldenNoiseMapParameters.setComputeLNight(false);
-//        ldenNoiseMapParameters.setComputeLDEN(true);
-//        ldenNoiseMapParameters.keepAbsorption = true;
-//        ldenNoiseMapParameters.setMaximumError(maxError);
-//
-//        NoiseMapMaker noiseMapMaker = new NoiseMapMaker(connection, ldenNoiseMapParameters);
-//
-//        ldenNoiseMapLoader.setPropagationProcessDataFactory(noiseMapMaker);
-//        ldenNoiseMapLoader.setComputeRaysOutFactory(noiseMapMaker);
-//
-//        RootProgressVisitor progressLogger = new RootProgressVisitor(1, true, 1);
-//
-//        ldenNoiseMapLoader.initialize(connection, new EmptyProgressVisitor());
-//
-//        ldenNoiseMapParameters.getPropagationProcessPathData(LdenNoiseMapParameters.TIME_PERIOD.DAY).setTemperature(20);
-//        ldenNoiseMapParameters.getPropagationProcessPathData(LdenNoiseMapParameters.TIME_PERIOD.EVENING).setTemperature(16);
-//        ldenNoiseMapParameters.getPropagationProcessPathData(LdenNoiseMapParameters.TIME_PERIOD.NIGHT).setTemperature(10);
-//
-//        ldenNoiseMapLoader.setGridDim(1);
-//
-//        // Set of already processed receivers
[…]
