### 2025-02-05 e026977
remove class inheritance for more clarity
