+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutput.java
@@ -70,226 +70,6 @@ public class AttenuationComputeOutput implements CutPlaneVisitorFactory {
     }
 
     /**
-     * Compute the Attenuation for each frequency with a given sourceId, sourceLi and sourceId
-     * @param data
-     * @param proPathParameters Cnossos paths
-     * @return double list of attenuation
-     */
-    public double[] computeCnossosAttenuation(AttenuationCnossosParameters data, CnossosPath proPathParameters) {
-        if (data == null) {
-            return new double[0];
-        }
-        // cache frequencies
-        double[] frequencies = new double[0];
-        if(scene != null) {
-            frequencies =  scene.profileBuilder.frequencyArray.stream().mapToDouble(value -> value).toArray();
-        }
-        // Compute receiver/source attenuation
-        if(exportAttenuationMatrix) {
-            proPathParameters.keepAbsorption = true;
-            proPathParameters.groundAttenuation.init(data.freq_lvl.size());
-            proPathParameters.init(data.freq_lvl.size());
-        }
-        AttenuationCnossos.init(data);
-        //ADiv computation
-        double[] aDiv = AttenuationCnossos.aDiv(proPathParameters, data);
-        //AAtm computation
-        double[] aAtm = AttenuationCnossos.aAtm(data.getAlpha_atmo(), proPathParameters.getSRSegment().d);
-        //Reflexion computation
-        double[] aRef = AttenuationCnossos.evaluateAref(proPathParameters, data);
-        //For testing purpose
-        if(exportAttenuationMatrix) {
-            proPathParameters.aRef = aRef.clone();
-        }
-        double[] aRetroDiff;
-        //ABoundary computation
-        double[] aBoundary;
-        double[] aGlobalMeteoHom = new double[data.freq_lvl.size()];
-        double[] aGlobalMeteoFav = new double[data.freq_lvl.size()];
-        double[] deltaBodyScreen = new double[data.freq_lvl.size()];
-
-        List<PointPath> ptList = proPathParameters.getPointList();
-
-        // todo get hRail from input data
-        double hRail = 0.5;
-        Coordinate src = ptList.get(0).coordinate;
-        PointPath pDif = ptList.stream().filter(p -> p.type.equals(PointPath.POINT_TYPE.DIFH)).findFirst().orElse(null);
-
-        if (pDif != null && !pDif.alphaWall.isEmpty()) {
-            if (pDif.bodyBarrier){
-
-                int n = 3;
-                Coordinate rcv = ptList.get(ptList.size() - 1).coordinate;
-                double[][] deltaGeo = new double[n+1][data.freq_lvl.size()];
-                double[][] deltaAbs = new double[n+1][data.freq_lvl.size()];
-                double[][] deltaDif = new double[n+1][data.freq_lvl.size()];
-                double[][] deltaRef = new double[n+1][data.freq_lvl.size()];
-                double[][] deltaRetroDifi = new double[n+1][data.freq_lvl.size()];
-                double[][] deltaRetroDif = new double[n+1][data.freq_lvl.size()];
-                double[] deltaL = new double[data.freq_lvl.size()];
-                Arrays.fill(deltaL, dBToW(0.0));
-
-                double db = pDif.coordinate.x;
-                double hb = pDif.coordinate.y;
-                Coordinate B = new Coordinate(db,hb);
-
-                double Cref = 1;
-                double dr = rcv.x;
-                double h0 = ptList.get(0).altitude+hRail;
-                double hs = ptList.get(0).altitude+src.y-hRail;
-                double hr = ptList.get(ptList.size()-1).altitude + ptList.get(ptList.size()-1).coordinate.y-h0;
-                double[] r = new double[4];
-                if (db<5*hb) {
-                    for (int idfreq = 0; idfreq < data.freq_lvl.size(); idfreq++) {
[…]
