### 2025-02-05 b955939
use only period covered by source emission (do not output attenuation if we don't find the source emission for a period)
