### 2025-02-05 6586051
issue with lden. we cant create a lden source, as attenuation depend on each period. i have to reword that special part.
