+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/NoiseMapByReceiverMakerTest.java
@@ -20,7 +20,9 @@ import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.io.WKTWriter;
+import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
 import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
+import org.noise_planet.noisemodelling.jdbc.output.NoiseMapWriter;
 import org.noise_planet.noisemodelling.jdbc.utils.CellIndex;
 import org.noise_planet.noisemodelling.pathfinder.*;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
@@ -34,6 +36,7 @@ import java.io.DataOutputStream;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.sql.Connection;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.*;
@@ -86,132 +89,172 @@ public class NoiseMapByReceiverMakerTest {
         }
     }
 
-//    @Test
-//    public void testNoiseMapBuilding() throws Exception {
-//        try(Statement st = connection.createStatement()) {
-//            st.execute(String.format("CALL SHPREAD('%s', 'LANDCOVER2000')", NoiseMapByReceiverMakerTest.class.getResource("landcover2000.shp").getFile()));
-//            st.execute(getRunScriptRes("scene_with_landcover.sql"));
-//            DelaunayReceiversMaker noisemap = new DelaunayReceiversMaker("BUILDINGS", "ROADS_GEOM");
-//            noisemap.setReceiverHasAbsoluteZCoordinates(false);
-//            noisemap.setSourceHasAbsoluteZCoordinates(false);
-//            noisemap.setHeightField("HEIGHT");
-//            noisemap.initialize(connection, new EmptyProgressVisitor());
-//
-//            AtomicInteger pk = new AtomicInteger(0);
-//            for(int i=0; i < noisemap.getGridDim(); i++) {
-//                for(int j=0; j < noisemap.getGridDim(); j++) {
-//                    noisemap.generateReceivers(connection, i, j, "NM_RECEIVERS", "TRIANGLES", pk);
-//                }
-//            }
-//            assertNotSame(0, pk.get());
-//        }
-//    }
-//
-//    private static String createSource(Geometry source, double lvl, Orientation sourceOrientation, int directivityId) {
-//        StringBuilder sb = new StringBuilder("CREATE TABLE ROADS_GEOM(PK SERIAL PRIMARY KEY, THE_GEOM GEOMETRY, YAW REAL, PITCH REAL, ROLL REAL, DIR_ID INT");
-//        StringBuilder values = new StringBuilder("(row_number() over())::int, ST_SETSRID('");
-//        values.append(new WKTWriter(3).write(source));
-//        values.append("', 2154) THE_GEOM, ");
-//        values.append(sourceOrientation.yaw);
-//        values.append(" YAW, ");
-//        values.append(sourceOrientation.pitch);
-//        values.append(" PITCH, ");
-//        values.append(sourceOrientation.roll);
-//        values.append(" ROLL, ");
-//        values.append(directivityId);
-//        values.append(" DIR_ID");
-//        AttenuationCnossosParameters data = new AttenuationCnossosParameters(false);
-//        for(String period : new String[] {"D", "E", "N"}) {
-//            for (int freq : data.freq_lvl) {
-//                String fieldName = "LW" + period + freq;
-//                sb.append(", ");
-//                sb.append(fieldName);
-//                sb.append(" real");
-//                values.append(", ");
[…]
