### 2025-02-06 c816606
rework, compute lden on finaliseReceiver step
