+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/input/SceneWithEmission.java
@@ -1,3 +1,11 @@
+/**
+ * NoiseModelling is a library capable of producing noise maps. It can be freely used either for research and education, as well as by experts in a professional use.
+ * <p>
+ * NoiseModelling is distributed under GPL 3 license. You can read a copy of this License in the file LICENCE provided with this software.
+ * <p>
+ * Official webpage : http://noise-planet.org/noisemodelling.html
+ * Contact: contact@noise-planet.org
+ */
 package org.noise_planet.noisemodelling.jdbc.input;
 
 import org.h2gis.utilities.SpatialResultSet;
@@ -41,7 +49,7 @@ public class SceneWithEmission extends SceneWithAttenuation {
     public void processTrafficFlowDEN(Long pk, SpatialResultSet rs) throws SQLException {
         // Source table PK, GEOM, LV_D, LV_E, LV_N ...
         double[][] lw = EmissionTableGenerator.computeLw(rs, sceneDatabaseInputSettings.coefficientVersion, sourceFieldNames);
-        // Will generate D E N and DEN emission
+        // Will generate D E N emission
         for (EmissionTableGenerator.STANDARD_PERIOD period : EmissionTableGenerator.STANDARD_PERIOD.values()) {
             addSourceEmission(pk, EmissionTableGenerator.STANDARD_PERIOD_VALUE[period.ordinal()], lw[period.ordinal()]);
         }
@@ -100,9 +108,7 @@ public class SceneWithEmission extends SceneWithAttenuation {
 
     private void processEmissionDEN(Long pk, SpatialResultSet rs) throws SQLException {
         List<Integer> frequencyArray = profileBuilder.frequencyArray;
-        double[] lden = new double[0];
-        for (EmissionTableGenerator.STANDARD_PERIOD period :
-                Arrays.copyOfRange(EmissionTableGenerator.STANDARD_PERIOD.values(),0,3)) {
+        for (EmissionTableGenerator.STANDARD_PERIOD period : EmissionTableGenerator.STANDARD_PERIOD.values()) {
             double[] lw = new double[profileBuilder.frequencyArray.size()];
             String periodFieldName = EmissionTableGenerator.STANDARD_PERIOD_VALUE[period.ordinal()];
             for (int i = 0, frequencyArraySize = frequencyArray.size(); i < frequencyArraySize; i++) {
@@ -112,9 +118,7 @@ public class SceneWithEmission extends SceneWithAttenuation {
                         rs.getDouble(tableFieldName));
             }
             addSourceEmission(pk, periodFieldName, lw);
-            lden = sumArray(lden, AcousticIndicatorsFunctions.multiplicationArray(lw, EmissionTableGenerator.RATIOS[period.ordinal()]));
         }
-        addSourceEmission(pk, EmissionTableGenerator.STANDARD_PERIOD_VALUE[3], lden);
     }
 
     public void addSourceEmission(Long pk, ResultSet rs) throws SQLException {
@@ -162,77 +166,4 @@ public class SceneWithEmission extends SceneWithAttenuation {
         }
     }
 
-    /**
-     * SceneWithEmission will read table according to this settings
-     */
-    public static class SceneDatabaseInputSettings {
-        public enum INPUT_MODE {
-            /** Read traffic from geometry source table */
-            INPUT_MODE_TRAFFIC_FLOW_DEN,
-            /** Read source emission noise level limited to DEN periods from source geometry table */
-            INPUT_MODE_LW_DEN,
-            /** Read traffic from emission source table for each period */
-            INPUT_MODE_TRAFFIC_FLOW,
-            /** Read source emission noise level from source emission table for each period */
-            INPUT_MODE_LW,
-            /** Compute only attenuation */
[…]
