+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/TimePeriodParametersLdenAttenuationOutputMultiThreadLoaderFactoryTest.java
@@ -8,16 +8,37 @@
  */
 package org.noise_planet.noisemodelling.jdbc;
 
+import org.h2gis.api.EmptyProgressVisitor;
 import org.h2gis.functions.factory.H2GISDBFactory;
+import org.h2gis.functions.io.dbf.DBFRead;
+import org.h2gis.functions.io.shp.SHPRead;
 import org.h2gis.utilities.JDBCUtilities;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.locationtech.jts.geom.LineString;
+import org.noise_planet.noisemodelling.emission.LineSource;
+import org.noise_planet.noisemodelling.emission.railway.RailWayParameters;
 import org.noise_planet.noisemodelling.emission.railway.cnossos.RailwayCnossos;
+import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
+import org.noise_planet.noisemodelling.jdbc.input.SceneDatabaseInputSettings;
+import org.noise_planet.noisemodelling.jdbc.railway.RailWayLWGeom;
+import org.noise_planet.noisemodelling.jdbc.railway.RailWayLWIterator;
+import org.noise_planet.noisemodelling.jdbc.utils.CellIndex;
+import org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
+import static org.junit.jupiter.api.Assertions.*;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.sumArray;
 import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.sumDbArray;
 
@@ -91,699 +112,308 @@ public class TimePeriodParametersLdenAttenuationOutputMultiThreadLoaderFactoryTe
 //            }
 //        }
 //    }
-//
-//    @Test
-//    public void testNoiseEmissionRailWay() throws SQLException, IOException {
-//        SHPRead.importTable(connection, TimePeriodParametersLdenAttenuationOutputMultiThreadLoaderFactoryTest.class.getResource("RailTrack.shp").getFile());
-//        DBFRead.importTable(connection, TimePeriodParametersLdenAttenuationOutputMultiThreadLoaderFactoryTest.class.getResource("RailTrain.dbf").getFile());
-//        int expectedNumberOfRows;
-//        try(ResultSet rs = connection.createStatement().executeQuery("SELECT COUNT(*) FROM RAILTRACK")) {
-//            assertTrue(rs.next());
-//            expectedNumberOfRows = rs.getInt(1);
-//        }
-//        RailWayLWIterator railWayLWIterator = new RailWayLWIterator(connection,"RAILTRACK", "RAILTRAIN","RailwayVehiclesCnossos.json","RailwayTrainsets.json", "RailwayCnossosSNCF_2021.json");
-//
-//        int numberOfRows = 0;
-//        while (railWayLWIterator.hasNext()) {
-//            RailWayLWGeom v = railWayLWIterator.next();
-//            assertNotNull(v);
-//            numberOfRows++;
-//        }
-//        assertEquals(expectedNumberOfRows, numberOfRows);
-//    }
-//
-//    @Test
-//    public void testNoiseEmissionRailWayTwoGeoms() throws SQLException, IOException {
[…]
