+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/NoiseMapByReceiverMakerTest.java
@@ -206,6 +206,7 @@ public class NoiseMapByReceiverMakerTest {
             noiseMapByReceiverMaker.getNoiseMapDatabaseParameters().exportRaysMethod = NoiseMapDatabaseParameters.ExportRaysMethods.TO_RAYS_TABLE;
             noiseMapByReceiverMaker.getNoiseMapDatabaseParameters().exportCnossosPathWithAttenuation = true;
             noiseMapByReceiverMaker.getNoiseMapDatabaseParameters().exportAttenuationMatrix = true;
+            noiseMapByReceiverMaker.getNoiseMapDatabaseParameters().mergeSources = true;
             noiseMapByReceiverMaker.setBodyBarrier(true);
 
             // Use train directivity functions instead of discrete directivity
@@ -238,14 +239,15 @@ public class NoiseMapByReceiverMakerTest {
             try(ResultSet rs = st.executeQuery("SELECT IDRECEIVER, PATH FROM " + parameters.raysTable + " WHERE PERIOD='D' ORDER BY IDRECEIVER")) {
                 assertTrue(rs.next());
                 assertEquals(1, rs.getInt(1));
-                String json = rs.getString(2);
-                CnossosPath cnossosPath = NoiseMapWriter.jsonToPropagationPath(json);
+                CnossosPath cnossosPath = NoiseMapWriter.jsonToPropagationPath(rs.getString(2));
                 // This is source orientation, not relevant to receiver position
                 assertOrientationEquals(new Orientation(45, 0.81, 0), cnossosPath.getSourceOrientation(), 0.01);
                 assertOrientationEquals(new Orientation(330.2084079818916,-5.947213381005439,0.0), cnossosPath.raySourceReceiverDirectivity, 0.01);
                 assertTrue(rs.next());
-                assertEquals(2, rs.getInt(1));
-                assertFalse(rs.next());
+                assertEquals(1, rs.getInt(1));
+                cnossosPath = NoiseMapWriter.jsonToPropagationPath(rs.getString(2));
+                assertOrientationEquals(new Orientation(45, 0.81, 0), cnossosPath.getSourceOrientation(), 0.01);
+                assertOrientationEquals(new Orientation(336.9922375343167,-4.684918495003125,0.0), cnossosPath.raySourceReceiverDirectivity, 0.01);
             }
         }
     }
@@ -256,159 +258,75 @@ public class NoiseMapByReceiverMakerTest {
                 new double[]{orientationB.yaw, orientationB.pitch, orientationB.roll}, epsilon, orientationA+" != "+orientationB);
     }
 
-//
-//    @Test
-//    public void testLineDirectivity() throws Exception {
-//        try (Statement st = connection.createStatement()) {
-//            st.execute("CREATE TABLE BUILDINGS(pk serial PRIMARY KEY, the_geom geometry, height real)");
-//            st.execute(createSource(new GeometryFactory().createLineString(
-//                    new Coordinate[]{new Coordinate(223915.72,6757480.22 ,5),
-//                            new Coordinate(223920.72,6757485.22, 5.1 )}), 91,
-//                    new Orientation(0,0,0),4));
-//            st.execute("create table receivers(id serial PRIMARY KEY, the_geom GEOMETRY(pointZ));\n" +
-//                    "insert into receivers(the_geom) values ('POINTZ (223922.55 6757495.27 4.0)');" +
-//                    "insert into receivers(the_geom) values ('POINTZ (223936.42 6757471.91 4.0)');");
-//            NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker("BUILDINGS", "ROADS_GEOM", "RECEIVERS");
[…]
