### 2025-02-06 f3be62d
use frequency prepend string on output
