+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/SceneWithAttenuation.java
@@ -57,6 +57,7 @@ public class SceneWithAttenuation extends Scene {
 
     public SceneWithAttenuation(ProfileBuilder profileBuilder) {
         super(profileBuilder);
+        defaultCnossosParameters.setFrequencies(profileBuilder.frequencyArray);
     }
 
     public SceneWithAttenuation() {

[…]
