### 2025-02-06 0d2094c
ISOContouring should now support PERIOD field. We generate a isocontour for each period in the same table.
