+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/utils/IsoSurface.java
@@ -21,6 +21,7 @@ import org.locationtech.jts.geom.*;
 import org.locationtech.jts.index.quadtree.Quadtree;
 import org.locationtech.jts.operation.union.CascadedPolygonUnion;
 import org.locationtech.jts.simplify.TopologyPreservingSimplifier;
+import org.noise_planet.noisemodelling.jdbc.NoiseMapDatabaseParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -38,13 +39,17 @@ import static org.noise_planet.noisemodelling.emission.utils.Utils.dbaToW;
 public class IsoSurface {
     Logger log = LoggerFactory.getLogger(IsoSurface.class);
     static final int BATCH_MAX_SIZE = 500;
-    String pointTable = "LDEN_RESULT";
+    String pointTable = NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME;
     String triangleTable = "TRIANGLES";
     String outputTable = "CONTOURING_NOISE_MAP";
-    String pointTableField = "LAEQ";
+    String pointTableField = "RECEIVERS";
     List<Double> isoLevels;
     List<String> isoLabels;
     boolean smooth = true;
+    /**
+     * Value table has a PERIOD field. Produce one iso-surface map per PERIOD distinct values.
+     */
+    boolean aggregateByPeriod = false;
 
     boolean mergeTriangles = true;
     double smoothCoefficient = 1.0;
@@ -367,7 +372,7 @@ public class IsoSurface {
      * @param cellId area id (aggregate polygons by large area in order to avoid memory overloading)
      * @param polys Polygons by isolevel
      */
-    void processCell(Connection connection, int cellId, Map<Short, ArrayList<Geometry>> polys) throws SQLException {
+    void processCell(Connection connection, int cellId, Map<Short, ArrayList<Geometry>> polys, String period) throws SQLException {
         // First step
         // Smoothing of polygons
         GeometryFactory factory = new GeometryFactory(new PrecisionModel(), srid);
@@ -423,8 +428,17 @@ public class IsoSurface {
         }
         // Second step insertion
         int batchSize = 0;
-        try(PreparedStatement ps = connection.prepareStatement("INSERT INTO " + TableLocation.parse(outputTable)
-                + "(cell_id, the_geom, ISOLVL, ISOLABEL) VALUES (?, ?, ?, ?);")) {
+        StringBuilder insertQuery = new StringBuilder().append("INSERT INTO ").append(TableLocation.parse(outputTable))
+                .append("(cell_id");
+        if(aggregateByPeriod) {
+            insertQuery.append(", PERIOD");
+        }
+        insertQuery.append(", the_geom, ISOLVL, ISOLABEL) VALUES (?");
+        if(aggregateByPeriod) {
+            insertQuery.append(", ?");
+        }
+        insertQuery.append(", ?, ?, ?);");
+        try(PreparedStatement ps = connection.prepareStatement(insertQuery.toString())) {
             for (Map.Entry<Short, ArrayList<Geometry>> entry : polys.entrySet()) {
                 ArrayList<Polygon> polygons = new ArrayList<>();
                 if(!smooth && mergeTriangles) {
@@ -471,6 +485,9 @@ public class IsoSurface {
                     }
                     int parameterIndex = 1;
                     ps.setInt(parameterIndex++, cellId);
+                    if(aggregateByPeriod) {
+                        ps.setString(parameterIndex++, period);
+                    }
                     ps.setObject(parameterIndex++, polygon);
                     ps.setInt(parameterIndex++, entry.getKey());
[…]
