### 2025-02-07 1707651
Iso contour test case with PERIOD field, ok
