+++ noisemodelling-tutorial-01/src/main/java/org/noise_planet/nmtutorial01/main.java
@@ -1,10 +1,7 @@
 package org.noise_planet.nmtutorial01;
 
-import org.cts.crs.CRSException;
-import org.cts.op.CoordinateOperationException;
 import org.h2.value.ValueBoolean;
 import org.h2gis.api.EmptyProgressVisitor;
-import org.h2gis.api.ProgressVisitor;
 import org.h2gis.functions.io.geojson.GeoJsonRead;
 import org.h2gis.functions.io.shp.SHPWrite;
 import org.h2gis.utilities.GeometryTableUtilities;
@@ -12,28 +9,17 @@ import org.h2gis.utilities.JDBCUtilities;
 import org.h2gis.utilities.TableLocation;
 import org.h2gis.utilities.dbtypes.DBTypes;
 import org.h2gis.utilities.dbtypes.DBUtils;
+import org.noise_planet.noisemodelling.jdbc.NoiseMapByReceiverMaker;
+import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
 import org.noise_planet.noisemodelling.jdbc.utils.IsoSurface;
-import org.noise_planet.noisemodelling.jdbc.utils.CellIndex;
-import org.noise_planet.noisemodelling.jdbc.LdenNoiseMapParameters;
-import org.noise_planet.noisemodelling.jdbc.NoiseMapMaker;
-import org.noise_planet.noisemodelling.jdbc.LdenNoiseMapLoader;
 import org.noise_planet.noisemodelling.jdbc.DelaunayReceiversMaker;
-import org.noise_planet.noisemodelling.pathfinder.CutPlaneVisitor;
 import org.noise_planet.noisemodelling.pathfinder.delaunay.LayerDelaunayError;
-import org.noise_planet.noisemodelling.pathfinder.profilebuilder.ProfileBuilder;
 import org.noise_planet.noisemodelling.pathfinder.utils.profiler.RootProgressVisitor;
-import org.noise_planet.noisemodelling.pathfinder.utils.profiler.JVMMemoryMetric;
-import org.noise_planet.noisemodelling.pathfinder.utils.documents.KMLDocument;
-import org.noise_planet.noisemodelling.pathfinder.utils.profiler.ProfilerThread;
-import org.noise_planet.noisemodelling.pathfinder.utils.profiler.ProgressMetric;
-import org.noise_planet.noisemodelling.pathfinder.utils.profiler.ReceiverStatsMetric;
-import org.noise_planet.noisemodelling.propagation.AttenuationComputeOutput;
+import org.noise_planet.noisemodelling.propagation.cnossos.AttenuationCnossosParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.xml.stream.XMLStreamException;
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Paths;
 import java.sql.Connection;
@@ -41,7 +27,6 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.time.LocalDateTime;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -110,105 +95,47 @@ class Main {
                 ValueBoolean.TRUE);
 
         // Init NoiseModelling
-        LdenNoiseMapLoader ldenNoiseMapLoader = new LdenNoiseMapLoader(tableBuildings.toString(),
+        NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker(tableBuildings.toString(),
[…]
