### 2025-02-10 766d67e
when mergeSources is enabled and source source is Day Evening Night, the expected rows on the results should be 4x the input receivers count.
