+++ noisemodelling-tutorial-01/src/test/java/org/noise_planet/nmtutorial01/TestDiffraction.java
@@ -1,2 +1,0 @@
-package org.noise_planet.nmtutorial01;public class TestDiffraction {
-}
