### 2025-02-10 5eba72a
read write atmospheric settings per period from the database
