+++ noisemodelling-tutorial-01/src/test/java/org/noise_planet/nmtutorial01/TutorialTest.java
@@ -7,7 +7,9 @@ import org.h2gis.utilities.TableLocation;
 import org.h2gis.utilities.dbtypes.DBTypes;
 import org.h2gis.utilities.dbtypes.DBUtils;
 import org.junit.jupiter.api.Test;
+import org.noise_planet.noisemodelling.jdbc.NoiseMapByReceiverMaker;
 import org.noise_planet.noisemodelling.jdbc.NoiseMapDatabaseParameters;
+import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
 import org.postgresql.util.PSQLException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,7 +36,7 @@ public class TutorialTest {
         try(Connection connection = JDBCUtilities.wrapConnection(dataSourceFactory.createDataSource(p).getConnection())) {
             connection.createStatement().execute("DROP TABLE IF EXISTS receivers_level");
             connection.createStatement().execute("DROP TABLE IF EXISTS contouring_noise_map");
-            Main.mainWithConnection(connection, "target");
+            NoiseMapByReceiverMaker map = Main.mainWithConnection(connection, "target");
             String receiverTable = TableLocation.capsIdentifier(
                     NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME, DBTypes.POSTGIS);
             assertTrue(JDBCUtilities.tableExists(connection.unwrap(Connection.class), receiverTable));
@@ -51,6 +53,11 @@ public class TutorialTest {
             // D E N and DEN, should be 4 more rows than receivers
             assertEquals(receiversRowCount * 4, resultRowCount);
 
+            assertEquals(3, ((DefaultTableLoader)map.getTableLoader()).getCnossosParametersPerPeriod().size());
+            assertEquals(20, ((DefaultTableLoader)map.getTableLoader()).getCnossosParametersPerPeriod().get("D").temperature);
+            assertEquals(16, ((DefaultTableLoader)map.getTableLoader()).getCnossosParametersPerPeriod().get("E").temperature);
+            assertEquals(10, ((DefaultTableLoader)map.getTableLoader()).getCnossosParametersPerPeriod().get("N").temperature);
+
         } catch (PSQLException psqlException) {
             if(!(psqlException.getCause() instanceof ConnectException)) {
                 throw psqlException;
@@ -66,7 +73,29 @@ public class TutorialTest {
         try(Connection connection = JDBCUtilities.wrapConnection(
                 H2GISDBFactory.createSpatialDataBase(TutorialTest.class.getSimpleName(),
                         true, ""));) {
-            Main.mainWithConnection(connection, "target");
+            NoiseMapByReceiverMaker map = Main.mainWithConnection(connection, "target");
+
+            String receiverTable = TableLocation.capsIdentifier(
+                    NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME, DBTypes.H2GIS);
+            assertTrue(JDBCUtilities.tableExists(connection.unwrap(Connection.class), receiverTable));
+            assertTrue(JDBCUtilities.hasField(connection.unwrap(Connection.class), receiverTable, "period"));
+            assertTrue(JDBCUtilities.tableExists(connection.unwrap(Connection.class), "contouring_noise_map"));
+            assertTrue(JDBCUtilities.hasField(connection.unwrap(Connection.class), "contouring_noise_map", "period"));
+
+
+            int receiversRowCount = JDBCUtilities.getRowCount(connection, "RECEIVERS");
+
+            int resultRowCount = JDBCUtilities.getRowCount(connection,
+                    receiverTable);
+
[…]
