### 2025-02-10 6796818
start fix wps scripts. Will migrate probalistic from dynamicTools (pierreomond:5.X)
