+++ wps_scripts/src/main/java/org/noisemodelling/runner/Main.java
@@ -35,6 +35,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.text.NumberFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -155,7 +156,7 @@ public class Main {
             String workingDir = "";
             String scriptPath = "";
             String databaseName = "h2gisdb";
-            Map<String, String> customParameters = new HashMap<>();
+            Map<String, Object> customParameters = new HashMap<>();
             boolean printVersion = false;
 
             CommandLineParser commandLineParser = new DefaultParser();
@@ -206,7 +207,18 @@ public class Main {
                     commandLine = commandLineParser.parse(options, args);
                     for (Iterator<Option> it = commandLine.iterator(); it.hasNext(); ) {
                         Option option = it.next();
-                        customParameters.put(option.getOpt(), option.getValue());
+                        if (option.getType() == String.class) {
+                            customParameters.put(option.getOpt(), option.getValue());
+                        } else if (option.getType() == Boolean.class) {
+                            customParameters.put(option.getOpt(), Boolean.valueOf(option.getValue()));
+                        } else if (option.getType() == Integer.class) {
+                            customParameters.put(option.getOpt(), Integer.valueOf(option.getValue()));
+                        } else if (option.getType() == Double.class) {
+                            customParameters.put(option.getOpt(),
+                                    NumberFormat.getInstance(Locale.ROOT).parse(option.getValue()).doubleValue());
+                        } else {
+                            throw new IllegalArgumentException("Unsupported type for option " + option.getOpt());
+                        }
                     }
                 } catch (ParseException ex) {
                     logger.info(ex.getMessage());
