+++ wps_scripts/src/test/java/org/noise_planet/noisemodelling/runner/MainTest.java
@@ -1,0 +1,13 @@
+package org.noise_planet.noisemodelling.runner;
+
+import org.junit.jupiter.api.Test;
+import org.noisemodelling.runner.Main;
+
+public class MainTest {
+
+    @Test
+    public void testDoubleArg() throws Exception {
+        String receiverPath = MainTest.class.getResource("../wps/receivers.shp").getPath()
+        Main.main("-w", "target", "-s", "test.groovy", "--pathFile", receiverPath);
+    }
+}
