### 2025-02-11 fd28cfd
fix wps scripts
