+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/DirectivityTableLoaderTest.java
@@ -15,6 +15,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.noise_planet.noisemodelling.emission.LineSource;
 import org.noise_planet.noisemodelling.emission.directivity.DirectivityRecord;
+import org.noise_planet.noisemodelling.emission.directivity.DirectivitySphere;
 import org.noise_planet.noisemodelling.emission.directivity.cnossos.RailwayCnossosDirectivitySphere;
 import org.noise_planet.noisemodelling.emission.directivity.DiscreteDirectivitySphere;
 import org.noise_planet.noisemodelling.jdbc.input.DefaultTableLoader;
@@ -80,14 +81,15 @@ public class DirectivityTableLoaderTest {
         }
 
         // Data is inserted now fetch it from the database
-        Map<Integer, DiscreteDirectivitySphere> directivities = DefaultTableLoader.fetchDirectivity(connection,
+        Map<Integer, DirectivitySphere> directivities = DefaultTableLoader.fetchDirectivity(connection,
                 "DIRTEST", 1);
 
         assertEquals(1, directivities.size());
 
         assertTrue(directivities.containsKey(1));
 
-        DiscreteDirectivitySphere d = directivities.get(1);
+        assertInstanceOf(DiscreteDirectivitySphere.class, directivities.get(1));
+        DiscreteDirectivitySphere d = (DiscreteDirectivitySphere)directivities.get(1);
         for(DirectivityRecord directivityRecord : d.getRecordsTheta()) {
             double[] attSpectrum = att.getAttenuationArray(freqTest, directivityRecord.getPhi(), directivityRecord.getTheta());
             assertArrayEquals(attSpectrum, directivityRecord.getAttenuation(), 1e-2);
