+++ wps_scripts/src/test/java/org/noise_planet/noisemodelling/runner/MainTest.java
@@ -7,7 +7,9 @@ public class MainTest {
 
     @Test
     public void testDoubleArg() throws Exception {
-        String receiverPath = MainTest.class.getResource("../wps/receivers.shp").getPath()
-        Main.main("-w", "target", "-s", "test.groovy", "--pathFile", receiverPath);
+        String receiverPath = MainTest.class.getResource("../wps/receivers.shp").getPath();
+        Main.main("-w", "target",
+                "-s", "groovy/org/noise_planet/noisemodelling/wps/Import_and_Export/Import_File.groovy",
+                "--pathFile", receiverPath);
     }
 }
