+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/NoiseEmissionMaker.java
@@ -1,179 +1,0 @@
-/**
- * NoiseModelling is a library capable of producing noise maps. It can be freely used either for research and education, as well as by experts in a professional use.
- * <p>
- * NoiseModelling is distributed under GPL 3 license. You can read a copy of this License in the file LICENCE provided with this software.
- * <p>
- * Official webpage : http://noise-planet.org/noisemodelling.html
- * Contact: contact@noise-planet.org
- */
-
-package org.noise_planet.noisemodelling.jdbc;
-
-import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.dBToW;
-
-/**
- * Read source database and compute the sound emission spectrum of roads sources
- */
-public class NoiseEmissionMaker {
-//
-//    public static final double DAY_RATIO = 12. / 24.;
-//    public static final double EVENING_RATIO = 4. / 24. * dbaToW(5.0);
-//    public static final double NIGHT_RATIO = 8. / 24. * dbaToW(10.0);
-//
-//    /**
-//     * Set of unique power spectrum. Key is the hash of the spectrum. Power spectrum in energetic e = pow(10, dbVal / 10.0)
-//     */
-//    public Map<String, double[]> powerSpectrum = new HashMap<>();
-//
-//    //  For each source index give the map between period and source power spectrum hash value
-//    public List<HashMap<String, String>> wjSources = new ArrayList<>();
-//
-//    public NoiseMapParameters noiseMapParameters;
-//
-//    /**
-//     *  Create NoiseEmissionMaker constructor
-//     * @param builder
-//     * @param noiseMapParameters
-//     */
-//    public NoiseEmissionMaker(ProfileBuilder builder, NoiseMapParameters noiseMapParameters) {
-//        super(builder);
-//        this.noiseMapParameters = noiseMapParameters;
-//    }
-//
-//    public static byte[] convertDoubleArrayToByteArray(double[] doubleArray) {
-//        ByteBuffer byteBuffer = ByteBuffer.allocate(doubleArray.length * Double.BYTES);
-//        for (double value : doubleArray) {
-//            byteBuffer.putDouble(value);
-//        }
-//        return byteBuffer.array();
-//    }
-//
-//
-//    String updatePowerSpectrumSet(double[] wj) {
-//        String spectrumHash = new String(DigestUtils.getSha256Digest().digest(convertDoubleArrayToByteArray(wj)),
-//                StandardCharsets.UTF_8);
-//        powerSpectrum.putIfAbsent(spectrumHash, wj);
-//        return spectrumHash;
-//    }
-//
-//    /**
-//     * Adds a noise source with its properties to the noise map.
-//     * @param pk Unique source identifier
-//     * @param geom Source geometry
-//     * @param rs Additional attributes fetched from database
-//     * @throws SQLException
-//     * @throws IOException
-//     */
-//    @Override
-//    public void addSource(Long pk, Geometry geom, SpatialResultSet rs) throws SQLException {
-//        super.addSource(pk, geom, rs);
-//        double[][] res = computeLw(rs);
-//        if(noiseMapParameters.computeLDay || noiseMapParameters.computeLDEN) {
-//            wjSourcesD.add(res[0]);
-//        }
-//        if(noiseMapParameters.computeLEvening || noiseMapParameters.computeLDEN) {
-//            wjSourcesE.add(res[1]);
-//        }
-//        if(noiseMapParameters.computeLNight || noiseMapParameters.computeLDEN) {
-//            wjSourcesN.add(res[2]);
-//        }
-//    }
-//
-//    /**
-//     * Retrieves the ground speed of the noise source at the specified index.
-//     * @param srcIndex
[…]
