+++ Docs/Dynamic_Tutorial.rst
@@ -1,0 +1,257 @@
+Dynamic Maps - groovy scripts
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Many publications have emerged showcasing the use of **NoiseModelling** to create dynamic maps (see `scientific production`_).
+
+.. _scientific production : https://noisemodelling.readthedocs.io/en/latest/Scientific_production.html
+
+If you'd like to achieve similar results but you feel a bit lost, this tutorial is here to help you navigate through the process.
+
+There are three main approaches to creating dynamic maps using NoiseModelling:
+
+1. **A road network with a single traffic flow**  
+   You have a road network and a single traffic flow associated with a specific time period (e.g., 24h). You want to compute dynamic indicators such as **L10**, **L90**, or the **number of events exceeding a threshold** or to get time series for the same time period.
+
+2. **A road network with traffic flows at regular intervals**  
+   You have a road network and traffic flow data available at regular intervals (e.g., hourly or every 15 minutes), and you want to generate a dynamic noise map every 15 min.
+
+3. **A road network with associated spatio-temporal data of moving sources**
+   You have spatio-temporal information about vehicles moving around a network (e.g., from traffic simulations such as Simuvya or SUMO; or from trajectories of drones). You want to compute **time-series data at each receiver** corresponding to the passage of these sources.
+
+A Word of Caution
+-----------------
+
+**Caution** : In all these cases, we assume that the **sound attenuation between the source and receiver remains constant throughout the calculation**. This is a strong approximation, and there are ways to account for variations, but this tutorial will not cover such specific cases.
+
+Dynamic mapping has its subtleties, and it's important to be aware of them to avoid errors. We recommend referring to the following documents for a better understanding of these concepts:
+
+- Can, A., & Aumond, P. (2018). Estimation of road traffic noise emissions: The influence of speed and acceleration. Transportation Research Part D: Transport and Environment, 58, 155-171.
+- Gozalo, G. R., Aumond, P., & Can, A. (2020). Variability in sound power levels: Implications for static and dynamic traffic models. Transportation Research Part D: Transport and Environment, 84, 102339.
+- Le Bescond, V., Can, A., Aumond, P., & Gastineau, P. (2021). Open-source modeling chain for the dynamic assessment of road traffic noise exposure. Transportation Research Part D: Transport and Environment, 94, 102793.
+
+This is an **advanced tutorial**, so you may need to open the Groovy scripts to see what’s under the hood. Assumptions are freely made, specific formats are expected, and so on. To understand the required data formats and check the expected structure of the input tables, please refer also to the example input tables and spatial layers!
+
+Study Cases
+---------------
+
+Case 1: A Road Network with a Single Traffic Flow
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. code-block:: groovy
+
+         // Import the road network (with predicted traffic flows) and buildings from an OSM file
+        new Import_OSM().exec(connection, [
+                "pathFile"      : TestImportExport.getResource("map.osm.gz").getPath(),
+                "targetSRID"    : 2154,
+                "ignoreGround"  : true,
+                "ignoreBuilding": false,
+                "ignoreRoads"   : false,
+                "removeTunnels" : true
+        ]);
+
+        // Create a receiver grid
+        new Regular_Grid().exec(connection,  ["buildingTableName": "BUILDINGS",
+                                              "sourcesTableName" : "ROADS",
+                                              "delta"            : 25])
+
[…]
