+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/NoiseMapByReceiverMakerTest.java
@@ -24,6 +24,8 @@ import org.noise_planet.noisemodelling.jdbc.input.SceneDatabaseInputSettings;
 import org.noise_planet.noisemodelling.jdbc.input.SceneWithEmission;
 import org.noise_planet.noisemodelling.jdbc.output.NoiseMapWriter;
 import org.noise_planet.noisemodelling.jdbc.utils.CellIndex;
+import org.noise_planet.noisemodelling.jdbc.utils.IsoSurface;
+import org.noise_planet.noisemodelling.pathfinder.utils.profiler.RootProgressVisitor;
 import org.noise_planet.noisemodelling.propagation.AttenuationParameters;
 import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.GroundAbsorption;
@@ -31,6 +33,7 @@ import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -302,6 +305,97 @@ public class NoiseMapByReceiverMakerTest {
             // receiver is on the left of the source
             assertEquals(new Orientation(360-90, 0, 0), pathParameters.getRaySourceReceiverDirectivity());
 
+        }
+    }
+
+
+
+    @Test
+    public void testEmissionTrafficTable() throws SQLException {
+        try (Statement st = connection.createStatement()) {
+            st.execute(String.format("CALL SHPREAD('%s', 'ROADS_TRAFF')", NoiseMapByReceiverMakerTest.class.getResource("roads_traff.shp").getFile()));
+            st.execute("CREATE TABLE SOURCES_GEOM(PK SERIAL PRIMARY KEY, THE_GEOM GEOMETRY) AS SELECT PK, THE_GEOM FROM ROADS_TRAFF");
+            st.execute("CREATE TABLE SOURCES_EMISSION(PERIOD VARCHAR, IDSOURCE INT, TV REAL, HV REAL, LV_SPD REAL, HV_SPD REAL, PVMT VARCHAR)");
+            st.execute("INSERT INTO SOURCES_EMISSION SELECT 'D', PK, TV_D, HV_D, LV_SPD_D, HV_SPD_D, PVMT FROM ROADS_TRAFF");
+            st.execute("INSERT INTO SOURCES_EMISSION SELECT 'E', PK, TV_E, HV_E, LV_SPD_E, HV_SPD_E, PVMT FROM ROADS_TRAFF");
+            st.execute("INSERT INTO SOURCES_EMISSION SELECT 'N', PK, TV_N, HV_N, LV_SPD_N, HV_SPD_N, PVMT FROM ROADS_TRAFF");
+
+            st.execute(String.format("CALL SHPREAD('%s', 'BUILDINGS')", NoiseMapByReceiverMakerTest.class.getResource("buildings.shp").getFile()));
+
+            int srid = org.h2gis.utilities.GeometryTableUtilities.getSRID(connection, "BUILDINGS");
+            IsoSurface isoSurface = new IsoSurface(IsoSurface.NF31_133_ISO, srid);
+            // Generate delaunay triangulation
+            DelaunayReceiversMaker delaunayReceiversMaker = new DelaunayReceiversMaker("BUILDINGS", "ROADS_TRAFF");
+            delaunayReceiversMaker.setMaximumArea(800);
+            delaunayReceiversMaker.setGridDim(1);
+            delaunayReceiversMaker.run(connection, "RECEIVERS", isoSurface.getTriangleTable());
+
+            // Create noise map for 4 periods
[…]
