+++ Docs/Get_Started_Script.rst
@@ -9,7 +9,7 @@ In this tutorial, we describe the different ways to pilot NoiseModelling thanks 
 .. _release page : https://github.com/Ifsttar/NoiseModelling/releases/latest
 .. _NoiseModelling_4.0.0_without_gui : https://github.com/Ifsttar/NoiseModelling/releases/download/v4.0.0/NoiseModelling_4.0.0_without_gui.zip
 
-From that point, NoiseModelling can be executed in 3 different maners: 
+From that point, NoiseModelling can be executed in 3 different manners:
 
 #. with simple command lines
 #. with Bash script
@@ -33,6 +33,8 @@ Requirements
 1. Simple command line
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+Using the terminal of your operating system (Java 11 must be in the system path)
+
 Below is an example of a bash instruction, executing the ``Noise_level_from_traffic.groovy`` WPS Script (located in the directory ``/noisemodelling/wps/``). This block has 5 arguments corresponding to the input table names (for buildings, roads, receivers, dem and ground type).
 
 .. literalinclude:: scripts/nm_terminal.bash
@@ -50,7 +52,7 @@ Below is an example of a bash instruction, executing the ``Noise_level_from_traf
 2. Bash script
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Below is an example of a sequence of simple .groovy scripts, using bash instructions and launching the differents steps described in the ":doc:`Get_Started_GUI`".
+Below is an example of a sequence of simple .groovy scripts, using bash instructions and launching the steps described in the ":doc:`Get_Started_GUI`".
 
 .. literalinclude:: scripts/get_started_tutorial_simple.bash
    :language: bash
