+++ Docs/Dynamic_Tutorial.rst
@@ -1,4 +1,4 @@
-Dynamic Maps - groovy scripts
+Dynamic Maps
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Many publications have emerged showcasing the use of **NoiseModelling** to create dynamic maps (see `scientific production`_).
@@ -29,7 +29,7 @@ Dynamic mapping has its subtleties, and it's important to be aware of them to av
 - Gozalo, G. R., Aumond, P., & Can, A. (2020). Variability in sound power levels: Implications for static and dynamic traffic models. Transportation Research Part D: Transport and Environment, 84, 102339.
 - Le Bescond, V., Can, A., Aumond, P., & Gastineau, P. (2021). Open-source modeling chain for the dynamic assessment of road traffic noise exposure. Transportation Research Part D: Transport and Environment, 94, 102793.
 
-This is an **advanced tutorial**, so you may need to open the Groovy scripts to see what’s under the hood. Assumptions are freely made, specific formats are expected, and so on. To understand the required data formats and check the expected structure of the input tables, please refer also to the example input tables and spatial layers!
+Assumptions are freely made, specific formats are expected, and so on. To understand the required data formats and check the expected structure of the input tables, please refer also to the example input tables and spatial layers!
 
 Study Cases
 ---------------
@@ -37,76 +37,104 @@ Study Cases
 Case 1: A Road Network with a Single Traffic Flow
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. code-block:: groovy
+Import the road network (with arbitrary traffic flows) and buildings from an OSM file
+------------------------------------------------------------------------------------------------
 
-         // Import the road network (with predicted traffic flows) and buildings from an OSM file
-        new Import_OSM().exec(connection, [
-                "pathFile"      : TestImportExport.getResource("map.osm.gz").getPath(),
-                "targetSRID"    : 2154,
-                "ignoreGround"  : true,
-                "ignoreBuilding": false,
-                "ignoreRoads"   : false,
-                "removeTunnels" : true
-        ]);
+Use Import_OSM WPS block
 
-        // Create a receiver grid
-        new Regular_Grid().exec(connection,  ["buildingTableName": "BUILDINGS",
-                                              "sourcesTableName" : "ROADS",
-                                              "delta"            : 25])
+#. ``Path of the osm file``: Enter the path of the provided Open Street map file (can be relative to NoiseModelling): ``data_dir/data/wpsdata/map.osm.gz``
+#. ``Target projection identifier``: Enter the official France projection for this tutorial files ``2154``
+#. ``Remove tunnels``: Check it
+#. ``Do not import surface``: Check it as we will not use this output
 
-        // Set a height to the receivers at 1.5 m
-        new Set_Height().exec(connection,
-                [ "tableName":"RECEIVERS",
-                  "height": 1.5
-                ])
+.. figure:: images/tutorial/dynamic/ImportOSM.png
+   :align: center
 
-        // From the network with traffic flow to individual trajectories with associated Lw using the Probabilistic method
-        // This method place randomly the vehicles on the network according to the traffic flow
-        new Flow_2_Noisy_Vehicles().exec(connection,
-                ["tableRoads": "ROADS",
-                 "method": "PROBA",
-                 "timestep": 1,
-                 "gridStep" : 10,
-                 "duration" : 60])
+Create a receiver grid using 25 meters step in a grid pattern
+------------------------------------------------------------------------------------------------
 
[…]
