### 2025-02-20 fd8b386
reflect change of output tables in quick start tutorial
