### 2025-02-20 bd0cb47
update tutorials with new output
