+++ Docs/Get_Started_GUI.rst
@@ -183,3 +183,4 @@ By clicking on each of the inputs or outputs, you will find a lot of information
 
 .. figure:: images/tutorial/Tutorial1_ImageLast.gif
    :align: center
+
