+++ .github/workflows/static.yml
@@ -40,7 +40,7 @@ jobs:
         with:
           java-version: 11
       - name: Generate Javadoc
-        run: mvn javadoc:aggregate
+        run: mvn javadoc:aggregate -Ddoclint=none
       - name: Generate Cnossos report
         working-directory: noisemodelling-tutorial-01
         run: mvn exec:java -Dmain.class=org.noise_planet.nmtutorial01.GenerateReferenceDeviation -Dexec.args="../Docs"
