+++ Docs/Dynamic_Tutorial.rst
@@ -58,18 +58,10 @@ Use Regular_Grid WPS block
 #. ``Table bounding box name``: Enter ``ROADS`` The receivers will use  the envelope of the ROADS table.
 #. ``Offset``: Enter ``15`` for 15 meters distance
 #. ``Output triangle table``: Check it in order to be able to generate the iso contours
+#. ``height``: Enter ``1.5``
 
 .. figure:: images/tutorial/dynamic/RegularGrid.png
    :align: center
-
-
-Set all receivers at 1.5 meters height
-------------------------------------------------------------------------------------------------
-
-Use Set_Height WPS block
-
-#. ``Table``: Enter ``RECEIVERS``
-#. ``Height``: Enter ``1.5``
 
 Convert traffic to dynamic traffic flow
 ------------------------------------------------------------------------------------------------
@@ -88,7 +80,7 @@ Use the ``Dynamic:Flow_2_Noisy_Vehicles`` WPS block:
 #. ``duration``: Enter ``60``
 #. ``gridStep``: Enter ``10``
 
-Compute the attenuation for each receiver-source points
+Compute noise level at receivers points for each receiver-period
 ------------------------------------------------------------------------------------------------
 
 Use the ``NoiseModelling:Noise_level_from_source`` WPS block
@@ -98,15 +90,14 @@ Use the ``NoiseModelling:Noise_level_from_source`` WPS block
 #. ``Source emission table name``: Enter ``SOURCES_EMISSION`` Contain for each source index and period the noise emission
 #. ``Receivers table name``: Enter ``RECEIVERS``
 #. ``Max Error (dB)``: Enter ``3`` Will skip further sources, reduce the computation time for this tutorial
-#. ``Maximum source receiver distance``: Enter ``150``
+#. ``Maximum source receiver distance``: Enter ``800``
 #. ``Diffraction on horizontal edges``: Check it
 #. ``Order of reflexion``: Enter ``0``
-#. ``Separate receiver level by source identifier``: Check it to have the SOURCEID column on the output
 
 Compute noise indicators
 ------------------------------------------------------------------------------------------------
 
-This step is optional, it compute the LA10, LA50 and LA90 at each receiver from the table LT_GEOM
+This step is optional, it compute the LA10, LA50 and LA90 at each receiver from the table RECEIVERS_LEVEL
 
 Use the ``Acoustic_Tools:DynamicIndicators`` wps block
 
@@ -122,155 +113,237 @@ Generate a dynamic iso-contour map for each time period based on the LAEQ of the
 Use the ``Acoustic_Tools:Create_Isosurface`` wps block
 
 #. ``Sound levels table``: Enter ``RECEIVERS_LEVEL``
+#. ``Smooth coefficient``: Enter ``0``
 
-In QGis in time window
+Export Map to QGis
+------------------------------------------------------------------------------------------------
+
+Using ``Export_Table`` block export the following tables as files in any folder.
+
+#. ``CONTOURING_NOISE_MAP``
+#. ``BUILDINGS``
+#. ``ROADS``
+
+Configure QGis to display time dependant map
+------------------------------------------------------------------------------------------------
+
+Load the 3 files in QGIS. ``Contouring_noise_map`` must be ordered as the last layer (rendered in the bottom)
+
+Load the style for contouring noise map:
+
+
+.. figure:: images/tutorial/dynamic/loadstyle.png
+   :align: center
+   :scale: 50 %
+   :alt: Load style menu in QGis
+
+Load the style located in the NoiseModelling folder ``Docs/styles/style_beate_tomio.sld``
+
+In QGis in time window paste the following formulae:
 
[…]
