### 2025-02-26 398566d
Create new wps script, to be able to generate default settings and edit the temperature and other settings for each time period.
