+++ .github/workflows/static.yml
@@ -41,6 +41,8 @@ jobs:
           java-version: 11
       - name: Generate Javadoc
         run: mvn javadoc:aggregate -Ddoclint=none
+      - name: Build NoiseModelling
+        run: mvn install -DskipTests
       - name: Generate Cnossos report
         working-directory: noisemodelling-tutorial-01
         run: mvn exec:java -Dmain.class=org.noise_planet.nmtutorial01.GenerateReferenceDeviation -Dexec.args="../Docs"
