### 2025-02-26 bc2c6ad
define sources_emission table
