### 2025-02-27 bbfff2e
docs emission table doc
