+++ Docs/Input_roads.rst
@@ -8,88 +8,158 @@ Below we describe the table ``ROADS``, dealing with the roads network.
 The other tables are accessible via the left menu in the ``Input tables & parameters`` section.
 
 .. figure:: images/Input_tables/roads_banner.png
-	:align: center
+    :align: center
 
-Table definition
+Table SOURCES_GEOM definition
 ---------------------
 
 .. warning::
-	* In the list below, the columns noted with ``*`` are mandatory
-	* This description is only valid for ``Noise_level_from_traffic`` and ``Road_Emission_from_Traffic`` WPS scripts. For the other WPS scripts, it is necessary to refer to the description of their input data
+    * In the list below, the columns noted with ``*`` are mandatory
+    * This description is only valid for ``Noise_level_from_traffic`` and ``Road_Emission_from_Traffic`` WPS scripts. For the other WPS scripts, it is necessary to refer to the description of their input data
 
 
 .. note::
-	In the list below, some columns are suffixed with the letters ``D``, ``E`` and ``N``. This correspond to ``Day`` (6-18h), ``Evening`` (18-22h) and ``Night`` (22-6h) periods. A column is expected for each of them.
+    In the list below, some columns are suffixed with the letters ``D``, ``E`` and ``N``. This correspond to ``Day`` (6-18h), ``Evening`` (18-22h) and ``Night`` (22-6h) periods. A column is expected for each of them.
 
 * ``THE_GEOM`` *
-	* Description: Geometry of the roads (``LINESTRING`` or ``MULTILINESTRING``)
-	* Type: Geometry
+    * Description: Geometry of the roads (``LINESTRING`` or ``MULTILINESTRING``)
+    * Type: Geometry
 * ``PK`` *
-	* Description: An identifier (PRIMARY KEY)
-	* Type:  Integer 
+    * Description: An identifier (PRIMARY KEY)
+    * Type:  Integer
 * ``LV_D``, ``LV_E``, ``LV_N``
-	* Description: Hourly average light vehicle count
-	* Type: Double
+    * Description: Hourly average light vehicle count
+    * Type: Double
 * ``MV_D``, ``MV_E``, ``MV_N``
-	* Description: Hourly average medium heavy vehicles, delivery vans > 3.5 tons, buses, touring cars, *etc.* with two axles and twin tyre mounting on rear axle count
-	* Type: Double
+    * Description: Hourly average medium heavy vehicles, delivery vans > 3.5 tons, buses, touring cars, *etc.* with two axles and twin tyre mounting on rear axle count
+    * Type: Double
 * ``HGV_D``, ``HGV_E``, ``HGV_N``
-	* Description: Hourly average heavy duty vehicles, touring cars, buses, with three or more axles count
-	* Type: Double
+    * Description: Hourly average heavy duty vehicles, touring cars, buses, with three or more axles count
+    * Type: Double
 * ``WAV_D``, ``WAV_E``, ``WAV_N``
-	* Description: Hourly average mopeds, tricycles or quads ≤ 50 cc count
-	* Type: Double
+    * Description: Hourly average mopeds, tricycles or quads ≤ 50 cc count
+    * Type: Double
 * ``WBV_D``, ``WBV_E``, ``WBV_N``
-	* Description: Hourly average motorcycles, tricycles or quads > 50 cc count
-	* Type: Double
+    * Description: Hourly average motorcycles, tricycles or quads > 50 cc count
+    * Type: Double
 * ``LV_SPD_D``, ``LV_SPD_E``, ``LV_SPD_N``
-	* Description: Hourly average light vehicle speed *(km/h)*
-	* Type: Double
[…]
