+++ Docs/Input_source.rst
@@ -11,15 +11,18 @@ SOURCES_GEOM Table definition
 ---------------------
 
 * ``PK`` *
-	* Description: An identifier (must be a PRIMARY KEY). The field name is not important, it could be IDSOURCE.
-	* Type:  Integer
+    * Description: An identifier (must be a PRIMARY KEY). The field name is not important, it could be IDSOURCE.
+    * Type:  Integer
 * ``THE_GEOM`` *
-	* Description: Geometry of the source (``LINESTRING`` , ``MULTILINESTRING`` or ``POINT``). The geometry **must** have a Z attribute.
-	* Type: Geometry
+    * Description: Geometry of the source (``LINESTRING`` , ``MULTILINESTRING`` or ``POINT``). The geometry **must** have a Z attribute.
+    * Type: Geometry
 
 .. warning::
-	* ``Zobject`` = ``Zdem + Zsource``
-	* If there is no DEM, the altitude will be equal to geometry Z attribute
+    * ``Zobject`` = ``Zdem + Zsource``
+    * If there is no DEM, the altitude will be equal to geometry Z attribute
+
+.. note::
+    For backward compatibility in this table you can define here the noise level for Day Evening and Night, for 500 Hz the expected field is ``LWD500`` ``LWE500`` ``LWN500``. NoiseModelling will output the level for this periods and also compute the global ``DEN`` period levels according to standard.
 
 If you provide only the geometry NoiseModelling will compute only the attenuation on the output table ``RECEIVERS_LEVEL``.
 
@@ -29,36 +32,36 @@ SOURCES_EMISSION Table definition
 ---------------------
 
 * ``IDSOURCE`` *
-	* Description: An identifier linked to the primary key of the ``SOURCES_GEOM`` table.
-	* Type:  Integer
+    * Description: An identifier linked to the primary key of the ``SOURCES_GEOM`` table.
+    * Type:  Integer
 * ``PERIOD`` *
-	* Description: Identifier of the time. ex. `8h00-9h00`
-	* Type:  String
+    * Description: Identifier of the time. ex. `8h00-9h00`
+    * Type:  String
 * ``LW63``
-	* Description: emission levels in dB for 63 Hz
-	* Type: Double
+    * Description: emission levels in dB for 63 Hz
+    * Type: Double
 * ``LW125``
-	* Description: emission levels in dB for 125 Hz
-	* Type: Double
+    * Description: emission levels in dB for 125 Hz
+    * Type: Double
 * ``LW250``
-	* Description: emission levels in dB for 250 Hz
-	* Type: Double
+    * Description: emission levels in dB for 250 Hz
+    * Type: Double
 * ``LW500``
-	* Description: emission levels in dB for 500 Hz
-	* Type: Double
+    * Description: emission levels in dB for 500 Hz
+    * Type: Double
 * ``LW1000``
-	* Description: emission levels in dB for 1000 Hz
-	* Type: Double
+    * Description: emission levels in dB for 1000 Hz
+    * Type: Double
 * ``LW2000``
-	* Description: emission levels in dB for 2000 Hz
-	* Type: Double
+    * Description: emission levels in dB for 2000 Hz
+    * Type: Double
 * ``LW4000``
-	* Description: emission levels in dB for 4000 Hz
-	* Type: Double
+    * Description: emission levels in dB for 4000 Hz
+    * Type: Double
 * ``LW8000``
-	* Description: emission levels in dB for 8000 Hz
-	* Type: Double
+    * Description: emission levels in dB for 8000 Hz
+    * Type: Double
 
 .. note::
[…]
