### 2025-02-27 5a575d1
documentation on maxerrordb
