+++ Docs/Noise_Map_From_Point_Source.rst
@@ -164,9 +164,6 @@ Use the ``NoiseModelling:Noise_level_from_source`` WPS script. Fill the three fo
 * ``Receivers table name`` : ``RECEIVERS``
 * ``Buildings table name`` : ``BUILDINGS``
 
-.. warning::
-   For this example, since we only added information for noise level during the day (field ``LWD500``), we have to skip the noise level calculation for AttenuatedPaths, LNIGHT and LEVENING. To do so, check the boxes for ``Do not compute LDEN_GEOM``, ``Do not compute LEVENING_GEOM`` and ``Do not compute LNIGHT_GEOM`` options.
-
 Once ready, click on ``Run Process`` button.
 
 You should then have this message: ``Calculation Done ! RECEIVERS_LEVEL table(s) have been created.``
