### 2025-02-27 e448254
fix concurrent modification of ArrayList
