+++ .github/workflows/CI.yml
@@ -37,7 +37,7 @@ jobs:
         env:
           MAVEN_USERNAME: ${{ secrets.OSSRH_USERNAME }}
           MAVEN_PASSWORD: ${{ secrets.OSSRH_TOKEN }}
-        if: env.MAVEN_USERNAME != null
+        if: ${{ (env.MAVEN_USERNAME != null) && (github.ref == 'refs/heads/main') }}
         run: mvn clean deploy -B -P maven-deploy
       - name: Clean
         run: rm -f $HOME/.gradle/caches/modules-2/modules-2.lock
