### 2025-03-05 b331ddf
test new method dbError
