+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/AttenuationOutputSingleThread.java
@@ -10,6 +10,7 @@
 package org.noise_planet.noisemodelling.jdbc.output;
 
 import org.h2gis.api.ProgressVisitor;
+import org.locationtech.jts.geom.Coordinate;
 import org.noise_planet.noisemodelling.jdbc.EmissionTableGenerator;
 import org.noise_planet.noisemodelling.jdbc.NoiseMapDatabaseParameters;
 import org.noise_planet.noisemodelling.jdbc.input.SceneDatabaseInputSettings;
@@ -52,9 +53,9 @@ public class AttenuationOutputSingleThread implements CutPlaneVisitor {
 
     /**
      * MaxError DB Processing variable
-     * Current, minimal power at receiver (minimal among time periods), only used to stop looking for far sources
+     * Current, power at receiver, only used to stop looking for far sources
      */
-    double[] wjAtReceiver = new double[0];
+    Map<String, Double> wjAtReceiver = new HashMap<>();
 
     /**
      * MaxError DB Processing variable
@@ -62,12 +63,14 @@ public class AttenuationOutputSingleThread implements CutPlaneVisitor {
      * Key source index
      * Value maximum expected noise level in w
      */
-    Map<Integer, Double> maximumWjExpectedSplAtReceiver = new HashMap<>();
+    Map<String, HashMap<Coordinate, Double>> maximumWjExpectedSplAtReceiver = new HashMap<>();
+
+
     /**
      * MaxError DB Processing variable
      * Next Free Field cumulated global power at receiver, only used to stop looking for far sources
      */
-    double sumMaximumRemainingWjExpectedSplAtReceiver = 0;
+    Map<String, Double> sumMaximumRemainingWjExpectedSplAtReceiver = new HashMap<>();
 
     public AtomicInteger cutProfileCount = new AtomicInteger(0);
 
@@ -82,14 +85,6 @@ public class AttenuationOutputSingleThread implements CutPlaneVisitor {
         this.multiThread = multiThreadParent;
         this.dbSettings = multiThreadParent.noiseMapDatabaseParameters;
         this.progressVisitor = progressVisitor;
-    }
-
-    private void addGlobalReceiverLevel(double[] wjLevel) {
-        if(wjAtReceiver.length != wjLevel.length) {
-            wjAtReceiver = wjLevel.clone();
-        } else {
-            wjAtReceiver = AcousticIndicatorsFunctions.sumArray(wjAtReceiver, wjLevel);
-        }
     }
 
     /**
@@ -187,8 +182,6 @@ public class AttenuationOutputSingleThread implements CutPlaneVisitor {
                 double[] defaultAttenuation = new double[0];
                 if(scene.wjSources.containsKey(sourcePk)) {
                     ArrayList<SceneWithEmission.PeriodEmission> emissions = scene.wjSources.get(sourcePk);
-                    double[] minimalPowerForAllPeriods = new double[0];
-                    double minimalPowerSum = Double.MAX_VALUE;
                     for (SceneWithEmission.PeriodEmission periodEmission : emissions) {
                         String period = periodEmission.period;
                         double [] attenuation = new double[0];
@@ -211,31 +204,48 @@ public class AttenuationOutputSingleThread implements CutPlaneVisitor {
                         processNoiseLevel(receiverNoiseLevel);
                         if(dbSettings.maximumError > 0) {
                             double powerSum = sumArray(levels);
-                            if(powerSum < minimalPowerSum) {
-                                minimalPowerSum = powerSum;
-                                minimalPowerForAllPeriods = levels;
-                            }
+                            wjAtReceiver.merge(period, powerSum, Double::sum);
                         }
-                    }
-                    if(dbSettings.maximumError > 0) {
-                        // Add minimal power
-                        addGlobalReceiverLevel(minimalPowerForAllPeriods);
                     }
[…]
