+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/SceneWithEmissionTest.java
@@ -35,7 +35,7 @@ import java.sql.Statement;
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.sumArray;
+import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicatorsFunctions.*;
 
 /**
  * Test class evaluation and testing attenuation values.
@@ -43,21 +43,30 @@ import static org.noise_planet.noisemodelling.pathfinder.utils.AcousticIndicator
 public class SceneWithEmissionTest {
     private static final double HUMIDITY = 70;
     private static final double TEMPERATURE = 10;
-
     private List<Long> testIgnoreNonSignificantSourcesParam(Connection connection, double maxError) throws SQLException, IOException {
+        return testIgnoreNonSignificantSourcesParam(connection, maxError, "BUILDINGS",
+                "LW_ROADS", "RECEIVERS", "");
+    }
+
+    private List<Long> testIgnoreNonSignificantSourcesParam(
+            Connection connection, double maxError, String buildingsTableName, String sourcesTableName,
+            String receiverTableName, String sourcesEmissionTableName) throws SQLException, IOException {
+
         // Init NoiseModelling
-        NoiseMapByReceiverMaker noiseMap = new NoiseMapByReceiverMaker("BUILDINGS",
-                "LW_ROADS", "RECEIVERS");
+        NoiseMapByReceiverMaker noiseMap = new NoiseMapByReceiverMaker(buildingsTableName,
+                sourcesTableName, receiverTableName);
 
         noiseMap.setMaximumPropagationDistance(5000.0);
         noiseMap.setSoundReflectionOrder(1);
         noiseMap.setThreadCount(1);
         noiseMap.setComputeHorizontalDiffraction(true);
         noiseMap.setComputeVerticalDiffraction(true);
+        if(!sourcesEmissionTableName.isEmpty()) {
+            noiseMap.setSourcesEmissionTableName(sourcesEmissionTableName);
+        }
         NoiseMapDatabaseParameters parameters = noiseMap.getNoiseMapDatabaseParameters();
         parameters.mergeSources = false;
         parameters.maximumError = maxError;
-        noiseMap.setInputMode(SceneDatabaseInputSettings.INPUT_MODE.INPUT_MODE_LW_DEN);
 
         // Building height field name
         noiseMap.setHeightField("HEIGHT");
@@ -94,6 +103,9 @@ public class SceneWithEmissionTest {
                                      "testReceiverOverBuilding", true, ""))) {
             try (Statement st = connection.createStatement()) {
                 st.execute(Utils.getRunScriptRes("scenario_skip_far_source.sql"));
+
+
+
                 List<Long> allSourcesPk = testIgnoreNonSignificantSourcesParam(connection, 0.);
                 List<Long> ignoreFarSourcesPk = testIgnoreNonSignificantSourcesParam(connection, maxError);
                 assertEquals(2, allSourcesPk.size());
@@ -102,6 +114,52 @@ public class SceneWithEmissionTest {
         }
     }
 
+
+    /**
+     * Test optimisation feature {@link NoiseMapDatabaseParameters#setMaximumError(double)}
+     * This feature is disabled and all sound sources are computed
+     */
+    @Test
+    public void testIgnoreNonSignificantSources2() throws Exception {
+        final double maxError = 3.0;
+
+        try (Connection connection =
+                     JDBCUtilities.wrapConnection(
+                             H2GISDBFactory.createSpatialDataBase(
+                                     "testIgnoreNonSignificantSources2", true, ""))) {
+            try (Statement st = connection.createStatement()) {
[…]
