+++ .github/workflows/CI.yml
@@ -56,5 +56,5 @@ jobs:
       - name: Archive production artifacts
         uses: actions/upload-artifact@v4
         with:
-          name:  NoiseModelling_without_gui.zip
+          name: NoiseModelling_without_gui
           path: wps_scripts/scriptrunner/
