### 2025-03-05 6c48941
test all periods in maxDbError unit test
