### 2025-03-05 0605c11
replace LW column by HZ for frequencies. Add a parameter frequency field name in order to be able to work with old files. TODO redo screenshots/input data in tutorials
