+++ noisemodelling-jdbc/src/test/java/org/noise_planet/noisemodelling/jdbc/NoiseMapByReceiverMakerTest.java
@@ -36,7 +36,6 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.noise_planet.noisemodelling.jdbc.Utils.getRunScriptRes;
@@ -69,7 +68,7 @@ public class NoiseMapByReceiverMakerTest {
             NoiseMapByReceiverMaker noiseMapByReceiverMaker = new NoiseMapByReceiverMaker("BUILDINGS", "ROADS_GEOM", "RECEIVERS");
             noiseMapByReceiverMaker.setHeightField("HEIGHT");
             noiseMapByReceiverMaker.setSoilTableName("LAND_G");
-            noiseMapByReceiverMaker.setLwFrequencyPrepend("DB_M");
+            noiseMapByReceiverMaker.setFrequencyFieldPrepend("DB_M");
             noiseMapByReceiverMaker.initialize(connection, new EmptyProgressVisitor());
 
             Set<Long> processedReceivers = new HashSet<>();
@@ -105,7 +104,7 @@ public class NoiseMapByReceiverMakerTest {
         AttenuationParameters data = new AttenuationParameters(false);
         for(String period : new String[] {"D", "E", "N"}) {
             for (int freq : data.getFrequencies()) {
-                String fieldName = "LW" + period + freq;
+                String fieldName = "HZ" + period + freq;
                 sb.append(", ");
                 sb.append(fieldName);
                 sb.append(" real");
@@ -149,7 +148,7 @@ public class NoiseMapByReceiverMakerTest {
 
             NoiseMapDatabaseParameters parameters = noiseMapByReceiverMaker.getNoiseMapDatabaseParameters();
 
-            try(ResultSet rs = st.executeQuery("SELECT LW63 FROM " + parameters.receiversLevelTable + " WHERE PERIOD='DEN' ORDER BY IDRECEIVER")) {
+            try(ResultSet rs = st.executeQuery("SELECT HZ63 FROM " + parameters.receiversLevelTable + " WHERE PERIOD='DEN' ORDER BY IDRECEIVER")) {
                 assertTrue(rs.next());
                 assertEquals(73.3, rs.getDouble(1), 0.1);
                 assertTrue(rs.next());
@@ -204,7 +203,7 @@ public class NoiseMapByReceiverMakerTest {
 
             NoiseMapDatabaseParameters parameters = noiseMapByReceiverMaker.getNoiseMapDatabaseParameters();
 
-            try(ResultSet rs = st.executeQuery("SELECT IDRECEIVER, LW63 FROM " + parameters.receiversLevelTable + " WHERE PERIOD='DEN' ORDER BY IDRECEIVER")) {
+            try(ResultSet rs = st.executeQuery("SELECT IDRECEIVER, HZ63 FROM " + parameters.receiversLevelTable + " WHERE PERIOD='DEN' ORDER BY IDRECEIVER")) {
                 assertTrue(rs.next());
                 assertEquals(1, rs.getInt(1));
                 assertEquals(68.3, rs.getDouble(2), 1);
@@ -360,7 +359,7 @@ public class NoiseMapByReceiverMakerTest {
         try (Statement st = connection.createStatement()) {
[…]
