### 2025-03-06 de1b2a2
about #297 explain range of isocontour
