### 2025-03-06 94ccd66
update documenation text and data with HZ frequency column name instead of LW. QGIS screenshots has not been updated.
