+++ Docs/requirements.txt
@@ -20,7 +20,7 @@ idna==3.7
     # via requests
 imagesize==1.4.1
     # via sphinx
-jinja2==3.1.5
+jinja2==3.1.6
     # via sphinx
 markupsafe==2.1.1
     # via jinja2
