### 2025-03-06 d4746da
fix titles in docs dynamic
