+++ Docs/Dynamic_Tutorial.rst
@@ -1,4 +1,4 @@
-Dynamic Maps
+Dynamic Tutorial - GUI
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Many publications have emerged showcasing the use of **NoiseModelling** to create dynamic maps (see `scientific production`_).
@@ -31,14 +31,11 @@ Dynamic mapping has its subtleties, and it's important to be aware of them to av
 
 Assumptions are freely made, specific formats are expected, and so on. To understand the required data formats and check the expected structure of the input tables, please refer also to the example input tables and spatial layers!
 
-Study Cases
----------------
-
 Case 1: A Road Network with a Single Traffic Flow
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+----------------------------------------------------
 
 Import the road network (with arbitrary traffic flows) and buildings from an OSM file
-------------------------------------------------------------------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Use Import_OSM WPS block
 
@@ -51,7 +48,7 @@ Use Import_OSM WPS block
    :align: center
 
 Create a receiver grid using 25 meters step in a grid pattern
-------------------------------------------------------------------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Use Regular_Grid WPS block
 
@@ -64,7 +61,7 @@ Use Regular_Grid WPS block
    :align: center
 
 Convert traffic to dynamic traffic flow
-------------------------------------------------------------------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 From the network with traffic flow to individual trajectories with associated Lw
 
@@ -81,7 +78,7 @@ Use the ``Dynamic:Flow_2_Noisy_Vehicles`` WPS block:
 #. ``gridStep``: Enter ``10``
 
 Compute noise level at receivers points for each receiver-period
-------------------------------------------------------------------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Use the ``NoiseModelling:Noise_level_from_source`` WPS block
 
@@ -95,7 +92,7 @@ Use the ``NoiseModelling:Noise_level_from_source`` WPS block
 #. ``Order of reflexion``: Enter ``0``
 
 Compute noise indicators
-------------------------------------------------------------------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 This step is optional, it compute the LA10, LA50 and LA90 at each receiver from the table RECEIVERS_LEVEL
 
@@ -106,7 +103,7 @@ Use the ``Acoustic_Tools:DynamicIndicators`` wps block
 
 
 Compute iso-surfaces for each time period
-------------------------------------------------------------------------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Generate a dynamic iso-contour map for each time period based on the LAEQ of the receivers.
 
@@ -116,7 +113,7 @@ Use the ``Acoustic_Tools:Create_Isosurface`` wps block
 #. ``Smooth coefficient``: Enter ``0``
 
 Export Map to QGis
[…]
