### 2025-03-12 67b49bb
publish wps scripts as maven library
