### 2025-03-14 cbafc3d
do not use timeout on end of processing, print stack size at the end.
