### 2025-03-18 afa4fcb
Add text in generator
