### 2025-03-18 f9c7a91
enable file logger on noisemodelling_without_gui fix demo scripts
