### 2025-03-18 7470643
Adding scientific production
