### 2025-03-19 8890a12
improve scriptrunner, logger file is in workspace dir now read maven pom in gradle project for the project version add release action workflow
