+++ .github/workflows/CI.yml
@@ -21,20 +21,17 @@ jobs:
       # Checkout the source code of the project
       - name: Checkout
         uses: actions/checkout@v4
-      # Install the GPG secret key
-      - name: Import GPG Key
-        run: |
-          echo "${{ secrets.GPG_SECRET_KEY }}" | gpg --import --no-tty --batch --yes
-        env:
-          GPG_SECRET_KEY: ${{ secrets.GPG_SECRET_KEY }}
-        if: ${{ (env.GPG_SECRET_KEY != null) && (github.ref == 'refs/heads/main') }}
-      - name: Set up JDK 11
-        uses: actions/setup-java@v1
+      # Setup the jdk using version 11 of Adoptium Temurin
+      - name: Setup java 11 using Adoptium Temurin
+        uses: actions/setup-java@v4
         with:
-          java-version: 11
+          distribution: 'temurin'
+          java-version: '11'
           server-id: ossrh
           server-username: MAVEN_USERNAME
           server-password: MAVEN_PASSWORD
+          gpg-private-key: ${{ secrets.GPG_SECRET_KEY }}
+          gpg-passphrase: MAVEN_GPG_PASSPHRASE
       - name: Building
         run: mvn test install -B
       - name: Setup Gradle
@@ -46,6 +43,7 @@ jobs:
         env:
           MAVEN_USERNAME: ${{ secrets.OSSRH_USERNAME }}
           MAVEN_PASSWORD: ${{ secrets.OSSRH_TOKEN }}
+          MAVEN_GPG_PASSPHRASE: ${{ secrets.GPG_PASSPHRASE }}
         if: ${{ (env.MAVEN_USERNAME != null) && (github.ref == 'refs/heads/main') }}
         run: mvn clean deploy -B -P maven-deploy
       - name: Clean
