### 2025-03-19 419b88d
fix gpg on wps gradle
