### 2025-03-20 57ca555
escape incompatible character javadoc
