### 2025-03-20 da70db5
escape incompatible character javadoc again
