+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/PathFinderTest.java
@@ -1036,7 +1036,7 @@ public class PathFinderTest {
     /**
      * TC19 - Complex object and 2 barriers on ground with spatially varying heights and
      * acoustic properties:
-     * erreur Cnossos: left path -> gPath table 207
+     * erreur Cnossos: left path -&gt; gPath table 207
      */
     @Test
     public void TC19() throws Exception {
