+++ .github/workflows/CI_Release.yml
@@ -39,6 +39,13 @@ jobs:
       - name: Build test
         run: mvn -ntp clean validate package test javadoc:test-javadoc javadoc:jar
 
+      - name: Install xmlstarlet (if needed)
+        run: sudo apt-get update && sudo apt-get install -y xmlstarlet
+      - name: Update version in build.gradle
+        run: |
+          bash update_version.sh
+          git commit -am "Update groovy wps version"
+        working-directory: wps_scripts
       # Create the release :
       #   - move from Snapshot version to Release
       #   - commit and tag release
@@ -62,11 +69,6 @@ jobs:
       # Export the last git tag into env.
       - name: Export env values
         run: echo "GIT_TAG=$(git describe --tags `git rev-list --tags --max-count=1`)" >> $GITHUB_ENV
-      - name: Setup Gradle
-        uses: gradle/actions/setup-gradle@v4
-      - name: Build and test WPS with Gradle
-        run:  ./gradlew build --info --stacktrace
-        working-directory: ./wps_scripts
       # Make the Github release from the last created tag. Write in its body the content of the changelog file.
       - name: Make Github release
         uses: ncipollo/release-action@v1
