### 2025-03-21 2323a92
Update JOSS branch, sync with main (#745)
* restore postgis test in github workflow

* fix npe

* fix compute deltaE ok <= TC11

* Update conf.py

* Update .readthedocs.yaml

* Add files via upload

* Update .readthedocs.yaml

* Update conf.py

* Bump pygments from 2.12.0 to 2.15.0 in /Docs

Bumps [pygments](https://github.com/pygments/pygments) from 2.12.0 to 2.15.0.
- [Release notes](https://github.com/pygments/pygments/releases)
- [Changelog](https://github.com/pygments/pygments/blob/master/CHANGES)
- [Commits](https://github.com/pygments/pygments/compare/2.12.0...2.15.0)

---
updated-depen
[…]
