+++ .github/workflows/CI.yml
@@ -56,10 +56,11 @@ jobs:
           MAVEN_GPG_PASSPHRASE: ${{ secrets.GPG_PASSPHRASE }}
         shell: bash
         run: |
-          echo "ossrhUsername=${MAVEN_USERNAME}" >> gradle.properties
-          echo "ossrhPassword=${MAVEN_PASSWORD}" >> gradle.properties
-          echo "signing.gnupg.passphrase=${MAVEN_GPG_PASSPHRASE}" >> gradle.properties
-        working-directory: ~/.gradle/
+          mkdir -p ~/.gradle/
+          echo "GRADLE_USER_HOME=${HOME}/.gradle" >> $GITHUB_ENV
+          echo "ossrhUsername=${MAVEN_USERNAME}" >> ~/.gradle/gradle.properties
+          echo "ossrhPassword=${MAVEN_PASSWORD}" >> ~/.gradle/gradle.properties
+          echo "signing.gnupg.passphrase=${MAVEN_GPG_PASSPHRASE}" >> ~/.gradle/gradle.properties
       - name: Deploy WPS
         if: ${{ (env.MAVEN_USERNAME != null) && (github.ref == 'refs/heads/main') }}
         run: ./gradlew uploadArchives
