### 2025-03-21 963037d
Update DefaultCutPlaneProcessing.java
Shutdown thread pool. NoiseModelling was hanging because one thread could avoid the jvm to stop.
