### 2025-04-17 4d4e304
train source position, read json configuration files (can be exterior file)
