+++ Docs/Data_Assimilation_Tutorial.rst
@@ -1,0 +1,224 @@
+Data Assimilation
+^^^^^^^^^^^^^^^^^^^^
+
+Introduction
+~~~~~~~~~~~~~~~
+
+In this tutorial, we’ll explore how to model and simulated noise in Geneva using static (road network, buildings) and dynamic (traffic, temperature) data.
+The main objective is to combine measurements from sensors located in Geneva with acoustic simulations to identify traffic configurations that best match measurements.
+
+Prerequisites
+~~~~~~~~~~~~~~~~~
+
+- You need to have a working installation of NoiseModelling (NM).
+- A folder named ``devices_data`` containing sensor measurement files in ``CSV`` format. Each file represents the measurements of a single sensor.
+- A ``CSV`` file named ``device_mapping_sf.csv`` containing the geometry and unique identifier of all sensors.
+- The the OpenStreetMap (OSM) data of the given area : ``geneva.osm.pbf``.
+
+The data
+~~~~~~~~~~~~~~~
+
+Each CSV file in the folder ``devices_data`` contains environmental noise measurements captured by individual sensors.
+The key columns are:
+  * ``deveui`` : Unique identifier of the sensor.
+  * ``epoch`` : Timestamp in epoch format (Unix time, ex:1724567400), representing the time of measurement.
+  * ``timestamp`` : Time in format <b>"%Y-%m-%d %H:%M:%S"</b>, representing the time of measurement.
+  * ``Leq`` : Equivalent continuous sound level in dB(A), calculated over a period (15 min).
+  * ``Temp`` : Temperature (°C) recorded by the sensor at the time of measurement.
+The ``device_mapping_sf.csv`` columns ka y are:
+  * ``deveui`` : Unique identifier of the sensor.
+  * ``The_GEOM`` : 3D point geometry in WKT (Well-Known Text) format — includes coordinates (X, Y) and altitude (Z) in the projected coordinate system.
+
+All values are sampled approximately every 15 minutes, but the exact spacing may vary slightly.
+Timestamps should therefore be aligned at fixed 15-minute intervals.
+
+Data Simulation
+-----------------
+This process prepares a training dataset from sensor measurements, importing it into a spatial database, and performing various calculations to determine noise levels.
+Maps are generated with all possible combinations, in order to identify the best road configuration in terms of noise levels generated.
+
+Step 1 : Generate all possible Configurations
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Generates all possible combinations of values from two given lists and inserts them into a sql table named ``ALL_CONFIGURATIONS``
+The generated combinations include variation (%) around standard values for type of roads primary, secondary, tertiary, others, and temperature.
+
+.. code-block:: groovy
+    new All_Possible_Configuration().exec(connection,[
+                   "trafficValues": [0.01,1.0, 2.0,3],
+                   "temperatureValues": [10,15,20]
+    ])
+
+
+Step 2 : Preparation of Sensor data
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+This part extracts sensor data for a given period.Then use ``trainingRation`` (%) for the training data which will represent the receivers.
+So at the end of this part, the ``SENSORS_MEASUREMENTS`` (representing all the data for this period), ``OBSERVATION`` (training data set) and ``RECEIVERS`` sql tables will be created.
[…]
