### 2025-04-29 #7500 @faroit
@kyleniemeyer While this sounds very interesting, I have very little expertise in JAVA, this isn't in my domain of expertise and I have still 4 submissions ongoing. Let me go through the list of unassigned papers this week and check if there is something closer to my domain that I can review instead. I can still offer my duties as a fallback if there is no other editor available within one week. Does that sound ok?
