### 2025-05-07 2f3e2c3
fix railway iterator, npe when traffic = 0
