### 2025-05-12 e57ae5d
fix: update gh workflows with java 21
