+++ .github/workflows/static.yml
@@ -35,10 +35,10 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      - name: Set up JDK 11
+      - name: Set up JDK 21
         uses: actions/setup-java@v1
         with:
-          java-version: 11
+          java-version: 21
       - name: Generate Javadoc
         run: mvn javadoc:aggregate -Ddoclint=none
       - name: Build NoiseModelling
