+++ .github/workflows/CI_Release.yml
@@ -17,12 +17,12 @@ jobs:
       # Checkout the source code of the project
       - name: Checkout
         uses: actions/checkout@v4
-      # Setup the jdk using version 11 of Adoptium Temurin
-      - name: Setup java 11 using Adoptium Temurin
+      # Setup the jdk using version 21 of Adoptium Temurin
+      - name: Setup java 21 using Adoptium Temurin
         uses: actions/setup-java@v4
         with:
           distribution: 'temurin'
-          java-version: '11'
+          java-version: '21'
           server-id: ossrh
           server-username: MAVEN_USERNAME
           server-password: MAVEN_PASSWORD
