+++ .github/workflows/CI.yml
@@ -21,8 +21,8 @@ jobs:
       # Checkout the source code of the project
       - name: Checkout
         uses: actions/checkout@v4
-      # Setup the jdk using version 21 of Adoptium Temurin
-      - name: Setup java 21 using Adoptium Temurin
+      # Setup the jdk using version 11 of Adoptium Temurin
+      - name: Setup java 11 using Adoptium Temurin
         uses: actions/setup-java@v4
         with:
           distribution: 'temurin'
