### 2025-05-14 1d641c1
fix: Deploy WPS is not run on push to main (#778)
