### 2025-05-14 bc24889
Revert "fix: Deploy WPS is not run on push to main (#778)"
This reverts commit 1d641c158ba399e86bb24ca501b39d3aa78cde3a.
