### 2025-05-14 1ca6223
Update CI.yml
publish instead of old task
