+++ Docs/Data_Assimilation_Tutorial.rst
@@ -10,7 +10,7 @@ The main objective is to combine measurements from sensors located in Geneva wit
 Prerequisites
 ~~~~~~~~~~~~~~~~~
 
-- You need to have a working installation of NoiseModelling (NM).
+- You need to have a working installation of NoiseModelling (NM) with version 5.
 - A folder named ``devices_data`` containing sensor measurement files in ``CSV`` format. Each file represents the measurements of a single sensor.
 - A ``CSV`` file named ``device_mapping_sf.csv`` containing the geometry and unique identifier of all sensors.
 - The the OpenStreetMap (OSM) data of the given area : ``geneva.osm.pbf``.
@@ -25,7 +25,7 @@ The key columns are:
   * ``timestamp`` : Time in format <b>"%Y-%m-%d %H:%M:%S"</b>, representing the time of measurement.
   * ``Leq`` : Equivalent continuous sound level in dB(A), calculated over a period (15 min).
   * ``Temp`` : Temperature (°C) recorded by the sensor at the time of measurement.
-The ``device_mapping_sf.csv`` columns ka y are:
+The ``device_mapping_sf.csv`` columns kay are:
   * ``deveui`` : Unique identifier of the sensor.
   * ``The_GEOM`` : 3D point geometry in WKT (Well-Known Text) format — includes coordinates (X, Y) and altitude (Z) in the projected coordinate system.
 
@@ -49,13 +49,24 @@ The generated combinations include variation (%) around standard values for type
                    "temperatureValues": [10,15,20]
     ])
 
+Step 2 : Import Sensor Positions
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Step 2 : Preparation of Sensor data
+Importing the location of the sensors into the database from the csv file ``device_mapping_sf.csv``.
+
+.. code-block:: groovy
+    new Import_File().exec(connection,[
+                    "pathFile" : workingFolder+"device_mapping_sf.csv",
+                    "inputSRID" : 2056,
+                    "tableName": "SENSORS_LOCATION"
+    ])
+
+
+Step 3 : Preparation of Sensor data
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 This part extracts sensor data for a given period.Then use ``trainingRation`` (%) for the training data which will represent the receivers.
-So at the end of this part, the ``SENSORS_MEASUREMENTS`` (representing all the data for this period), ``OBSERVATION`` (training data set) and ``RECEIVERS`` sql tables will be created.
-And the ``device_mapping_sf.csv`` file is transformed into an sql table named ``SENSORS``.
+So at the end of this part, the ``SENSORS_MEASUREMENTS`` (representing all the data for this period), ``SENSORS_MEASUREMENTS_TRAINING`` (training data set representing the receiver) sql tables will be created.
 
 .. code-block:: groovy
     new Prepare_Sensors().exec(connection,[
@@ -66,7 +77,7 @@ And the ``device_mapping_sf.csv`` file is transformed into an sql table named ``
                     "targetSRID": 2056
     ])
 
-Step 3 : Import Buildings and Roads
+Step 4: Import Buildings and Roads
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Import buildings and road network (with predicted traffic flows) from an OSM file.
@@ -81,144 +92,115 @@ Import buildings and road network (with predicted traffic flows) from an OSM fil
                     "removeTunnels" : true
     ])
 
-Step 4 : Generate Traffic Emissions
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
[…]
