+++ Docs/Scientific_production.rst
@@ -6,65 +6,98 @@ Scientific production
 Below is a non-exhaustive list of articles or presentations in which NoiseModelling is used.
 
 SOTO-MOLINA, Ignacio, AUSEJO PRIETO, Miguel, et ARCE RUIZ, Rosa Maria. `Enhancing noise assessment accuracy: FEM as an advanced complement to CNOSSOS-EU <https://journals.sagepub.com/doi/10.1177/09574565251319271>.`_ Noise & Vibration Worldwide, 2025, p. 09574565251319271.
-Contact: - rosa.arce.ruiz@upm.es
+
+Contact: 
+         - rosa.arce.ruiz@upm.es
          - miguelausejo@gmail.com
          - ignacio.soto.molina@gmail.com
 
 TAGUSARI, Junta. `H-RISK with NoiseModelling: a QGIS plugin to predict environmental noise and estimate health risks. <https://joss.theoj.org/papers/10.21105/joss.06023.pdf>`_ Journal of Open Source Software, 2024, vol. 9, no 93, p. 6023.
+
 Contact: j.tagusari@eng.hokudai.ac.jp
 
 MONTENEGRO, A. L., MELLUSO, D., STASI, G., PANCI, A., BOLOGNESE, M., PALAZZUOLI, D., ... & LICITRA, G., `An open-source pipeline in noise modelling and noise exposure reduction in a port city.<https://www.researchgate.net/profile/Alexandra-Montenegro/publication/382072874_An_open-source_pipeline_in_noise_modelling_and_noise_exposure_reduction_in_a_port_city/links/668bdf59f3b61c4e2cb7e962/An-open-source-pipeline-in-noise-modelling-and-noise-exposure-reduction-in-a-port-city.pdf>`_, In Proceedings of the 30th International Congress on Sound and Vibration (ICSV30), 2024.
-Contact: - alexandra.montenegro@phd.unipi.it
+
+Contact: 
+         - alexandra.montenegro@phd.unipi.it
          - g.licitra@arpat.toscana.it
 
 TAGUSARI, Junta. `Small-scale noise mapping for arbitrary regions using open source noise prediction module. <https://www.ingentaconnect.com/content/ince/incecp/2024/00000270/00000009/art00070>`_ In : INTER-NOISE and NOISE-CON Congress and Conference Proceedings. Institute of Noise Control Engineering, 2024. p. 2621-2627.
+
 Contact: j.tagusari@eng.hokudai.ac.jp
 
+BISSON, Jérôme, CÔTÉ, Sabrina, CURRY, Henri, et al. `Setting up light and noise maps to define the contours of a reserve of darkness and silence in Sherbrooke, Québec, Canada <https://www.researchgate.net/publication/384881081_Setting_up_light_and_noise_maps_to_define_the_contours_of_a_reserve_of_darkness_and_silence_in_Sherbrooke_Quebec_Canada>_`. In : INTER-NOISE and NOISE-CON Congress and Conference Proceedings. Institute of Noise Control Engineering, 2024. p. 5334-5342.
+
+Contact:
+         -  jerome.bisson@usherbrooke.ca
+
+
+
[…]
