### 2025-05-21 b2496aa
Update Scientific_production.rst
