+++ Docs/NoiseModelling.bib
@@ -1,0 +1,726 @@
+
+@inproceedings{baclet_machine_2023,
+	address = {Glasgow, Scotland},
+	title = {A machine learning-and compressed sensing-based approach for surrogate modelling in environmental acoustics: towards fast evaluation of building façade road traffic noise levels},
+	url = {https://www.ingentaconnect.com/content/ince/incecp/2023/00000265/00000001/art00006},
+	doi = {10.3397/IN_2022_0898},
+	abstract = {State-of-the-art urban road traffic noise propagation simulation methods such as the CNOSSOS-EU framework rely on ray tracing to estimate noise levels at specific locations on façades, so-called receiver points; this method is computationally expensive and its cost increases with the number of receiver points, which limits the spatial accuracy of such simulations in the context of real-time or near-real-time urban noise simulation applications. This contribution aims to investigate the applicability of multiple data-driven methods to the surrogate modelling of traffic noise propagation for fast façade noise calculation as an alternative to these traditional, ray-tracing-based methods. The proposed approach uses compressed sensing to select a small subset of receiver points from which the data set of the entire façade may be reconstructed, associated with a Kriging model and neural networks, used to predict noise levels for these sensors. The prediction performance of each of these steps is evaluated on an academic test case, with two levels of complexity based on the dimensionality of the problem.},
+	booktitle = {{INTER}-{NOISE} and {NOISE}-{CON} {Congress} and {Conference} {Proceedings}},
+	author = {Baclet, Sacha and Venkataraman, Siddharth and Gomez, Erik and Bouchouireb, Hamza},
+	year = {2023},
+	keywords = {✅ Check, Road traffic, Surrogate Model},
+	pages = {6040--6051},
+}
+
+@inproceedings{baclet_methodology_2021,
+	title = {A methodology to assess the impact of driving noise from individual vehicles in an urban environment},
+	url = {https://kth.diva-portal.org/smash/get/diva2:1641387/FULLTEXT01.pdf},
+	booktitle = {27th {International} {Congress} on {Sound} and {Vibration}},
+	author = {Baclet, Sacha and Venkataraman, Siddharth and Rumpler, Romain},
+	year = {2021},
+	keywords = {⛔ No DOI found, ✅ Check, Exposure assessment, Traffic dynamics, traffic noise pollution},
+}
+
+@inproceedings{montenegro_open-source_2024,
+	address = {Amsterdam, Netherlands},
+	title = {An open-source pipeline in noise modelling and noise exposure reduction in a port city},
+	url = {https://iiav.org/content/archives_icsv_last/2024_icsv30/content/papers/papers/full_paper_183_20240415175901591.pdf},
+	booktitle = {Proceedings of the 30th {International} {Congress} on {Sound} and {Vibration} ({ICSV30})},
+	author = {Montenegro, Alexandra L. and Melluso, Dulia and Stasi, Gianmarco and Panci, Andrea and Bolognese, Matteo and Palazzuoli, Diego and Cerchiai, Mauro and Licitra, Gaetano},
+	year = {2024},
+	keywords = {⛔ No DOI found, Traffic dynamics, ✅ Check},
+}
+
+@article{can_approche_2018,
+	title = {Approche dynamique pour l’étude de l’emprise spatiale du bruit de trafic routier aux heures de pointe},
+	volume = {2018},
+	url = {https://hal.science/hal-02482315/},
+	doi = {10.1016/j.retrec.2018.07.001},
+	abstract = {Une chaine de modélisation dynamique, couplant le modèle de trafic dynamique Symuvia (LICIT, Ifsttar / ENTPE) et l’outil de prédiction des niveaux sonores Noisemodelling (UMRAE, Ifsttar / Cerema), est constituée dans cet article pour évaluer l’impact du trafic routier sur
[…]
