+++ Docs/NoiseModelling.bib
@@ -608,7 +608,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 	file = {Full Text PDF:/home/aumond/Zotero/storage/FLKBB8ZI/Nourmohammadi et al. - 2021 - Mapping Urban Environmental Performance with Emerging Data Sources A Case of Urban Greenery and Tra.pdf:application/pdf},
 }
 
-@misc{spitznagel_urban_2024,
+@article{spitznagel_urban_2024,
 	title = {Urban {Sound} {Propagation}: a {Benchmark} for 1-{Step} {Generative} {Modeling} of {Complex} {Physical} {Systems}},
 	shorttitle = {Urban {Sound} {Propagation}},
 	url = {https://arxiv.org/abs/2403.10904v2},
@@ -680,6 +680,7 @@ thus aims to assess the evolution of population exposure to noise pollution, usi
 	url = {https://openaccess.thecvf.com/content/CVPR2025/html/Spitznagel_PhysicsGen_Can_Generative_Models_Learn_from_Images_to_Predict_Complex_CVPR_2025_paper.html},
 	language = {en},
 	urldate = {2025-06-10},
+	booktitle = {Proceedings of the {Computer} {Vision} and {Pattern} {Recognition} {Conference}},
 	author = {Spitznagel, Martin and Vaillant, Jan and Keuper, Janis},
 	year = {2025},
 	keywords = {✅ Check, Surrogate Model},
