+++ README.md
@@ -11,9 +11,7 @@ NoiseModelling
 NoiseModelling is a library capable of producing noise maps. 
 It can be freely used either for research and education, as well as by experts in a professional use.
 
-A general overview of the model (v3.4.5 - September 2020) can be found in [this video](https://www.youtube.com/watch?v=V1-niMT9cYE&t=1s).
-
-* for **more information** on NoiseModelling, [visit the offical NoiseModelling website](https://noise-planet.org/noisemodelling.html)
+* for **more information** on NoiseModelling, [visit the online documentation](https://noisemodelling.readthedocs.io/en/latest/)
 * to **contribute to NoiseModelling** source code, follow the ["Get Started Dev"](https://noisemodelling.readthedocs.io/en/latest/Get_Started_Dev.html) page
 * to **contact the support / development team**, 
     - open an issue : https://github.com/Universite-Gustave-Eiffel/NoiseModelling/issues or a write a message : https://github.com/Universite-Gustave-Eiffel/NoiseModelling/discussions *(we prefer these two options)*
@@ -43,26 +41,8 @@ Licence
 
 NoiseModelling and its documentation are distributed for free under [GPL v3](https://noisemodelling.readthedocs.io/en/latest/License.html). 
 
-Publications
+Publications, Fundings, Tutorials...
 ---------------------------
 
-NoiseModelling was initially developed in a research context, which has led to numerous scientific publications. For more information, have a look to the ["Scientific production"](https://noisemodelling.readthedocs.io/en/v4.0.0/Scientific_production.html) page. 
-
-To quote this tool, please use the following bibliographic reference: Erwan Bocher, Gwenaël Guillaume, Judicaël Picaut, Gwendall Petit, Nicolas Fortin. *NoiseModelling: An Open Source GIS Based Tool to Produce Environmental Noise Maps*. ISPRS International Journal of Geo-Information, MDPI, 2019, 8 (3), pp.130. - [10.3390/ijgi8030130](https://www.mdpi.com/2220-9964/8/3/130)
-
-Fundings
----------------------------
-
-*Research projects:*
-- ANR [Eval-PDU](https://anr.fr/Projet-ANR-08-VILL-0005) (ANR-08-VILL-0005) 2008-2011
-- ANR [Veg-DUD](https://anr.fr/Projet-ANR-09-VILL-0007) (ANR-09-VILL-0007) 2009-2014
-- ANR [CENSE](https://anr.fr/Projet-ANR-16-CE22-0012) (ANR-16-CE22-0012) 2017-2021
-- [Nature4cities](https://www.nature4cities.eu/) (N4C) project, funded by European Union’s Horizon 2020 research and innovation programme under grant agreement No 730468
-- [PlaMADE](https://www.cerema.fr/fr/projets/plamade-plate-forme-mutualisee-aide-au-diagnostic) 2020-2022
-
-*Institutional (public) fundings:*
-- [Université Gustave Eiffel](https://www.univ-gustave-eiffel.fr/) (formerly Ifsttar, formerly LCPC), [CNRS](https://www.cnrs.fr), [Cerema](https://www.cerema.fr/), [Université Bretagne Sud](https://www.univ-ubs.fr/), [Ecole Centrale de Nantes](https://www.ec-nantes.fr/)
-
-*Private fundings:*
-- Airbus Urban Mobility
+Please consult the online documentation [here](https://noisemodelling.readthedocs.io/en/latest/).
 
