+++ Docs/Data_Assimilation_Tutorial.rst
@@ -1,6 +1,9 @@
 Data Assimilation
 ^^^^^^^^^^^^^^^^^^^^
 
+.. image:: ./images/Data_Assimilation/data_assimilation_banner.png
+        :alt: Data assimilation
+
 Introduction
 ~~~~~~~~~~~~~~~
 
@@ -141,11 +144,11 @@ Result
 The generated combinations include values for type of roads primary, secondary, tertiary, others, and temperature. The resulting table ``ALL_CONFIGURATIONS`` has the following columns:
 
 * ``IT``: Unique identifier of the combination (Primary Key - Integer)
-* ``PRIMARY_VAL``: percentage of primary roads traffic, given by trafficValues  (Double)
-* ``SECONDARY_VAL``: (Double)
-* ``TERTIARY_VAL``: (Double)
-* ``OTHERS_VAL``: (Double)
-* ``TEMP_VAL``: (Integer)
+* ``PRIMARY_VAL``: percentage of primary roads traffic, given by trafficValues (Double)
+* ``SECONDARY_VAL``: ... of secondary roads ... (Double)
+* ``TERTIARY_VAL``: ... of tertiary roads ...(Double)
+* ``OTHERS_VAL``: ... of other roads ...(Double)
+* ``TEMP_VAL``: temperature (Double)
 
 
 The first 10 lines of this table are shown below:
@@ -159,7 +162,7 @@ The first 10 lines of this table are shown below:
 .. warning::
     The total number of combinations can be huge. This value is defined as: (number of ``trafficValues`` elements) ^ 4  * (number of ``temperatureValues`` elements). 
 
-    In our example, we have ``"trafficValues": [0.01, 1.0, 2.0]`` and ``"temperatureValues": [10,15,20]``, so the number of combinations = 3 ^ 4 * 3 = 243 *(before filtration - see ‘note‘ tab just below)*
+    In our example, we have ``"trafficValues": "0.01, 1.0, 2.0"`` and ``"temperatureValues": "10,15,20"``, so the number of combinations = 3 ^ 4 * 3 = 243 *(before filtration - see ‘note‘ tab just below)*
 
     Even though this table is very important, only part of it will be used for all the maps to be simulated (see :ref:`Step 5 <Generate_all_traffic_emissions_and_maps>`)
 
@@ -230,9 +233,8 @@ For this tutorial, you can fill with these informations:
 * ``startDate`` : ``2024-08-25 06:30:00``
 * ``endDate`` : ``2024-08-25 07:30:00``
 * ``trainingRatio`` : ``0.8``
-* ``workingFolder`` : ``data_dir/data/wpsdata/dataAssimilation/``
+* ``workingFolder`` : ``.../data_dir/data/wpsdata/dataAssimilation/`` (enter the full URL e.g ``/home/myUserName/Documents/NoiseModelling_5.0.0/NoiseModelling_5.0.0/data_dir/data/wpsdata/dataAssimilation/``)
 * ``targetSRID`` : ``2056``
-
 
 If you are using the Groovy script
 
@@ -242,7 +244,7 @@ If you are using the Groovy script
                     "startDate":"2024-08-25 06:30:00",
                     "endDate": "2024-08-25 07:30:00",
                     "trainingRatio": 0.8,
-                    "workingFolder": "data_dir/data/wpsdata/dataAssimilation/",
+                    "workingFolder": ".../data_dir/data/wpsdata/dataAssimilation/",
                     "targetSRID": 2056
     ])
 
@@ -259,8 +261,7 @@ Once executed, two tables are created:
     * ``IDRECEIVER`` : The receiver's unique Id
     * ``EPOCH`` : Time of measurement (Epoch format - Unix time, ex:1724567400),
[…]
