### 2025-06-24 #7500 @kyleniemeyer
@faroit can we return to this submission? We really do not have any other editors with any closer expertise to the topic area, and I think you could rely on reviewers with Java expertise (since as editor you don't really need to dive into the code). Thank you!
