+++ Docs/Community.rst
@@ -20,41 +20,20 @@ Former and/or actual users
 
 NoiseModelling benefits from a large community of users. Some of them are listed below:
 
-Academic labs
+Academic labs 🧪
 ------------------
 
-* `UMRAE`_ (Université Gustave Eiffel - Cerema) (France)
-* `KTH`_ Royal Institue of Technology (Stockholm - Sweden)
-* `Utrecht University`_ (Netherlands)
-* `Université de Sherbrooke`_ (Canada)
-* Groupe WAVES (`Ghent University`_ - Belgium)
-* `Dipartimento di Ingegneria Civile/DICIV`_ (Università Degli Studi di Salerno - Italy)
-* `HFT Stuttgart Technical University of Applied Science`_ (Germany)
-* `LETG - CNRS`_ (Nantes, France)
-* `LAboratoire de Géographie et d’Aménagement de Montpellier`_ (LAGAM) - Université Paul Valéry (Montpellier, France)
-
-Private company
-------------------
-
-* `Archimethod SA`_ (Switzerland)
-* `OrbiWise`_ (Switzerland)
-* `Neovya`_ (France)
-* `Quiet Places Ltd`_ (United-Kingdom)
-
-Teaching
-------------------
-
-* `ENSIM`_ (Le Mans, France)
-* `Université de Sherbrooke`_ (Canada)
-* `INSA Rennes`_ (France)
-* `Master Acoustique`_ Université du Mans (France)
-* `École Centrale de Nantes`_ (France)
-
-Other
-------------------
-
-* `Centre National d’Incendie et de Secours`_ - CNIS (Luxembourg)
-
+* `UMRAE`_ (Université Gustave Eiffel - Cerema) (🇫🇷)
+* `KTH`_ Royal Institue of Technology (Stockholm - 🇸🇪)
+* `Utrecht University`_ (🇳🇱)
+* `Université de Sherbrooke`_ (🇨🇦)
+* Groupe WAVES (`Ghent University`_ - 🇧🇪)
+* `Dipartimento di Ingegneria Civile/DICIV`_ (Università Degli Studi di Salerno - 🇮🇹)
+* `HFT Stuttgart Technical University of Applied Science`_ (🇩🇪)
+* `LETG - CNRS`_ (Nantes, 🇫🇷)
+* `LAboratoire de Géographie et d’Aménagement de Montpellier`_ (LAGAM) - Université Paul Valéry (Montpellier, 🇫🇷)
+* `South China University of Technology`_ (🇨🇳)
+* `Università di Pisa - ARPAT`_ (🇮🇹)
 
 .. _UMRAE: https://www.umrae.fr
 .. _KTH: https://www.kth.se/en
@@ -65,24 +44,66 @@ Other
 .. _HFT Stuttgart Technical University of Applied Science: https://www.hft-stuttgart.com/
 .. _LETG - CNRS: https://letg.cnrs.fr/
 .. _LAboratoire de Géographie et d’Aménagement de Montpellier: https://lagam.xyz/
+.. _South China University of Technology: https://www.scut.edu.cn/en/
+.. _Università di Pisa - ARPAT: https://www.arpat.toscana.it/notizie/arpatnews/2021/231-21/arpat-e-universita-di-pisa-insieme-per-iniziative-congiunte-nel-campo-della-protezione-dell-ambiente/?searchterm=None
+
+
+Private company 🏭
+-------------------
+
+* `Archimethod SA`_ (🇨🇭)
+* `OrbiWise`_ (🇨🇭)
+* `Neovya`_ (🇫🇷)
+* `Quiet Places Ltd`_ (🇬🇧)
+* `MobiLysis`_ (🇨🇭)
 
 .. _Archimethod SA: https://archi-method.ch/
[…]
