### 2025-07-09 a43921d
Update JOSS branch to NM 5 (#817)
* About #751 remove empty geometry in clean buildings wps process.

* modified:   wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/NoiseModelling/Noise_level_from_source.groovy
AttenuationCnossosParameters -> AttenuationParameters

* modified:   wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/NoiseModelling/Noise_level_from_source.groovy

* Update Noise_level_from_source.groovy

* modified:   wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/NoiseModelling/Noise_level_from_source.groovy

* DataAssim !!

* Fix Data_A
[…]
