+++ Docs/Input_source.rst
@@ -1,5 +1,5 @@
 Sound source
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^
 
 NoiseModelling is a tool for producing noise maps. To do so, at different stages of the process, the application needs input data, respecting a strict formalism.
 
@@ -8,7 +8,7 @@ Below we describe the table ``SOURCES_GEOM`` and ``SOURCES_EMISSION``, expected 
 The goal is to define any sound source with a custom noise spectrum level.
 
 SOURCES_GEOM Table definition
----------------------
+-------------------------------
 
 * ``PK`` *
     * Description: An identifier (must be a PRIMARY KEY). The field name is not important, it could be IDSOURCE.
@@ -29,7 +29,7 @@ If you provide only the geometry NoiseModelling will compute only the attenuatio
 To provide the emission level you must define an emission table for each time period.
 
 SOURCES_EMISSION Table definition
----------------------
+-----------------------------------
 
 * ``IDSOURCE`` *
     * Description: An identifier linked to the primary key of the ``SOURCES_GEOM`` table.
@@ -64,4 +64,3 @@ SOURCES_EMISSION Table definition
 
 .. note::
     You can define partially the bands. We defined here only octave bands but NoiseModelling is supporting third-octaves ``HZ50, HZ63, HZ80, HZ100, HZ125, HZ160, HZ200, HZ250, HZ315, HZ400, HZ500, HZ630, HZ800, HZ1000, HZ1250, HZ1600, HZ2000, HZ2500, HZ3150, HZ4000, HZ5000, HZ6300, HZ8000, HZ10000``.
-
