+++ Docs/Data_Assimilation_Tutorial.rst
@@ -1,0 +1,681 @@
+Data Assimilation
+^^^^^^^^^^^^^^^^^^^^
+
+.. image:: ./images/Data_Assimilation/data_assimilation_banner.png
+        :alt: Data assimilation
+
+Introduction
+~~~~~~~~~~~~~~~
+
+Data assimilation is a technique that combines observations with a numerical model to improve the accuracy of forecasts or analyses. Applied to acoustics and noise maps, data assimilation makes it possible to integrate real noise measurements (*e.g* coming from sensors) into noise propagation models to produce more accurate and reliable noise maps.
+
+In this tutorial, we will see how to model and simulate noise in `Geneva`_ 🇨🇭 using static (road network, buildings) and dynamic (traffic, temperature) data.
+The main objective is to **combine measurements from sensors** located in Geneva **with acoustic simulations** to **identify traffic configurations that best match measurements**.
+
+.. _Geneva: https://www.openstreetmap.org/relation/1685488
+
+Requirements
+~~~~~~~~~~~~~~~~~
+
+To play with this tutorial, you will need:
+
+* a working installation of NoiseModelling (NM) with at least version 5.0.1. If needed, get the last release on the official `GitHub repository`_,
+* the tutorial's datasets, stored in the folder ``.../NoiseModelling_5.0.1/data_dir/data/wpsdata/dataAssimilation/``,
+* the dedicated WPS ``.groovy`` scripts, stored in the folder ``.../NoiseModelling_5.0.1/data_dir/scripts/wps/DataAssimilation/``.
+
+.. _GitHub repository: https://github.com/Universite-Gustave-Eiffel/NoiseModelling/releases
+
+The data
+~~~~~~~~~~~~~~~~~
+
+In the the tutorial's folder, you have the following data:
+
+    1. sensor measurement ``.csv`` files, stored in the ``devices_data/`` folder (One file per sensor),
+    2. ``device_mapping_sf.geojson`` containing two columns : the point geometry (``THE_GEOM``) and a sensor's unique identifier (``DEVEUI```). This file is also provided in ``.csv`` format,
+    3. the `OpenStreetMap`_ (OSM) dataset of the studied area : ``geneva.osm.pbf``.
+
+.. _OpenStreetMap: https://www.openstreetmap.org/
+
+
+Sensor measurements
+-------------------------------------------------
+
+Each ``.csv`` files in the folder ``devices_data/`` contains environmental noise measurements recorded by individual sensors.
+The columns are:
+
+* ``deveui`` : Unique identifier of the sensor,
+* ``epoch`` : Time of measurement (Epoch format - Unix time, ex:1724567400),
+* ``Leq`` : Equivalent continuous sound level in dB(A), calculated over a period (15 min),
+* ``Temp`` : Temperature (°C) recorded by the sensor at the time of measurement,
+* ``timestamp`` : Time of measurement (timestamp format ``"YYYY-MM-DD HH:MM:SS"``).
+
+Below is an illustration with the file ``4a6.csv``. Here we can see that 5 measurements were taken, every 15 minutes, between 6:30 and 7:30 am. During this period, noise levels rose slightly, while temperatures remained stable overall (except for the first measurement).
+
+.. csv-table:: Informations stored in the sensor's file ``4a6.csv``
+   :file: ./data/4a6.csv
+   :widths: 15, 15, 15, 15, 30
+   :header-rows: 1
+
+.. warning::
+    In this tutorial all the values, coming from the 7 sensors, are sampled approximately every 15 minutes, but the exact spacing may vary slightly.
+
+    We could also have had data with a time step of 5 minutes, 10 minutes, 1 hour, etc.
+
[…]
