+++ Docs/Scientific_production.rst
@@ -1,90 +1,104 @@
 .. _My target:
 
-Scientific production 
+Scientific production
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Below is a non-exhaustive list of articles or presentations in which NoiseModelling is used.
+Below is a **non-exhaustive** list of scientific publications in which NoiseModelling has been used.
 
-Standard Noise maps
-~~~~~~~~~~~~~~~~~~~~~~~~~
-
-
-GRAZIUSO G., FRANCAVILLA A. B., MANCINI S., GUARNACCIA C., `Open-source software tools for strategic noise mapping: a case study <https://iopscience.iop.org/article/10.1088/1742-6596/2162/1/012014>`_, Journal of Physics: Conference Series, 2022, vol. 2162, 012014
-
-AUMOND P., BOCHER E., ECOTIERE D., FORTIN N., GAUVREAU B., GUILLAUME G., PETIT G., `Improvement of city noise map production processes and sensitivity analysis to noise models inputs <http://www.sea-acustica.es/fileadmin/Madeira21/ID122.pdf>`_, Euronoise Conference Proceedings, 2021, p. 1128
-
-NOURMOHAMMADI Z., LILASATHAPORNKIT T., ASHFAQ M., et al., `Mapping Urban Environmental Performance with Emerging Data Sources: A Case of Urban Greenery and Traffic Noise in Sydney, Australia <https://www.mdpi.com/2071-1050/13/2/605>`_, Sustainability, 2021, vol. 13, n° 2, p. 605
-
-BAEZA J. L., SIEVERT J. L., LANDWEHR A., et al., `CityScope Platform for Real-Time Analysis and Decision-Support in Urban Design Competitions <https://www.igi-global.com/article/cityscope-platform-for-real-time-analysis-and-decision-support-in-urban-design-competitions/278826>`_, International Journal of E-Planning Research (IJEPR), 2021, vol. 10, n° 4, p. 1-17
-
-WANG Z., NOVACK T., YAN Y., ZIPF A., `Quiet Route Planning for Pedestrians in Traffic Noise Polluted Environments <https://ieeexplore.ieee.org/document/9139350/>`_, IEEE Transactions on Intelligent Transportation Systems, 2020
-
-AUMOND P., FORTIN N., CAN A., `Overview of the NoiseModelling open-source software version 3 and its applications <https://www.ingentaconnect.com/contentone/ince/incecp/2020/00000261/00000004/art00003>`_, INTER-NOISE and NOISE-CON Congress and Conference Proceedings, 2020, vol. 261, n°4, p. 2005-2011
+* Articles (co-)written by members of the NoiseModelling team, belonging to the `Joint Research Unit in Environmental Acoustics`_, are in the ``UMRAE`` tab.
+* Articles written outside the UMRAE are in the ``Not UMRAE`` tab.
 
 
-Dynamic Noise maps
-~~~~~~~~~~~~~~~~~~~~~~~~~
+.. tip::
+    If you have done work with NoiseModelling and your article is not referenced in this list, please let us know (contact@noise-planet.org).
 
[…]

[…]
