### 2025-07-30 047de4a
Update Community.rst
add DGPR
