+++ Docs/Community.rst
@@ -80,9 +80,11 @@ Teaching 📚
 Other
 ------------------
 
+* `Direction générale de la prévention des risques (DGPR)`_ - (🇫🇷)
 * `Centre National d’Incendie et de Secours`_ - CNIS (🇱🇺)
 * `Ministerio para la Transición Ecológica y el Reto Demográfico`_ (🇪🇸)
 
+.. _Direction générale de la prévention des risques (DGPR): https://www.ecologie.gouv.fr/direction-generale-prevention-risques-dgpr
 .. _Centre National d’Incendie et de Secours: https://112.public.lu/fr/organisation/cnis0.html
 .. _Ministerio para la Transición Ecológica y el Reto Demográfico: https://www.miteco.gob.es
 
