### 2025-07-31 #7500 @faroit
@hendriks73 would you be able to review this for when you are back from vacation?
