### 2025-09-02 7f373d6
process point source with relative height TODO, add unit test
