### 2025-09-03 5e21b4f
Add unit test for source with relative height (will compute altitude at the begining of the simulation). Source is also rejected if inside a building, with a warning
