### 2025-09-05 d626289
do not use profile, sign only on deploy phase
