+++ .github/workflows/CI_Release.yml
@@ -58,7 +58,6 @@ jobs:
             -ntp \
             --batch-mode \
             -Dmaven.test.skip=true \
-            -P maven-deploy \
             release:prepare release:perform \
             -Dusername=$GITHUB_ACTOR -Dpassword=$GITHUB_TOKEN ${VERSION:+"-DdevelopmentVersion="$VERSION"-SNAPSHOT"}
         env:
