### 2025-09-05 dcc4b18
update gradle publish configuration
