### 2025-09-08 04badfc
use TC28 a base to check for modified paths/planes in favorable conditions
