+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/CurvedProfileTest.java
@@ -1,0 +1,133 @@
+/**
+ * NoiseModelling is a library capable of producing noise maps. It can be freely used either for research and education, as well as by experts in a professional use.
+ * <p>
+ * NoiseModelling is distributed under GPL 3 license. You can read a copy of this License in the file LICENCE provided with this software.
+ * <p>
+ * Official webpage : http://noise-planet.org/noisemodelling.html
+ * Contact: contact@noise-planet.org
+ */
+
+
+package org.noise_planet.noisemodelling.pathfinder;
+
+import org.junit.jupiter.api.Test;
+import org.locationtech.jts.geom.Coordinate;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.*;
+import org.noise_planet.noisemodelling.pathfinder.utils.geometry.CurvedProfileGenerator;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+
+public class CurvedProfileTest {
+
+
+    @Test
+    public void curvedProfile() {
+        // Test the implementation with some dummy data
+        List<CutPoint> flatProfile = new ArrayList<>();
+        flatProfile.add(new CutPointSource(new Coordinate(0, 0, 4)));
+        flatProfile.add(new CutPoint(new Coordinate(500, 500, 1)));
+        flatProfile.add(new CutPoint(new Coordinate(1000, 1000, 1)));
+        flatProfile.add(new CutPoint(new Coordinate(1500, 1500, 1)));
+        flatProfile.add(new CutPointReceiver(new Coordinate(2000, 2000, 4)));
+
+        flatProfile.get(0).setZGround(0);
+        flatProfile.get(1).setZGround(0);
+        flatProfile.get(2).setZGround(0);
+        flatProfile.get(3).setZGround(0);
+        flatProfile.get(4).setZGround(0);
+
+        double distance = flatProfile.get(0).getCoordinate().distance(flatProfile.get(4).getCoordinate());
+        List<CutPoint> curvedProfile = CurvedProfileGenerator.applyTransformation(flatProfile);
+        assertInstanceOf(CutPointSource.class, curvedProfile.get(0));
+        assertInstanceOf(CutPointReceiver.class, curvedProfile.get(curvedProfile.size() - 1));
+        assertEquals(2828.427, distance, 1e-3);
+        assertEquals(0, new Coordinate(0, 0, 4).distance3D(curvedProfile.get(0).getCoordinate()), 1e-6);
+        assertEquals(0, new Coordinate(2000, 2000, 4).distance3D(curvedProfile.get(curvedProfile.size() - 1).getCoordinate()), 1e-6);
+        assertEquals(-15.57, curvedProfile.get(1).getCoordinate().z, 1e-2);
+        assertEquals(-21.09, curvedProfile.get(2).getCoordinate().z, 1e-2);
+        assertEquals(-15.57, curvedProfile.get(3).getCoordinate().z, 1e-2);
+        assertEquals(-16.57, curvedProfile.get(1).zGround, 1e-2);
+        assertEquals(-22.09, curvedProfile.get(2).zGround, 1e-2);
+        assertEquals(-16.57, curvedProfile.get(3).zGround, 1e-2);
+    }
+
+    @Test
+    public void testTC28CurvedProfile() {
+
+        //Create obstruction test object
+        ProfileBuilder builder = new ProfileBuilder();
+
+        // Add building
+        builder.addBuilding(new Coordinate[]{
+                        new Coordinate(113, 10, 0),
+                        new Coordinate(127, 16, 0),
[…]
