### 2025-09-08 ec6609a
Maybe fixed gamma, not sure Add unit test, tc28 hull look ok
