### 2025-09-09 d7a6d61
Check for T28 SO length, ok
