### 2025-09-09 930e13e
build path of left and right propagation altered by the offset of buildings roofs by the curved coordinate system method for favorable conditions.
