### 2025-09-11 9b749e7
we do not need height as this not a parameter to the curve equation (we use only altitude, the same reference point for source and receiver)
Change methods according to that, add inverse option too
