+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/utils/geometry/CurvedProfileGenerator.java
@@ -27,21 +27,6 @@ import static java.lang.Math.max;
  */
 public class CurvedProfileGenerator {
 
-    // Method to transform a list of flat coordinates to curved coordinates
-
-
-    public static double computeZp(Coordinate cs, Coordinate cr, double C0, Coordinate p) {
-        double xpp = p.x - (cs.x+cr.x)/2;
-        double ypp = p.y - (cs.y+cr.y)/2;
-        double zpp = p.z - (cs.z+cr.z)/2;
-
-        double xpp2 = xpp*xpp;
-        double ypp2 = ypp*ypp;
-        double zpp2 = zpp*zpp;
-
-        return (C0*(xpp2+ypp2+zpp2+zpp*C0))/(xpp2+ypp2+((C0+zpp)*(C0+zpp)));
-    }
-
     /**
      * Eq.2.5.24 and Eq. 2.5.25
      * @param mn Length of ray
@@ -52,49 +37,21 @@ public class CurvedProfileGenerator {
         return 2*max(1000, 8*d)* asin(mn/(2*max(1000, 8*d)));
     }
 
-/**
-* Salomons, E., Van Maercke, D., Defrance, J., &amp; De Roo, F. (2011). The Harmonoise sound propagation model. Acta acustica united with acustica, 97(1), 62-74.
-* @param cs Source coordinate
-* @param cr Receiver coordinate
-* @param hs Height of source above ground
-* @param hr Height of receiver above ground
-* @param flatProfile Array of coordinates representing the flat profile (should be discretized with segments distance &lt; 50 m)
-* @return Array of coordinates representing the curved profile
-*/
-    public static Coordinate[] applyTransformation(Coordinate cs, Coordinate cr, double hs, double hr,  Coordinate[] flatProfile) {
-        Coordinate[] curvedProfile = new Coordinate[flatProfile.length];
-
-        // Radius of curvature
-        double R_c = Math.max(1000, 8 * cs.distance3D(cr));
-        double C0 = 2*(((hs+hr)/2)+R_c);
-        double zs = computeZp(cs, cr, C0, cs) - cs.z;
-
-        for (int i = 0; i < flatProfile.length; i++) {
-            double zp = computeZp(cs, cr, C0, flatProfile[i]);
-            curvedProfile[i] = new Coordinate(flatProfile[i].x, flatProfile[i].y, zp - zs); // Adjust z to keep the endpoints consistent
-        }
-        return curvedProfile;
-    }
-
     /**
      * Salomons, E., Van Maercke, D., Defrance, J.,&amp;De Roo, F. (2011). The Harmonoise sound propagation model. Acta acustica united with acustica, 97(1), 62-74.
      * @param flatProfile
      * @return
      */
-
-    public static List<CutPoint> applyTransformation(List<CutPoint> flatProfile) {
-
+    public static List<CutPoint> applyTransformation(List<CutPoint> flatProfile, boolean inversed) {
         // Get chord endpoints
         CutPoint sourcePoint = flatProfile.get(0);
         CutPoint receiverPoint = flatProfile.get(flatProfile.size() - 1);
         Coordinate cs = sourcePoint.getCoordinate();
         Coordinate cr = receiverPoint.getCoordinate();
-        double hs = cs.getZ() - sourcePoint.zGround;
-        double hr = cr.getZ() - receiverPoint.zGround;
 
         List<CutPoint> curvedProfile = new ArrayList<>();
-        Coordinate[] curvedCoords = applyTransformation(cs, cr, hs, hr, flatProfile.stream().map(CutPoint::getCoordinate).toArray(Coordinate[]::new));
[…]
