### 2025-09-11 3b79127
TC28 good diffraction path in unit test
