### 2025-09-11 784d515
TC28 check corners
