### 2025-09-11 a2b86b2
better check inverse operation
