### 2025-09-11 9aaca1a
computed curved path left and right in main code, unit test fail (obviously at this step)
