### 2025-09-15 29eeddb
restore coordinates from curved coordinate system, add attributes on cutplane for better understanding
