### 2025-09-16 eb56f17
add left/right curved paths in TC28 unit test
