+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/PathFinderTest.java
@@ -62,6 +62,27 @@ public class PathFinderTest {
      */
     public static final double DELTA_PLANES = 0.1;
 
+    private void assertCutProfiles(String utName, Collection<CutProfile> cutProfiles) throws IOException {
+        for(CutProfile cutProfile : cutProfiles) {
+            StringBuilder sb = new StringBuilder(utName);
+            switch (cutProfile.getProfileType()) {
+                case LEFT:
+                    sb.append("_Left");
+                    break;
+                case RIGHT:
+                    sb.append("_Right");
+                    break;
+                default:
+                    sb.append("_Direct");
+                    break;
+            }
+            if(cutProfile.curvedPath) {
+                sb.append("_Curved");
+            }
+            assertCutProfile(sb.toString(), cutProfile);
+        }
+    }
+
     private void assertCutProfile(String utName, CutProfile cutProfile) throws IOException {
         String testCaseFileName = utName + ".json";
         if(overwriteTestCase) {
@@ -88,6 +109,7 @@ public class PathFinderTest {
 
     public static void assertCutProfile(InputStream expected, CutProfile got) throws IOException {
         ObjectMapper mapper = new ObjectMapper();
+        assertNotNull(expected);
         CutProfile cutProfile = mapper.readValue(expected, CutProfile.class);
         assertCutProfile(cutProfile, got);
     }
@@ -96,43 +118,45 @@ public class PathFinderTest {
         assertNotNull(expected);
         assertNotNull(got);
         assertEquals(expected.cutPoints.size(), got.cutPoints.size(), "Not the same number of cut points");
+        assertEquals(expected.profileType, got.profileType, "Not the same profile type");
+        assertEquals(expected.curvedPath, got.curvedPath, "Not the same curved path attribute value");
+
         for (int i = 0; i < expected.cutPoints.size(); i++) {
             CutPoint expectedCutPoint = expected.cutPoints.get(i);
             CutPoint gotCutPoint = got.cutPoints.get(i);
             assertInstanceOf(expectedCutPoint.getClass(), gotCutPoint);
-            assert3DCoordinateEquals(expectedCutPoint+"!="+gotCutPoint, expectedCutPoint.coordinate,
+            assert3DCoordinateEquals(expectedCutPoint + "!=" + gotCutPoint, expectedCutPoint.coordinate,
                     gotCutPoint.coordinate, DELTA_COORDS);
             assertEquals(expectedCutPoint.zGround, gotCutPoint.zGround, 0.01, "zGround");
             assertEquals(expectedCutPoint.groundCoefficient, gotCutPoint.groundCoefficient, 0.01, "groundCoefficient");
 
-            if(expectedCutPoint instanceof CutPointSource) {
+            if (expectedCutPoint instanceof CutPointSource) {
                 CutPointSource expectedCutPointSource = (CutPointSource) expectedCutPoint;
                 CutPointSource gotCutPointSource = (CutPointSource) gotCutPoint;
-                assertEquals(expectedCutPointSource.li, gotCutPointSource.li,0.01);
-                assertEquals(expectedCutPointSource.orientation.yaw, gotCutPointSource.orientation.yaw,0.01);
-                assertEquals(expectedCutPointSource.orientation.pitch, gotCutPointSource.orientation.pitch,0.01);
-                assertEquals(expectedCutPointSource.orientation.roll, gotCutPointSource.orientation.roll,0.01);
+                assertEquals(expectedCutPointSource.li, gotCutPointSource.li, 0.01);
[…]
