### 2025-09-16 6ea0bfa
Ignore exporting boolean values in json files for cutPlane not used in attenuation computation
