### 2025-09-16 c4d82b7
remove homogenous and favorable variables in cnossospath
