+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/CnossosPath.java
@@ -13,51 +13,42 @@ package org.noise_planet.noisemodelling.propagation.cnossos;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 
 /**
- * All the datas Path of Cnossos
+ * Attenuation computed from vertical Profile and scene settings following CNOSSOS-EU method.
  */
 public class CnossosPath extends Path {
     public  double[] aAtm = new double[0];
     public  double[] aDiv = new double[0];
     public  double[] aRef = new double[0];
-    public  double[] double_aBoundaryH = new double[0];
-    public  double[] double_aBoundaryF = new double[0];
-    public  double[] aRetroDiffH = new double[0]; // Alpha Retro Diffraction homogenous
-    public  double[] aRetroDiffF = new double[0]; // Alpha Retro Diffraction favorable
-    public  double[] aGlobalH = new double[0];
-    public double[] aGlobalF = new double[0];
-    public double[] aDifH = new double[0];
-    public double[] aDifF = new double[0];
-    public double[] aGlobal = new double[0];
+    public  double[] double_aBoundary = new double[0];
+    public  double[] aRetroDiff = new double[0]; // Alpha Retro Diffraction
+    /**
+     * aGlobalRaw but with attenuation (dB) from the ponderation of
+     * - the directivity attenuation (favorable or homogeneous atmospheric conditions)
+     * - The source directivity attenuation
+     */
+    public  double[] aGlobal = new double[0];
+    public  double[] aGlobalRaw = new double[0];
+    public double[] aDif = new double[0];
     public double[] aSource = new double[0]; // directivity attenuation
-    public double deltaH = Double.MAX_VALUE;
-    public double deltaF= Double.MAX_VALUE;
-    public double deltaPrimeH= Double.MAX_VALUE;
-    public double deltaPrimeF= Double.MAX_VALUE;
-    public double deltaSPrimeRH= Double.MAX_VALUE;
-    public double deltaSRPrimeH= Double.MAX_VALUE;
-    public ABoundary aBoundaryH = new ABoundary();
-    public ABoundary aBoundaryF = new ABoundary();
+    public double delta = Double.MAX_VALUE;
+    public double deltaPrime= Double.MAX_VALUE;;
+    public double deltaSPrimeR= Double.MAX_VALUE;
+    public double deltaSRPrime= Double.MAX_VALUE;
+    public ABoundary aBoundary = new ABoundary();;
     public GroundAttenuation groundAttenuation = new GroundAttenuation();
-    public double deltaSPrimeRF= Double.MAX_VALUE;
-    public double deltaSRPrimeF= Double.MAX_VALUE;
     public double e=0;
-    public double deltaRetroH= Double.MAX_VALUE;
-    public double deltaRetroF= Double.MAX_VALUE;
+    public double deltaRetro= Double.MAX_VALUE;
 
     public void init(int size) {
         this.aAtm = new double[size];
         this.aDiv = new double[size];
         this.aRef = new double[size];
-        this.double_aBoundaryH = new double[size];
-        this.double_aBoundaryF = new double[size];
-        this.aGlobalH = new double[size];
-        this.aGlobalF = new double[size];
-        this.aDifH = new double[size];
-        this.aDifF = new double[size];
+        this.double_aBoundary = new double[size];;
+        this.aGlobal = new double[size];;
+        this.aDif = new double[size];
         this.aGlobal = new double[size];
         this.aSource = new double[size];
-        this.aRetroDiffH = new double[size];
-        this.aRetroDiffF = new double[size];
+        this.aRetroDiff = new double[size];
     }
 
     public CnossosPath() {
[…]

[…]
