### 2025-09-17 d4914e2
fix last code, keep TC28 unit test for favorable updates fix
