+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/AttenuationCnossos.java
@@ -159,18 +159,10 @@ public class AttenuationCnossos {
 
             //For testing purpose
             if(pathParameters.keepAbsorption) {
-                if(pathParameters.isFavorable()) {
-                    pathParameters.groundAttenuation.wF[idfreq] += w;
-                    pathParameters.groundAttenuation.cfF[idfreq] = cf;
-                    pathParameters.groundAttenuation.aGroundF[idfreq] = aGround[idfreq];
-                }
-                else{
-                    pathParameters.groundAttenuation.wH[idfreq] += w;
-                    pathParameters.groundAttenuation.cfH[idfreq] = cf;
-                    pathParameters.groundAttenuation.aGroundH[idfreq] = aGround[idfreq];
-                }
+                pathParameters.groundAttenuation.w[idfreq] += w;
+                pathParameters.groundAttenuation.cf[idfreq] = cf;
+                pathParameters.groundAttenuation.aGround[idfreq] = aGround[idfreq];
             }
-
         }
         return aGround;
     }
@@ -251,12 +243,7 @@ public class AttenuationCnossos {
 
             //For testing purpose
             if(pathParameters.keepAbsorption) {
-                if(pathParameters.isFavorable()) {
-                    pathParameters.groundAttenuation.aGroundF = aGround;
-                }
-                else{
-                    pathParameters.groundAttenuation.aGroundH = aGround;
-                }
+                pathParameters.groundAttenuation.aGround = aGround;
             }
             return aGround;
         }
@@ -483,16 +470,13 @@ public class AttenuationCnossos {
 
     /**
      *
-     * @param pp
-     * @param freq
-     * @param favorable
-     * @return
+     * @param pp Cnossos path
+     * @param freq Frequency
+     * @return true if the r-criterion is valid
      */
-    private static boolean isValidRcrit(CnossosPath pp, int freq, boolean favorable) {
+    private static boolean isValidRcrit(CnossosPath pp, int freq) {
         double lambda = 340.0/freq;
-        return favorable ?
-                pp.deltaF > -lambda / 20 && pp.deltaF > lambda / 4 - pp.deltaPrimeF || pp.deltaF > 0 :
-                pp.deltaH > -lambda / 20 && pp.deltaH > lambda / 4 - pp.deltaPrimeH || pp.deltaH > 0 ;
+        return pp.delta > -lambda / 20 && pp.delta > lambda / 4 - pp.deltaPrime || pp.delta > 0;
     }
 
     /**
@@ -507,12 +491,11 @@ public class AttenuationCnossos {
         List<PointPath> diffPts = path.getPointList().stream().
                 filter(pointPath -> pointPath.type.equals(DIFH_RCRIT) || pointPath.type.equals(DIFH)
                         || pointPath.type.equals(DIFV)).collect(Collectors.toList());
-        path.aBoundaryH.init(data.getFrequencies().size());
-        path.aBoundaryF.init(data.getFrequencies().size());
+        path.aBoundary.init(data.getFrequencies().size());
         // Without diff
         for(int i=0; i<data.getFrequencies().size(); i++) {
             int finalI = i;
-            boolean isValidRCriterion = isValidRcrit(path, data.getFrequencies().get(finalI), path.isFavorable());
+            boolean isValidRCriterion = isValidRcrit(path, data.getFrequencies().get(finalI));
             PointPath first = diffPts.stream()
[…]
