### 2025-09-17 a600990
TC28 3dB error in all frequencies, looking for A ground issues ?
