+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/CnossosPathBuilder.java
@@ -246,69 +246,22 @@ public class CnossosPathBuilder {
         pathParameters.setSegmentList(segments);
         pathParameters.setSRSegment(srPath);
         pathParameters.init(exactFrequencyArray.size());
-        // Extract the first and last points to define the line segment
-        Coordinate firstPt = pts2D.get(0);
-        Coordinate lastPt = pts2D.get(pts2D.size() - 1);
-
-        // Filter out points that are below the line segment
-        List<Coordinate> convexHullInput = new ArrayList<>();
-        // Add source position
-        convexHullInput.add(pts2D.get(0));
-        // Add valid diffraction point, building/walls/dem
-        for (int idPoint=1; idPoint < cutProfilePoints.size() - 1; idPoint++) {
-            CutPoint currentPoint = cutProfilePoints.get(idPoint);
-            // We only add the point at the top of the wall, not the point at the bottom of the wall
-            if(currentPoint instanceof CutPointTopography
-                    || (currentPoint instanceof CutPointWall
-                    && Double.compare(currentPoint.getCoordinate().z, currentPoint.getzGround()) != 0)) {
-                convexHullInput.add(pts2D.get(idPoint));
-            }
+        List<Coordinate> hullPts2D = pts2D;
+        if(favorable) {
+            // Compute the altered profile for favorable path
+            hullPts2D = cutProfile.computePts2D(true);
         }
-        // Add receiver position
-        convexHullInput.add(pts2D.get(pts2D.size() - 1));
+        // Compute convex hull of the profile
+        List<Integer> hullPointsIndices = cutProfile.getConvexHullIndices(hullPts2D);
 
-        // Compute the convex hull using JTS
-        List<Coordinate> convexHullPoints = new ArrayList<>();
-        if(convexHullInput.size() > 2) {
-            GeometryFactory geomFactory = new GeometryFactory();
-            Coordinate[] coordsArray = convexHullInput.toArray(new Coordinate[0]);
-            ConvexHull convexHull = new ConvexHull(coordsArray, geomFactory);
-            Coordinate[] convexHullCoords = convexHull.getConvexHull().getCoordinates();
-            int indexFirst = Arrays.asList(convexHull.getConvexHull().getCoordinates()).indexOf(firstPt);
-            int indexLast = Arrays.asList(convexHull.getConvexHull().getCoordinates()).lastIndexOf(lastPt);
-            if(indexFirst == -1 || indexLast == -1 || indexFirst > indexLast) {
-                throw new IllegalArgumentException("Wrong input data " + cutProfile.toString());
-            }
-            convexHullCoords = Arrays.copyOfRange(convexHullCoords, indexFirst, indexLast + 1);
-            CoordinateSequence coordSequence = geomFactory.getCoordinateSequenceFactory().create(convexHullCoords);
-            Geometry geom = geomFactory.createLineString(coordSequence);
-            Geometry uniqueGeom = geom.union(); // Removes duplicate coordinates
-            convexHullCoords = uniqueGeom.getCoordinates();
-            // Convert the result back to your format (List<Point2D> pts)
-            if (convexHullCoords.length == 3) {
-                convexHullPoints = Arrays.asList(convexHullCoords);
-            } else {
-                for (int j = 0; j < convexHullCoords.length; j++) {
-                    // Check if the y-coordinate is valid (not equal to Double.MAX_VALUE and not infinite)
-                    if (convexHullCoords[j].y == Double.MAX_VALUE || Double.isInfinite(convexHullCoords[j].y)) {
-                        continue; // Skip this point as it's not part of the hull
-                    }
-                    convexHullPoints.add(convexHullCoords[j]);
-                }
[…]
