### 2025-09-17 f266ab1
clean unused methods
