+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/AttenuationCnossos.java
@@ -36,45 +36,8 @@ import static org.noise_planet.noisemodelling.propagation.cnossos.PointPath.POIN
  */
 
 public class AttenuationCnossos {
-    private static double[] freq_lambda;
-    private static double[] aGlobal;
-
-    public static double[] getaGlobal() {
-        return aGlobal;
-    }
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AttenuationCnossos.class);
-
-    /**
-     * Eq 2.5.21: calculate the value of DeltaDif
-     * @param srpath
-     * @param data
-     * @return double list with the value of DeltaDif
-     */
-    public static double[] getDeltaDif(SegmentPath srpath, AttenuationParameters data) {
-        double[] DeltaDif = new double[data.getFrequencies().size()];
-        double cprime;
-
-        for (int idfreq = 0; idfreq < data.getFrequencies().size(); idfreq++) {
-            double Ch = 1; // Eq 2.5.21
-            if (srpath.eLength > 0.3) {
-                double gammaPart = pow((5 * freq_lambda[idfreq]) / srpath.eLength, 2);
-                cprime = (1. + gammaPart) / (1./3. + gammaPart); // Eq. 2.5.23
-            } else {
-                cprime = 1.;
-            }
-
-            //(7.11) NMP2008 P.32
-            double testForm = 40 / freq_lambda[idfreq]* cprime * srpath.delta;
-
-            double deltaDif = 0.;
-            if (testForm >= -2.) {
-                deltaDif = 10 * Ch * log10(Math.max(0, 3 + testForm));
-            }
-            DeltaDif[idfreq] = Math.max(0, deltaDif);
-        }
-        return DeltaDif;
-    }
 
     /**
      * Compute attenuation of sound energy by distance. Minimum distance is one
@@ -168,24 +131,6 @@ public class AttenuationCnossos {
     }
 
     /**
-     * Formulae Eq. 2.5.31 - Eq. 2.5.32
-     * @param aGround        Asol(O,R) or Asol(S,O) (sol mean ground)
-     * @param deltaDifPrim Δdif(S,R') if Asol(S,O) is given or Δdif(S', R) if Asol(O,R)
-     * @param deltaDif     Δdif(S, R)
-     * @return Δsol(S, O) if Asol(S,O) is given or Δsol(O,R) if Asol(O,R) is given
-     */
-    private static double getDeltaGround(double aGround, double deltaDifPrim, double deltaDif) {
-        double attArg = 1 + (pow(10, -aGround / 20) - 1) * pow(10, -(deltaDifPrim - deltaDif) / 20);
-        if(Double.isNaN(attArg)){
-            attArg = Double.MAX_VALUE;
-        }
-        else if (attArg < 0) {
-            attArg = 0;
-        }
-        return -20 * log10(attArg);
-    }
-
-    /**
      * Compute ARef
      * @param pathParameters
      * @param data
@@ -220,23 +165,7 @@ public class AttenuationCnossos {
         if (!(segmentPath.gPath == 0 && data.isgDisc())) {
             return getAGroundCore(pathParameters, segmentPath, data);
         } else {
-            double aGroundMin;
-            //For testing purpose
-            if(pathParameters.keepAbsorption) {
-                //Used to calculate value ignored like Cf
-                getAGroundCore(pathParameters, segmentPath, data);
-            }
-
-            if (pathParameters.isFavorable()) {
-                // The lower bound of Aground,F (calculated with unmodified heights) depends on the geometry of the path
-                if (segmentPath.testFormF <= 1) {
-                    aGroundMin = -3 * (1 - segmentPath.gm);
-                } else {
-                    aGroundMin = -3 * (1 - segmentPath.gm) * (1 + 2 * (1 - (1 / segmentPath.testFormF)));
-                }
[…]
