### 2025-09-17 ca86496
unit test TC28 ERROR_EPSILON_LOWEST now !!
