### 2025-09-18 700a7c7
TC28 issue ab plane, w, Cf , on left side, removing building not sufficant ? but Z ground profile is good ?
