+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -5476,29 +5476,23 @@ public class AttenuationComputeOutputCnossosTest {
                 new Coordinate(1022.31, 0.0));
 
         /* Table 348 */
-        double [][] segmentsMeanPlanes0 = new double[][]{
+        double [][] segmentsMeanPlanesDirect = new double[][]{
                 //  a     b     zs    zr      dp    Gp   Gp'
                 {0.0, 0.25, 3.75, 9.09, 169.37, 0.45, 0.48},
                 {0.0, 0.0, 8.0, 1.0, 10.34, 0.5, NaN}
         };
 
-        double [][] segmentsMeanPlanes1 = new double[][]{ // Right
+        double [][] segmentsMeanPlanesRight = new double[][]{ // Right
                 //  a     b     zs    zr      dp    Gp   Gp'
                 {0.0, 0.0, 4.0, 1.0, 1028.57, 0.5, 0.5}
         };
 
-        double [][] segmentsMeanPlanes2 = new double[][]{ // left
+        double [][] segmentsMeanPlanesLeft = new double[][]{ // left
                 //  a     b     zs    zr      dp    Gp   Gp'
                 {0.0, 0.68, 3.32, 1.12, 1022.31, 0.49, 0.49}
         };
 
         assertEquals(6, propDataOut.getPropagationPaths().size());
-
-        // CnossosPath pathLeft = propDataOut.getPropagationPaths().get(2);
-        // Error in CNOSSOS unit test, left diffraction is going over a building but not in their 3D view !
-        // Why the weird left path in homogeneous ? it is not explained.
-        // assertZProfil(expectedZProfileLeft, Arrays.asList(pathLeft.getSRSegment().getPoints2DGround()));
-        // assertPlanes(segmentsMeanPlanes2,propDataOut.getPropagationPaths().get(2).getSRSegment()); // if b = 0.68: -> z2 = 0.32. In Cnossos z2 = 1.32 if b = 0.68
 
         //Expected values
         //Path0 : vertical plane
@@ -5522,7 +5516,7 @@ public class AttenuationComputeOutputCnossosTest {
         assertZProfil(expectedZProfileOR, Arrays.asList(
                 cnossosPath.getSegmentList().get(cnossosPath.getSegmentList().size() - 1).getPoints2DGround()));
 
-        assertPlanes(segmentsMeanPlanes0,cnossosPath.getSegmentList());
+        assertPlanes(segmentsMeanPlanesDirect,cnossosPath.getSegmentList());
 
         double[] actualAlphaAtm = propDataOut.scene.defaultCnossosParameters.getAlpha_atmo();
         double[] actualAAtm = cnossosPath.aAtm;
@@ -5552,6 +5546,87 @@ public class AttenuationComputeOutputCnossosTest {
         assertDoubleArrayEquals("ABoundaryF - vertical plane", expectedABoundaryF, actualABoundaryF, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("L - vertical plane", expectedL, actualL, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("LA - vertical plane", expectedLA, actualLA, ERROR_EPSILON_LOWEST);
+
+        ///   ///////////////////////////////////////////////////////////////////////
+        // Test Right Path in homogeneous conditions
+
[…]
