+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -5408,6 +5408,8 @@ public class AttenuationComputeOutputCnossosTest {
     public void TC28() throws IOException {
         AttenuationComputeOutput propDataOut =  computeCnossosPath("TC28_Direct", "TC28_Right", "TC28_Right_Curved", "TC28_Left", "TC28_Left_Curved");
 
+        double[] LW_SOURCE = new double[]{150, 150, 150, 150, 150, 150, 150, 150};
+
         /* Table 346 */
         List<Coordinate> expectedZProfile = Arrays.asList(
                 new Coordinate(0.0, 0.0),
@@ -5522,7 +5524,7 @@ public class AttenuationComputeOutputCnossosTest {
         double[] actualAAtm = cnossosPath.aAtm;
         double[] actualADiv = cnossosPath.aDiv;
         double[] actualABoundaryH = cnossosPath.double_aBoundary;
-        double[] actualLH = addArray(cnossosPath.aGlobalRaw, new double[]{150,150,150,150,150,150,150,150});
+        double[] actualLH = addArray(cnossosPath.aGlobalRaw, LW_SOURCE);
         CnossosPath homogeneousPath = cnossosPath;
 
         assertDoubleArrayEquals("AlphaAtm - vertical plane", expectedAlphaAtm, actualAlphaAtm, ERROR_EPSILON_LOWEST);
@@ -5605,9 +5607,6 @@ public class AttenuationComputeOutputCnossosTest {
         double[] expectedLeftH_CF = {1117.47, 1192.83, 751.24, 141.43, 14.08, 2.56, 0.50, 0.10};
         double[] expectedLeftH_Aground = {-1.53, -1.53, -1.25, 15.70, 21.95, 13.29, 5.98, -0.45};
 
-        double[] expectedLeftF_W = {0.00, 0.00, 0.00, 0.01, 0.06, 0.33, 1.73, 8.47};
-        double[] expectedLeftF_CF = {1099.07, 1188.81, 818.96, 176.07, 16.94, 2.99, 0.58, 0.12};
-        double[] expectedLeftF_Aground = {-4.53, -4.53, -4.53, -4.53, -4.53, -4.53, -4.53, -4.53};
 
         // left non favorable path
         cnossosPath = propDataOut.getPropagationPaths().get(4);
@@ -5615,29 +5614,44 @@ public class AttenuationComputeOutputCnossosTest {
         assertEquals(CutProfile.PROFILE_TYPE.LEFT, cnossosPath.getCutProfile().getProfileType());
 
         assertZProfil(expectedZProfileLeft, Arrays.asList(cnossosPath.getSRSegment().getPoints2DGround()));
+        // TODO Weird plane, A B not at 0 (there is no DEM)
         //assertPlanes(segmentsMeanPlanesLeft, cnossosPath.getSRSegment());
 
-        assertDoubleArrayEquals("w (H) - lateral left", expectedLeftH_W, cnossosPath.groundAttenuation.w, ERROR_EPSILON_LOWEST);
-        assertDoubleArrayEquals("Cf (H) - lateral left", expectedLeftH_CF, cnossosPath.groundAttenuation.cf, ERROR_EPSILON_VERY_LOW);
-        assertDoubleArrayEquals("AGround (H) - lateral left", expectedLeftH_Aground, cnossosPath.groundAttenuation.aGround, ERROR_EPSILON_VERY_LOW);
+        assertDoubleArrayEquals("w (H) - lateral left", expectedLeftH_W, cnossosPath.groundAttenuation.w, ERROR_EPSILON_LOW);
[…]
