### 2025-09-19 dfa74f0
restore more tc test
