+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -141,305 +141,382 @@ public class AttenuationComputeOutputCnossosTest {
 
         return propDataOut;
     }
-//
-//    /**
-//     * Test TC01 -- Reflecting ground (G = 0)
-//     */
-//    @Test
-//    public void TC01() throws IOException {
-//
-//        AttenuationComputeOutput propDataOut = computeCnossosPath("TC01_Direct");
-//
-//        //Expected values
-//        double[][][] pts = new double[][][]{
-//                {{0.00, 1.00}, {194.16, 4.00}} //Path 1 : direct
-//        };
-//        double[][] gPaths = new double[][]{
-//                {0.0} //Path 1 : direct
-//        };
-//
-//        //Assertion
-//        assertPaths(pts, gPaths, propDataOut.getPropagationPaths());
-//
-//        //Expected values
-//        double[] expectedWH = new double[]{0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00};
-//        double[] expectedCfH = new double[]{194.16, 194.16, 194.16, 194.16, 194.16, 194.16, 194.16, 194.16};
-//        double[] expectedAGroundH = new double[]{-3.00, -3.00, -3.00, -3.00, -3.00, -3.00, -3.00, -3.00};
-//        double[] expectedWF = new double[]{0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00};
-//        double[] expectedCfF = new double[]{194.16, 194.16, 194.16, 194.16, 194.16, 194.16, 194.16, 194.16};
-//        double[] expectedAGroundF = new double[]{-4.36, -4.36, -4.36, -4.36, -4.36, -4.36, -4.36, -4.36};
-//
-//        double[] expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
-//        double[] expectedAAtm = new double[]{0.02, 0.08, 0.20, 0.37, 0.71, 1.88, 6.36, 22.70};
-//        double[] expectedADiv = new double[]{56.76, 56.76, 56.76, 56.76, 56.76, 56.76, 56.76, 56.76};
-//        double[] expectedABoundaryH = new double[]{-3.00, -3.00, -3.00, -3.00, -3.00, -3.00, -3.00, -3.00};
-//        double[] expectedABoundaryF = new double[]{-4.36, -4.36, -4.36, -4.36, -4.36, -4.36, -4.36, -4.36};
-//        double[] expectedLH = new double[]{39.21, 39.16, 39.03, 38.86, 38.53, 37.36, 32.87, 16.54};
-//        double[] expectedLF = new double[]{40.58, 40.52, 40.40, 40.23, 39.89, 38.72, 34.24, 17.90};
-//        double[] expectedL = new double[]{39.95, 39.89, 39.77, 39.60, 39.26, 38.09, 33.61, 17.27};
-//        double[] expectedLA = sumArray(expectedL,A_WEIGHTING);
-//
-//
-//        //Actual values
-//        double[] actualWH = propDataOut.getPropagationPaths().get(0).groundAttenuation.wH;
-//        double[] actualCfH = propDataOut.getPropagationPaths().get(0).groundAttenuation.cfH;
-//        double[] actualAGroundH = propDataOut.getPropagationPaths().get(0).groundAttenuation.aGroundH;
-//        double[] actualWF = propDataOut.getPropagationPaths().get(0).groundAttenuation.wF;
[…]
