### 2025-09-22 a50f985
restore tc05 and TC06
