### 2025-09-22 15db71a
Use linear interpolation for Z values on left and right diffraction path as Cnossos ignore the real path of the ray Update test case dump of vertical cuts
