+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -754,432 +754,475 @@ public class AttenuationComputeOutputCnossosTest {
 
     }
 
-//    /**
-//     * Test TC07 -- Flat ground with spatially varying acoustic properties and long barrier
-//     */
-//    @Test
-//    public void TC07() throws IOException {
-//
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut = computeCnossosPath("TC07_Direct");
-//
-//        //Expected values
-//
-//        /* Table 33 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.00, 0.00));
-//        expectedZProfile.add(new Coordinate(170.23, 0.00));
-//        expectedZProfile.add(new Coordinate(170.23, 6.00));
-//        expectedZProfile.add(new Coordinate(170.23, 0.00));
-//        expectedZProfile.add(new Coordinate(194.16, 0.00));
-//
-//        assertZProfil(expectedZProfile,
-//                Arrays.asList(cnossosPath.getSRSegment().getPoints2DGround()));
-//
-//        /* Table 34 */
-//        Coordinate expectedSPrime =new Coordinate(0.00,-1.00);
-//        Coordinate expectedRPrime =new Coordinate(194.16,-4.00);
-//
-//        assertMirrorPoint(expectedSPrime,
-//                expectedRPrime,cnossosPath.getSegmentList().get(0).sPrime,
-//                cnossosPath.getSegmentList().
-//                        get(cnossosPath.getSegmentList().size()-1).rPrime);
-//
-//
-//
-//        double[][] gPaths = new double[][]{
-//                {0.55, 0.20},{0.61,  NaN} //Path 1 : direct
-//        };
-//
-//        /* Table 35 */
-//        double [][] segmentsMeanPlanes = new double[][]{
-//                //  a     b    zs    zr      dp    Gp   Gp'
-//                {0.00, 0.00, 1.00, 6.00, 170.23, 0.55, 0.61},
-//                {0.00, 0.00, 6.00, 4.00, 023.93, 0.20,  NaN}
-//        };
-//
-//
-//        //Assertion
-//        assertPlanes(segmentsMeanPlanes, cnossosPath.getSegmentList());
-//
-//        //Expected values
-//        double[] expectedDeltaDiffSRH = new double[]{6.01, 6.96, 8.41, 10.36, 12.72, 15.37, 18.19, 21.10};
-//        double[] expectedAGroundSOH = new double[]{-1.16, -1.16, -1.16, -1.16, 1.45, -1.16, -1.16, -1.16};
-//        double[] expectedAGroundORH = new double[]{-2.40, -2.40, -2.40, -2.40, -2.40, -2.40, -2.40, -2.40};
-//        double[] expectedDeltaDiffSPrimeRH = new double[]{6.24, 7.32, 8.92, 11.00, 13.46, 16.16, 19.01, 21.94};
-//        double[] expectedDeltaDiffSRPrimeH = new double[]{12.54, 15.13, 17.94, 20.85, 23.80, 26.78, 29.78, 32.78};
-//        double[] expectedDeltaGroundSOH = new double[]{-1.13, -1.11, -1.09, -1.08, 1.32, -1.06, -1.06, -1.06};
[…]
