+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/CnossosPathBuilder.java
@@ -240,12 +240,12 @@ public class CnossosPathBuilder {
         srPath.setPoints2DGround(pts2DGround);
         srPath.dc = CGAlgorithms3D.distance(cutProfile.getReceiver().getCoordinate(),
                 cutProfile.getSource().getCoordinate());
-        CnossosPath pathParameters = new CnossosPath(cutProfile);
-        pathParameters.setFavorable(favorable);
-        pathParameters.setPointList(points);
-        pathParameters.setSegmentList(segments);
-        pathParameters.setSRSegment(srPath);
-        pathParameters.init(exactFrequencyArray.size());
+        CnossosPath cnossosPath = new CnossosPath(cutProfile);
+        cnossosPath.setFavorable(favorable);
+        cnossosPath.setPointList(points);
+        cnossosPath.setSegmentList(segments);
+        cnossosPath.setSRSegment(srPath);
+        cnossosPath.init(exactFrequencyArray.size());
         List<Coordinate> hullPts2D = pts2D;
         if(favorable) {
             // Compute the altered profile for favorable path
@@ -323,7 +323,7 @@ public class CnossosPathBuilder {
                         cutProfilePoints.get(i0).getCoordinate(), targetPosition);
                 points.get(0).orientation = emissionDirection;
                 // TODO what about favorable path with curved profile ?
-                pathParameters.raySourceReceiverDirectivity = emissionDirection;
+                cnossosPath.raySourceReceiverDirectivity = emissionDirection;
                 src = pts2D.get(i0);
             }
             // Add reflection/vertical edge diffraction points/segments between i0 i1
@@ -407,29 +407,29 @@ public class CnossosPathBuilder {
                 // Check for Rayleigh criterion for segments computation
                 LineSegment dSR = new LineSegment(firstPts2D, lastPts2D);
                 // Look for diffraction over edge on free field (frequency dependent)
-                computeRayleighDiff(srPath, cutProfile, pathParameters, dSR, rayleighSegments, rayleighPoints, pts2D,
+                computeRayleighDiff(srPath, cutProfile, cnossosPath, dSR, rayleighSegments, rayleighPoints, pts2D,
                         pts2DGround, cut2DGroundIndex, exactFrequencyArray);
             }
             if(rayleighSegments.isEmpty()) {
                 // We don't have a Rayleigh diffraction over DEM. Only direct SR path
                 if(segments.isEmpty()) {
-                    segments.add(pathParameters.getSRSegment());
+                    segments.add(cnossosPath.getSRSegment());
                 }
                 // Compute cumulated distance between the first diffraction and the last diffraction point
-                pathParameters.e = 0;
+                cnossosPath.e = 0;
                 List<PointPath> diffPoints = points.stream().filter(pointPath -> pointPath.type != REFL).collect(Collectors.toList());
                 for(int idPoint = 1; idPoint < diffPoints.size() - 2; idPoint++) {
-                    pathParameters.e += diffPoints.get(idPoint).coordinate.distance(diffPoints.get(idPoint+1).coordinate);
[…]

[…]
