### 2025-09-22 0fb0a82
maybe linear interpolation not good.. (should be capped to wall height)
