+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -1256,495 +1256,531 @@ public class AttenuationComputeOutputCnossosTest {
 //                .addTopographicLine(205, 75, 10, 225, 80, 0)
 //                .addTopographicLine(205, -5, 10, 225, -20, 0);
 //    }
-//
-//    /**
-//     * Test TC09 -- Ground with spatially varying heights and and acoustic properties and short barrier
-//     */
-//    @Test
-//    public void TC09() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC09_Direct", "TC09_Right", "TC09_Left");
-//
-//        assertEquals(3, propDataOut.getPropagationPaths().size());
-//
-//        //Expected values
-//
-//        /* Table 59 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.00, 0.00));
-//        expectedZProfile.add(new Coordinate(112.41, 0.00));
-//        expectedZProfile.add(new Coordinate(170.49, 8.74));
-//        expectedZProfile.add(new Coordinate(170.49, 16.63));
-//        expectedZProfile.add(new Coordinate(170.49, 8.74));
-//        expectedZProfile.add(new Coordinate(178.84, 10.00));
-//        expectedZProfile.add(new Coordinate(194.16, 10.00));
-//
-//        /* Table 61 */
-//        Coordinate expectedSPrime =new Coordinate(0.24,-4.92);
-//        Coordinate expectedRPrime =new Coordinate(194.48,6.59);
-//
-//        assertMirrorPoint(expectedSPrime,expectedRPrime,cnossosPath.
-//                getSegmentList().get(0).sPrime,cnossosPath.getSegmentList().
-//                get(cnossosPath.getSegmentList().size()-1).rPrime);
-//
-//
-//        /* Table 60 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b    zs    zr      dp    Gp   Gp'
-//                {0.04, -1.96, 2.96, 11.68, 170.98, 0.55, 0.76},
-//                {0.04, 1.94, 7.36, 3.71, 23.54, 0.20,  0.20}
-//        };
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a     b    zs    zr      dp    Gp   Gp'
-//                {0.06, -3.10, 4.09, 3.77, 221.62, 0.46, 0.49}
-//        };
-//        double [][] segmentsMeanPlanes2 = new double[][]{
-//                //  a     b    zs    zr      dp    Gp   Gp'
-//                {0.05, -2.82, 3.81, 6.23, 195.20, 0.51, 0.64}
-//        };
-//
-//        //Assertion
-//        assertZProfil(expectedZProfile, Arrays.asList(cnossosPath.getSRSegment()
-//                .getPoints2DGround()));
-//        assertPlanes(segmentsMeanPlanes0, cnossosPath.getSegmentList());
-//        assertPlanes(segmentsMeanPlanes1, propDataOut.getPropagationPaths().get(1).getSRSegment());
[…]
