### 2025-09-27 18015c5
restore original computation of Z of left and right path. On the curved path, first compute on the curved coordinate and skip the building if it is under the propagation ray, use the original coordinates if the building still intersects with the homogenous propgation plane
