+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/profilebuilder/BuildingIntersectionPathVisitor.java
@@ -119,8 +119,14 @@ public final class BuildingIntersectionPathVisitor implements ItemVisitor {
             }
             if(curved) {
                 // Adjust the altitude of the building roof points to be in the curved coordinate system
-                roofPoints = Arrays.asList(CurvedProfileGenerator.applyTransformation(p1, p2,
+                List<Coordinate> curvedRoofPoints = Arrays.asList(CurvedProfileGenerator.applyTransformation(p1, p2,
                         roofPoints.toArray(new Coordinate[0]), false));
+                // Create a cut of the building volume with the roof points z moved to the bottom following
+                // the curve coordinate system formulae
+                if(cutRoofPointsWithPlane(cutPlane, curvedRoofPoints).isEmpty()) {
+                    // The building roof is below the curved ray
+                    return;
+                }
             }
             // Create a cut of the building volume
             roofPoints = cutRoofPointsWithPlane(cutPlane, roofPoints);
@@ -155,8 +161,14 @@ public final class BuildingIntersectionPathVisitor implements ItemVisitor {
             List<Coordinate> roofPoints = Arrays.asList(extendedP0, extendedP1);
             if(curved) {
                 // Adjust the altitude of the building roof points to be in the curved coordinate system
-                roofPoints = Arrays.asList(CurvedProfileGenerator.applyTransformation(p1, p2,
+                List<Coordinate> curvedRoofPoints = Arrays.asList(CurvedProfileGenerator.applyTransformation(p1, p2,
                         roofPoints.toArray(new Coordinate[0]), false));
+                // Create a cut of the building volume with the roof points z moved to the bottom following
+                // the curve coordinate system formulae
+                if(cutRoofPointsWithPlane(cutPlane, curvedRoofPoints).isEmpty()) {
+                    // The building roof is below the curved ray
+                    return;
+                }
             }
             // Create a cut of the building volume
             roofPoints = cutRoofPointsWithPlane(cutPlane, roofPoints);
