### 2025-09-27 3d62a2e
forget removing block of code in curved condition
