### 2025-09-27 d2c2fc4
delta value is like homogenous if the cut profile is not direct.
Have to check why
