+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -2047,258 +2047,274 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(  new double[]{21.28,28.39,32.47,34.51,34.54,33.37,32.14,27.73},L, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * Test TC12 -- Flat ground with homogeneous acoustic properties and polygonal object – receiver at low height
-//     */
-//    @Test
-//    public void TC12() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC12_Direct", "TC12_Right", "TC12_Left");
-//
-//        /* Table 100 */
-//        List<Coordinate> expectedZProfile = Arrays.asList(
-//                new Coordinate(0.00, 0.00),
-//                new Coordinate(12.26, 0.00),
-//                new Coordinate(12.26, 10.00),
-//                new Coordinate(18.82, 10),
-//                new Coordinate(18.82, 0),
-//                new Coordinate(31.62, 0));
-//
-//        List<Coordinate> expectedZProfileSO = Arrays.asList(
-//                new Coordinate(0.00, 0.00),
-//                new Coordinate(12.26, 0.00));
-//
-//        List<Coordinate> expectedZProfileOnR = Arrays.asList(
-//                new Coordinate(18.82, 0),
-//                new Coordinate(31.62, 0));
-//
-//        /* Table 101 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b    zs    zr     dp    Gp   Gp'
-//                {0.00, 0.00, 1.00, 10.0, 12.26, 0.50, 0.50},
-//                {0.00, 0.00, 10.0, 6.00, 12.80, 0.50,  NaN}
-//        };
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a     b    zs    zr     dp    Gp    Gp'
-//                {0.00, 0.00, 1.00, 6.00, 32.11, 0.50, 0.50}
-//        };
-//        double [][] segmentsMeanPlanes2 = new double[][]{
-//                //  a     b    zs    zr     dp    Gp    Gp'
-//                {0.00, 0.00, 1.00, 6.00, 32.66, 0.50, 0.50}
-//        };
-//
-//        //Assertion
-//        assertEquals(3, propDataOut.getPropagationPaths().size());
-//
-//        CnossosPath directPath = propDataOut.getPropagationPaths().get(0);
-//        assertZProfil(expectedZProfile, Arrays.asList(directPath.getSRSegment().getPoints2DGround()));
-//        assertZProfil(expectedZProfileSO, Arrays.asList(directPath.getSegmentList().get(0).getPoints2DGround()));
-//        assertZProfil(expectedZProfileOnR, Arrays.asList(directPath.getSegmentList().
-//                get(directPath.getSegmentList().size() - 1).getPoints2DGround()));
-//
-//        assertPlanes(segmentsMeanPlanes0, propDataOut.getPropagationPaths().get(0).getSegmentList());
-//
[…]
