+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -2315,241 +2315,262 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(new double[]{21.81, 29.66, 34.31, 36.14, 35.57, 33.72, 31.12, 25.37},L, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * Test TC13 -- Ground with spatially varying heights and acoustic properties and polygonal
-//     * building
-//     */
-//    @Test
-//    public void TC13() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC13_Direct", "TC13_Right", "TC13_Left");
-//
-//
-//        //Expected values
-//
-//        /* Table 117 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.00, 0.00));
-//        expectedZProfile.add(new Coordinate(112.41, 0.00));
-//        expectedZProfile.add(new Coordinate(164.07, 7.8));
-//        expectedZProfile.add(new Coordinate(164.07, 30.00));
-//        expectedZProfile.add(new Coordinate(181.83, 30));
-//        expectedZProfile.add(new Coordinate(181.83, 10));
-//        expectedZProfile.add(new Coordinate(194.16, 10));
-//
-//        /* Table 118 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a      b    zs     zr      dp    Gp   Gp'
-//                {0.04, -1.68, 2.68, 25.86, 164.99, 0.71, 0.54},
-//                {0.00, 10.00, 20.0, 18.50, 12.33, 0.20,  NaN}
-//        };
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a      b    zs     zr      dp    Gp    Gp'
-//                {0.06, -2.99, 3.98, 19.83, 201.30, 0.61, 0.53}
-//        };
-//        double [][] segmentsMeanPlanes2 = new double[][]{
-//                //  a      b    zs     zr      dp    Gp    Gp'
-//                {0.05, -2.82, 3.82, 20.69, 196.29, 0.63, 0.54}
-//        };
-//
-//        //Assertion
-//        assertZProfil(expectedZProfile, Arrays.asList(propDataOut.getPropagationPaths().get(0)
-//                .getSRSegment().getPoints2DGround()));
-//        assertPlanes(segmentsMeanPlanes0, propDataOut.getPropagationPaths().get(0).getSegmentList());
-//        assertPlanes(segmentsMeanPlanes1, propDataOut.getPropagationPaths().get(1).getSRSegment());
-//        assertPlanes(segmentsMeanPlanes2, propDataOut.getPropagationPaths().get(2).getSRSegment());
-//
-//        //Expected values
-//        //Path0 : vertical plane
-//        double[] expectedDeltaDiffSRH = new double[]{9.76, 13.15, 17.15, 20.71, 23.88, 26.92, 29.93, 32.94};
-//        double[] expectedAGroundSOH = new double[]{-1.38, -1.38, -1.38, -1.38, -1.38, -1.38, -1.38, -1.38};
[…]
