+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -2378,7 +2378,7 @@ public class AttenuationComputeOutputCnossosTest {
         };
 
         //Assertion
-        assertZProfil(expectedZProfile, Arrays.asList(propDataOut.getPropagationPaths().get(0)
+        assertZProfil(expectedZProfile, Arrays.asList(cnossosPathDirectH
                 .getSRSegment().getPoints2DGround()));
         assertPlanes(segmentsMeanPlanes0, cnossosPathDirectH.getSegmentList());
         assertPlanes(segmentsMeanPlanes1, cnossosPathRightH.getSRSegment());
@@ -2571,251 +2571,271 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(  new double[]{5.14,12.29,16.39,18.47,18.31,15.97,9.72,-9.92},L, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * Test TC14 -- Flat ground with homogeneous acoustic properties and polygonal building –
-//     * receiver at large height
-//     */
-//    @Test
-//    public void TC14() throws IOException {
-//
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC14_Direct", "TC14_Right", "TC14_Left");
-//
-//
-//        //Expected values
-//
-//        /* Table 132 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.00, 0.00));
-//        expectedZProfile.add(new Coordinate(5.39, 0.00));
-//        expectedZProfile.add(new Coordinate(5.39, 10.00));
-//        expectedZProfile.add(new Coordinate(11.49, 10.0));
-//        expectedZProfile.add(new Coordinate(11.49, 0.0));
-//        expectedZProfile.add(new Coordinate(19.72, 0));
-//
-//        /* Table 133 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a       b    zs     zr    dp    Gp   Gp'
-//                {0.00,   0.00, 1.00, 10.00, 5.39, 0.20, 0.20},
-//                {-1.02, 17.11, 1.08, 18.23, 0.72, 0.11,  NaN} // Fix Cnossos document Zs is 1.08 not 0
-//        };
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a      b    zs     zr     dp    Gp    Gp'
-//                {-0.02, 1.13, 0.10, 22.32, 19.57, 0.18, 0.20} // Fix Cnossos document Zs is 0.1 not 0
-//        };
-//        double [][] segmentsMeanPlanes2 = new double[][]{
-//                //  a     b    zs     zr      dp    Gp    Gp'
-//                {0.00, 1.35, 0.32, 21.69, 22.08, 0.17, 0.20} // Fix Cnossos document Zs is 0.32 not 0
-//        };
-//
-//
-//        //Assertion
-//        // Wrong value of z1 in Cnossos document for the 3 paths
-//        assertZProfil(expectedZProfile, Arrays.asList(propDataOut.getPropagationPaths().get(0)
-//                .getSRSegment().getPoints2DGround()));
[…]
