+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -2836,226 +2836,250 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(  new double[]{25.61,34.06,39.39,42.04,41.86,39.42,35.26,27.57},L, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * Test TC15 -- Flat ground with homogeneous acoustic properties and four buildings
-//     */
-//    @Test
-//    public void TC15() throws IOException {
-//
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC15_Direct", "TC15_Right", "TC15_Left");
-//
-//        /* Table 148 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.00, 0.00));
-//        expectedZProfile.add(new Coordinate(5.02, 0.00));
-//        expectedZProfile.add(new Coordinate(5.02, 8.00));
-//        expectedZProfile.add(new Coordinate(15.07, 8.0));
-//        expectedZProfile.add(new Coordinate(15.08, 0.0));
-//        expectedZProfile.add(new Coordinate(24.81, 0.0));
-//        expectedZProfile.add(new Coordinate(24.81, 12.0));
-//        expectedZProfile.add(new Coordinate(30.15, 12.00));
-//        expectedZProfile.add(new Coordinate(30.15, 0.00));
-//        expectedZProfile.add(new Coordinate(37.19, 0.0));
-//        expectedZProfile.add(new Coordinate(37.19, 10.0));
-//        expectedZProfile.add(new Coordinate(41.52, 10.0));
-//        expectedZProfile.add(new Coordinate(41.52, 0.0));
-//        expectedZProfile.add(new Coordinate(50.25, 0.0));
-//
-//        /* Table 149 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b     zs    zr    dp    Gp   Gp'
-//                {0.00, 0.00,  1.00, 8.00, 5.02, 0.50, 0.50},
-//                {0.00, 0.00, 10.00, 5.00, 8.73, 0.50,  NaN}
-//        };
-//        double [][] segmentsMeanPlanes1 = new double[][]{ // right
-//                //  a      b    zs    zr     dp    Gp    Gp'
-//                {0.08, -1.19, 2.18, 2.01, 54.80, 0.46, 0.48}
-//        };
-//        double [][] segmentsMeanPlanes2 = new double[][]{ // left
-//                //  a     b    zs    zr     dp    Gp    Gp'
-//                {0.00, 0.00, 1.00, 5.00, 53.60, 0.50, 0.50}
-//        };
-//
-//
-//        //Assertion
-//        assertZProfil(expectedZProfile, Arrays.asList(cnossosPathDirectH.
-//                getSRSegment().getPoints2DGround()));
-//        assertPlanes(segmentsMeanPlanes0, cnossosPathDirectH.getSegmentList());
-//        assertPlanes(segmentsMeanPlanes2, propDataOut.getPropagationPaths().get(2).getSRSegment()); // left
-//        //assertPlanes(segmentsMeanPlanes1, propDataOut.getPropagationPaths().get(1).getSRSegment()); // right : error in value of b cnossos
-//
-//        //Expected values
-//        //Path0 : vertical plane
[…]
