+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -3080,192 +3080,206 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(new double[]{10.75,16.57,20.81,24.51,26.55,26.78,25.04,18.50},L, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * Reflecting barrier on ground with spatially varying heights and acoustic properties
-//     */
-//    @Test
-//    public void TC16() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC16_Direct", "TC16_Reflection");
-//
-//        /* Table 163 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.0, 0.0));
-//        expectedZProfile.add(new Coordinate(112.41, 0.0));
-//        expectedZProfile.add(new Coordinate(178.84, 10));
-//        expectedZProfile.add(new Coordinate(194.16, 10));
-//
-//        /* Table 169 */
-//        List<Coordinate> expectedZProfileReflection = new ArrayList<>();
-//        expectedZProfileReflection.add(new Coordinate(0.0, 0.0));
-//        expectedZProfileReflection.add(new Coordinate(117.12, 0.0));
-//        expectedZProfileReflection.add(new Coordinate(129.75, 1.82));
-//        expectedZProfileReflection.add(new Coordinate(129.75, 1.82));
-//        expectedZProfileReflection.add(new Coordinate(129.75, 1.82));
-//        expectedZProfileReflection.add(new Coordinate(183.01, 10));
-//        expectedZProfileReflection.add(new Coordinate(198.04, 10));
-//
-//        /* Table 166 */
-//        Coordinate expectedSPrime =new Coordinate(0.42,-6.64);
-//        Coordinate expectedRPrime =new Coordinate(194.84,1.70);
-//
-//        assertMirrorPoint(expectedSPrime,expectedRPrime,cnossosPathDirectH.getSRSegment().
-//                sPrime,cnossosPathDirectH.getSRSegment().rPrime);
-//
-//
-//        /* Table 165 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b     zs    zr      dp    Gp   Gp'
-//                {0.05, -2.83, 3.83, 6.16, 194.59, 0.54, 0.64}
-//        };
-//
-//        /* Table 171 */
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a     b     zs    zr      dp    Gp   Gp'
-//                {0.05, -2.80, 3.80, 6.37, 198.45, 0.51, 0.65}
-//        };
-//
-//        //Assertion
-//
-//        // Check SR direct line
-//        List<Coordinate> result = cnossosPathDirectH.getCutProfile().computePts2DGround();
-//        assertZProfil(expectedZProfile,result);
-//        assertEquals(2, propDataOut.getPropagationPaths().size());
-//        assertPlanes(segmentsMeanPlanes0, cnossosPathDirectH.getSRSegment());
-//
-//        // Check reflection path
-//        result = propDataOut.getPropagationPaths().get(1).getCutProfile().computePts2DGround();
[…]
