+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -3280,196 +3280,212 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(new double[]{13.62, 23.58, 30.71, 35.68, 38.27, 38.01, 32.98, 15.00}, LA, ERROR_EPSILON_LOWEST);
     }
 
-//    /**
-//     * TC17 - Reflecting barrier on ground with spatially varying heights and acoustic properties
-//     * reduced receiver height
-//     */
-//    @Test
-//    public void TC17() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC17_Direct", "TC17_Reflection");
-//
-//        // Expected Values
-//
-//        /* Table 178 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.0, 0.0));
-//        expectedZProfile.add(new Coordinate(112.41, 0.0));
-//        expectedZProfile.add(new Coordinate(178.84, 10));
-//        expectedZProfile.add(new Coordinate(194.16, 10));
-//
-//        //Assertion
-//        assertZProfil(expectedZProfile, Arrays.asList(cnossosPathDirectH.
-//                getSRSegment().getPoints2DGround()));
-//
-//        //Expected values
-//        //Path0 : vertical plane
-//        double[] expectedWH = new double[]{0.00, 0.00, 0.00, NaN, NaN, 0.53, 2.70, 12.70};
-//        double[] expectedCfH = new double[]{200.89, 217.45, 220.41, NaN, NaN, 1.88, 0.37, 0.08};
-//        double[] expectedAGroundH = new double[]{-1.32, -1.32, -1.32, NaN, NaN, -1.32, -1.32, -1.32};
-//        double[] expectedWF = new double[]{0.00, 0.00, 0.00, 0.01, 0.08, 0.42, 2.16, 10.35};
-//        double[] expectedCfF = new double[]{199.59, 214.11, 225.39, 131.90, 22.89, 2.42, 0.46, 0.10};
-//        double[] expectedAGroundF = new double[]{-1.32, -1.32, -1.29, -1.05, -1.32, -1.32, -1.32, -1.32};
-//
-//        double[] expectedDeltaDiffSR = new double[]{0., 0., 0., 3.16, 0.56, 0., 0., 0.};
-//        double[] expectedAGroundSO = new double[]{0., 0., 0., 2.74, -1.21, 0., 0., 0.};
-//        double[] expectedAGroundOR = new double[]{0., 0., 0., -2.40, -2.40, 0., 0., 0.};
-//        double[] expectedDeltaDiffSPrimeR = new double[]{0., 0., 0., 4.71, 4.65, 0., 0., 0.};
-//        double[] expectedDeltaDiffSRPrime = new double[]{0., 0., 0., 10.83, 13.26, 0., 0., 0.};
-//        double[] expectedDeltaGroundSO = new double[]{0., 0., 0., 2.23, -0.77, 0., 0., 0.};
-//        double[] expectedDeltaGroundOR = new double[]{0., 0., 0., -1.07, -0.62, 0., 0., 0.};
-//        double[] expectedADiff = new double[]{0., 0., 0., 4.31, -0.83, 0., 0., 0.};
-//
-//        double[] expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
[…]
