### 2025-09-29 5570f7a
TC18 ok, fixed, only Direct and Reflection use curved computed delta on favorable
