+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -3486,229 +3486,237 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(  new double[]{14.02, 23.84, 30.95, 33.86, 38.37, 38.27, 33.25, 15.28},LA, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * TC18 - Screening and reflecting barrier on ground with spatially varying heights and
-//     * acoustic properties
-//     */
-//    @Test
-//    public void TC18() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC18_Direct", "TC18_Reflection");
-//
-//        assertEquals(2, propDataOut.getPropagationPaths().size());
-//
-//        assertEquals(2, propDataOut.getPropagationPaths().size());
-//
-//        // Expected Values
-//
-//        /* Table 193  Z Profile SR */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.0, 0.0));
-//        expectedZProfile.add(new Coordinate(112.41, 0.0));
-//        expectedZProfile.add(new Coordinate(178.84, 10));
-//        expectedZProfile.add(new Coordinate(194.16, 10));
-//
-//        CutProfile cutProfile = cnossosPathDirectH.getCutProfile();
-//        List<Coordinate> result = cutProfile.computePts2DGround();
-//        assertZProfil(expectedZProfile, result);
-//
-//
-//        /* Table 194 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a      b    zs    zr      dp    Gp    Gp'
-//                {0.05, -2.83, 3.83, 4.16, 194.48, 0.51, 0.58}
-//        };
-//
-//        /* Table 197 */
-//        double [][] segmentsMeanPlanes1 = new double[][]{
-//                //  a      b    zs    zr      dp    Gp    Gp'
-//                {0.0, 0.0, 1.0, 12.0, 85.16, 0.7, 0.86},
-//                {0.11, -12.03, 14.16, 1.29, 112.14, 0.37,  NaN}
-//        };
-//
-//
-//
-//
-//        // S-R (not the rayleigh segments SO OR)
-//        assertPlanes(segmentsMeanPlanes0, cnossosPathDirectH.getSRSegment());
-//
-//        CnossosPath reflectionPath = propDataOut.getPropagationPaths().get(1);
-//        // Check reflexion mean planes
-//        assertPlanes(segmentsMeanPlanes1, reflectionPath.getSegmentList());
-//
-//        assertEquals(4,  reflectionPath.getPointList().size());
-//
-//        PointPath reflectionPoint = reflectionPath.getPointList().get(2);
-//        assertEquals(PointPath.POINT_TYPE.REFL, reflectionPoint.type);
-//
-//        assert3DCoordinateEquals("Reflection position TC18 ",
-//                new Coordinate(129.75,12), reflectionPoint.coordinate, DELTA_COORDS);
-//
-//        double[] L = addArray(propDataOut.getVerticesSoundLevel().get(0).levels, new double[]{93-26.2,93-16.1,93-8.6,93-3.2,93,93+1.2,93+1.0,93-1.1});
-//
-//        // Table 195
[…]
