+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -3717,325 +3717,340 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(  new double[]{11.69,21.77,28.93,32.71,36.83,36.83,32.12,13.66},L, ERROR_EPSILON_VERY_LOW);
     }
 
-//    /**
-//     * TC19 - Complex object and 2 barriers on ground with spatially varying heights and
-//     * acoustic properties
-//     */
-//    //TODO : the error is due to the left VDiff path which z-path seems to be false in the document
-//    @Test
-//    public void TC19() throws IOException {
-//        //Out and computation settings
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC19_Direct", "TC19_Right");
-//
-//        // Should be 3 but left path could not be defined
-//        // assertEquals(3, propDataOut.pathParameters.size());
-//
-//
-//        //Expected values
-//
-//        /* Table 208 */
-//        List<Coordinate> expectedZProfile = Arrays.asList(
-//                new Coordinate(0.00, 0.00),
-//                new Coordinate(100.55, 0.00),
-//                new Coordinate(100.55, 7.00),
-//                new Coordinate(108.60, 7.00),
-//                new Coordinate(108.60, 0.0),
-//                new Coordinate(110.61, 0.0),
-//                new Coordinate(145.34, 5.31),
-//                new Coordinate(145.34, 14.00),
-//                new Coordinate(145.34, 5.31),
-//                new Coordinate(171.65, 9.34),
-//                new Coordinate(171.66, 14.50),
-//                new Coordinate(171.66, 9.34),
-//                new Coordinate(175.97, 10),
-//                new Coordinate(191.05, 10));
-//
-//        /* Table 205 */
-//        List<Coordinate> expectedZProfileSO = Arrays.asList(
-//                new Coordinate(0.00, 0.00),
-//                new Coordinate(100.55, 0.00),
-//                new Coordinate(100.55, 7.00),
-//                new Coordinate(108.60, 7.00),
-//                new Coordinate(108.60, 0.0),
-//                new Coordinate(110.61, 0.0),
-//                new Coordinate(145.34, 5.31),
-//                new Coordinate(145.34, 14.00),
-//                new Coordinate(145.34, 5.31));
-//
-//        List<Coordinate> expectedZProfileOR = Arrays.asList(
-//                new Coordinate(171.66, 9.34),
-//                new Coordinate(175.97, 10),
-//                new Coordinate(191.05, 10));
-//
-//        List<Coordinate> expectedZProfileRight = Arrays.asList(
-//                new Coordinate(0, 0),
-//                new Coordinate(110.03, 0),
-//                new Coordinate(135.03, 3.85),
-//                new Coordinate(176.56, 10),
-//                new Coordinate(179.68, 10),
-//                new Coordinate(195.96, 10));
-//
-//        List<Coordinate> expectedZProfileLeft = Arrays.asList(
-//                new Coordinate(0, 0),
-//                new Coordinate(93.44, 0),
[…]
