+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -4051,380 +4051,385 @@ public class AttenuationComputeOutputCnossosTest {
 
     }
 
-//    /**
-//     * TC20 - Ground with spatially varying heights and acoustic properties
-//     */
-//    @Test
-//    public void TC20() throws IOException {
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC20_Direct");
-//
-//        //Expected values
-//
-//        /* Table 221 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.0, 0.0));
-//        expectedZProfile.add(new Coordinate(110.34, 0.0));
-//        expectedZProfile.add(new Coordinate(175.54, 10));
-//        expectedZProfile.add(new Coordinate(190.59, 10));
-//
-//        /* Table 230 S -> R TC21 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b     zs    zr      dp    Gp   Gp'
-//                {0.06, -2.84, 3.84, 6.12, 191.02, 0.50, 0.65}
-//        };
-//
-//        //Assertion
-//        assertZProfil(expectedZProfile, Arrays.asList(cnossosPathDirectH
-//                .getSRSegment().getPoints2DGround()));
-//        assertPlanes(segmentsMeanPlanes0, cnossosPathDirectH.getSegmentList());
-//
-//        //Expected values
-//        //Path0 : vertical plane
-//        double[] expectedWH = new double[]{0.00, 0.00, 0.00, 0.03, 0.15, 0.76, 3.76, 17.01};
-//        double[] expectedCfH = new double[]{199.65, 218.28, 203.92, 80.61, 9.46, 1.31, 0.27, 0.06};
-//        double[] expectedAGroundH = new double[]{-1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06};
-//        double[] expectedWF = new double[]{0.00, 0.00, 0.00, 0.01, 0.08, 0.42, 2.16, 10.35};
-//        double[] expectedCfF = new double[]{195.99, 210.11, 221.65, 131.04, 23.06, 2.42, 0.46, 0.10};
-//        double[] expectedAGroundF = new double[]{-1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06};
-//
-//        double[] expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
-//        double[] expectedAAtm = new double[]{0.02, 0.08, 0.20, 0.37, 0.70, 1.85, 6.26, 22.33};
-//        double[] expectedADiv = new double[]{56.62, 56.62, 56.62, 56.62, 56.62, 56.62, 56.62, 56.62};
-//        double[] expectedABoundaryH = new double[]{-1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06};
-//        double[] expectedABoundaryF = new double[]{-1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06, -1.06};
-//        double[] expectedLH = new double[]{37.41, 37.35, 37.23, 37.06, 36.73, 35.59, 31.17, 15.10};
-//        double[] expectedLF = new double[]{37.41, 37.35, 37.23, 37.06, 36.73, 35.59, 31.17, 15.10};
-//        double[] expectedL = new double[]{37.41, 37.35, 37.23, 37.06, 36.73, 35.59, 31.17, 15.10};
[…]
