+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -4017,8 +4017,8 @@ public class AttenuationComputeOutputCnossosTest {
         //actualAGroundF = proPath.groundAttenuation.aGround;
         //
         //actualAlphaAtm = propDataOut.scene.defaultCnossosParameters.getAlpha_atmo();
-        //actualAAtm = proPath.aAtm;
-        //actualADiv = proPath.aDiv;
+        //actualAAtm = cnossosPathDirectH.aAtm;
+        //actualADiv = cnossosPathDirectH.aDiv;
         //actualABoundaryH = proPath.double_aBoundary;
         //actualABoundaryF = proPath.double_aBoundary;
         //actualLH = addArray(proPath.aGlobal, SOUND_POWER_LEVELS);
@@ -4430,495 +4430,419 @@ public class AttenuationComputeOutputCnossosTest {
         assertArrayEquals(new double[]{10.44, 20.58, 27.78, 33.09, 35.84, 35.73, 30.91, 12.48}, LA, ERROR_EPSILON_LOWEST);
     }
 
-//    /**
-//     * TC22 - Building with receiver backside on ground with spatially varying heights and
-//     * acoustic properties
-//     */
-//    @Test
-//    public void TC22() throws IOException {
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC22_Direct", "TC22_Right", "TC22_Left");
-//
-//        // Expected Values
-//
-//        /* Table 248 */
-//        List<Coordinate> expectedZProfile = Arrays.asList(
-//                new Coordinate(0.0, 0.0),
-//                new Coordinate(110.39, 0.0),
-//                new Coordinate(169.60, 9.08),
-//                new Coordinate(169.61, 20),
-//                new Coordinate(177.63, 20),
-//                new Coordinate(177.64, 10),
-//                new Coordinate(177.68, 10));
-//
-//        /* Table 245 */
-//        List<Coordinate> expectedZProfileSO1 = Arrays.asList(
-//                new Coordinate(0.0, 0.0),
-//                new Coordinate(110.39, 0.0),
-//                new Coordinate(169.60, 9.08));
-//
-//
-//        List<Coordinate> expectedZProfileOnR = Arrays.asList(
-//                new Coordinate(177.64, 10),
-//                new Coordinate(177.68, 10));
-//
-//        List<Coordinate> expectedZProfileRight = Arrays.asList(
-//                new Coordinate(0, 0),
-//                new Coordinate(110.04, 0),
-//                new Coordinate(175.06, 10),
-//                new Coordinate(187.07, 10),
-//                new Coordinate(193.08, 10),
-//                new Coordinate(203.80, 10));
-//
-//        List<Coordinate> expectedZProfileLeft = Arrays.asList(
-//                new Coordinate(0, 0),
-//                new Coordinate(111.29, 0),
-//                new Coordinate(170.99, 9.08),
-//                new Coordinate(176.99, 10),
-//                new Coordinate(188.99, 10),
-//                new Coordinate(195.00, 10),
-//                new Coordinate(206.14, 10));
-//
-//        /* Table 249 */
[…]
