+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -4843,795 +4843,842 @@ public class AttenuationComputeOutputCnossosTest {
         assertDoubleArrayEquals("LA - vertical plane", expectedLA, actualLA, ERROR_EPSILON_LOWEST);
     }
 
-//    /**
-//     * TC24 – Two buildings behind an earth-berm on flat ground with homogeneous acoustic properties – receiver position modified
-//     */
-//    @Test
-//    public void TC24() throws IOException {
-//        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC24_Direct", "TC24_Reflection");
-//
-//        /* Table 279 */
-//        List<Coordinate> expectedZProfile = new ArrayList<>();
-//        expectedZProfile.add(new Coordinate(0.0, 0.0));
-//        expectedZProfile.add(new Coordinate(14.46, 0.0));
-//        expectedZProfile.add(new Coordinate(23.03, 5.0));
-//        expectedZProfile.add(new Coordinate(24.39, 5.0));
-//        expectedZProfile.add(new Coordinate(32.85, 0.0));
-//        expectedZProfile.add(new Coordinate(45.10, 0.0));
-//        expectedZProfile.add(new Coordinate(45.10, 6.0));
-//        expectedZProfile.add(new Coordinate(60.58, 6.0));
-//        expectedZProfile.add(new Coordinate(60.58, 0.0));
-//        expectedZProfile.add(new Coordinate(68.15, 0.0));
-//
-//
-//        /* Table 287 Z-Profile SO */
-//        List<Coordinate> expectedZProfileSO = new ArrayList<>();
-//        expectedZProfileSO.add(new Coordinate(0.0, 0.0));
-//        expectedZProfileSO.add(new Coordinate(14.13, 0.0));
-//        expectedZProfileSO.add(new Coordinate(22.51, 5.0));
-//
-//        List<Coordinate> expectedZProfileOR = new ArrayList<>();
-//        expectedZProfileOR.add(new Coordinate(22.51, 5.0));
-//        expectedZProfileOR.add(new Coordinate(23.84, 5.0));
-//        expectedZProfileOR.add(new Coordinate(32.13, 0.0));
-//        expectedZProfileOR.add(new Coordinate(43.53, 0.0));
-//        expectedZProfileOR.add(new Coordinate(70.74, 0.0));
-//
-//        List<Coordinate> result = cnossosPathDirectH.getCutProfile().computePts2DGround();
-//        assertZProfil(expectedZProfile,result);
-//
-//        /* Table 280 */
-//        double [][] segmentsMeanPlanes0 = new double[][]{
-//                //  a     b     zs    zr      dp    Gp   Gp'
-//                {0.18, -1.17, 2.13, 1.94, 23.37, 0.37, 0.07},
-//                {0.0, 0.0, 6.0, 4.0, 7.57, 0.00, NaN}
-//        };
-//        assertPlanes(segmentsMeanPlanes0,cnossosPathDirectH.getSegmentList());
-//
-//        //Expected values
-//        //Path0 : vertical plane
-//        double[] expectedDeltaDiffSRH = new double[]{10.18, 13.64, 16.95, 20.02, 23.02, 26.01, 29.00, 32.01};
[…]
