+++ noisemodelling-tutorial-01/src/main/java/org/noise_planet/nmtutorial01/GenerateReferenceDeviation.java
@@ -41,7 +41,7 @@ public class GenerateReferenceDeviation {
             AcousticIndicatorsFunctions.asOctaveBands(ProfileBuilder.DEFAULT_FREQUENCIES_THIRD_OCTAVE));
     private static final double[] SOUND_POWER_LEVELS = new double[]{93, 93, 93, 93, 93, 93, 93, 93};
     private static final double[] A_WEIGHTING = new double[]{-26.2, -16.1, -8.6, -3.2, 0.0, 1.2, 1.0, -1.1};
-
+    private static final String[] PATH_NAMES = new String[] {"Vertical Plane", "Left Lateral", "Right Lateral", "Reflection"};
     private static final double HUMIDITY = 70;
     private static final double TEMPERATURE = 10;
     private static final String CHECKED = "☑";
@@ -78,7 +78,7 @@ public class GenerateReferenceDeviation {
         String testCaseFileName = utName + ".json";
         try(InputStream inputStream = PathFinder.class.getResourceAsStream("test_cases/"+testCaseFileName)) {
             if(inputStream == null) {
-                throw new IOException("Document " + testCaseFileName + " not found");
+                return null;
             }
             ObjectMapper mapper = new ObjectMapper();
             return mapper.readValue(inputStream, CutProfile.class);
@@ -119,12 +119,14 @@ public class GenerateReferenceDeviation {
         PathFinder.ReceiverPointInfo lastReceiver = new PathFinder.ReceiverPointInfo(-1,-1,new Coordinate());
         for (String utName : utNames) {
             CutProfile cutProfile = loadCutProfile(utName);
-            attenuationVisitor.onNewCutPlane(cutProfile);
-            if(lastReceiver.receiverPk != -1 && cutProfile.getReceiver().receiverPk != lastReceiver.receiverPk) {
-                // merge attenuation per receiver
-                attenuationVisitor.finalizeReceiver(new PathFinder.ReceiverPointInfo(cutProfile.getReceiver()));
+            if(cutProfile != null) {
+                attenuationVisitor.onNewCutPlane(cutProfile);
+                if (lastReceiver.receiverPk != -1 && cutProfile.getReceiver().receiverPk != lastReceiver.receiverPk) {
+                    // merge attenuation per receiver
+                    attenuationVisitor.finalizeReceiver(new PathFinder.ReceiverPointInfo(cutProfile.getReceiver()));
+                }
+                lastReceiver = new PathFinder.ReceiverPointInfo(cutProfile.getReceiver());
             }
-            lastReceiver = new PathFinder.ReceiverPointInfo(cutProfile.getReceiver());
         }
         // merge attenuation per receiver
         attenuationVisitor.finalizeReceiver(lastReceiver);
@@ -202,8 +204,20 @@ public class GenerateReferenceDeviation {
                 utName));
     }
 
-    private static void addUTDeviationDetails(String utName, StringBuilder sb, JsonNode expectedValues, CnossosPath actual, double[] powerLevel) {
-        double[] actualLH = addArray(actual.aGlobalH, powerLevel);
+    private static CnossosPath fetchPath(List<CnossosPath> paths, CutProfile.PROFILE_TYPE profileType, boolean favorable) {
+        for(CnossosPath path : paths) {
+            if(path.getCutProfile().profileType == profileType && path.isFavorable() == favorable) {
+                return path;
+            }
+        }
+        return null;
+    }
+
+    private static void addUTDeviationDetails(CutProfile.PROFILE_TYPE profileType, StringBuilder sb, JsonNode expectedValues, List<CnossosPath> paths, double[] powerLevel) {
+        CnossosPath homogenous = fetchPath(paths, profileType, false);
+        String utName = PATH_NAMES[profileType.ordinal()];
[…]
