### 2025-10-03 2901266
In ray table, give information about favorable or not favorable cnossosPath as a column
