+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/NoiseMapWriter.java
@@ -121,6 +121,7 @@ public class NoiseMapWriter implements Callable<Boolean> {
         if(exportPeriod) {
             query.append(", PERIOD");
         }
+        query.append(", FAVORABLE");
         query.append(") VALUES (?, ?, ?");
         if(databaseParameters.exportCnossosPathWithAttenuation) {
             query.append(", ?");
@@ -131,6 +132,7 @@ public class NoiseMapWriter implements Callable<Boolean> {
         if(exportPeriod) {
             query.append(", ?");
         }
+        query.append(", ?");
         query.append(");");
         // PK, GEOM, ID_RECEIVER, ID_SOURCE
         PreparedStatement ps;
@@ -165,6 +167,7 @@ public class NoiseMapWriter implements Callable<Boolean> {
             if(exportPeriod) {
                 ps.setString(parameterIndex++, row.getTimePeriod());
             }
+            ps.setBoolean(parameterIndex++, row.isFavorable());
             ps.addBatch();
             batchSize++;
             if (batchSize >= BATCH_MAX_SIZE) {
@@ -379,6 +382,7 @@ public class NoiseMapWriter implements Callable<Boolean> {
             if(exportPeriod) {
                 sb.append(", PERIOD VARCHAR");
             }
+            sb.append(", FAVORABLE BOOLEAN");
             sb.append(");");
             processQuery(sb.toString());
         }
