### 2025-10-03 8940879
fix type, favorable -> favourable
