+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/AttenuationCnossos.java
@@ -17,7 +17,6 @@ import org.noise_planet.noisemodelling.propagation.SceneWithAttenuation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -92,7 +91,7 @@ public class AttenuationCnossos {
             double cf = dp * (1 + 3 * w * dp * pow(Math.E, -sqrt(w * dp))) / (1 + w * dp);
             //NF S 31-113 page 41 A sol
 
-            if (pathParameters.isFavorable()) {
+            if (pathParameters.isFavourable()) {
                 if (data.isPrime2520()) {
                     if (segmentPath.testFormF <= 1) {
                         aGroundMin = -3 * (1 - segmentPath.gm);
@@ -181,7 +180,7 @@ public class AttenuationCnossos {
     private static double getAGroundMin(SegmentPath segmentPath, CnossosPath pathParameters) {
         double aGroundMin;
 
-        if (pathParameters.isFavorable()) {
+        if (pathParameters.isFavourable()) {
             // The lower bound of Aground,F (calculated with unmodified heights) depends on the geometry of the path
             if (segmentPath.testFormF <= 1) {
                 aGroundMin = -3 * (1 - segmentPath.gm);
@@ -257,7 +256,7 @@ public class AttenuationCnossos {
                             (pp.type.equals(DIFH_RCRIT) && isValidRCriterion ))
                     .findFirst()
                     .orElse(null);
-            aGround[i] = path.isFavorable() ?
+            aGround[i] = path.isFavourable() ?
                     aGroundF(path, path.getSRSegment(), data, i) :
                     aGroundH(path, path.getSRSegment(), data, i);
             if(path.groundAttenuation != null && path.groundAttenuation.aGround != null) {
@@ -317,7 +316,7 @@ public class AttenuationCnossos {
                 //Get the point on the top of the obstacle
                 Coordinate p = new Coordinate(pointPath.coordinate.x, pointPath.obstacleZ);
                 double ch = 1.;
-                if (reflect.isFavorable()) {
+                if (reflect.isFavourable()) {
                     double SP = s.distance(p);
                     double PR = p.distance(r);
                     double gamma = 2 * max(1000, 8 * SR);
@@ -406,8 +405,8 @@ public class AttenuationCnossos {
         testForm = 40/lambda*cSecond*_delta;
         double deltaDiffSRPrime = testForm>=-2 ? 10*ch*log10(3+testForm) : 0;
 
-        double aGroundSO = proPathParameters.isFavorable() ? aGroundF(proPathParameters, first, data, frequencyIndex) : aGroundH(proPathParameters, first, data, frequencyIndex);
-        double aGroundOR = proPathParameters.isFavorable() ? aGroundF(proPathParameters, last, data, frequencyIndex, true) : aGroundH(proPathParameters, last, data, frequencyIndex, true);
+        double aGroundSO = proPathParameters.isFavourable() ? aGroundF(proPathParameters, first, data, frequencyIndex) : aGroundH(proPathParameters, first, data, frequencyIndex);
+        double aGroundOR = proPathParameters.isFavourable() ? aGroundF(proPathParameters, last, data, frequencyIndex, true) : aGroundH(proPathParameters, last, data, frequencyIndex, true);
 
         //If the source or the receiver are under the mean plane, change the computation of deltaDffSR and deltaGround
         double deltaGroundSO = -20*log10(1+(pow(10, -aGroundSO/20)-1)*pow(10, -(deltaDiffSPrimeR-deltaDiffSR)/20));
@@ -464,7 +463,7 @@ public class AttenuationCnossos {
[…]
