### 2025-10-08 bcaf404
add joss paper
