+++ paper/paper.md
@@ -1,0 +1,155 @@
+---
+title: 'NoiseModelling: an open-source software for outdoor noise mapping'
+tags:
+  - Acoustics
+  - JAVA
+  - Noise emission
+  - Noise propagation
+  - Noise mapping
+  - Road
+  - Railway
+authors:
+  - name: Pierre Aumond # code contributor, NoiseModelling technical comitee
+    orcid: 0000-0001-8834-4251
+    affiliation: 1
+  - name: Nicolas Fortin # code contributor, NoiseModelling technical comitee, noise-planet steering comitee
+    affiliation: 1
+  - name: Valentin Le Bescond # code contributor
+    orcid: 0000-0001-5287-9668
+    affiliation: 1
+    corresponding: true
+  - name: Gwendall Petit # code contributor, NoiseModelling technical comitee, noise-planet steering comitee
+    orcid: 0000-0002-4750-9600
+    affiliation: 1
+  - name: Judicaël Picaut # NoiseModelling technical comitee, noise-planet steering comitee
+    orcid: 0000-0002-0586-2414
+    affiliation: 1
+  - name: Erwan Bocher # NoiseModelling technical comitee, noise-planet steering comitee
+    orcid: 0000-0002-4936-7079
+    affiliation: 2
+    
+affiliations:
+ - name:  Univ Gustave Eiffel, CEREMA, UMRAE, F-44344 Bouguenais, France
+   index: 1
+ - name: CNRS, Lab-STICC laboratory UMR 6285, Vannes, France
+   index: 2
+date: 11 October 2024
+bibliography: paper.bib
+---
+
+![NoiseModelling Header](assets/header.png "NoiseModelling Header")
+
+# Summary
+
+NoiseModelling is a library capable of producing noise maps at a variety of scales, from neighbourhood and metropolitan level to large infrastructure covering entire countries.
+
+The tool integrates the CNOSSOS-EU road and rail traffic noise emission and CNOSSOS-EU propagation models [@noauthor_commission_2015]. 
+
+It is closely associated with the H2GIS [@bocher_h2gis_2015] and PostGIS [@mckenna_postgis_2022] spatial databases in order to handle spatial features and to facilitate cross-analysis with other spatial data.
+
+The NoiseModelling java library can be used independently or controlled through a graphical user interface.
+
+It is freely available for use by researchers, educational institutions and private companies.
+
+# Statement of need
+
+Noise has significant impacts on public health, confort and biodiversity. Prolonged exposure to high noise levels can lead to health issues such as stress, sleep disturbances, and cardiovascular diseases [@chen_environmental_2023]. Additionally, noise can disrupt ecosystems, affecting wildlife and altering the natural behaviors of species [@sordello_evidence_2020]. Including noise in environmental assessments is essential to better prevent its negative effects and supports the development of more sustainable planning strategies.
+
+It is therefore crucial to simulate and predict noise propagation in various environments. This provides a quantitative understanding of sound sources, enabling, for example, the assessment of noise mitigation measures and regulatory compliance.
+
+Since 2019 and Directive (EU) 2015/996 [@noauthor_commission_2015], the CNOSSOS-UE emission and propagation method must be used to produce strategic noise maps as described in Directive 2002/49/EC [@noauthor_commission_2002].
+
+There is software that fully implements the CNOSSOS-EU method, but this is commercial closed-source software.
+
+The NoiseModelling library has been developed jointly by Gustave Eiffel University and CNRS in an effort to provide the engineering and scientific communities around the world with a fully free and open-source software to compute and explore noise maps and experiment with the underlying models.
+
+To date, NoiseModelling has been applied in a variety of research projects and practical scenarios, including, but not limited to:
+
+- Standard noise maps of the main transport infrastructures [@aumond_strategic_2023].
[…]
