### 2025-10-08 248c037
take only account of buildings intersecting direct receiver/source ray for left and right diffraction.
