### 2025-10-08 61ca679
issue with the crossing of buildings, not far from solution
