### 2025-10-09 33424a2
go through ignored buildings when doing left and right diffraction
