+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -3722,7 +3722,7 @@ public class AttenuationComputeOutputCnossosTest {
     @Test
     public void TC19() throws IOException {
         //Out and computation settings
-        AttenuationComputeOutput propDataOut =  computeCnossosPath("TC19_Direct", "TC19_Right",
+        AttenuationComputeOutput propDataOut = computeCnossosPath("TC19_Direct", "TC19_Right",
                 "TC19_Right_Curved", "TC19_Left", "TC19_Left_Curved");
 
         assertEquals(6, propDataOut.getPropagationPaths().size());
@@ -3807,16 +3807,16 @@ public class AttenuationComputeOutputCnossosTest {
                 new Coordinate(192.38, 10));
 
         /* Table 209 */
-        double [][] segmentsMeanPlanes0 = new double[][]{
+        double[][] segmentsMeanPlanes0 = new double[][]{
                 //  a      b     zs     zr      dp    Gp   Gp'
-                {0.03, -1.09,  2.09, 10.89, 145.65, 0.57, 0.78},
-                {0.02,  6.42,  4.76,  3.89,  19.38, 0.20,  NaN}
+                {0.03, -1.09, 2.09, 10.89, 145.65, 0.57, 0.78},
+                {0.02, 6.42, 4.76, 3.89, 19.38, 0.20, NaN}
         };
-        double [][] segmentsMeanPlanes1 = new double[][]{
+        double[][] segmentsMeanPlanes1 = new double[][]{
                 //  a      b    zs    zr      dp    Gp    Gp'
                 {0.06, -2.92, 3.92, 5.66, 196.38, 0.50, 0.62}
         };
-        double [][] segmentsMeanPlanes2 = new double[][]{
+        double[][] segmentsMeanPlanes2 = new double[][]{
                 //  a      b    zs    zr      dp    Gp    Gp'
                 {0.06, -2.01, 3.00, 5.00, 192.81, 0.46, 0.55}
         };
@@ -3995,29 +3995,28 @@ public class AttenuationComputeOutputCnossosTest {
         expectedAlphaAtm = new double[]{0.12, 0.41, 1.04, 1.93, 3.66, 9.66, 32.77, 116.88};
         expectedAAtm = new double[]{0.02, 0.08, 0.20, 0.37, 0.71, 1.86, 6.32, 22.54};
         expectedADiv = new double[]{56.64, 56.64, 56.64, 56.64, 56.64, 56.64, 56.64, 56.64};
-        expectedABoundaryH = new double[]{-1.35, -1.35, -1.35, 1.32, -1.35, -1.35, -1.35, -1.35};
-        expectedABoundaryF = new double[]{-1.35, -1.35, -1.35, -1.35, -1.35, -1.35, -1.35, -1.35};
         expectedLH = new double[]{26.60, 24.10, 21.27, 15.57, 14.99, 10.86, 3.41, -15.80};
         expectedLF = new double[]{26.60, 24.10, 21.27, 18.25, 14.99, 10.86, 3.41, -15.80};
-        expectedLA = new double[]{0.40, 8.00, 12.67, 13.91, 14.99, 12.06, 4.41, -9.38};
+        expectedLA = new double[]{0.40, 8.00, 12.67, 13.91, 14.99, 12.06, 4.41, -16.90};
+        expectedADiffH = new double[]{11.08, 13.52, 16.24, 19.09, 22.02, 24.99, 27.98, 30.98};
+        expectedADiffF = new double[]{11.08, 13.52, 16.24, 19.09, 22.02, 24.99, 27.98, 30.98};
 
[…]
