+++ Docs/Cnossos_Report.rst
@@ -11,7 +11,7 @@ It is important to note that the ISO standard provides recommendations rather th
 
 During our analysis, we identified several issues within the standard that hinder a complete and reliable comparison. Notably, we observed inconsistencies between 2D and 3D visualizations, preventing us from achieving a coherent assessment. Additionally, discrepancies exist between the geometric description of the scene and the corresponding acoustic response, raising concerns about the accuracy and reliability of the standard’s methodology.
 
-Furthermore, with respect to favorable rays, our findings indicate a different implementation of CNOSSOS compared to the approach suggested by the standard. This divergence may have implications for the interpretation and reproducibility of results, necessitating further clarification and alignment.
+Furthermore, with respect to favourable rays, our findings indicate a different implementation of CNOSSOS compared to the approach suggested by the standard. This divergence may have implications for the interpretation and reproducibility of results, necessitating further clarification and alignment.
 
 
 Conformity table
@@ -19,7 +19,7 @@ Conformity table
 | Conform
 
 * Do not the deviate more than ±0,1 dB
-* Percentage of conformity : 86% (24/28)
+* Percentage of conformity : 90% (25/28)
 
 | NLD Conform
 
@@ -87,7 +87,7 @@ Conformity table
    * - TC11
      - ☑
      - ☑
-     - 0.02 dB @ 125 Hz
+     - 0.01 dB @ 125 Hz
      - `TC11`_
    * - TC12
      - ☑
@@ -125,9 +125,9 @@ Conformity table
      - 0.02 dB @ 500 Hz
      - `TC18`_
    * - TC19
-     - □
      - ☑
-     - 0.77 dB @ 250 Hz
+     - ☑
+     - 0.02 dB @ 63 Hz
      - `TC19`_
    * - TC20
      - ☑
@@ -526,11 +526,11 @@ Right Lateral
      - Maximum Difference
      - Frequency
    * - Lʜ
-     - 0.05 dB
-     - 250
+     - 0.03 dB
+     - 500
    * - Lꜰ
-     - 0.05 dB
-     - 250
+     - 0.03 dB
+     - 500
 
 
 Left Lateral 
@@ -544,11 +544,11 @@ Left Lateral
      - Maximum Difference
      - Frequency
    * - Lʜ
-     - 0.05 dB
-     - 250
+     - 0.03 dB
+     - 500
    * - Lꜰ
-     - 0.05 dB
-     - 250
+     - 0.03 dB
+     - 500
 
 TC12
 ^^^^
@@ -946,10 +946,10 @@ Left Lateral
      - Maximum Difference
      - Frequency
    * - Lʜ
-     - 6.46 dB
-     - 8000
+     - 0.04 dB
+     - 500
    * - Lꜰ
-     - 6.46 dB
+     - 0.01 dB
      - 8000
 
 TC20
