### 2025-10-10 7b26fcb
reflection, curved path over beam
