+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -4790,8 +4790,6 @@ public class AttenuationComputeOutputCnossosTest {
         double[] expectedL = new double[]{38.90, 37.17, 36.26, 34.68, 31.42, 27.54, 22.75, 15.02};
         double[] expectedLA = new double[]{12.70, 21.07, 27.66, 31.48, 31.42, 28.74, 23.75, 13.92};
 
-        CnossosPath proPath = cnossosPathDirectH;
-
         double[] actualDeltaDiffSRH = cnossosPathDirectH.aBoundary.deltaDiffSR;
         double[] actualAGroundSOH = cnossosPathDirectH.aBoundary.aGroundSO;
         double[] actualAGroundORH = cnossosPathDirectH.aBoundary.aGroundOR;
@@ -4913,6 +4911,22 @@ public class AttenuationComputeOutputCnossosTest {
                 {0.0, 0.0, 6.0, 4.0, 7.57, 0.00, NaN}
         };
         assertPlanes(segmentsMeanPlanes0,cnossosPathDirectH.getSegmentList());
+
+        /* Table 290 */
+        double[][] segmentsMeanPlanesRefH = new double[][]{
+                //  a     b     zs    zr      dp    Gp   Gp'
+                {0.19, -1.17, 2.13, 1.94, 22.86, 0.37, 0.07},
+                {-0.05, 2.8, 3.38, 4.72, 46.90, 0.18, NaN}
+        };
+        assertPlanes(segmentsMeanPlanesRefH,cnossosPathReflectionH.getSegmentList());
+
+        /* Table 291 */
+        double[][] segmentsMeanPlanesRefF = new double[][]{
+                //  a     b     zs    zr      dp    Gp   Gp'
+                {0.19, -1.17, 2.13, 1.94, 22.86, 0.37, 0.07},
+                {-0.06, 3.41, 2.96, 4.90, 48.20, 0.20, NaN}
+        };
+        assertPlanes(segmentsMeanPlanesRefF,cnossosPathReflectionF.getSegmentList());
 
         //Expected values
         //Path0 : vertical plane
@@ -5066,6 +5080,9 @@ public class AttenuationComputeOutputCnossosTest {
         assertDoubleArrayEquals("DeltaGroundORH reflection plane", expectedDeltaGroundORH, actualDeltaGroundORH, ERROR_EPSILON_VERY_LOW);
         assertDoubleArrayEquals("actualADiffH reflection plane", expectedADiffH, actualADiffH, ERROR_EPSILON_VERY_LOW);
 
+        assertEquals(0.29, cnossosPathReflectionF.delta, ERROR_EPSILON_LOWEST);
+        // curved ray goes above beam in favorable
+        assertEquals(0, cnossosPathReflectionF.e, ERROR_EPSILON_LOWEST);
         assertDoubleArrayEquals("DeltaDiffSRF reflection plane", expectedDeltaDiffSRF, actualDeltaDiffSRF, ERROR_EPSILON_VERY_LOW);
         assertDoubleArrayEquals("AGroundSOF reflection plane", expectedAGroundSOF, actualAGroundSOF, ERROR_EPSILON_VERY_LOW);
         assertDoubleArrayEquals("AGroundORF reflection plane", expectedAGroundORF, actualAGroundORF, ERROR_EPSILON_VERY_LOW);
@@ -5073,7 +5090,7 @@ public class AttenuationComputeOutputCnossosTest {
[…]
