### 2025-10-10 f3f2efc
apply curved path for hull on reflection too.
