### 2025-10-14 c21c2e3
curved path formula should using the distance portion of the current point ignoring the height (so 2d distance)
