### 2025-10-14 b4b5512
d should be in 2d too
