### 2025-10-14 be7a590
curved path formula fix (#842)
* curved path formula should using the distance portion of the current point ignoring the height (so 2d distance)

* d should be in 2d too
