### 2025-10-20 fc99a7f
refactor: update matsim scripts to use the new dynamic tools (#845)
* refactor: update matsim scripts to use the new dynamic tools

* update docs

* update matsim tutorial
