+++ Docs/Matsim_Tutorial.rst
@@ -31,6 +31,7 @@ The data folder should contain the following files :
 
 - ``nantes_mini.osm.pbf`` : the Openstreetmap data of the area. We'll use it to import buildings into NoiseModelling.
 - ``detailed_network.csv`` : A file containing the 'true' geometries of the road segments (called "links" in MATSim)
+- ``output_allVehicles.xml.gz`` : A file containing the various vehicles used by the agents in the simulation.
 - ``output_events.xml.gz`` : A file containing the list of MATSim events from the simulation.
 - ``output_facilities.xml.gz`` : A file containing the list of facilities, the agent's activity locations.
 - ``output_network.xml.gz`` : A file containing the MATSim road network, a list of nodes and links.
@@ -79,7 +80,7 @@ You can explore the other options by reading their descriptions. Here we are goi
 - Network CSV file: ``/path/to/your/scenario_matsim/detailed_network.csv``
 - Export additional traffic data ? : ``true``
 - Path of the Matsim output folder: ``/path/to/your/scenario_matsim``
-- populationFactor: ``0.001``
+- populationFactor: ``0.01``
 - The size of time bins in seconds: ``900``
 - outTableName: "" (not set, use default)
 - Skip unused links ?: ``true``
@@ -154,71 +155,32 @@ You can inspect the results to see where each activity is placed now.
    :align: center
 
 
-Step 5 : Calculate Noise Attenuation Matrix
+Step 5 : Calculate Noise Receiver Levels
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-In this step, we want to calculate and store the noise propagation part of NoiseModelling.
-We need this because we actually have several power spectrum for every road segment, one for every timestep of 15min.
-In the end we want to have a noise map every 15 minutes (96 maps in total). If we do that directly, by calling something like ``Noise_level_from_source`` WPS bloc 96 times, we would be calculating the exact same noise propagation 96 times.
+In this step, we want to calculate a noise level for every receiver, every 15 minutes (96 maps in total).
 
-So the process is as follows :
-
-1. we use the ``MATSIM_ROADS`` table as a SOURCE table. It only contains geometries, without any power levels, so the ``Noise_level_from_source`` bloc will compute acoustic attenuation by virtually attributing a power level of 0dB for every road link.
-2. We use the ``Noise_level_from_source`` WPS bloc and setting the ``confExportSourceId`` input paramter.
-
-The ``confExportSourceId`` parameter will actually ouput, for every recevier, the list of sources that contribute to the resulting levels, with the source-receiver noise attenuation.
-
-We'll then use this attenuation matrix in the next steps to get the 96 noise maps.
-
-Calculate the attenuation matrix
-----------------------------------
+We'll use the ``Noise_level_from_source`` WPS bloc. When the using the ``Sources emission table name``, the WPS bloc will automatically take into account the time bins defined in the emission table.
+For more details about the different parameters, browse the NoiseModelling general documentation.
 
 Let's use the previously generated table to launch our propagation calculation.
-
-As explained before, we'll use the ``Noise_level_from_source`` WPS bloc with the ``Separate receiver level by source identifier`` parameter enabled.
-For more details about the different parameters, browse the NoiseModelling general documentation.
 
 The parameters we will use are the following :
 
 - Buildings table name: ``BUILDINGS``
 - Receivers table name: ``ACTIVITY_RECEIVERS``
[…]
