### 2025-11-04 17ab6c9
Fix TC26 Reflexion without fav condition.
