### 2025-11-05 f222144
Fix TC26 to have all tests passing
