+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/CnossosPathBuilder.java
@@ -225,14 +225,9 @@ public class CnossosPathBuilder {
         List<SegmentPath> segments = new ArrayList<>();
         List<PointPath> points = new ArrayList<>();
 
-        // Get transformed cutPoints if favourable (without modifying the original cutProfile)
-        List<CutPoint> cutProfilePoints = new ArrayList<>();
-
-        List<Coordinate> pts2D = cutProfile.computePts2D(favourable, cutProfilePoints);
-        // If not favourable, use original cutPoints
-        if (!favourable || cutProfilePoints.isEmpty()) {
-            cutProfilePoints = cutProfile.cutPoints;
-        }
+        // Use original cutPoints for all calculations (no transformation)
+        final List<CutPoint> cutProfilePoints = cutProfile.cutPoints;
+        List<Coordinate> pts2D = cutProfile.computePts2D();
         if(pts2D.size() != cutProfilePoints.size()) {
             throw new IllegalArgumentException("The two arrays size should be the same");
         }
@@ -267,33 +262,40 @@ public class CnossosPathBuilder {
         Coordinate src = cutProfile.getSource().getCoordinate();
         
         // For reflection paths without diffraction points, check if reflection is valid
-        // In favorable conditions, the wall is transformed (lowered), and we need to verify
-        // the reflection point is actually on the wall, not above it
-        if (cutProfile.profileType == CutProfile.PROFILE_TYPE.REFLECTION && hullPointsIndices.size() == 2) {
+        // In favorable conditions, the wall is effectively lowered, and we need to verify
+        // the reflection point is actually on the transformed wall, not above it
+        if (favourable && cutProfile.profileType == CutProfile.PROFILE_TYPE.REFLECTION && hullPointsIndices.size() == 2) {
             // No diffraction points, just source -> reflection -> receiver
-            for (int i = 0; i < cutProfilePoints.size(); i++) {
-                CutPoint cutPoint = cutProfilePoints.get(i);
-                if (cutPoint instanceof CutPointReflection &&
-                        Double.compare(cutPoint.getCoordinate().z, cutPoint.getzGround()) != 0) {
-                    CutPointReflection cutPointReflection = (CutPointReflection) cutPoint;
-                    
-                    // Get the wall altitude at the reflection point (now properly transformed in favorable conditions)
-                    double wallAltitudeAtReflection = Vertex.interpolateZ(cutPoint.coordinate,
-                            cutPointReflection.wall.p0, cutPointReflection.wall.p1);
-                    
-                    // Get the direct line altitude from source to receiver at this reflection X position
-                    Coordinate srcCoord = cutProfilePoints.get(0).getCoordinate();
-                    Coordinate rcvCoord = cutProfilePoints.get(cutProfilePoints.size() - 1).getCoordinate();
-                    double reflectionX = pts2D.get(i).x;
-                    double srcX = pts2D.get(0).x;
-                    double rcvX = pts2D.get(cutProfilePoints.size() - 1).x;
-                    double t = (reflectionX - srcX) / (rcvX - srcX);
-                    double directLineAltitude = srcCoord.z + t * (rcvCoord.z - srcCoord.z);
-                    
-                    // If the direct line clears the wall (goes above it), reject the reflection
-                    if (directLineAltitude > wallAltitudeAtReflection + EPSILON) {
-                        // Ray passes over the wall, reflection is not valid
-                        return null;
+            // Compute transformed coordinates for favorable conditions to check reflection validity
+            List<CutPoint> transformedCutPoints = new ArrayList<>();
+            cutProfile.computePts2D(true, transformedCutPoints);
+            
+            if (!transformedCutPoints.isEmpty()) {
[…]
