+++ noisemodelling-propagation/src/test/java/org/noise_planet/noisemodelling/propagation/AttenuationComputeOutputCnossosTest.java
@@ -5706,7 +5706,7 @@ public class AttenuationComputeOutputCnossosTest {
 
         double[] L = addArray(propDataOut.getVerticesSoundLevel().get(0).levels, new double[]{93 - 26.2, 93 - 16.1, 93 - 8.6, 93 - 3.2, 93, 93 + 1.2, 93 + 1.0, 93 - 1.1});
 
-        assertArrayEquals(expectedFullLA, L, ERROR_EPSILON_HIGH);
+        assertArrayEquals(expectedFullLA, L, ERROR_EPSILON_VERY_LOW);
     }
 
     /**
