### 2025-12-02 99395fe
Update Java installation instructions for Linux and macOS
