### 2025-12-04 19f78c0
update readme and docs index
