### 2025-12-22 Universite-Gustave-Eiffel/NoiseModelling#859 @hendriks73
JOSS authorship question
Hey there,

based in commit counts, the lead author of the repo is @nicolas-f . The submitting author @Nitnelav seems to have played a marginal role (purely based on # of commits).

Could you please explain?

Thanks.

See https://github.com/openjournals/joss-reviews/issues/9171
