### 2025-12-22 Universite-Gustave-Eiffel/NoiseModelling#860 @hendriks73
JOSS contribute
Hey,

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

It seems to me you have covered 2. and 3. pretty well. I don't see any *explicit* information regarding 1. though.

So, if I wanted to contribute, what would I need to do? Fork, PR, discuss via issue first, ...

Perhaps you'd like to add a `CONTRIBUTING.md` file to make this as clear and easy as possible.

See also https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

See https://github.com/openjournals/joss-reviews/issues/9171
