### 2026-01-07 1b4bfb4
read/write dump data with configurable file paths set delaunay epsilon value
