+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/delaunay/LayerTinfour.java
@@ -13,6 +13,8 @@ package org.noise_planet.noisemodelling.pathfinder.delaunay;
 import org.locationtech.jts.algorithm.Orientation;
 import org.locationtech.jts.geom.*;
 import org.locationtech.jts.index.quadtree.Quadtree;
+import org.locationtech.jts.io.ParseException;
+import org.locationtech.jts.io.WKTReader;
 import org.locationtech.jts.io.WKTWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,10 +22,7 @@ import org.tinfour.common.*;
 import org.tinfour.standard.IncrementalTin;
 import org.tinfour.utils.TriangleCollector;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.util.*;
 
 public class LayerTinfour implements LayerDelaunay {
@@ -127,11 +126,29 @@ public class LayerTinfour implements LayerDelaunay {
         return new Coordinate( cx, cy, cz);
     }
 
-    public void dumpData() {
+    public void readDump(File dumpPath) throws IOException, LayerDelaunayError, ParseException {
+        WKTReader wktReader = new WKTReader();
+        try(BufferedReader reader = new BufferedReader(new FileReader(dumpPath))) {
+            String line;
+            int index = 0;
+            while ((line = reader.readLine()) != null) {
+                Geometry obj = wktReader.read(line);
+                if(obj instanceof Point) {
+                    addVertex(obj.getCoordinate());
+                } else if(obj instanceof Polygon) {
+                    addPolygon((Polygon)obj, index++);
+                } else if (obj instanceof LineString) {
+                    addLineString((LineString)obj, index++);
+                }
+            }
+        }
+    }
+
+    public void dumpData(File destinationCsv) {
         GeometryFactory factory = new GeometryFactory();
         WKTWriter wktWriter = new WKTWriter(3);
         try {
-            try (BufferedWriter writer = new BufferedWriter(new FileWriter(new File(dumpFolder, "tinfour_dump.csv")))) {
+            try (BufferedWriter writer = new BufferedWriter(new FileWriter(destinationCsv))) {
                 for(Object vObj : ptsIndex.queryAll()) {
                     if(vObj instanceof Vertex) {
                         final Vertex v = (Vertex)vObj;
@@ -189,7 +206,7 @@ public class LayerTinfour implements LayerDelaunay {
         List<SimpleTriangle> simpleTriangles = new ArrayList<>();
         do {
             // Triangulate
-            tin = new IncrementalTin();
+            tin = new IncrementalTin(epsilon);
             // Add points
             tin.add(meshPoints, null);
             // Add constraints
@@ -199,7 +216,7 @@ public class LayerTinfour implements LayerDelaunay {
                 // Got error
                 // Dump input data
                 if(!dumpFolder.isEmpty()) {
-                    dumpData();
+                    dumpData(new File(dumpFolder, "tinfour_dump.csv"));
                 }
                 throw new LayerDelaunayError(ex);
             }
