### 2026-01-08 2571f44
Add a CONTRIBUTING.md file Fixes #860
