+++ CONTRIBUTING.md
@@ -1,0 +1,62 @@
+# Contributing to NoiseModelling
+
+First off, thanks for taking the time to contribute!
+
+The following is a set of guidelines for contributing to NoiseModelling. These are mostly guidelines, not rules. Use your best judgment, and feel free to propose changes to this document in a pull request.
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the [NoiseModelling Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code.
+
+## How Can I Contribute?
+
+### Reporting Bugs
+
+This section guides you through submitting a bug report for NoiseModelling. Following these guidelines helps maintainers and the community understand your report, reproduce the behavior, and find related reports.
+
+*   **Check the [issues](https://github.com/Universite-Gustave-Eiffel/NoiseModelling/issues)** to see if the problem has already been reported.
+*   **Open a new issue**. Explain the problem and include additional details to help maintainers reproduce the problem:
+    *   Use a clear and descriptive title for the issue to identify the problem.
+    *   Describe the exact steps to reproduce the problem in as many details as possible.
+    *   Describe the behavior you observed after following the steps and point out what exactly is the problem with that behavior.
+    *   Explain which behavior you expected to see instead and why.
+    *   Include screenshots and animated GIFs which show you following the reproduction steps.
+
+### Suggesting Enhancements
+
+This section guides you through submitting an enhancement suggestion for NoiseModelling, including completely new features and minor improvements to existing functionality.
+
+*   **Check the [issues](https://github.com/Universite-Gustave-Eiffel/NoiseModelling/issues) and [discussions](https://github.com/Universite-Gustave-Eiffel/NoiseModelling/discussions)** to see if the enhancement has already been suggested.
+*   **Open a new issue** or start a **new discussion**. Describe the enhancement and provide as much detail as possible.
+
+### Your First Code Contribution
+
+Unsure where to begin contributing to NoiseModelling? You can start by looking through these `good first issue` and `help wanted` issues:
+
+*   [Good first issues](https://github.com/Universite-Gustave-Eiffel/NoiseModelling/labels/good%20first%20issue) - issues which should only require a few lines of code, and a test or two.
+*   [Help wanted issues](https://github.com/Universite-Gustave-Eiffel/NoiseModelling/labels/help%20wanted) - issues which should be a bit more involved than `good first issue`.
+
+### Pull Requests
+
+*   **Fork the repository** and clone it locally.
+*   Create a branch for your edits.
+*   Make sure your code lints and tests pass.
+*   **Submit a pull request** to the `main` branch.
+*   If you're adding a new feature, add tests for it.
+
+## Styleguides
+
+### Java Styleguide
+
+*   Please follow the existing coding style in the project.
+*   Ensure that your code is well-documented with Javadoc where appropriate.
+
+### Documentation Styleguide
+
+*   Documentation is written in reStructuredText (.rst) and Markdown (.md).
+*   Correct spelling and grammar are appreciated.
+
+## Additional Resources
+
+*   [Get Started Dev Documentation](https://noisemodelling.readthedocs.io/en/latest/Get_Started_Dev.html)
+*   [NoiseModelling Website](http://noise-planet.org/noisemodelling.html)
