### 2026-01-08 1d5f91e
Add a CONTRIBUTING.md file (#863)
Fixes #860
