### 2026-01-15 add18e4
Refine GUI documentation for NoiseModelling
Updated text for clarity and consistency in the GUI documentation.
