### 2026-01-15 494694e
Update NoiseModelling GUI launch instructions
Clarified instructions for launching NoiseModelling GUI and updated tips regarding its operation.
