### 2026-01-15 #9171 @svchb
@Nitnelav also your requirements page is not listing the "optional?" requirements for libgeos-c1v5, libnetcdf19, libgdal-java gdal-bin
