### 2026-01-19 c6d26b3
Clarify getstarted instructions (#865)
* Refine GUI documentation for NoiseModelling

Updated text for clarity and consistency in the GUI documentation.

* Update NoiseModelling GUI launch instructions

Clarified instructions for launching NoiseModelling GUI and updated tips regarding its operation.
