+++ Docs/Get_Started_GUI.rst
@@ -1,24 +1,24 @@
 Get Started - GUI
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Below we present a simple application case, allowing you to discover NoiseModelling through its Graphical User Interface .
+Below we present a simple example to help you discover NoiseModelling through its Graphical User Interface (GUI).
 
 Step 1: Download NoiseModelling
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Download the latest realease of NoiseModelling on `Github`_. 
+Download the latest release of NoiseModelling on `Github`_.
 
-* Windows: you can directly download and execute the ``NoiseModelling_5.x.x_install.exe`` installer file *(or you can also follow Linux / Mac instructions below)*
+* Windows: you can directly download and run the ``NoiseModelling_5.x.x_install.exe`` installer file *(or you can also follow the Linux / Mac instructions below)*
 * Linux or Mac: download the ``NoiseModelling_5.x.x.zip`` file and unzip it into a chosen directory
 
 .. warning::
-    The chosen directory can be anywhere, but be sure that you have write access. If you are using the computer of your company, the Program Files folder is probably not a good idea.
+    The chosen directory can be anywhere, but make sure you have write access. If you are using a company computer, the Program Files folder is probably not a good idea.
 
 .. warning::
-    For **Linux** and **Mac** users, please make sure your Java environment is well setted. For more information, please read the page :doc:`Requirements`. **Windows** users who are using the ``.exe`` file are not concerned since the Java Runtime Environment is **already embeded**.
+    For **Linux** and **Mac** users, please make sure your Java environment is correctly set up. For more information, please read the page :doc:`Requirements`. **Windows** users who are using the ``.exe`` file are not concerned, since the Java Runtime Environment is **already embedded**.
 
 .. note::
-    Only from version 3.3, NoiseModelling releases include the user interface described in this tutorial. 
+    Starting from version 3.3, NoiseModelling releases include the user interface described in this tutorial.
 
 .. _Github : https://github.com/Ifsttar/NoiseModelling/releases
 
@@ -27,27 +27,29 @@ Download the latest realease of NoiseModelling on `Github`_.
 Step 2: Start NoiseModelling GUI
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-As seen in the page ":doc:`Architecture`", NoiseModelling can be used through a Graphic User Interface (GUI), thanks to Geoserver and WPS Builder bricks.
+As described on the page ":doc:`Architecture`", NoiseModelling can be used through a Graphical User Interface (GUI), thanks to the GeoServer and WPS Builder components.
 
-In this tutorial, we will use the default and already configured H2GIS database.
+In this tutorial, we will use the default, already configured H2GIS database.
 
-Those tools (Geoserver, WPS Builder and H2GIS) are already included in the archive. So you don't have to install them before.
+These tools (GeoServer, WPS Builder and H2GIS) are already included in the archive, so you don't have to install them separately.
 
-To launch NoiseModelling with GUI, please execute :
+To launch NoiseModelling with the GUI, start it from a command prompt (terminal). This will start a local server on your computer, which provides the GUI as a web application.
+
+Please execute:
 
 * Windows: ``NoiseModelling.exe`` or ``NoiseModelling_xxx\bin\startup_windows.bat``
-* Linux or Mac: ``NoiseModelling_xxx/bin/startup_linux_mac.sh`` *(check authorize file execution in property of this file before)*
[…]
