### 2026-01-23 de11549
Fix Railway Roughness evaluation with and without impact on line
