### 2026-01-26 c34ef69
Fix Test RailwayCnossos : correction due to period on Test_Cnossos_Rail_emission_OC2
