+++ noisemodelling-emission/src/test/java/org/noise_planet/noisemodelling/emission/railway/RailwayCnossosTest.java
@@ -520,7 +520,7 @@ public class RailwayCnossosTest {
         railwayCnossos.setTrainSetDataFile("RailwayTrainsets.json");
         railwayCnossos.setRailwayDataFile("RailwayCnossosSNCF_2021.json");
         // Expected values
-        double[] expectedValuesLWRolling = new double[]{43.8903627967019, 44.9655697489009, 48.9478577301374, 55.5479496503529, 55.4699076555030, 56.4112551136621, 57.9704698930737, 60.2216492989323, 61.6201352980830, 63.0726300511446, 63.4414710623726, 67.9470850371138, 72.2126730845647, 71.1272989535570, 68.1142891000406, 65.1061856218274, 65.2001156877479, 65.9588236417732, 65.2117259673999, 63.2782993972508, 59.9386370040542, 52.6037946949367, 50.7526799075511, 48.1468415469288};
+        double[] expectedValuesLWRolling = new double[]{43.8903627967019, 44.9655697489009, 48.9478577301374, 55.5479496503529, 55.435644216130996, 56.28918949883372, 57.76517536242085, 59.857189090598624, 60.9673183839427, 61.989589714413825, 61.667715524758066, 64.86050810840895, 67.79089584625551, 65.62626998386897, 62.18314392899565, 59.57968136675302, 61.08640983016812, 62.20761878325611, 61.49143701879498, 59.61042465277447, 56.53747558356267, 50.41901289738583, 49.0315804220768, 46.79550354867328};
         double[] expectedValuesLW_Traction_A = new double[]{49.8691001300806, 49.7691001300806, 49.6691001300806, 50.1691001300806, 47.0691001300806, 54.1691001300806, 50.7691001300806, 44.2691001300806, 46.8691001300806, 47.0691001300806, 45.6691001300806, 44.9691001300806, 43.4691001300806, 44.5691001300806, 41.5691001300806, 41.1691001300806, 38.9691001300806, 37.3691001300806, 36.4691001300806, 35.3691001300806, 31.7691001300806, 27.6691001300806, 22.4691001300806, 20.4691001300806};
         double[] expectedValuesLW_Traction_B = new double[]{-39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572, -39.030899869572};
         double[][] expectedValues={expectedValuesLWRolling,expectedValuesLW_Traction_A,expectedValuesLW_Traction_B};
@@ -831,11 +831,7 @@ public class RailwayCnossosTest {
         railwayCnossos.setRailwayDataFile("RailwayCnossosSNCF_2021.json");
 
         // Expected values
-        double[] expectedValuesLWRollingD = new double[]{-140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140};
-        double[] expectedValuesLW_Traction_A_D = new double[]{-140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140, -140};
[…]
