### 2026-01-26 Universite-Gustave-Eiffel/NoiseModelling#866 @hendriks73
Apple Silicon support?
The installation instructions for macOS don't mention Apple Silicon (M1, M2, ... essentially arm64) at all.

Does the software run on arm? If not, please clarify the instructions or support arm. Apple has started shipping arm-based laptops in 2020 and stopped shipping x86-based computers in 2023. Any software published for macOS today should support arm natively.
