### 2026-01-26 #9171 @hendriks73
Unclear installation instructions for macOS.
https://github.com/Universite-Gustave-Eiffel/NoiseModelling/issues/866
