### 2026-01-26 2ba8a6c
Fix Test RailwayCnossos : correction due to period on Test_Cnossos_Rail_emission_section_1
