+++ noisemodelling-emission/src/test/java/org/noise_planet/noisemodelling/emission/railway/RailwayCnossosTest.java
@@ -73,7 +73,7 @@ public class RailwayCnossosTest {
         railwayCnossos.setTrainSetDataFile("RailwayTrainsets.json");
         railwayCnossos.setRailwayDataFile("RailwayCnossosEU_2020.json");
 
-        double[] expectedValuesLWRolling = new double[]{46.6292393707431, 47.5930887825746, 49.4885539543655, 50.8452515062722, 48.2904168235536, 47.5599118826212, 48.3659880536885, 53.6850278385126, 55.1794804329062, 56.4435891375742, 57.3811010252223, 58.0623572653650, 59.8295159717300, 59.4546602463242, 56.5427214224336, 52.1790459841615, 54.5278138699040, 53.2809012600130, 51.1582349752733, 48.8021326510878, 49.2065949846659, 48.6821233578964, 48.6765663190977, 50.2184119171958};
+        double[] expectedValuesLWRolling = new double[]{46.6292393707431, 47.5930887825746, 49.4885539543655, 50.8452515062722, 48.2904168235536, 47.5599118826212, 48.3659880536885, 53.6850278385126, 55.1794804329062, 56.4435891375742, 57.3811010252223, 58.0623572653650, 59.8295159717300, 59.4546602463242, 56.5427214224336, 52.1790459841615, 54.51129722944201, 53.2570463370446, 51.118153521942574, 48.73174211849708, 49.13517759611298, 48.60046539576308, 48.58746317902302, 50.14566621392022};
         double[] expectedValuesLWTractionA = new double[]{46.8200805302231, 42.7200805302231, 40.5200805302231, 42.5200805302231, 40.7200805302231, 40.7200805302231, 40.9200805302231, 42.7200805302231, 42.5200805302231, 43.6200805302231, 43.5200805302231, 46.5200805302231, 43.1200805302231, 43.0200805302231, 43.0200805302231, 42.0200805302231, 42.0200805302231, 47.3200805302231, 40.4200805302231, 37.4200805302231, 34.9200805302231, 32.0200805302231, 29.4200805302231, 27.1200805302231};
         double[] expectedValuesLWTractionB = new double[]{51.1200805302231, 47.9200805302231, 43.4200805302231, 41.9200805302231, 41.2200805302231, 41.5200805302231, 40.8200805302231, 40.6200805302231, 40.3200805302231, 40.7200805302231, 40.7200805302231, 44.7200805302231, 40.6200805302231, 40.9200805302231, 40.8200805302231, 41.0200805302231, 41.1200805302231, 46.2200805302231, 39.4200805302231, 36.6200805302231, 33.9200805302231, 31.3200805302231, 28.8200805302231, 26.6200805302231};
         String[] typeNoise = new String[] {"ROLLING", "TRACTIONA", "TRACTIONB"};
