### 2026-01-26 #9171 @Nitnelav
> [@Nitnelav](https://github.com/Nitnelav) So I guess its still only working with Java11?

Yes ! very probably, @nicolas-f can confirm
