### 2026-01-26 #9171 @Nitnelav
> [@Nitnelav](https://github.com/Nitnelav) also your requirements page is not listing the "optional?" requirements for libgeos-c1v5, libnetcdf19, libgdal-java gdal-bin

Ah maybe, what page are you referring to ?
