### 2026-01-26 #9171 @svchb
> > [@Nitnelav](https://github.com/Nitnelav) also your requirements page is not listing the "optional?" requirements for libgeos-c1v5, libnetcdf19, libgdal-java gdal-bin
>
> Ah maybe, what page are you referring to ?

Here is nothing about the requirements:
https://noisemodelling.readthedocs.io/en/latest/Requirements.html

Also the first box says 4.x is not working with Java 11.x but I guess 5.x isn't as well.
