### 2026-01-26 #9171 @nicolas-f
Hi,

There is two release of NoiseModelling:

- NoiseModelling in console mode without GeoServer, there is no restriction on the Java version (still must be >= Java 11)
- NoiseModelling with Geoserver (web graphical interface), this version of GeoServer is compatible with only Java 11

But this things will change in the near future (We will remove Geoserver)
