### 2026-01-26 #9171 @svchb
> Hi,
>
> There is two release of NoiseModelling:
>
> * NoiseModelling in console mode without GeoServer, there is no restriction on the Java version (still must be >= Java 11)
> * NoiseModelling with Geoserver (web graphical interface), this version of GeoServer is compatible with only Java 11
>
> But this things will change in the near future (We will remove Geoserver)

Please clarify this version requirement on your requirements page. If there is a strict version requirement, it would be best to show a clear error message when it is not met. In the current state, users simply end up with a non-functional web interface.

Also, please document the optional libraries and explain what impact they have when installed (or not installed).
