### 2026-01-26 c98b39f
Fix Test RailwayCnossos : correction expected value
