### 2026-01-29 6173621
Fix issue with leak of constraint areas by using an internal RTree for triangle polygon index retrieval instead
