+++ noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/LayerTinfourTest.java
@@ -100,8 +100,7 @@ public class LayerTinfourTest {
         layerTinfour.processDelaunay();
         List<org.noise_planet.noisemodelling.pathfinder.delaunay.Triangle> triangleList = layerTinfour.getTriangles();
         List<Coordinate> vertices = layerTinfour.getVertices();
-        // Test dump
-        layerTinfour.dumpData();
+
         Point hole1 =  factory.createPoint(new Coordinate(222690.860,6758520.184));
         Point hole2 = factory.createPoint(new Coordinate(222711.177,6758532.233));
         Point inGeom = merged.getInteriorPoint();
