### 2026-01-29 c69c73f
add two options on DelaunayReceiversMaker Do not export triangles geometries Skip cells without sources at a certain distance
