### 2026-01-30 f57be10
Fix issue, do not insert point inside polygons when refining according to maxArea
