### 2026-01-30 f648640
Fix groovy script according to java update
