### 2026-02-02 2b736cd
Refactor Delaunay triangulation to use streams for improved performance and readability; update triangle polygon association logic
