### 2026-02-02 600086b
Faster point in polygon test thanks to PreparedPolygon
