### 2026-02-02 261761b
Add logging for cell processing in Delaunay receiver generation and implement method to retrieve unique time periods from receivers level table
