### 2026-02-02 39f7140
fix json missing reflection on tc27
