### 2026-02-05 7a23058
Fix github action release script
