### 2026-02-12 #9171 @svchb
@nicolas-f @Nitnelav Please implement these changes:

> Please clarify this version requirement on your requirements page. If there is a strict version requirement, it would be best to show a clear error message when it is not met. In the current state, users simply end up with a non-functional web interface.
> Also, please document the optional libraries and explain what impact they have when installed (or not installed).
