### 2026-02-12 #9171 @faroit
@svchb @hendriks73 can you give us a timeline for when you have finished the review? we typically aim for a review turn-around of 6 weeks max, so i would love to finish the review in time.

@svchb please open individual issues in the repo and refer back to [this issue](https://github.com/openjournals/joss-reviews/issues/9171) to avoid lengthy discussions in this thread.
