### 2026-03-18 #9171 @Nitnelav
Hello, I've updated the Requirements page in the documentation here : https://github.com/Universite-Gustave-Eiffel/NoiseModelling/pull/884

Concerning the optional libraries, there shouldn't be any. Java should be the only thing you have to install.

my theory is that you used a too-recent version of Java and ended up with a broken web interface. You then looked at the console, saw some suspicious/odd error messages, and interpreted them as missing optional dependencies. Does it sound plausible ?
