### 2026-03-18 #9171 @Nitnelav
About handling a too-recent version of JAVA more gracefully :
we are in the process of getting rid of Geoserver and that it will greatly improve the installation process along with the JAVA version tolerance (>= 11) and so we will not fix this issue in the mean time.
