### 2026-03-24 #9171 @faroit
@hendriks73 @svchb can you please give us an ETA for when you can complete the review?

Please do not add review comments to this thread but instead open separate issues on https://github.com/Universite-Gustave-Eiffel/NoiseModelling with reference links back to https://github.com/openjournals/joss-reviews/issues/9171 to avoid that this thread getting clogged up.
