### 2026-04-09 #9171 @Nitnelav
Hello @hendriks73 @svchb maybe it will lower the barrier to review if I may remind you that you can review our work without the need to install and run NoiseModelling on your machine (even though we'd be happy if you do of course). You can rest assured that our software is indeed working :) by looking at the list of scientific productions we're aware of : https://noisemodelling.readthedocs.io/en/latest/Scientific_production.html.
Cheers
